% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{add_new_metadata}
\alias{add_new_metadata}
\title{Replace the meta data}
\usage{
add_new_metadata(object, new_metadata, ...)
}
\arguments{
\item{object}{The class}

\item{new_metadata}{data.table. The new meta data you wish to add.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
The object with updated metadata.
}
\description{
This function will replace the meta data within the given object
}
\examples{
# swap in a metadata table that carries an extra batch column
set.seed(42)
counts <- matrix(rpois(60, 20), nrow = 10, ncol = 6)
rownames(counts) <- sprintf("gene_\%i", 1:10)
colnames(counts) <- sprintf("sample_\%i", 1:6)
meta <- data.table::data.table(
  sample_id = colnames(counts),
  case_control = rep(c("case", "control"), each = 3)
)
object <- BulkDge(raw_counts = counts, meta_data = meta)
new_meta <- data.table::copy(meta)[, batch := rep(c("b1", "b2"), 3)]
object <- add_new_metadata(object, new_metadata = new_meta)
head(S7::prop(object, "meta_data"))
}
