% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{add_nhoods_info}
\alias{add_nhoods_info}
\alias{add_nhoods_info.miloR}
\title{Add neighbourhood info on majority cell type}
\usage{
add_nhoods_info(x, cell_info)

\method{add_nhoods_info}{miloR}(x, cell_info)
}
\arguments{
\item{x}{\code{miloR} object on which to tag on additional neighbourhood
information.}

\item{cell_info}{Character vector. Represents the cell type annotations
you wish to add to the different neighbourhoods. Must be the same length
as the number of cells (rows) in the nhoods matrix.}
}
\value{
Modified \code{miloR} object with updated \code{nhoods_info} containing
\code{majority_celltype} and \code{majority_prop} columns.
}
\description{
This function adds cell type composition information to the \code{nhoods_info}
slot within the \code{miloR} object. For each neighbourhood, it calculates
the proportion of the majority cell type and identifies which cell type
is most abundant. This is useful for annotating differential abundance
results with the cellular composition of each neighbourhood.
}
\examples{
# tag each neighbourhood with its majority cell type
sc <- demo_single_cells(
  syn_data_params = params_sc_synthetic_data(
    n_cells = 500L,
    n_genes = 50L,
    n_samples = 6L,
    sample_bias = "even"
  )
)
milo <- get_miloR_abundances_sc(
  sc,
  sample_id_col = "sample_id",
  miloR_params = params_sc_miloR(k_refine = 10L),
  .verbose = FALSE
)
design_df <- data.frame(
  grp = rep(c("a", "b"), each = 3),
  row.names = sprintf("sample_\%i", 1:6)
)
milo <- test_nhoods(milo, design = ~grp, design_df = design_df)
milo <- add_nhoods_info(milo, cell_info = get_sc_obs(sc)$cell_grp)
head(get_differential_abundance_res(milo))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
