% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{add_sc_new_obs}
\alias{add_sc_new_obs}
\title{Add an obs table derived from a method to the SingleCells.}
\usage{
add_sc_new_obs(object, obs_data)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{obs_data}{data.table. A data.table you generated with \code{\link[=get_data]{get_data()}}
on some sub class.}
}
\value{
The class with updated obs table in the DuckDB
}
\description{
Add an obs table derived from a method to the SingleCells.
}
\examples{
# cluster memberships carry their own cell_idx, so they join straight on
sc <- demo_single_cells()
clusters <- fast_cluster_sc(sc, resolutions = 1.0, .verbose = FALSE)
sc <- add_sc_new_obs(sc, get_data(clusters))
head(get_sc_obs(sc), 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
