% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R
\name{add_spatial_sample}
\alias{add_spatial_sample}
\title{Add a SpatialSample to a spatial object}
\usage{
add_spatial_sample(object, sample)
}
\arguments{
\item{object}{A \code{\link[=SpatialSpot]{SpatialSpot()}}.}

\item{sample}{A \code{\link[=new_spatial_sample]{new_spatial_sample()}}.}
}
\value{
The updated object.
}
\description{
Add a SpatialSample to a spatial object
}
