% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{all_limma_contrasts}
\alias{all_limma_contrasts}
\title{Create all limma contrasts (based on combination of everything)}
\usage{
all_limma_contrasts(limma_fit, contrast_grps)
}
\arguments{
\item{limma_fit}{The fitted limma model, i.e., output of \code{\link[limma:lmFit]{limma::lmFit()}}.}

\item{contrast_grps}{String vector. The contrast groups of interest. If NULL
all co-variate comparisons will be returned.}
}
\value{
The Limma contrasts for further usage.
}
\description{
Create all limma contrasts (based on combination of everything)
}
