% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_helpers.R
\name{anndata_parser}
\alias{anndata_parser}
\title{Class for Anndata}
\description{
This class helps dealing with h5ad objects from Python AnnData.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-anndata_parser-new}{\code{anndata_parser$new()}}
\item \href{#method-anndata_parser-get_obs_data}{\code{anndata_parser$get_obs_data()}}
\item \href{#method-anndata_parser-get_raw_counts}{\code{anndata_parser$get_raw_counts()}}
\item \href{#method-anndata_parser-clone}{\code{anndata_parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-anndata_parser-new"></a>}}
\if{latex}{\out{\hypertarget{method-anndata_parser-new}{}}}
\subsection{Method \code{new()}}{
Initialises the Anndata Parser.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{anndata_parser$new(h5_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h5_path}}{String. Path to the h5 file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the initialised class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-anndata_parser-get_obs_data"></a>}}
\if{latex}{\out{\hypertarget{method-anndata_parser-get_obs_data}{}}}
\subsection{Method \code{get_obs_data()}}{
Returns the observation table with all the data from the
h5ad file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{anndata_parser$get_obs_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table. The found observations are returned. The pandas index
will be named \code{sample_id}. Remaining columns will be returned as factors
due to the way the data is stored in h5.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-anndata_parser-get_raw_counts"></a>}}
\if{latex}{\out{\hypertarget{method-anndata_parser-get_raw_counts}{}}}
\subsection{Method \code{get_raw_counts()}}{
Returns the counts that are stored in \code{X} slot of the
anndata object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{anndata_parser$get_raw_counts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the count matrix with samples = columns and rows =
features.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-anndata_parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-anndata_parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{anndata_parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
