% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cPCA.R
\name{apply_contrastive_pca}
\alias{apply_contrastive_pca}
\title{Apply contrastive PCA.}
\usage{
apply_contrastive_pca(object, alpha, no_pcs)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{alpha}{Alpha parameter to use.}

\item{no_pcs}{Number of contrastive PCs to generate.}
}
\value{
\code{bulk_coexp} with additional data in the slots
}
\description{
Applies the contrastive PCA algorithm given a specified alpha and a number of
contrastive principal components to extract.
}
\references{
Abid, et al., Nature Communications, 2018
}
