% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertCellMarkerList}
\alias{assertCellMarkerList}
\title{Assert cell marker list}
\usage{
assertCellMarkerList(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to assert.}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in checkmate.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns \code{x} if the assertion is successful.
}
\description{
Checkmate extension for asserting a cell marker list as returned
by \code{\link[=prepare_cell_markers]{prepare_cell_markers()}}.
}
\keyword{internal}
