% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertCellsExist}
\alias{assertCellsExist}
\title{Assert that the cell names exist in the object}
\usage{
assertCellsExist(x, cell_names, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The \code{SingleCells} or \code{SingleCellsSubset} object to check/assert.}

\item{cell_names}{String. The provided cell names.}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in checkmate.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is successful.
}
\description{
Checkmate extension for asserting if the provided cell names
exist in the object.
}
\keyword{internal}
