% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc_lr.R
\name{assertExpTable}
\alias{assertExpTable}
\title{Assert an expression info table has the expected columns and types}
\usage{
assertExpTable(x, name)
}
\arguments{
\item{x}{The candidate \code{data.table}.}

\item{name}{Variable name for error messages.}
}
\description{
Assert an expression info table has the expected columns and types
}
\keyword{internal}
