% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertLrDb}
\alias{assertLrDb}
\title{Assert a ligand-receptor database}
\usage{
assertLrDb(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The data.table to assert.}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in checkmate.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns \code{x} if the assertion is successful.
}
\description{
Checkmate extension for asserting a ligand-receptor database
as returned by \code{\link[=get_cellphonedb_db]{get_cellphonedb_db()}}.
}
\keyword{internal}
