% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertScDblFinder}
\alias{assertScDblFinder}
\title{Assert scDblFinder parameters}
\usage{
assertScDblFinder(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to check/assert.}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages.}
}
\value{
Invisibly returns the checked object if successful.
}
\description{
Checkmate extension for asserting scDblFinder parameters.
}
\keyword{internal}
