% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertScSyntheticCellsweep}
\alias{assertScSyntheticCellsweep}
\title{Assert synthetic CellSweep data parameters}
\usage{
assertScSyntheticCellsweep(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to check/assert}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in checkmate.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is successful.
}
\description{
Checkmate assertion for the parameters for the generation of
synthetic CellSweep data.
}
\keyword{internal}
