% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertSpH5adIo}
\alias{assertSpH5adIo}
\title{Assert spatial h5ad ingest parameters}
\usage{
assertSpH5adIo(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to check/assert}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is
successful.
}
\description{
Checkmate assertion for the spatial h5ad ingest parameters
built by \code{\link[=params_sp_h5ad_io]{params_sp_h5ad_io()}}.
}
\keyword{internal}
