% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertSpSparkx}
\alias{assertSpSparkx}
\title{Assert SPARK-X parameters}
\usage{
assertSpSparkx(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to check/assert}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is
successful.
}
\description{
Checkmate assertion for the SPARK-X parameters built by
\code{\link[=params_sp_sparkx]{params_sp_sparkx()}}.
}
\keyword{internal}
