% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertSpVisiumIo}
\alias{assertSpVisiumIo}
\title{Assert Visium ingest parameters}
\usage{
assertSpVisiumIo(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to check/assert}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is
successful.
}
\description{
Checkmate assertion for the Visium ingest parameters built by
\code{\link[=params_sp_visium_io]{params_sp_visium_io()}}.
}
\keyword{internal}
