% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{auto_streaming}
\alias{auto_streaming}
\title{Helper to set streaming to TRUE on large data sets}
\usage{
auto_streaming(n_cells, streaming = NULL, .verbose = TRUE)
}
\arguments{
\item{n_cells}{Integer. Number of cells in the data set.}

\item{streaming}{Optional boolean. Parsed from the main function.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Boolean if streaming should be used.
}
\description{
Helper to set streaming to TRUE on large data sets
}
\keyword{internal}
