% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R
\name{bixverse_base_class}
\alias{bixverse_base_class}
\title{bixverse base class}
\usage{
bixverse_base_class()
}
\value{
Returns the S7 object for further operations.
}
\description{
Generic base class that is used for inheritance in certain common methods
across other classes.
}
\section{Properties}{

\describe{
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{A data.table that will contain the final results.}
}
}

