% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R
\name{BixverseBaseClass}
\alias{BixverseBaseClass}
\title{BixverseBaseClass}
\usage{
BixverseBaseClass()
}
\value{
Returns the S7 object for further operations.
}
\description{
Generic base class that is used for inheritance in certain common methods
across other classes.
}
\section{Properties}{

\describe{
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{A data.table that will contain the final results.}
}
}

\examples{
# every analysis class inherits the base class getters
object <- SimilarityNetworkFusion(snf_params = params_snf(k = 3L))
inherits(object, "bixverse::BixverseBaseClass")
}
\keyword{internal}
