% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bulk_coexp}
\alias{bulk_coexp}
\title{Bulk RNAseq co-expression modules (deprecated)}
\usage{
bulk_coexp(raw_data, meta_data, variable_info = NULL)
}
\arguments{
\item{raw_data}{The raw count matrix. Rows = samples, columns = features.}

\item{meta_data}{data.table. Metadata information on the samples. Expects to
have a \code{sample_id} column.}

\item{variable_info}{data.table. Metadata information on the features. This
is an optional table.}
}
\value{
Returns a \code{\link[=BulkCoExp]{BulkCoExp()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=BulkCoExp]{BulkCoExp()}}.
}
\keyword{internal}
