% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bulk_dge}
\alias{bulk_dge}
\title{Bulk RNAseq differential gene expression class (deprecated)}
\usage{
bulk_dge(
  raw_counts,
  meta_data,
  variable_info = NULL,
  alternative_gene_id = NULL
)
}
\arguments{
\item{raw_counts}{matrix. The raw count matrix. Rows = genes, columns =
samples.}

\item{meta_data}{data.table. Metadata information on the samples. Expects to
have a \code{sample_id} column.}

\item{variable_info}{data.table. Metadata information on the features.
Defaults to \code{NULL}.}

\item{alternative_gene_id}{String. Optional alternative gene identifier to
be used. Must be a column of \code{variable_info}.}
}
\value{
Returns a \code{\link[=BulkDge]{BulkDge()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=BulkDge]{BulkDge()}}.
}
\keyword{internal}
