% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{BulkCoExp}
\alias{BulkCoExp}
\title{Bulk RNAseq co-expression modules}
\usage{
BulkCoExp(raw_data, meta_data, variable_info = NULL)
}
\arguments{
\item{raw_data}{The raw count matrix. Rows = samples, columns = features.}

\item{meta_data}{data.table Metadata information on the samples. Expects to
have a \code{sample_id} column.}

\item{variable_info}{data.table. Metadata information on the features. This
is an optional table.}
}
\value{
Returns the \code{BulkCoExp} class for further operations.
}
\description{
Class for applying various co-expression module detection methods on top of
bulk RNAseq data.
}
\section{Properties}{

\describe{
\item{raw_data}{A numerical matrix of the provided raw data.}
\item{meta_data}{A data.table with the meta-information about the samples.}
\item{variable_info}{An optional data.table containing the variable info.}
\item{processed_data}{A list in which various types of processed data will
be stored.}
\item{outputs}{A list in which key outputs will be stored.}
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{A data.table that will contain the final results.}
}
}

\examples{
# co-expression class over synthetic bulk counts (samples x genes)
syn <- synthetic_bulk_cor_matrix()
mat <- log1p(t(syn$counts))
meta <- data.table::data.table(sample_id = rownames(mat))
object <- BulkCoExp(raw_data = mat, meta_data = meta)
object
}
