% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{calc_knn_metrics}
\alias{calc_knn_metrics}
\title{Calculate recall at k and distance ratio}
\usage{
calc_knn_metrics(ref_knn, query_knn)
}
\arguments{
\item{ref_knn}{The reference \code{SingleCellNearestNeighbour}.}

\item{query_knn}{The query \code{SingleCellNearestNeighbour}.}
}
\value{
A list with:
\itemize{
\item matches - The intersecting indices between the reference and query
kNN for each sample. In an ideal match up should be equal to k.
\item distance_ratio - The distance ratio. Calculates
\code{sum(dist_query) / sum(dist_ref)} per sample. Indicates how much worse the
reference is.
\item final_recall - The final recall across all samples.
\item final_ratio - The final distance ratio across all samples.
}
}
\description{
Helper function to compare the results of two \code{SingleCellNearestNeighbour}
against each other. The first one can serve as a reference (ground truth)
and you can compare against the second one.
}
\examples{
# recall of an annoy index against the default one
set.seed(42L)
embd <- matrix(rnorm(500 * 10), nrow = 500)
rownames(embd) <- sprintf("cell_\%03d", 1:500)
ref <- generate_sc_knn(embd, .verbose = FALSE)
query <- generate_sc_knn(
  embd,
  neighbours_params = params_sc_neighbours(
    knn = list(knn_method = "annoy")
  ),
  .verbose = FALSE
)
calc_knn_metrics(ref, query)$final_recall
}
