% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mc_processing.R
\name{calc_manifold_metrics}
\alias{calc_manifold_metrics}
\title{Calculate manifold metrics}
\usage{
calc_manifold_metrics(object)
}
\arguments{
\item{object}{\code{MetaCells} class for which to calculate the different metrics.}
}
\value{
The class with the compactness and separation scores added.
}
\description{
This function will calculate the compactness and separation of your metacells
on the manifold (defined by the diffusion map). You must have run
\code{\link[=calc_diffusion_coordinates]{calc_diffusion_coordinates()}} before calling this function. The idea is
that compactness indicates how tight the metacell spans the manifold, whereas
separation indicates how well the different metacells span the manifold.
}
\examples{
# compactness and separation on top of the diffusion coordinates
sc <- demo_single_cells()
mc <- generate_bt_meta_cells_sc(
  sc,
  sc_meta_cell_params = params_sc_bt_metacells(target_no_metacells = 50L),
  .verbose = FALSE
)
mc <- calc_diffusion_coordinates(
  mc,
  knn_data = get_knn_obj(sc),
  .verbose = FALSE
)
mc <- calc_manifold_metrics(mc)
head(mc[[c("compactness", "separation")]])

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\references{
Persad, et al. Nat Biotechnol, 2023
}
