% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{calc_mitch}
\alias{calc_mitch}
\title{Calculate a mitch gene set enrichments on contrast}
\usage{
calc_mitch(contrast_mat, gene_set_list, min_size = 5L)
}
\arguments{
\item{contrast_mat}{Numerical matrix. The rows represent the gene statistic
per contrast and each column represents the contrast.}

\item{gene_set_list}{Named list. Contains the pathways you wish to test
against.}

\item{min_size}{Integer. Minimum size of the gene set to be included.}
}
\value{
A data.table with the Mitch enrichment results.
}
\description{
Rust-based version of the mitch multi-contrast enrichment, see Kaspi and
Ziemann. Takes in a matrix representing the contrasts you wish to test
against.
}
\examples{
# multi-contrast enrichment over two contrasts
set.seed(10L)
contrast_mat <- matrix(
  rnorm(300 * 2),
  ncol = 2,
  dimnames = list(sprintf("gene_\%03i", 1:300), c("contrast_a", "contrast_b"))
)
contrast_mat[1:30, ] <- contrast_mat[1:30, ] + 1.5
gene_sets <- list(
  hit_set = sprintf("gene_\%03i", 1:30),
  bg_set = sprintf("gene_\%03i", 100:150)
)
res <- calc_mitch(contrast_mat, gene_sets)
res[, c("pathway_names", "manova_pval", "s_dist")]
}
\references{
Kaspi and Ziemann, Bmc Genomics, 2020
}
