% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{calculate_all_dges}
\alias{calculate_all_dges}
\title{Calculate all possible DGE variants (DEPRECATED!)}
\usage{
calculate_all_dges(
  object,
  contrast_column,
  contrast_list = NULL,
  filter_column = NULL,
  co_variates = NULL,
  small_sample_correction = NULL,
  ...,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=BulkDge]{BulkDge()}}.}

\item{contrast_column}{String. The contrast column in which the groupings
are stored. Needs to be found in the meta_data within the properties.}

\item{contrast_list}{Optional string vector. A vectors that contains the
contrast formatted as \code{"contrast1-contrast2"}. Default \code{NULL} will create
all possible contrast automatically.}

\item{filter_column}{Optional String. If there is a column you wish to use as
sub groupings, this can be provided here. An example could be different
sampled tissues and you wish to run the DGE analyses within each tissue
separately.}

\item{co_variates}{Optional string vector. Any co-variates you wish to
consider during the Limma Voom modelling.}

\item{small_sample_correction}{Can be NULL (automatic determination if a
small sample size correction should be applied) or a Boolean.}

\item{...}{Additional parameters that used to go to limma's \code{eBayes()} or
\code{voom()}.}

\item{.verbose}{Controls verbosity of the function.}
}
\value{
Throws an error, as it is deprecated.
}
\description{
This is a deprecated method and will raise an error. Please use
\code{\link[=calculate_dge_limma]{calculate_dge_limma()}} and \code{\link[=calculate_dge_hedges]{calculate_dge_hedges()}}.
}
