% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{calculate_diffusion_auc}
\alias{calculate_diffusion_auc}
\title{Calculate the AUROC for a diffusion score}
\usage{
calculate_diffusion_auc(
  object,
  hit_nodes,
  auc_iters = 10000L,
  random_aucs = 1000L,
  permutation_test = FALSE,
  seed = 42L
)
}
\arguments{
\item{object}{\code{NetworkDiffusions} object. The underlying class
\code{\link[=NetworkDiffusions]{NetworkDiffusions()}}.}

\item{hit_nodes}{String vector. Which nodes in the graph are considered a
'hit'.}

\item{auc_iters}{Integer. How many iterations to run to approximate the
AUROC.}

\item{random_aucs}{Integer. How many random AUROCs to calculate to estimate
the Z-score. Only of relevance if permutation test is set to \code{TRUE}.}

\item{permutation_test}{Boolean. Shall a permutation based Z-score be
calculated.}

\item{seed}{Integer. Random seed.}
}
\value{
List with AUC and Z-score as the two named elements if permutations
test set to TRUE; otherwise just the AUC.
}
\description{
This functions can take a given \code{NetworkDiffusions} object and calculates an
AUC and generates a Z-score based on random permutation of \code{random_aucs} for
test for statistical significance if desired.
}
\examples{
# AUROC of the diffusion score against two known hit nodes
set.seed(42)
g <- igraph::sample_pa(15, directed = FALSE)
edges <- data.table::setDT(igraph::as_data_frame(g))[, `:=`(
  from = sprintf("node_\%i", from),
  to = sprintf("node_\%i", to)
)]
object <- NetworkDiffusions(edges, weighted = FALSE, directed = FALSE)
object <- diffuse_seed_nodes(object, c(node_1 = 1, node_3 = 1), "max")
calculate_diffusion_auc(object, hit_nodes = c("node_2", "node_4"))
}
