% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{calculate_effect_size}
\alias{calculate_effect_size}
\title{Calculate the Hedge's G effect between two matrices}
\usage{
calculate_effect_size(
  mat_a,
  mat_b,
  small_sample_correction = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{mat_a}{Numerical matrix. Contains the values for group a. Assumes that
rows = samples, and columns = features.}

\item{mat_b}{Numerical matrix. Contains the values for group b.}

\item{small_sample_correction}{Can be NULL (automatic determination if a
small sample size correction should be applied) or Boolean.}

\item{.verbose}{Boolean that controls verbosity of the function.}
}
\value{
x, robustly scaled.
}
\description{
This function takes two matrices in and calculate on a per column basis the
Hedge's G effect size and the standard error. These results can be
subsequently used for meta-analyses or other approaches.
}
\examples{
# Hedge's G between two groups of ten samples
set.seed(42)
mat_a <- matrix(rnorm(100), nrow = 10, ncol = 10)
mat_b <- matrix(rnorm(100, mean = 1), nrow = 10, ncol = 10)
colnames(mat_a) <- colnames(mat_b) <- sprintf("gene_\%i", 1:10)
res <- calculate_effect_size(mat_a, mat_b, .verbose = FALSE)
head(res$effect_sizes)
}
