% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ontology.R
\name{calculate_information_content}
\alias{calculate_information_content}
\title{Calculate the information content for each ontology term}
\usage{
calculate_information_content(ancestor_list)
}
\arguments{
\item{ancestor_list}{List. Named list of terms with ancestor identifiers as
their values}
}
\value{
A named list of each term and their information content as values.
}
\description{
this function will calculate the information content of each
provided term based on a list of ancestors, which is a named list of terms
with ancestor identifiers as their values. Can be calculated using
\code{\link[=get_ontology_ancestry]{get_ontology_ancestry()}}. The information content is calculated
as \verb{-log2(number descendant/total terms in the ontology)}. More information
can be found \href{https://yulab-smu.top/biomedical-knowledge-mining-book/semantic-similarity-overview.html}{here}.
}
\examples{
# information content from the descendant counts of a toy ontology
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f")
)
ancestry <- get_ontology_ancestry(onto)
unlist(calculate_information_content(ancestry$descendants))
}
