% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_multimodal.R
\name{calculate_pca_adt_sc}
\alias{calculate_pca_adt_sc}
\title{Calculate the PCA on top of the normalised ADT counts}
\usage{
calculate_pca_adt_sc(
  object,
  no_pcs,
  features = NULL,
  randomised_svd = FALSE,
  seed = 42L
)
}
\arguments{
\item{object}{\code{SingleCellsMultiModal} class with ADT counts added.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{features}{Optional string vector. If you want to subset to a specific
set of ADT probes (for example to exclude isotypes).}

\item{randomised_svd}{Boolean. Shall randomised SVD be used. Faster, but
less precise.}

\item{seed}{Integer. Controls reproducibility. Only relevant if
\code{randomised_svd = TRUE}.}
}
\value{
The function will add the PCA factors, loadings and singular values
for the ADT data to the object.
}
\description{
This function will run PCA - via (randomised) SVD - on the normalised counts
and add the PCA results to the ScCache for the ADT counts.
}
\examples{
# PCA over the CLR-normalised protein counts
rna <- generate_single_cell_test_data()
adt <- generate_single_cell_test_data_adt()
dir <- tempfile("bixverse_mm")
dir.create(dir)
object <- load_r_data(
  SingleCellsMultiModal(dir_data = dir),
  counts = rna$counts,
  obs = rna$obs,
  var = rna$var,
  sc_qc_param = params_sc_min_quality(min_unique_genes = 5L),
  .verbose = FALSE
)
object <- add_adt_counts_sc(object, adt_counts = adt$counts, method = "clr")
object <- calculate_pca_adt_sc(object, no_pcs = 10L)
dim(get_pca_factors(object, modality = "adt"))

unlink(dir, recursive = TRUE, force = TRUE)
}
