% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ontology.R
\name{calculate_semantic_sim}
\alias{calculate_semantic_sim}
\title{Calculate the Resnik or Lin semantic similarity}
\usage{
calculate_semantic_sim(
  terms,
  similarity_type,
  ancestor_list,
  ic_list,
  add_self = FALSE
)
}
\arguments{
\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}

\item{similarity_type}{String. One of \code{c("resnik", "lin", "combined")}.}

\item{ancestor_list}{List. Names being the term and the elements in the
list the names of the ancestors, see \code{\link[=get_ontology_ancestry]{get_ontology_ancestry()}}.}

\item{ic_list}{List. The names being the term and the elements the
information content of this given term. Needs to be a single float! See
\code{\link[=calculate_information_content]{calculate_information_content()}}.}

\item{add_self}{Boolean. Shall self-similarities be added. Defaults to
\code{FALSE}.}
}
\value{
A data.table with the calculated similarities.
}
\description{
This function calculates the semantic similarities based on
Resnik or Lin similarity for a given ontology. Has also the option to
calculate a combined version of the two.
}
\examples{
# Lin similarity for a subset of terms only
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f")
)
ancestry <- get_ontology_ancestry(onto)
ic <- calculate_information_content(ancestry$descendants)
calculate_semantic_sim(
  terms = c("c", "d", "f"),
  similarity_type = "lin",
  ancestor_list = ancestry$ancestors,
  ic_list = ic
)
}
