% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{calculate_semantic_sim_onto}
\alias{calculate_semantic_sim_onto}
\title{Calculate the Resnik or Lin semantic similarity for an ontology.}
\usage{
calculate_semantic_sim_onto(
  object,
  sim_type = c("resnik", "lin", "combined"),
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{OntologySim} class. See \code{\link[=OntologySim]{OntologySim()}}.}

\item{sim_type}{String. One of \code{c("resnik", "lin", "combined")}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
The class with added semantic similarities to the properties.
}
\description{
This function calculates the specified semantic similarities for
the whole ontology and adds it to the class.
}
\examples{
# Resnik similarities for the whole ontology stored in the class
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f"),
  type = c("part_of", "part_of", "part_of", "is_a", "is_a")
)
onto_obj <- pre_process_sim_onto(
  OntologySim(onto, .verbose = FALSE),
  .verbose = FALSE
)
calculate_semantic_sim_onto(onto_obj, sim_type = "resnik", .verbose = FALSE)
}
