% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ontology.R
\name{calculate_wang_sim}
\alias{calculate_wang_sim}
\title{Calculate the Wang similarities between terms}
\usage{
calculate_wang_sim(terms, parent_child_dt, weights, add_self = FALSE)
}
\arguments{
\item{terms}{String vector. The terms for which to calculate the Wang
similarity.}

\item{parent_child_dt}{data.table. The data.table with column parent and
child. You also need to have a type column for the Wang similarity to provide
the weights for the relationships.}

\item{weights}{Named numeric. The relationship of type to weight for this
specific edge. For example \code{c("part_of" = 0.8, "is_a" = 0.6)}.}

\item{add_self}{Boolean. Shall self-similarities be added. Defaults to
\code{FALSE}.}
}
\value{
A data.table with the calculated similarities.
}
\description{
This function calculates the Wang similarity for a set of temrs,
based on the DAG for a given ontology.
}
\examples{
# Wang similarity for a subset of terms only
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f"),
  type = c("part_of", "part_of", "part_of", "is_a", "is_a")
)
weights <- c("part_of" = 0.8, "is_a" = 0.6)
calculate_wang_sim(c("c", "d", "f"), onto, weights = weights)
}
