% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ontology.R
\name{calculate_wang_sim_mat}
\alias{calculate_wang_sim_mat}
\title{Calculate the Wang similarity matrix}
\usage{
calculate_wang_sim_mat(parent_child_dt, weights)
}
\arguments{
\item{parent_child_dt}{data.table. The data.table with column parent and
child. You also need to have a type column for the Wang similarity to provide
the weights for the relationships.}

\item{weights}{Named numeric. The relationship of type to weight for this
specific edge. For example \code{c("part_of" = 0.8, "is_a" = 0.6)}.}
}
\value{
The symmetric Wang similarity matrix.
}
\description{
This function calculates the Wang similarity, based on the DAG
for a given ontology. This function will return the full similarity matrix.
}
\examples{
# Wang similarity matrix with relationship-specific weights
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f"),
  type = c("part_of", "part_of", "part_of", "is_a", "is_a")
)
weights <- c("part_of" = 0.8, "is_a" = 0.6)
round(calculate_wang_sim_mat(onto, weights = weights), 3)
}
