% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{calculate_wang_sim_onto}
\alias{calculate_wang_sim_onto}
\title{Calculate the Wang similarity for an ontology.}
\usage{
calculate_wang_sim_onto(object, weights, .verbose = TRUE)
}
\arguments{
\item{object}{\code{OntologySim} class. See \code{\link[=OntologySim]{OntologySim()}}.}

\item{weights}{Named numeric. The relationship of type to weight for this
specific edge. For example \code{c("part_of" = 0.8, "is_a" = 0.6)}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
The class with added semantic similarities to the properties.
}
\description{
This function calculates the Wang similarity for the whole
ontology and adds it to the class in a memory-efficient format for
subsequent usage.
}
\examples{
# Wang similarities for the whole ontology stored in the class
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f"),
  type = c("part_of", "part_of", "part_of", "is_a", "is_a")
)
onto_obj <- OntologySim(onto, .verbose = FALSE)
calculate_wang_sim_onto(
  onto_obj,
  weights = c(part_of = 0.8, is_a = 0.6),
  .verbose = FALSE
)
}
