% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_cycle_genes}
\alias{cell_cycle_genes}
\title{Cell cycle genes}
\format{
\subsection{\code{cell_cycle_genes}}{

A data.table
\describe{
\item{hgnc_symbol}{The (HGNC) gene symbol}
\item{ensembl_gene_id}{The corresponding ensembl identifiers.}
\item{set}{To which cell cycle phase the gene belongs}
}
}
}
\usage{
cell_cycle_genes
}
\description{
(Human) cell cycle genes for scoring in single cell, please see
\code{\link[=module_scores_sc]{module_scores_sc()}} and Tirosh et al, Science (2016).
}
\keyword{datasets}
