% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{check_dim}
\alias{check_dim}
\title{Helper to assert same number of samples}
\usage{
check_dim(object, no_samples)
}
\arguments{
\item{object}{\code{snf} class.}

\item{no_samples}{Integer. New number of samples}
}
\value{
The object. If no dimensions were found in the object yet, it will
be added. If the addition of the new data has dimensions that do not fit
with what has been stored in the class so far, it will throw an error.
}
\description{
Helper to assert same number of samples
}
