% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkBinaryHeatmapParams}
\alias{checkBinaryHeatmapParams}
\title{Check binary heatmap params}
\usage{
checkBinaryHeatmapParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_binary_heatmap]{params_binary_heatmap()}}.
}
\keyword{internal}
