% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkCellMarkerList}
\alias{checkCellMarkerList}
\title{Check cell marker list}
\usage{
checkCellMarkerList(x)
}
\arguments{
\item{x}{The list to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking a cell marker list as returned
by \code{\link[=prepare_cell_markers]{prepare_cell_markers()}}.
}
\keyword{internal}
