% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkDialogueHlm}
\alias{checkDialogueHlm}
\title{Check DIALOGUE mixed model parameters}
\usage{
checkDialogueHlm(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the stage two DIALOGUE
parameters, see \code{\link[=params_dialogue_hlm]{params_dialogue_hlm()}}.
}
\keyword{internal}
