% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkDialogueHlmParams}
\alias{checkDialogueHlmParams}
\title{Check DIALOGUE mixed model params}
\usage{
checkDialogueHlmParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_dialogue_hlm]{params_dialogue_hlm()}}.
}
\keyword{internal}
