% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkDialoguePmd}
\alias{checkDialoguePmd}
\title{Check DIALOGUE decomposition parameters}
\usage{
checkDialoguePmd(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the stage one DIALOGUE
parameters, see \code{\link[=params_dialogue_pmd]{params_dialogue_pmd()}}.
}
\keyword{internal}
