% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkDialogueRefine}
\alias{checkDialogueRefine}
\title{Check DIALOGUE refinement parameters}
\usage{
checkDialogueRefine(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the stage three DIALOGUE
parameters, see \code{\link[=params_dialogue_refine]{params_dialogue_refine()}}.
}
\keyword{internal}
