% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkEdgeRQlParams}
\alias{checkEdgeRQlParams}
\title{Check edgeR QL params}
\usage{
checkEdgeRQlParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_edger_ql]{params_edger_ql()}}.
}
\keyword{internal}
