% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkFastClusterDefaultParams}
\alias{checkFastClusterDefaultParams}
\title{Check FastCluster params}
\usage{
checkFastClusterDefaultParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_fast_cluster_default]{params_fast_cluster_default()}}.
}
\keyword{internal}
