% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkFilesExist}
\alias{checkFilesExist}
\title{Check that files exist}
\usage{
checkFilesExist(x, file_names)
}
\arguments{
\item{x}{String. The directory to check the files for.}

\item{file_names}{String. Vector of names of the expected files in this
directory.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking if files exist in the
directory.
}
