% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkIcaIterParams}
\alias{checkIcaIterParams}
\title{Check ICA randomisation params}
\usage{
checkIcaIterParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_ica_randomisation]{params_ica_randomisation()}}.
}
\keyword{internal}
