% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkIcaNcomps}
\alias{checkIcaNcomps}
\title{Check ICA no of component parameters}
\usage{
checkIcaNcomps(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the ICA number of component
parameters.
}
