% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkKMeansParams}
\alias{checkKMeansParams}
\title{Check k-means params}
\usage{
checkKMeansParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_kmeans_defaults]{params_kmeans_defaults()}}.
}
\keyword{internal}
