% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkLigandTarget}
\alias{checkLigandTarget}
\title{Check ligand-target influence parameters}
\usage{
checkLigandTarget(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error
message.
}
\description{
Checkmate extension for checking the ligand to target
influence parameters.
}
\keyword{internal}
