% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkLigandTargetParams}
\alias{checkLigandTargetParams}
\title{Check ligand-target params}
\usage{
checkLigandTargetParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_ligand_target]{params_ligand_target()}}.
}
\keyword{internal}
