% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkLimmaVoomParams}
\alias{checkLimmaVoomParams}
\title{Check limma-voom params}
\usage{
checkLimmaVoomParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_limma_voom]{params_limma_voom()}}.
}
\keyword{internal}
