% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkLrDb}
\alias{checkLrDb}
\title{Check a ligand-receptor database}
\usage{
checkLrDb(x)
}
\arguments{
\item{x}{The data.table to check. Needs \code{interaction_id}, \code{partner_a} and
\code{partner_b} as character columns and \code{genes_a} and \code{genes_b} as list
columns of non-empty character vectors (the subunits).}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking a ligand-receptor database
as returned by \code{\link[=get_cellphonedb_db]{get_cellphonedb_db()}}.
}
\keyword{internal}
