% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkModuleMembershipParams}
\alias{checkModuleMembershipParams}
\title{Check module membership params}
\usage{
checkModuleMembershipParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_module_membership]{params_module_membership()}}.
}
\keyword{internal}
