% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkNmfConsensus}
\alias{checkNmfConsensus}
\title{Check consensus NMF parameters}
\usage{
checkNmfConsensus(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the consensus NMF parameters.
}
\keyword{internal}
