% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkNmfConsensusParams}
\alias{checkNmfConsensusParams}
\title{Check NMF consensus params}
\usage{
checkNmfConsensusParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_nmf_consensus]{params_nmf_consensus()}}.
}
\keyword{internal}
