% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkNmfHals}
\alias{checkNmfHals}
\title{Check NMF HALS parameters}
\usage{
checkNmfHals(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the NMF HALS parameters for
single cell.
}
\keyword{internal}
