% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkNmfHalsParams}
\alias{checkNmfHalsParams}
\title{Check NMF HALS params}
\usage{
checkNmfHalsParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_nmf_hals]{params_nmf_hals()}}.
}
\keyword{internal}
