% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScBbknnParams}
\alias{checkScBbknnParams}
\title{Check BBKNN params}
\usage{
checkScBbknnParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_sc_bbknn]{params_sc_bbknn()}}.
}
\keyword{internal}
