% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScBranchSelectionParams}
\alias{checkScBranchSelectionParams}
\title{Check branch selection params}
\usage{
checkScBranchSelectionParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_sc_branch_selection]{params_sc_branch_selection()}}.
}
\keyword{internal}
