% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScCellPhoneDbParams}
\alias{checkScCellPhoneDbParams}
\title{Check CellPhoneDB params}
\usage{
checkScCellPhoneDbParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_sc_cellphonedb]{params_sc_cellphonedb()}}.
}
\keyword{internal}
