% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScCellsweep}
\alias{checkScCellsweep}
\title{Check CellSweep parameters}
\usage{
checkScCellsweep(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the CellSweep parameters.
}
\keyword{internal}
