% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScDblFinder}
\alias{checkScDblFinder}
\title{Check scDblFinder parameters}
\usage{
checkScDblFinder(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking scDblFinder parameters.
}
\keyword{internal}
