% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScenicBinariseParams}
\alias{checkScenicBinariseParams}
\title{Check SCENIC binarisation params}
\usage{
checkScenicBinariseParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_scenic_binarise]{params_scenic_binarise()}}.
}
\keyword{internal}
