% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScHarmonyParamsV2}
\alias{checkScHarmonyParamsV2}
\title{Check Harmony v2 parameters}
\usage{
checkScHarmonyParamsV2(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking Harmony v2 parameters.
}
\keyword{internal}
