% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScHarmonyV2Params}
\alias{checkScHarmonyV2Params}
\title{Check Harmony v2 params}
\usage{
checkScHarmonyV2Params(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_sc_harmony_v2]{params_sc_harmony_v2()}}.
}
\keyword{internal}
