% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScKnn}
\alias{checkScKnn}
\title{Check KNN generation parameters}
\usage{
checkScKnn(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the KNN generation parameters
for single cell.
}
