% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScMetacells}
\alias{checkScMetacells}
\title{Check metacell generation parameters}
\usage{
checkScMetacells(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the metacell generation
parameters.
}
