% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScMinQC}
\alias{checkScMinQC}
\title{Check SC minimum QC parameters}
\usage{
checkScMinQC(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the minimum QC parameters
in single cell.
}
