% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScMinQCParams}
\alias{checkScMinQCParams}
\title{Check single cell QC params}
\usage{
checkScMinQCParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}.
}
\keyword{internal}
