% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScMtxIO}
\alias{checkScMtxIO}
\title{Check SC MTX load parameters}
\usage{
checkScMtxIO(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking MTX loading parameters
}
