% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScMtxIOParams}
\alias{checkScMtxIOParams}
\title{Check MTX IO params}
\usage{
checkScMtxIOParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_sc_mtx_io]{params_sc_mtx_io()}}.
}
\keyword{internal}
