% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScNeighbours}
\alias{checkScNeighbours}
\title{Check neighbour generation parameters}
\usage{
checkScNeighbours(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the neighbour generation
parameters for single cell.
}
