% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScPca}
\alias{checkScPca}
\title{Check PCA parameters for single cell}
\usage{
checkScPca(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the PCA parameters for single
cell.
}
\keyword{internal}
