% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScScrubletParams}
\alias{checkScScrubletParams}
\title{Check Scrublet params}
\usage{
checkScScrubletParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_scrublet]{params_scrublet()}}.
}
\keyword{internal}
