% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScSeuratCca}
\alias{checkScSeuratCca}
\title{Check Seurat CCA parameters}
\usage{
checkScSeuratCca(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the Seurat CCA parameters.
}
\keyword{internal}
