% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScSyntheticCellsweep}
\alias{checkScSyntheticCellsweep}
\title{Check synthetic CellSweep data parameters}
\usage{
checkScSyntheticCellsweep(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the parameters for the
generation of synthetic CellSweep data.
}
\keyword{internal}
