% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScSyntheticCellsweepParams}
\alias{checkScSyntheticCellsweepParams}
\title{Check synthetic CellSweep params}
\usage{
checkScSyntheticCellsweepParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_sc_synthetic_cellsweep]{params_sc_synthetic_cellsweep()}}.
}
\keyword{internal}
