% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScSyntheticDataAdt}
\alias{checkScSyntheticDataAdt}
\title{Check synthetic ADT data parameters}
\usage{
checkScSyntheticDataAdt(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the synthetic ADT data
parameters.
}
\keyword{internal}
