% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkScSyntheticDataAdtParams}
\alias{checkScSyntheticDataAdtParams}
\title{Check synthetic ADT params}
\usage{
checkScSyntheticDataAdtParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_sc_synthetic_data_adt]{params_sc_synthetic_data_adt()}}.
}
\keyword{internal}
