% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScSyntheticDialogue}
\alias{checkScSyntheticDialogue}
\title{Check synthetic DIALOGUE data parameters}
\usage{
checkScSyntheticDialogue(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the parameters for the
generation of synthetic DIALOGUE data.
}
\keyword{internal}
