% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkSctypeCellParams}
\alias{checkSctypeCellParams}
\title{Check ScType cell params}
\usage{
checkSctypeCellParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_sctype_cells]{params_sctype_cells()}}.
}
\keyword{internal}
