% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpatialSample}
\alias{checkSpatialSample}
\title{Check that an object is a valid SpatialSample}
\usage{
checkSpatialSample(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message
string.
}
\description{
Checkmate extension for the \code{SpatialSample} S3 class built by
\code{\link[=new_spatial_sample]{new_spatial_sample()}}.
}
\keyword{internal}
