% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpGraph}
\alias{checkSpGraph}
\title{Check spatial graph parameters}
\usage{
checkSpGraph(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the spatial graph parameters
built by \code{\link[=params_sp_graph]{params_sp_graph()}}.
}
\keyword{internal}
