% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpH5adIo}
\alias{checkSpH5adIo}
\title{Check spatial h5ad ingest parameters}
\usage{
checkSpH5adIo(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the spatial h5ad ingest
parameters built by \code{\link[=params_sp_h5ad_io]{params_sp_h5ad_io()}}.
}
\keyword{internal}
