% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpImage}
\alias{checkSpImage}
\title{Check histology image feature parameters}
\usage{
checkSpImage(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the image feature parameters
built by \code{\link[=params_sp_image]{params_sp_image()}}.
}
\keyword{internal}
