% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpNhood}
\alias{checkSpNhood}
\title{Check neighbourhood enrichment parameters}
\usage{
checkSpNhood(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the neighbourhood enrichment
parameters built by \code{\link[=params_sp_nhood]{params_sp_nhood()}}.
}
\keyword{internal}
