% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpSparkx}
\alias{checkSpSparkx}
\title{Check SPARK-X parameters}
\usage{
checkSpSparkx(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the SPARK-X parameters built by
\code{\link[=params_sp_sparkx]{params_sp_sparkx()}}.
}
\keyword{internal}
