% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpSvg}
\alias{checkSpSvg}
\title{Check Moran's I parameters}
\usage{
checkSpSvg(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the Moran's I parameters built
by \code{\link[=params_sp_svg]{params_sp_svg()}}.
}
\keyword{internal}
