% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSpVisiumIo}
\alias{checkSpVisiumIo}
\title{Check Visium ingest parameters}
\usage{
checkSpVisiumIo(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the Visium ingest parameters
built by \code{\link[=params_sp_visium_io]{params_sp_visium_io()}}.
}
\keyword{internal}
