% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSymphonyMap}
\alias{checkSymphonyMap}
\title{Check Symphony map parameters}
\usage{
checkSymphonyMap(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the Symphony query mapping
parameters.
}
\keyword{internal}
