% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkSyntheticBulkParams}
\alias{checkSyntheticBulkParams}
\title{Check synthetic bulk params}
\usage{
checkSyntheticBulkParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of
\code{\link[=params_synthetic_bulk_rnaseq]{params_synthetic_bulk_rnaseq()}}.
}
\keyword{internal}
