% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{component_mutual_information}
\alias{component_mutual_information}
\title{Assess the NMI of the ICA components}
\usage{
component_mutual_information(centrotype)
}
\arguments{
\item{centrotype}{Numeric matrix. The calculated centrotypes for a given
number of component.}
}
\value{
Returns the average normalised mutual information across all
components
}
\description{
Assesses the normalised mutual information of the components.
}
