% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{component_orthogonality}
\alias{component_orthogonality}
\title{Assess the orthogonality of the ICA components}
\usage{
component_orthogonality(centrotype)
}
\arguments{
\item{centrotype}{Numeric matrix. The calculated centrotypes for a given
number of component.}
}
\value{
Returns the average orthogonality between all the pairs.
}
\description{
Assesses the orthogonality of the discovered independent components. It leverages
the cosine similarity to calculate this. Orthogonality is defined as:
\code{1 - abs(cos) / length(cos)}.
}
