% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_c_pca.R
\name{contrastive_pca_processing}
\alias{contrastive_pca_processing}
\title{Prepare class for contrastive PCA}
\usage{
contrastive_pca_processing(
  object,
  background_matrix,
  scale = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}

\item{background_matrix}{Numeric matrix. The background matrix you wish to
remove. You should apply any data transformation to this matrix, too!}

\item{scale}{Boolean. Shall the data be scaled. Defaults to FALSE.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{BulkCoExp} with the needed data for contrastive PCA in the
properties of the class.
}
\description{
This function will prepare the \code{BulkCoExp} for subsequent usage of the
contrastive PCA functions. This is based on the work of Abid, et al.
}
\examples{
# covariance matrices of the target and the background
cpca_data <- synthetic_c_pca_data()
target <- t(cpca_data$target)
background <- t(cpca_data$background)
meta <- data.table::data.table(sample_id = rownames(target))
obj <- BulkCoExp(target, meta)
obj <- preprocess_bulk_coexp(obj, .verbose = FALSE)
obj <- contrastive_pca_processing(obj, background, .verbose = FALSE)
dim(obj@processed_data$target_covar)
}
\references{
Abid, et al., Nature Communications, 2018
}
