% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{cor_cluster_quality}
\alias{cor_cluster_quality}
\title{Assess the quality of clusters}
\usage{
cor_cluster_quality(community_df, correlation_res)
}
\arguments{
\item{community_df}{A data.table. Contains node_name (name of the node in the
graph) and membership (membership of the node to the respective community).}

\item{correlation_res}{A data.table. Contains feature_a (first node name),
feature_b (second node name) and the r_abs (absolute correlation coefficient)
columns.}
}
\value{
A data.table with the followoing columns:
}
\description{
Helper function that will assess the quality (defined as the
median absolute correlation within the cluster) of a given set of clusters.
}
