% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{cor_graph_params}
\alias{cor_graph_params}
\title{Wrapper function for graph generation}
\usage{
cor_graph_params(
  kernel_bandwidth = 0.2,
  min_affinity = 0.001,
  min_cor = 0.2,
  fdr_threshold = 0.05
)
}
\arguments{
\item{kernel_bandwidth}{Float. Defines the bandwidth of the Gaussian
kernel to generate the affinities. Relevant for single correlation-based
graphs.}

\item{min_affinity}{Float. Minimum affinity before the edge gets dropped.
Relevant for single correlation-based graphs.}

\item{min_cor}{Float. Minimum absolute correlation that needs to be
observed in either data set. Only relevant for differential correlation-based
graphs.}

\item{fdr_threshold}{Float. Maximum FDR for the differential correlation
p-value.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function for graph generation
}
