% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_check_epsilon}
\alias{cor_module_check_epsilon}
\title{Iterate through different epsilon parameters}
\usage{
cor_module_check_epsilon(
  object,
  rbf_func = c("bump", "gaussian", "inverse_quadratic"),
  epsilons = c(0.25, seq(from = 0.5, to = 10, by = 0.5)),
  .verbose = TRUE
)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}. You need to run
\code{\link[=cor_module_processing]{cor_module_processing()}} before running this function.}

\item{rbf_func}{The type of RBF function you want to apply. A choice of
\code{c('bump', 'gaussian', 'inverse_quadratic')}.}

\item{epsilons}{Vector of floats. The different epsilon parameters you
would like to run.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added data to the properties for subsequent usage.
}
\description{
This functions iterates through a set of provided epsilons and checks for
each one to what extend the underlying affinity matrix will follow a power
law distribution.
}
\examples{
# scan epsilons of a Gaussian RBF for scale-free topology
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- cor_module_processing(obj, cor_method = "spearman", .verbose = FALSE)
obj <- cor_module_check_epsilon(
  obj, rbf_func = "gaussian", .verbose = FALSE
)
head(get_epsilon_res(obj))
}
