% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_coremo_cor_sign}
\alias{cor_module_coremo_cor_sign}
\title{Split CoReMo modules by correlation sign}
\usage{
cor_module_coremo_cor_sign(object, min_stability = NULL, .verbose = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}

\item{min_stability}{Optional float. The minimum stability for the gene
you wish to filter for based on the leave-one-out resampling. If \code{NULL},
no filtering will be applied.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with updated correlation module names.
}
\description{
This function will split the identified modules by their correlation sign.
In certain cases, positive and negatively correlated genes can be part of
the same module. This function will annotate them with \code{"_pos"} and \code{"_neg"}
respectively.
}
\examples{
# split the modules into positively and negatively correlated genes
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- cor_module_processing(obj, cor_method = "spearman", .verbose = FALSE)
obj <- cor_module_coremo_clustering(obj, .verbose = FALSE)
obj <- cor_module_coremo_cor_sign(obj, .verbose = FALSE)
table(obj@outputs$final_modules$sign)
}
\references{
Srivastava, et al., Nat. Commun., 2018; Francois, Romagnolo,
et al., Nat. Commun., 2024.
}
