% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_coremo_eigengene}
\alias{cor_module_coremo_eigengene}
\title{Calculate Eigengenes for CoReMo modules}
\usage{
cor_module_coremo_eigengene(object, min_stability = NULL, .verbose = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}

\item{min_stability}{Optional float. The minimum stability for the gene
you wish to filter for based on the leave-one-out resampling. If \code{NULL},
no filtering will be applied.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added correlations to the modules and the values for
a given eigengene per sample as a data.table.
}
\description{
This function will calculate the eigengene values for the modules on a per
sample basis and add correlations of the gene expression of a given gene
within the module to its eigengene.
}
\examples{
# eigengenes per module and the gene to eigengene correlations
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- cor_module_processing(obj, cor_method = "spearman", .verbose = FALSE)
obj <- cor_module_coremo_clustering(obj, .verbose = FALSE)
obj <- cor_module_coremo_eigengene(obj, .verbose = FALSE)
head(get_modules(get_results(obj)))
}
\references{
Srivastava, et al., Nat. Commun., 2018; Francois, Romagnolo,
et al., Nat. Commun., 2024.
}
