% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_coremo_stability}
\alias{cor_module_coremo_stability}
\title{Assesses CoReMo-based gene module stability}
\usage{
cor_module_coremo_stability(object, chunk_size = 15L, .verbose = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}

\item{chunk_size}{Integer. Chunk size in which to process the data. Defaults
to \code{15L}, i.e., 15 samples are being processed in one go. You can use bigger
values here, but be aware of memory pressure.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added data to the properties for subsequent usage.
}
\description{
The function assesses the stability of the CoReMo modules, leveraging a
leave-one-out sample method. In each iteration, one of the samples is left
out and the clustering is redone. Subsequently, the Jaccard similarity (as
a surrogate for membership stability) is calculated for each feature across
the different resamplings and added to the \code{final_modules} data.table.
}
\examples{
# leave-one-out stability of the CoReMo modules
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- cor_module_processing(obj, cor_method = "spearman", .verbose = FALSE)
obj <- cor_module_coremo_clustering(obj, .verbose = FALSE)
obj <- cor_module_coremo_stability(obj, .verbose = FALSE)
summary(obj@outputs$final_modules$stability)
}
\references{
Srivastava, et al., Nat. Commun., 2018; Francois, Romagnolo,
et al., Nat. Commun., 2024.
}
