% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_final_modules}
\alias{cor_module_final_modules}
\title{Identify correlation-based gene modules via graphs.}
\usage{
cor_module_final_modules(
  object,
  resolution = NULL,
  min_size = 10L,
  max_size = 500L,
  subclustering = TRUE,
  random_seed = 123L,
  .graph_params = params_cor_graph(),
  .max_iters = 100L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{resolution}{The Leiden resolution parameter you wish to use. If NULL,
it will use the optimal one identified by \code{\link[=cor_module_check_res]{cor_module_check_res()}}.
If nothing can be found, will default to 1.}

\item{min_size}{Integer. Minimum size of the communities.}

\item{max_size}{Integer. Maximum size of the communities.}

\item{subclustering}{Boolean. Shall after a first clustering communities that
are too large be further sub clustered. Defaults to \code{TRUE}.}

\item{random_seed}{Integer. Random seed.}

\item{.graph_params}{List. Parameters for the generation of the (differential)
correlation graph, see \code{\link[=params_cor_graph]{params_cor_graph()}}. Contains:
\itemize{
\item Epsilon - Defines the epsilon parameter for the radial basis
function. Defaults to 2, but should be ideally optimised.
\item min_cor - Float. Minimum absolute correlation that needs to be
observed in either data set. Only relevant for differential correlation-based
graphs.
\item fdr_threshold - Float. Maximum FDR for the differential correlation
p-value.
\item verbose - Boolean. Controls verbosity of the graph generation.
}
This parameter is only relevant if you did \emph{not} run
\code{\link[=cor_module_check_res]{cor_module_check_res()}}.}

\item{.max_iters}{Integer. If sub clustering is set to \code{TRUE}, what shall be the
maximum number of iterations. Defaults to 100L.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
The class with added data to the properties.
}
\description{
This function leverages graph-based clustering to identify gene co-expression
modules. The class has the option to sub-cluster large communities within
their respective sub graphs, akin to the approach taken by Barrio-Hernandez,
et al.
}
\references{
Barrio-Hernandez, et al., Nat Genet, 2023.
}
