% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{cor_module_identification}
\alias{cor_module_identification}
\title{Identify correlation-based gene modules via graph methods.}
\usage{
cor_module_identification(bulk_coexp, ...)

...
}
\arguments{
\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} class. You need to run either
\code{\link[=cor_module_processing]{cor_module_processing()}} or ... .
}
\description{
This is the generic function for identifying gene modules via graph-based
methods.

...
}
