% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_processing}
\alias{cor_module_processing}
\title{Prepare correlation-based module detection}
\usage{
cor_module_processing(
  object,
  cor_method = c("pearson", "spearman"),
  .verbose = TRUE
)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}. Ideally, you
should run \code{\link[=preprocess_bulk_coexp]{preprocess_bulk_coexp()}} before applying this function.}

\item{cor_method}{String. Option of \code{c("pearson", "spearman")}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added data to the properties for subsequent usage.
}
\description{
This function will calculate the correlation coefficients between the genes,
using the highly variable genes (if available, otherwise the function will
use the raw data). The data will be stored in a memory-efficient format
in the properties of the class.
}
\examples{
# spearman correlations over the 300 most variable genes
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- cor_module_processing(obj, cor_method = "spearman", .verbose = FALSE)
dim(obj@processed_data$correlation_res$get_sym_matrix(
  .verbose = FALSE
))
}
