% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{cor_module_resolutions}
\alias{cor_module_resolutions}
\title{Identify correlation-based gene modules via graph methods.}
\usage{
cor_module_resolutions(bulk_coexp, ...)

...
}
\arguments{
\item{bulk_coexp}{description}

\item{kernel_bandwidth}{description}

\item{min_res}{description}

\item{max_res}{description}

\item{by}{description}

\item{random_seed}{description}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} class. You need to run either
\code{\link[=cor_module_processing]{cor_module_processing()}} or ... .
}
\description{
This is the generic function for identifying gene modules via graph-based
methods.

...
}
