% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{coremo_cluster_quality}
\alias{coremo_cluster_quality}
\title{Measures the quality of the clusters}
\usage{
coremo_cluster_quality(modules, cor_mat, random_seed = 10101L)
}
\arguments{
\item{modules}{Named vector. The names reflect the gene of the associated
module.}

\item{cor_mat}{Numeric matrix. The original correlation matrix.}

\item{random_seed}{Integer. Random seed to ensure consistency if sampling is
used.}
}
\value{
A data.table with the quality measures of the cluster.
}
\description{
Utility functions to give back the summary stats (median and mean R²) for
each of the identified clusters.
}
