% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_methods_cPCA.R
\name{cPCA_preprocessing}
\alias{cPCA_preprocessing}
\title{Pre-process the transcriptomics data for cPCA}
\usage{
cPCA_preprocessing(bulk_coexp, ...)
}
\arguments{
\item{bulk_coexp}{\code{bulk_coexp} class.}

\item{background_matrix}{Numeric matrix.}

\item{scale}{Boolean. Shall the data be scaled. Defaults to FALSE.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} with additional data in the slots
}
\description{
...
}
\references{
Abid, et al., Nature Communications, 2018
}
