% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{create_dist_obj}
\alias{create_dist_obj}
\title{Create distance object from a vector}
\usage{
create_dist_obj(x, size)
}
\arguments{
\item{x}{Numerical vector. The upper-triangle values for which to generate
the \code{dist} object.}

\item{size}{Integer. Nrow (or ncol) of the symmetric matrix.}
}
\value{
Returns the distance object
}
\description{
Create distance object from a vector
}
