% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{create_sparse_matrix}
\alias{create_sparse_matrix}
\title{Helper function transform Rust counts into sparse matrices}
\usage{
create_sparse_matrix(count_data, return_format)
}
\arguments{
\item{count_data}{A list. Output of \code{get_counts_from_rust()}.}

\item{return_format}{String. One of \code{c("cell", "gene")}.}
}
\value{
The sparse matrix in CSR or CSC format, pending the choice in
\code{return_format}
}
\description{
Helper function transform Rust counts into sparse matrices
}
\keyword{internal}
