% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{create_synthetic_cPCA_data}
\alias{create_synthetic_cPCA_data}
\title{Generates synthetic data for cPCA exploration.}
\usage{
create_synthetic_cPCA_data(seed = 10101L)
}
\arguments{
\item{seed}{Integer. Initial random seed for generation of the synthetic data. Default: 10101L.}
}
\value{
A list. First elements is the target matrix, second element the background matrix and the
third one being a factor with the cluster labels of the target matrix.
}
\description{
Generates three elements: target matrix, background matrix
and labels for the target matrix.
}
