% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_diff_class_methods.R
\name{diffuse_seed_genes}
\alias{diffuse_seed_genes}
\title{Diffuse seed genes over network}
\usage{
diffuse_seed_genes(
 network_diffusions,
 diffusion_vector,
 summarisation = c("max", "mean", "harmonic_sum")
)
}
\arguments{
\item{network_diffusions}{The underlying \code{network_diffusions} class.}

\item{diffusion_vector}{A named vector with values to use for the reset parameter in the
personalised page-rank diffusion. Names should represent node names of the graph.}

\item{summarisation}{If there are duplicated names in the \code{diffusion_vector} how to summarise
these.}
}
\description{
This function takes a diffusion vector and leverages personalised page-rank diffusion
to identify influential nodes. These can be used subsequently for community detection
or check AUROC values given a set of genes.
}
