% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{diffuse_seed_nodes}
\alias{diffuse_seed_nodes}
\title{Diffuse seed genes over a network}
\usage{
diffuse_seed_nodes(
  object,
  diffusion_vector,
  summarisation = c("max", "mean", "harmonic_sum")
)
}
\arguments{
\item{object}{\code{NetworkDiffusions} object. The underlying class
\code{\link[=NetworkDiffusions]{NetworkDiffusions()}}.}

\item{diffusion_vector}{Named nuermic. A named vector with values to use for
the reset parameter in the personalised page-rank diffusion. Names should
represent node names of the graph.}

\item{summarisation}{String. If there are duplicated names in the
\code{diffusion_vector} how to summarise the scores.}
}
\value{
The class with added diffusion score based on a single set of seed
genes. Additionally, the seed genes are stored in the class.
}
\description{
This function takes a diffusion vector and leverages personalised page-rank
diffusion to identify influential nodes. These can be used subsequently for
community detection or check AUROC values given a set of genes.
}
\examples{
# personalised page-rank diffusion from three seed nodes
set.seed(42)
g <- igraph::sample_pa(15, directed = FALSE)
edges <- data.table::setDT(igraph::as_data_frame(g))[, `:=`(
  from = sprintf("node_\%i", from),
  to = sprintf("node_\%i", to)
)]
object <- NetworkDiffusions(edges, weighted = FALSE, directed = FALSE)
object <- diffuse_seed_nodes(
  object,
  c(node_1 = 1, node_3 = 1, node_10 = 1),
  summarisation = "max"
)
head(get_diffusion_vector(object))
}
