% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{dim.single_cell_exp}
\alias{dim.single_cell_exp}
\title{dim Method for single_cell_exp object}
\arguments{
\item{x}{An object of class \code{single_cell_exp}.}
}
\value{
An integer vector of length 2 with the number of cells and genes.
}
\description{
Returns the dimensions of a single_cell_exp object.
}
