% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_subset.R
\name{dim.SingleCellsSubset}
\alias{dim.SingleCellsSubset}
\title{dim Method for SingleCellsSubset object}
\usage{
## S7 method for class <bixverse::SingleCellsSubset>
dim(x)
}
\arguments{
\item{x}{An object of class \code{SingleCellsSubset}.}
}
\value{
An integer vector of length 2 with the number of cells and genes.
}
\description{
dim Method for SingleCellsSubset object
}
\keyword{internal}
