% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{.add_is_obs_attr}
\alias{.add_is_obs_attr}
\title{Add "is_obs" attribute}
\usage{
.add_is_obs_attr(x)
}
\arguments{
\item{x}{data.table. The data.table to which to add the attribute.}
}
\value{
\code{x} with added attribute
}
\description{
Add "is_obs" attribute
}
\keyword{internal}
