% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{.adt_feature_names}
\alias{.adt_feature_names}
\title{Helper to get the ADT feature names}
\usage{
.adt_feature_names(x)
}
\arguments{
\item{x}{The class from which to extract the ADT feature names.}
}
\value{
The feature names if available
}
\description{
Helper to get the ADT feature names
}
\keyword{internal}
