% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{.cache_slot_from_modality}
\alias{.cache_slot_from_modality}
\title{Helper to get the cache slot}
\usage{
.cache_slot_from_modality(modality)
}
\arguments{
\item{modality}{String. One of \code{c("rna", "adt", "atac")}. Assumed already
validated by the caller via \code{match.arg}.}
}
\value{
The name of the cache property.
}
\description{
Helper to get the cache slot
}
\keyword{internal}
