% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc_residuals.R
\name{.calculate_pca_residual_mc}
\alias{.calculate_pca_residual_mc}
\title{Residual-based PCA for meta cells}
\usage{
.calculate_pca_residual_mc(
  object,
  no_pcs,
  pca_params,
  selected_hvg,
  seed,
  .verbose
)
}
\arguments{
\item{object}{\code{MetaCells} class.}

\item{no_pcs}{Integer. Number of PCs.}

\item{pca_params}{List. See \code{\link[=params_sc_pca]{params_sc_pca()}}.}

\item{selected_hvg}{Integer. The 1-based genes to use.}

\item{seed}{Integer. Random seed.}

\item{.verbose}{Boolean or Integer. Controls verbosity.}
}
\value{
The object with the PCA attached.
}
\description{
Residual-based PCA for meta cells
}
\keyword{internal}
