% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{.community_stability}
\alias{.community_stability}
\title{Assess stability of ICA components}
\usage{
.community_stability(no_comp, s, return_centrotype)
}
\arguments{
\item{no_comp}{Integer. Number of components that were set for this run.}

\item{s}{Numeric matrix. Combined S matrices from the ICA runs. Assumes
rows = features and column individual independent components.}

\item{return_centrotype}{Boolean. Shall the centrotype be returned for this
run.}
}
\value{
A list containing:
\itemize{
\item stability_scores - The stability scores for the given runs.
\item centrotype - The centrotype component of each cluster.
}
}
\description{
Assesses the stability of a component given some random iterations (in terms
of random initialisations and potentially additional cross-validation). It
will calculate the (absolute) Pearson's correlation between the different
generated independent components and run hierarchical clustering on top of
these and assess the cluster stability. If desired, it can also return the
centrotype of each cluster.
}
