% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{.demote_verbosity}
\alias{.demote_verbosity}
\title{Reduce verbosity by one level}
\usage{
.demote_verbosity(.verbose)
}
\arguments{
\item{.verbose}{\code{TRUE}, \code{FALSE}, or an integer.}
}
\value{
Integer verbosity level, floored at 0.
}
\description{
Reduce verbosity by one level
}
\keyword{internal}
