% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc_residuals.R
\name{.find_hvg_residual_mc}
\alias{.find_hvg_residual_mc}
\title{Residual-based HVG selection for meta cells}
\usage{
.find_hvg_residual_mc(object, hvg_no, .verbose = TRUE)
}
\arguments{
\item{object}{\code{MetaCells} class.}

\item{hvg_no}{Integer. Variable features per group.}

\item{.verbose}{Boolean or Integer. Controls verbosity.}
}
\value{
The object with the HVGs set.
}
\description{
Residual-based HVG selection for meta cells
}
\keyword{internal}
