% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{.flip_ica_loading_signs}
\alias{.flip_ica_loading_signs}
\title{Flips the ICA source sign}
\usage{
.flip_ica_loading_signs(x)
}
\arguments{
\item{x}{Numeric vector. The source signal for that independent component.}
}
\value{
Returns a consistent ICA feature loading
}
\description{
Makes sure that the largest absolute value in the ICA source matrix S will
be positive.
}
