% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{.generate_adt_var}
\alias{.generate_adt_var}
\title{Helper function to generate ADT vars}
\usage{
.generate_adt_var(adt_counts)
}
\arguments{
\item{adt_counts}{Numeric matrix. Of shape samples x features with the ADT
raw counts.}
}
\value{
A data.table with the feature_idx and feature_names of the ADT data.
}
\description{
Helper function to generate ADT vars
}
\keyword{internal}
