% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_go_object.R
\name{.get_go_genes_biomind}
\alias{.get_go_genes_biomind}
\title{Wrapper to clean up the GO identifiers to ensembl identifer table.}
\usage{
.get_go_genes_biomind(dt)
}
\arguments{
\item{dt}{The data.table containing the pathway to gene edges from BioMind.}
}
\value{
data.table with go_id and a list_col of collapsed, associated ensembl identifiers.
}
\description{
Wrapper to clean up the GO identifiers to ensembl identifer table.
}
