% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_go_obj.R
\name{.get_go_ontology_depth_ancestry_biomind}
\alias{.get_go_ontology_depth_ancestry_biomind}
\title{Wrapper to get the ancestry and the ontology depth level.}
\usage{
.get_go_ontology_depth_ancestry_biomind(dt)
}
\arguments{
\item{dt}{The data.table containing the ontology edges from BioMind.}
}
\value{
A list with two data.tables with the first containing the depth information
and the second the ancestry information
}
\description{
Wrapper to get the ancestry and the ontology depth level.
}
