% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_plots_helpers.R
\name{.get_manifoldsr_knn_from_wnn}
\alias{.get_manifoldsr_knn_from_wnn}
\title{Placeholder: manifoldsR nearest neighbours from a WNN graph}
\usage{
.get_manifoldsr_knn_from_wnn(x)
}
\arguments{
\item{x}{\code{SingleCellsMultiModal} object holding a WNN graph.}
}
\value{
A \code{manifoldsR} nearest neighbours object.
}
\description{
Placeholder: manifoldsR nearest neighbours from a WNN graph
}
\keyword{internal}
