% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_meta_cell.R
\name{.merge_mc_obs}
\alias{.merge_mc_obs}
\title{Row-bind the observation tables of meta cell objects}
\usage{
.merge_mc_obs(inputs, source_ids, prefix_ids)
}
\arguments{
\item{inputs}{List of \code{MetaCells} objects.}

\item{source_ids}{Character vector with the source identifiers.}

\item{prefix_ids}{Boolean. Prefix the meta cell identifiers.}
}
\value{
A data.table with the merged observations, re-indexed.
}
\description{
Row-bind the observation tables of meta cell objects
}
\keyword{internal}
