% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{.N_CELLS_STREAMING_THRESHOLD}
\alias{.N_CELLS_STREAMING_THRESHOLD}
\title{Auto-streaming fires up with more than 100,000 cells}
\usage{
.N_CELLS_STREAMING_THRESHOLD
}
\description{
Auto-streaming fires up with more than 100,000 cells
}
\keyword{internal}
