% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_annotations_reference.R
\name{.read_label_columns}
\alias{.read_label_columns}
\title{Read named obs columns in cells_to_keep order, return a data.table.}
\usage{
.read_label_columns(sc_object, columns)
}
\arguments{
\item{sc_object}{\code{SingleCells} class}

\item{columns}{String. The columns to extract}
}
\value{
A data.table with the desired columns.
}
\description{
Read named obs columns in cells_to_keep order, return a data.table.
}
\keyword{internal}
