% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{.remap_communities_by_size}
\alias{.remap_communities_by_size}
\title{Remap communities by size}
\usage{
.remap_communities_by_size(labels)
}
\arguments{
\item{labels}{Integer vector that represents the community assignment.}
}
\value{
The remapped communities by size
}
\description{
Remap community assignment by size
}
\keyword{internal}
