% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mc_analysis.R
\name{.resolve_mc_nmf_selection}
\alias{.resolve_mc_nmf_selection}
\title{Resolve cell and gene selection for NMF on MetaCells}
\usage{
.resolve_mc_nmf_selection(object, cell_ids, gene_ids)
}
\arguments{
\item{object}{\code{MetaCells} class.}

\item{cell_ids}{Optional string vector. The cells to include.}

\item{gene_ids}{Optional string vector. The genes to include.}
}
\value{
A list with the following items:
\itemize{
\item cell_indices - The Rust-based cell indices.
\item gene_indices - The Rust-based gene indices.
\item cell_ids - The cell identifiers (1-based)
\item gene_ids - The gene identifiers (1-based).
}
}
\description{
Resolve cell and gene selection for NMF on MetaCells
}
\keyword{internal}
