% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_pathways.R
\name{.resolve_sc_nmf_selection}
\alias{.resolve_sc_nmf_selection}
\title{Resolve cell and gene selection for NMF on SingleCells}
\usage{
.resolve_sc_nmf_selection(object, cell_ids, gene_ids)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{cell_ids}{Optional string vector. The cells to include.}

\item{gene_ids}{Optional string vector. The genes to include.}
}
\value{
A list with the following items:
\itemize{
\item cell_indices - The Rust-based cell indices.
\item gene_indices - The Rust-based gene indices.
\item cell_ids - The cell identifiers.
\item gene_ids - The gene identifiers.
}
}
\description{
Resolve cell and gene selection for NMF on SingleCells
}
\keyword{internal}
