% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc_cache_state.R
\name{.sc_cache_put}
\alias{.sc_cache_put}
\title{Write the cache for a given modality back onto the object}
\usage{
.sc_cache_put(x, cache, modality = "rna")
}
\arguments{
\item{x}{The object owning the cache.}

\item{cache}{The cache list to write.}

\item{modality}{String. One of \code{c("rna", "adt", "atac", "wnn")}.}
}
\value{
The object with the cache written back.
}
\description{
Write the cache for a given modality back onto the object
}
\keyword{internal}
