% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_sc.R
\name{.sc_metacell_generator}
\alias{.sc_metacell_generator}
\title{Internal: the generic behind a meta cell method.}
\usage{
.sc_metacell_generator(method)
}
\arguments{
\item{method}{String. To which of the methods to dispatch.}
}
\value{
The method
}
\description{
Internal: the generic behind a meta cell method.
}
\keyword{internal}
