% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{.scale_free_fit}
\alias{.scale_free_fit}
\title{Calculate the goodness of fit for a power law distribution.}
\usage{
.scale_free_fit(k, breaks = 50L, plot = FALSE)
}
\arguments{
\item{k}{Numeric vector. The vector of}

\item{breaks}{Integer. Number of breaks for fitting the data.}

\item{plot}{Boolean. Shall the log-log plot be generated.}
}
\value{
The R2 value of of the goodness of fit.
}
\description{
Calculate the goodness of fit for a power law distribution.
}
