% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell_lr.R
\name{.scaling_zscore}
\alias{.scaling_zscore}
\title{Z-score scaling}
\usage{
.scaling_zscore(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Returns \code{(x - mean(x)) / sd(x)}. Returns zeros if \code{sd(x)} is zero or \code{NA}.
}
\keyword{internal}
