% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_cd34_data}
\alias{download_cd34_data}
\title{Download the CD34 example data from SEACells}
\usage{
download_cd34_data(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean. If the download shall be quiet.}
}
\value{
String. The path to CD34 SEACells data set.
}
\description{
This function downloads the CD34 data set from the SEACells paper into the
temperorary directory.
}
\examples{
\dontrun{
# pulls the archive into the session tempdir()
path <- download_cd34_data()
get_h5ad_dimensions(path)$dims
}
}
\references{
Persad, et al., Nat. Biotechnol., 2023
}
