% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_demuxlet_pbmc}
\alias{download_demuxlet_pbmc}
\title{Download PBMCs with demuxlet doublet information}
\usage{
download_demuxlet_pbmc(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean. If the download shall be quiet.}
}
\value{
String. The path to the extracted doublet detection data.
}
\description{
This function downloads a PBMC data set with demuxlet information to test
doublet detection methods.
}
\examples{
\dontrun{
# pulls the archive into the session tempdir()
path <- download_demuxlet_pbmc()
list.files(path)
}
}
