% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_marrow_cd34}
\alias{download_marrow_cd34}
\title{Download the marrow CD34 example data from Palantir}
\usage{
download_marrow_cd34(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean. If the download shall be quiet.}
}
\value{
String. The path to the marrow CD34 data set.
}
\description{
This function downloads the bone marrow CD34 data set from the Palantir
paper into the temporary directory.
}
\examples{
\dontrun{
# pulls the archive into the session tempdir()
path <- download_marrow_cd34()
get_h5ad_dimensions(path)$dims
}
}
\references{
Setty, et al., Nat. Biotechnol., 2019
}
