% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_pbmc3k}
\alias{download_pbmc3k}
\title{Download PBMC3K data from Zenodo}
\usage{
download_pbmc3k(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean. If the download shall be quiet.}
}
\value{
String. The path to the extracted PBMC3K data.
}
\description{
This function downloads the PBMC3K dataset from 10x Genomics (uploaded on
Zenodo) and extracts it and returns the (temporary) paths.
}
\examples{
\dontrun{
# pulls the archive into the session tempdir()
path <- download_pbmc3k()
list.files(path)
}
}
