% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_pbmc_batches}
\alias{download_pbmc_batches}
\title{Download two different PBMC data sets for batch correction testing}
\usage{
download_pbmc_batches(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean. If the download shall be quiet.}
}
\value{
String. The path to the directory with the PBMC h5ad files.
}
\description{
This function downloads two different h5ad files for testing batch correction
methods into the temporary directory.
}
\examples{
\dontrun{
# pulls the archive into the session tempdir()
path <- download_pbmc_batches()
list.files(path)
}
}
