% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_pbmc_totalseq_data}
\alias{download_pbmc_totalseq_data}
\title{Download the PBMC TotalSeq data with ADT counts}
\usage{
download_pbmc_totalseq_data(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean. If the download shall be quiet.}
}
\value{
String. The path to the TotalSeq data.
}
\description{
This function downloads the PBMC TotalSeq data set.
}
\examples{
\dontrun{
# pulls the archive into the session tempdir()
path <- download_pbmc_totalseq_data()
read_tenx_h5_metadata(path)$feature_types
}
}
