% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_enrichment.R
\name{fgsea_simple}
\alias{fgsea_simple}
\title{Bixverse implementation of the simple fgsea algorithm}
\usage{
fgsea_simple(stats, pathways, nperm = 2000L, gsea_params = params_gsea())
}
\arguments{
\item{stats}{...}

\item{pathways}{...}

\item{nperm}{...}

\item{min_size}{...}

\item{max_size}{...}

\item{gsea_param}{...}
}
\value{
...
}
\description{
Bixverse implementation of the simple fgsea algorithm
}
