% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{finalise_matrix}
\alias{finalise_matrix}
\title{Finalise the count matrix}
\usage{
finalise_matrix(matrix, return_format, cell_indices, gene_indices, sc_map)
}
\arguments{
\item{matrix}{The sparse matrix to finalise}

\item{return_format}{String. One of \code{c("cell", "gene")}.}

\item{cell_indices}{Optional integer. The cell indices to return.}

\item{gene_indices}{Optional integer. The gene indices to return.}

\item{sc_map}{A \code{sc_mapper} class. Contains various mapping information.}
}
\value{
The finalised matrix.
}
\description{
Finalise the count matrix
}
