% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_dge_diffabund.R
\name{find_all_markers_sc}
\alias{find_all_markers_sc}
\title{Find all markers}
\usage{
find_all_markers_sc(
  object,
  column_of_interest,
  method = "wilcox",
  alternative = c("greater", "less", "twosided"),
  min_prop = 0.05,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} or \code{SingleCellsSubset} class.}

\item{column_of_interest}{String. The column you wish to use to identify
the markers between all combination. Needs to be in the obs table}

\item{method}{String. Which method to use for the calculations of the DGE.
At the moment the only option is \code{"wilcox"}, but the parameter is reserved
for future features.}

\item{alternative}{String. Test alternative. One of
\code{c("twosided", "greater", "less")}. This function will default to
\code{"greater"}, i.e., genes upregulated in the group.}

\item{min_prop}{Numeric. The minimum proportion of cells that need to express
the gene to be tested in the group or in the rest.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
data.table with the DGE results from the test.
}
\description{
This function can be used to run differential gene expression for every
group of an unsupervised clustering method for example. You specify a column
and the function calculates differential gene expression of the first
cluster vs. everything else, second cluster vs. everything else, etc. All
groups are tested in a single pass over the genes, so memory stays bounded
however many cells there are.
}
\examples{
# one versus rest across the Leiden clusters
sc <- demo_single_cells()
sc <- find_clusters_sc(sc, res = 1.0)
res <- find_all_markers_sc(
  sc,
  column_of_interest = "leiden_clustering",
  .verbose = FALSE
)
head(res)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
