% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{find_clusters_sc}
\alias{find_clusters_sc}
\title{Graph-based clustering of cells on the sNN graph}
\usage{
find_clusters_sc(
  object,
  cluster_algorithm = c("leiden", "louvain"),
  res = 1,
  name = "leiden_clustering",
  modality = c("rna", "adt", "wnn"),
  seed = 42L
)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} (or potentially other) class.}

\item{cluster_algorithm}{String. One of \code{c("leiden", "louvain")}.}

\item{res}{Numeric. The resolution parameter for \code{\link[igraph:cluster_leiden]{igraph::cluster_leiden()}}
or \code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}.}

\item{name}{String. The name to add to the obs table in the DuckDB.}

\item{modality}{String. On which modality to run the UMAP. One of
\code{c("rna", "adt", "wnn")}. The two latter options are only available for
multi-modal versions with the added data.}

\item{seed}{Integer. For reproducibility.}
}
\value{
The object with added clustering in the obs table.
}
\description{
This function will apply Leiden clustering on the sNN graph with the
given resolution and add a column to the obs table.
}
\examples{
# Leiden on the cached sNN graph
sc <- demo_single_cells()
sc <- find_clusters_sc(sc, res = 1.0, name = "clusters")
table(get_sc_obs(sc)$clusters)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
