% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell.R
\name{find_hvg}
\alias{find_hvg}
\title{Identify HVGs}
\usage{
find_hvg(object, gene_set_list)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{hvg_method}{String. One of \code{c("vst", "meanvarbin", "dispersion")}.
Atm, only \code{"vst"} is implemented.}

\item{loess_span}{Float. The span for the loess function,}
}
\value{
It will add the columns based on the names in the \code{gene_set_list} to
the obs table.
}
\description{
This is a helper function to identify highly variable genes. At the moment
the implementation has only
}
