% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{find_threshold_otsu}
\alias{find_threshold_otsu}
\title{Find a threshold via the Otsu method}
\usage{
find_threshold_otsu(x, bins = 100L)
}
\arguments{
\item{x}{Numerical vector. The vector for which to find the threshold via
the Otsu method}

\item{bins}{Integer. Number of bins to use.}
}
\value{
The threshold.
}
\description{
Find a threshold via the Otsu method
}
\examples{
# threshold separating two Gaussian modes
set.seed(42)
find_threshold_otsu(c(rnorm(100), rnorm(100, mean = 5)))
}
