% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{fix_contrast_names}
\alias{fix_contrast_names}
\title{Fixes contrast names for DGEs}
\usage{
fix_contrast_names(x)
}
\arguments{
\item{x}{Vector of strings or factors.}
}
\value{
Vector with fixed naming based on R conventions.
}
\description{
Fixes contrast names for DGEs
}
