% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gene_ontology_data}
\alias{gene_ontology_data}
\title{Gene Ontology data (deprecated)}
\usage{
gene_ontology_data(go_data_dt, min_genes)
}
\arguments{
\item{go_data_dt}{A data.table that contains the gene ontology information.
This can be extracted with for example \code{\link[=get_go_data_human]{get_go_data_human()}}.}

\item{min_genes}{Integer. The minimum number of genes in a gene ontology
term to conduct the test.}
}
\value{
Returns a \code{\link[=GeneOntologyElim]{GeneOntologyElim()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=GeneOntologyElim]{GeneOntologyElim()}}.
}
\keyword{internal}
