% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell.R
\name{gene_set_proportions}
\alias{gene_set_proportions}
\title{Calculate the proportions of reads for specific gene sets}
\usage{
gene_set_proportions(object, gene_set_list)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{gene_set_list}{A named list with each element containing the gene
identifiers of that set. These should be the same as
\code{get_gene_names(object)}!}
}
\value{
It will add the columns based on the names in the \code{gene_set_list} to
the obs table.
}
\description{
This is a helper function that calculates proportions of reads belonging to
given gene sets. This can be used for example for the calculation of
percentage mitochondrial reads per cell. These will be automatically added
to the obs table
}
