% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{gene_set_proportions_sc}
\alias{gene_set_proportions_sc}
\title{Calculate the proportions of reads for specific gene sets}
\usage{
gene_set_proportions_sc(
  object,
  gene_set_list,
  streaming = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{gene_set_list}{A named list with each element containing the gene
identifiers of that set. These should be the same as
\code{get_gene_names(object)}!}

\item{streaming}{Optional Boolean. Shall the data be streamed in. Useful for
larger data sets where you wish to avoid loading in the whole data. If
\code{NULL}, will automatically detect.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
It will add the columns based on the names in the \code{gene_set_list} to
the obs table.
}
\description{
This is a helper function that calculates proportions of reads belonging to
given gene sets. This can be used for example for the calculation of
percentage mitochondrial reads per cell. These will be automatically added
to the obs table
}
\examples{
# read proportion of a gene set, the mitochondrial percentage pattern
sc <- demo_single_cells(prepped = FALSE)
sc <- gene_set_proportions_sc(
  sc,
  gene_set_list = list(set_a = c("gene_01", "gene_02", "gene_03")),
  .verbose = FALSE
)
head(unlist(sc[["set_a"]]))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
