% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{GeneOntologyElim}
\alias{GeneOntologyElim}
\title{Gene Ontology data}
\usage{
GeneOntologyElim(go_data_dt, min_genes)
}
\arguments{
\item{go_data_dt}{A data.table that contains the gene ontology information.
This can be extract with for example \code{\link[=get_go_data_human]{get_go_data_human()}}.}

\item{min_genes}{data.frame. Meta-data information in form of a data.frame.}
}
\value{
Returns the class for subsequent usage.
}
\description{
This class is used to store the gene ontology information for usage in GSE
elimination methods.
}
\section{Properties}{

\describe{
\item{go_info}{data.table. Contains the gene ontology identifiers and
names.}
\item{go_to_genes}{List. Contains the genes within each gene ontology
term.}
\item{ancestry}{List. Contains the ancestors for each gene ontology term.}
\item{levels}{List. Which gene ontology terms sit at which level.}
\item{min_genes}{Integer, the minimum genes in the gene ontology term to
conduct the test.}
}
}

\examples{
\donttest{
# human GO restricted to terms with at least 25 genes
go_obj <- GeneOntologyElim(
  get_go_data_human(.verbose = FALSE),
  min_genes = 25L
)
go_obj
}
}
