% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell_io.R
\name{generate_gene_index_map}
\alias{generate_gene_index_map}
\title{Generate an updated gene index mapping}
\usage{
generate_gene_index_map(gene_mask)
}
\arguments{
\item{gene_mask}{Boolean vector.}
}
\value{
A \code{gene_column_index_map} class that can be used to readjust
column indices to genes that were kept.
}
\description{
Generate an updated gene index mapping
}
