% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synthetic_bulk.R
\name{generate_gene_module_data}
\alias{generate_gene_module_data}
\title{Generates synthetic gene module data.}
\usage{
generate_gene_module_data(
  n_samples = 24L,
  n_genes = 60L,
  n_modules = 4L,
  overlap_fraction = 0.25,
  seed = 10101L
)
}
\arguments{
\item{n_samples}{Integer. Number of samples.}

\item{n_genes}{Integer. Number of genes.}

\item{n_modules}{Integer. Number of active gene modules. To a maximum of 10.}

\item{overlap_fraction}{Float. How much the same modules can be active
in the same samples.}

\item{seed}{Integer. Initial random seed for generation of the synthetic
data. Default: 10101L.}
}
\value{
A \code{synthetic_matrix_modules} class with the following items:
\itemize{
\item data - The data matrix.
\item meta_data - The sample metadata.
}
}
\description{
Generates an artifical matrix of active modules (for a maximum of 8) in
samples being active. Also allows for overlap between the modules. Designed
to test some of the algorithms.
}
\examples{
# four partially overlapping modules over 24 samples
mods <- generate_gene_module_data(
  n_samples = 24L,
  n_genes = 60L,
  n_modules = 4L
)
dim(mods$data)
head(mods$meta_data)
}
