% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{generate_hotspot_membership}
\alias{generate_hotspot_membership}
\alias{generate_hotspot_membership.Hotspot}
\title{Identify hotspot gene clusters}
\usage{
generate_hotspot_membership(x, fdr_threshold = 0.05, min_size = 10L)

\method{generate_hotspot_membership}{Hotspot}(x, fdr_threshold = 0.05, min_size = 10L)
}
\arguments{
\item{x}{An object to generate the hotspot gene clusters for.}

\item{fdr_threshold}{Numeric. The maximum FDR for a given gene-gene local
correlation to be included.}

\item{min_size}{Integer. Minimum cluster size.}
}
\description{
Identify hotspot gene clusters
}
\examples{
# cluster the local gene-gene correlations into modules
sc <- demo_single_cells()
hs <- hotspot_gene_cor_sc(sc, .verbose = FALSE)
generate_hotspot_membership(hs)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
