% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_pathways.R
\name{generate_null_perm_gs}
\alias{generate_null_perm_gs}
\title{Generate random gene sets for VISION p-value calculations}
\usage{
generate_null_perm_gs(
  gs_list,
  expr_genes,
  n_perm = 500L,
  n_comp = 5L,
  random_seed = 42L,
  no_cores = NULL
)
}
\arguments{
\item{gs_list}{Named nested list for which to calculate the local
auto-correlations. The elements have the gene identifiers of the respective
gene sets and have the option to have a \code{"pos"} and \code{"neg"} gene sets. The
names need to be part of the variables of the \code{single_cell_exp} class.}

\item{expr_genes}{Character vector. Represents the genes expressed in the
experiment.}

\item{n_perm}{Integer. Number of random permutations to generate.}

\item{n_comp}{Integer. Number of k-means cluster to identify.}

\item{random_seed}{Integer. For reproducibility purposes.}

\item{no_cores}{Optional integer. Number of sessions to use for the
\code{\link[mirai:mirai_map]{mirai::mirai_map()}} approach during generation of the random gene sets.
If not provided, will default to half of the available cores with a maximum
of \code{8L}.}
}
\value{
A list with the following elements:
\itemize{
\item random_signatures - Nested list representing the random permutations.
\item clusters - Association of original gene set to random permutation
set.
}
}
\description{
This function will generate random gene sets given the provided gs_list.
Under the hood, it uses the same approach as in DeTomaso, et al. and does
not generate a random signature per given signature, but \code{n_comp}
representative ones based on size and balance (positive and negative genes)
via k-means clustering. The original gene sets are than matched to the
closest cluster. The authors observed that this sufficed to estimate
significance, see DeTomaso, et al.
}
