% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{generate_perm_diffusion_vecs}
\alias{generate_perm_diffusion_vecs}
\title{Generates random permutation vectors}
\usage{
generate_perm_diffusion_vecs(
  graph,
  diffusion_vec,
  bins = 25L,
  iters = 1000L,
  random_seed = 10101L
)
}
\arguments{
\item{graph}{igraph. The graph for which to generate the random diffusion
vectors}

\item{diffusion_vec}{Named numeric. The initial diffusion vector.}

\item{bins}{Integer. Number of bins to use for node degree aware sampling.}

\item{iters}{Integer. Number of random permutations to generate.}

\item{random_seed}{Integer. Random seed.}
}
\value{
List with the permutations.
}
\description{
Generates random permutation vectors
}
