% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{generate_rbh_graph}
\alias{generate_rbh_graph}
\title{Generate an RBH graph.}
\usage{
generate_rbh_graph(
  object,
  minimum_similarity,
  k_best = 1L,
  overlap_coefficient = FALSE,
  spearman = FALSE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=RbhGraph]{RbhGraph()}}.}

\item{minimum_similarity}{The minimum similarity to create an edge.}

\item{k_best}{Integer. Number of best neighbours to consider. If set to
\code{1L}, this behaves as the traditional reciprocal best hit. If you set this
to \code{3L} you consider edges if the modules is in the top 3 best modules
by similarity for each other.}

\item{overlap_coefficient}{Boolean. Shall the overlap coefficient be used
instead of Jaccard similarity. Only relevant if the underlying class is set
to set similarity.}

\item{spearman}{Boolean. Shall Spearman correlation be used. Only relevant
if the underlying class is set to correlation-based similarity.}
}
\value{
The class with added properties.
}
\description{
This function will generate an RBH graph based on set similarity between
gene modules. You have the option to use an overlap coefficient instead of
Jaccard similarity and to specify a minimum similarity.
}
\examples{
# Jaccard-based reciprocal best hits between two module sets
set.seed(123)
modules <- data.table::data.table(
  origin = rep(c("set_a", "set_b"), each = 20),
  module = rep(c("m1", "m2", "m3", "m4"), each = 10),
  gene = unlist(replicate(4, sample(letters, 10), simplify = FALSE))
)
object <- RbhGraph(
  modules,
  rbh_type = "set",
  dataset_col = "origin",
  module_col = "module",
  value_col = "gene"
)
object <- generate_rbh_graph(object, minimum_similarity = 0)
head(get_rbh_res(object))
}
