% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_aggregations.R
\name{generate_seacells_sc}
\alias{generate_seacells_sc}
\title{Generate meta cells based on SEACells and return a \code{MetaCells} object}
\usage{
generate_seacells_sc(
  object,
  seacell_params = params_sc_seacells(),
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  cells_to_use = NULL,
  regenerate_knn = FALSE,
  target_size = 1e+05,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} or \code{SingleCellsSubset} class.}

\item{seacell_params}{List. Output of \code{\link[=params_sc_seacells]{params_sc_seacells()}}.
A list with the following items:
\itemize{
\item n_sea_cells - Number of SEA cells to detect.
\item max_fw_iters - Maximum iterations for the Franke-Wolfe algorithm per
matrix update.
\item convergence_epsilon - Convergence threshold. Algorithm stops when
RSS change < epsilon * RSS(0).
\item max_iter - Maximum iterations to run SEACells for.
\item min_iter - Minimum iterations to run SEACells for.
\item greedy_threshold - Maximum number of cells before defaulting to rapid
random selection of archetypes.
\item graph_building - Graph building method.
\item pruning - Boolean. Shall small values be pruned during the Franke-
Wolfe iterations.
\item pruning_threshold - The threshold below which pruning shall be
applied during Franke-Wolfe iterations.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{embd_to_use}{String. The embedding to use. Atm, the only option is
\code{"pca"}. Only relevant if you set regenerate_knn to \code{TRUE}.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used. Only relevant if you set regenerate_knn to
\code{TRUE}.}

\item{cells_to_use}{Optional string. Names of the cells to use for the
generation of the SEACells.}

\item{regenerate_knn}{Boolean. Shall a kNN graph be regenerated. If not,
the internal one will be used.}

\item{target_size}{Numeric. The library target size to normalise the meta
cells to.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
A \code{\link[=MetaCells]{MetaCells()}} with the data generated by this meta cell
aggregation method.
}
\description{
This function implements the meta cell aggregation from Persad et al., and
returns the resulting SEACells. Compared to other algorithms, a kernel
archetype analysis is used to identify the metacells. For details, please
refer to the publication.
}
\examples{
# archetypal analysis based meta cells
sc <- demo_single_cells()
mc <- generate_seacells_sc(
  sc,
  seacell_params = params_sc_seacells(
    n_sea_cells = 50L,
    min_iter = 5L,
    knn = list(k = 10L)
  ),
  .verbose = FALSE
)
mc

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\references{
Persad, et al. Nat Biotechnol, 2023
}
