% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_aggregations.R
\name{generate_supercells_sc}
\alias{generate_supercells_sc}
\title{Generate SuperCells and return a \code{MetaCells} object}
\usage{
generate_supercells_sc(
  object,
  sc_supercell_params = params_sc_supercell(),
  regenerate_knn = FALSE,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  cells_to_use = NULL,
  target_size = 1e+05,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} or \code{SingleCellsSubset} class.}

\item{sc_supercell_params}{List. Output of \code{\link[=params_sc_supercell]{params_sc_supercell()}}.
A list with the following items:
\itemize{
\item walk_length - Integer. Walk length for the Walktrap algorithm.
\item graining_factor - Numeric. Graining level of data (proportion of
number of single cells in the initial dataset to the number of metacells in
the final dataset)
\item use_kernel - Boolean. Shall a kernel be applied.
\item kith_neighbour - Optional integer. Which neighbour to use for the
sigma estimation.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{regenerate_knn}{Boolean. Shall a kNN graph be regenerated.}

\item{embd_to_use}{String. The embedding to use. Only relevant if you set
regenerate_knn to \code{TRUE}.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used. Only relevant if you set regenerate_knn to
\code{TRUE}.}

\item{cells_to_use}{Optional string. Names of the cells to use for the
generation of the meta-cells. If provided, this function will regenerate the
kNN graph no matter what.}

\item{target_size}{Numeric. The library target size to normalise the meta
cells to.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
A \code{\link[=MetaCells]{MetaCells()}} with the data generated by this meta cell
aggregation method.
}
\description{
This function implements the meta cell aggregation from Bilous, et al.
The core idea is to use the walktrap community detection on the kNN graph.
For details, please refer to the paper.
}
\examples{
# walktrap based super cells at a graining factor of 10
sc <- demo_single_cells()
mc <- generate_supercells_sc(
  sc,
  sc_supercell_params = params_sc_supercell(
    graining_factor = 10,
    knn = list(k = 10L)
  ),
  regenerate_knn = TRUE,
  .verbose = FALSE
)
mc

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\references{
Bilous, et al. BMC Bioinform., 2022
}
