% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_methods.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Return ancestor terms from an ontology}
\usage{
get_ancestors(ontology)
}
\arguments{
\item{ontology}{Dataframe. The dataframe with column parent and child.}
}
\value{
a named list with ancestor terms as values
}
\description{
this function will return all ancestor terms based on a provided dataframe with parent-child terms
}
