% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_best_run}
\alias{get_best_run}
\alias{get_best_run.StabilisedNmfResult}
\title{Get the best run from a stabilised NMF result}
\usage{
get_best_run(x)

\method{get_best_run}{StabilisedNmfResult}(x)
}
\arguments{
\item{x}{\code{StabilisedNmfResult} object.}
}
\value{
An \code{NmfResult} containing the W/H of the best run.
}
\description{
Extracts the run with the lowest reconstruction loss and returns it as a
single-run \code{NmfResult} for downstream methods.
}
\examples{
# the restart with the lowest reconstruction loss
sc <- demo_single_cells()
res <- stabilised_nmf_sc(sc, k = 5L, n_runs = 5L, .verbose = FALSE)
get_best_run(res)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
