% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_c_pca.R
\name{get_c_pca_factors}
\alias{get_c_pca_factors}
\title{Get the contrastive PCA factors}
\usage{
get_c_pca_factors(object)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}
}
\value{
The sample scores of the contrastive PCA run. If not found,
returns a warning and NULL.
}
\description{
Getter function for the feature factors of the contrastive PCA
}
\examples{
# sample scores of the contrastive components
cpca_data <- synthetic_c_pca_data()
target <- t(cpca_data$target)
background <- t(cpca_data$background)
meta <- data.table::data.table(sample_id = rownames(target))
obj <- BulkCoExp(target, meta)
obj <- preprocess_bulk_coexp(obj, .verbose = FALSE)
obj <- contrastive_pca_processing(obj, background, .verbose = FALSE)
obj <- contrastive_pca(obj, alpha = 2.5, no_pcs = 5L)
head(get_c_pca_factors(obj)[, 1:3])
}
