% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{get_cell_indices}
\alias{get_cell_indices}
\alias{get_cell_indices.MetaCells}
\alias{get_cell_indices.ScMap}
\alias{get_cell_indices.SingleCells}
\alias{get_cell_indices.SingleCellsSubset}
\title{Get the index position for a gene}
\usage{
get_cell_indices(x, cell_ids, rust_index)

## S7 method for class <bixverse::MetaCells>
get_cell_indices(x, cell_ids, rust_index)

\method{get_cell_indices}{ScMap}(x, cell_ids, rust_index)

## S7 method for class <bixverse::SingleCells>
get_cell_indices(x, cell_ids, rust_index)

## S7 method for class <bixverse::SingleCellsSubset>
get_cell_indices(x, cell_ids, rust_index)
}
\arguments{
\item{x}{An object to get the gene index from.}

\item{cell_ids}{String vector. The cell ids to search for.}

\item{rust_index}{Bool. Shall rust-based indexing be returned.}
}
\value{
The indices of the cells
}
\description{
Returns the index for a given gene based on the internal gene mapping. This
is used for the single cell-related classes and methods.
}
\examples{
# R-based positions of two cells
sc <- demo_single_cells(prepped = FALSE)
get_cell_indices(sc, c("cell_001", "cell_002"), rust_index = FALSE)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
