% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R,
%   R/classes_single_cell_subset.R
\name{get_cell_names}
\alias{get_cell_names}
\alias{get_cell_names.ScMap}
\alias{get_cell_names.SingleCells}
\alias{get_cell_names.SingleCellsSubset}
\title{Get the cell names}
\usage{
get_cell_names(x, filtered = FALSE)

\method{get_cell_names}{ScMap}(x, filtered = FALSE)

## S7 method for class <bixverse::SingleCells>
get_cell_names(x, filtered = FALSE)

## S7 method for class <bixverse::SingleCellsSubset>
get_cell_names(x, filtered = FALSE)
}
\arguments{
\item{x}{An object to get the cell names from.}

\item{filtered}{Boolean. Shall, if found only the cell names of the
\code{cells_to_keep} be returned (see \code{\link[=set_cells_to_keep]{set_cells_to_keep()}}. Defaults
to \code{FALSE}}
}
\value{
The cell names (barcodes)
}
\description{
Returns the cell names (usually barcodes).
}
\examples{
# barcodes of the cells that passed quality control
sc <- demo_single_cells(prepped = FALSE)
head(get_cell_names(sc, filtered = TRUE), 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
