% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_cell_names.sc_mapper}
\alias{get_cell_names.sc_mapper}
\title{Get the indices of the cells to keep}
\usage{
\method{get_cell_names}{sc_mapper}(x)
}
\arguments{
\item{x}{An \code{sc_mapper} object}
}
\value{
The cell indices (0-based for Rust)
}
\description{
Get the indices of the cells to keep
}
