% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{get_cor_graph}
\alias{get_cor_graph}
\title{Get correlation-based graph}
\usage{
get_cor_graph(object, epsilon, .verbose)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}

\item{epsilon}{Float. The epsilon parameter for the RBF function, in this
case the bump function.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item graph - The igraph
\item params - A list that contains the parameters of the graph generation
and general graph information (node, edge numbers).
}
}
\description{
Helper function to get a correlation-based igraph from the class
}
\examples{
# igraph from the stored correlations at a fixed epsilon
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- cor_module_processing(obj, cor_method = "spearman", .verbose = FALSE)
graph_res <- get_cor_graph(obj, epsilon = 2, .verbose = FALSE)
graph_res$params
}
