% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{get_cor_graph_single}
\alias{get_cor_graph_single}
\title{Get correlation-based graph}
\usage{
get_cor_graph_single(bulk_coexp, kernel_bandwidth, min_affinity, .verbose)
}
\arguments{
\item{kernel_bandwidth}{Numerical. The bandwidth to use for the affinity
kernel}

\item{min_affinity}{Numerical. Minimum affinity needed to keep the edge.}

\item{.verbose}{Boolean. Controls verbosity of the function.}

\item{`bulk_coexp`}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
igraph given the provided parameters.
}
\description{
Helper function to get a correlation-based igraph from the class
}
