% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_cores}
\alias{get_cores}
\title{Helper function for core detection.}
\usage{
get_cores(abs_max_workers = 8L)
}
\arguments{
\item{abs_max_workers}{Integer. Absolute maximum number of cores to use.
Defaults to \code{8L}.}
}
\value{
Integer. Number of cores to use.
}
\description{
Identifies the number of cores/workers to use for parallel tasks. Will
default to half of the available cores, to a maximum of \code{abs_max_workers}.
}
