% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_counts_from_rust}
\alias{get_counts_from_rust}
\title{Helper function to get counts from Rust}
\usage{
get_counts_from_rust(
  rust_con,
  assay,
  return_format,
  cell_indices,
  gene_indices,
  sc_map,
  .verbose = TRUE
)
}
\arguments{
\item{rust_con}{\code{SingeCellCountData} class. The connector to Rust.}

\item{assay}{String. One of \code{c("raw", "norm")}.}

\item{return_format}{String. One of \code{c("cell", "gene")}.}

\item{cell_indices}{Optional integer vector. The index positions of cells to
return.}

\item{gene_indices}{Optional integer vector. The index positions of genes to
return.}

\item{sc_map}{A \code{sc_mapper} class. Contains various mapping information.}

\item{.verbose}{Boolean. Controls verbosity}
}
\value{
Returns a list with:
\itemize{
\item indptr - The index pointers of the compressed sparse format.
\item indices - The indices of the data.
\item data - The underlying data.
\item no_cells - Number of cells.
\item no_genes - Number of genes.
}
}
\description{
Helper function to get counts from Rust
}
