% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_diffusion_perms}
\alias{get_diffusion_perms}
\title{Get the diffusion permutations}
\usage{
get_diffusion_perms(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=NetworkDiffusions]{NetworkDiffusions()}}.}
}
\value{
The diffusion Z scores if found. Otherwise \code{NULL}.
}
\description{
Returns the diffusion Z-scores if you ran
\code{\link[=permute_seed_nodes]{permute_seed_nodes()}}.
}
\examples{
# Z-scores from node-degree adjusted permutations
set.seed(42)
g <- igraph::sample_pa(15, directed = FALSE)
edges <- data.table::setDT(igraph::as_data_frame(g))[, `:=`(
  from = sprintf("node_\%i", from),
  to = sprintf("node_\%i", to)
)]
object <- NetworkDiffusions(edges, weighted = FALSE, directed = FALSE)
object <- diffuse_seed_nodes(object, c(node_1 = 1, node_3 = 1), "max")
object <- permute_seed_nodes(object, perm_iters = 100L, .verbose = FALSE)
head(get_diffusion_perms(object))
}
