% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_diffusion_vector}
\alias{get_diffusion_vector}
\title{Get the diffusion vector}
\usage{
get_diffusion_vector(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=NetworkDiffusions]{NetworkDiffusions()}}.}
}
\value{
The diffusion vector if found. If you did not run either diffusion
functions, it will return \code{NULL} and a warning.
}
\description{
Returns the diffusion vector if you ran \code{\link[=tied_diffusion]{tied_diffusion()}}
or \code{\link[=diffuse_seed_nodes]{diffuse_seed_nodes()}}.
}
\examples{
# diffusion scores after a single seed node diffusion
set.seed(42)
g <- igraph::sample_pa(15, directed = FALSE)
edges <- data.table::setDT(igraph::as_data_frame(g))[, `:=`(
  from = sprintf("node_\%i", from),
  to = sprintf("node_\%i", to)
)]
object <- NetworkDiffusions(edges, weighted = FALSE, directed = FALSE)
object <- diffuse_seed_nodes(object, c(node_1 = 1, node_3 = 1), "max")
head(get_diffusion_vector(object))
}
