% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_feature_mat}
\alias{get_feature_mat}
\alias{get_feature_mat.ScDblFinderRes}
\title{Get the feature matrix used for the classifier}
\usage{
get_feature_mat(x, ...)

\method{get_feature_mat}{ScDblFinderRes}(x, ...)
}
\arguments{
\item{x}{An object to get the feature matrix from. This will only include
the values of the observed cells.}

\item{...}{Additional parameters to forward to the method.}
}
\description{
Get the feature matrix used for the classifier
}
\examples{
# the features the scDblFinder classifier was trained on
sc <- demo_single_cells(prepped = FALSE)
res <- scdblfinder_sc(
  sc,
  scdblfinder_params = params_scdblfinder(
    pca = list(no_pcs = 10L),
    n_genes = 25L,
    cxds_genes = 25L
  ),
  return_features = TRUE,
  .verbose = FALSE
)
dim(get_feature_mat(res))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
