% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_gene_lengths.bulk_coexp}
\alias{get_gene_lengths.bulk_coexp}
\title{Get gene set lengths for a bulk_coexp class}
\arguments{
\item{x}{An object of class \code{bulk_coexp}.}

\item{species}{String. One of \code{c("human", "mouse", "rat")}.}

\item{...}{Additional parameters. Not in use atm.}
}
\value{
Named numeric representing the gene lengths.
}
\description{
Get gene lengths for a bulk_coexp object by extracting the count matrix
and delegating to the matrix method.
}
