% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_gene_lengths.matrix}
\alias{get_gene_lengths.matrix}
\title{Get gene set lengths for a matrix}
\usage{
\method{get_gene_lengths}{matrix}(x, species = c("human", "mouse", "rat"), ...)
}
\arguments{
\item{x}{Numerical matrix. Assumes genes x samples as format and Ensembl
identifiers as gene ids.}

\item{species}{String. One of \code{c("human", "mouse", "rat")}.}

\item{...}{Additional parameters. Not in use atm.}
}
\value{
Named numeric representing the gene lengths.
}
\description{
Get gene set lengths for a matrix
}
\examples{
\dontrun{
# the matrix method, dispatched on Ensembl rownames
counts <- matrix(
  1:4,
  nrow = 2,
  dimnames = list(c("ENSG00000141510", "ENSG00000012048"), c("s1", "s2"))
)
get_gene_lengths.matrix(counts, species = "human")
}

}
\keyword{internal}
