% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_gene_lengths}
\alias{get_gene_lengths}
\alias{get_gene_lengths,bixverse::BulkCoExp-method}
\title{Get the gene lengths}
\usage{
get_gene_lengths(x, species = c("human", "mouse", "rat"), ...)

## S7 method for class <bixverse::BulkCoExp>
get_gene_lengths(x, species = c("human", "mouse", "rat"), ...)
}
\arguments{
\item{x}{Object to extract gene lengths from. Can be a matrix or
BulkCoExp object.}

\item{species}{String. One of \code{c("human", "mouse", "rat")}.}

\item{...}{Additional parameters passed to methods.}
}
\description{
Get the gene lengths
}
\examples{
\dontrun{
# median transcript length per Ensembl gene, queried from Ensembl
counts <- matrix(
  1:4,
  nrow = 2,
  dimnames = list(c("ENSG00000141510", "ENSG00000012048"), c("s1", "s2"))
)
get_gene_lengths(counts, species = "human")
}
}
