% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_gene_names.sc_mapper}
\alias{get_gene_names.sc_mapper}
\title{Get the gene names}
\usage{
\method{get_gene_names}{sc_mapper}(x)
}
\arguments{
\item{x}{An \code{sc_mapper} object}
}
\value{
The gene names
}
\description{
Get the gene names
}
