% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R,
%   R/classes_single_cell_subset.R
\name{get_gene_names_from_idx}
\alias{get_gene_names_from_idx}
\alias{get_gene_names_from_idx.ScMap}
\alias{get_gene_names_from_idx.SingleCells}
\alias{get_gene_names_from_idx.SingleCellsSubset}
\title{Get the gene names based on the gene idx}
\usage{
get_gene_names_from_idx(x, gene_idx, rust_based = TRUE)

\method{get_gene_names_from_idx}{ScMap}(x, gene_idx, rust_based = TRUE)

## S7 method for class <bixverse::SingleCells>
get_gene_names_from_idx(x, gene_idx, rust_based = TRUE)

## S7 method for class <bixverse::SingleCellsSubset>
get_gene_names_from_idx(x, gene_idx, rust_based = TRUE)
}
\arguments{
\item{x}{An object to get the gene names from.}

\item{gene_idx}{Integer. The original gene indices for which to return
the gene names.}

\item{rust_based}{Boolean. Is it Rust-based, i.e., 0-index or R-based, i.e.,
1-indexed.}
}
\description{
Get the gene names based on the gene idx
}
\examples{
# Rust indices translated back into gene identifiers
sc <- demo_single_cells(prepped = FALSE)
get_gene_names_from_idx(sc, gene_idx = 0:2, rust_based = TRUE)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
