% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_go_human_data}
\alias{get_go_human_data}
\title{Get the Gene Ontology data human}
\usage{
get_go_human_data()
}
\value{
data.table with the data ready for \code{\link[=gene_ontology_data]{gene_ontology_data()}}.
}
\description{
Get the Gene Ontology data human
}
