% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_go_levels}
\alias{get_go_levels}
\title{Helper to get the Gene Ontology levels}
\usage{
get_go_levels(edge_dt)
}
\arguments{
\item{edge_dt}{data.table. The gene ontology edge data (i.e., term
connections between the different terms)}
}
\value{
A data.table with the identifier and depth.
}
\description{
Gets the ontology depth based on the edge dt
}
