% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_h}
\alias{get_h}
\alias{get_h.NmfResult}
\alias{get_h.StabilisedNmfResult}
\alias{get_h.ConsensusNmfResult}
\title{Get the H (cell activations) matrix}
\usage{
get_h(x)

\method{get_h}{NmfResult}(x)

\method{get_h}{StabilisedNmfResult}(x)

\method{get_h}{ConsensusNmfResult}(x)
}
\arguments{
\item{x}{An object holding NMF results.}
}
\description{
Get the H (cell activations) matrix
}
\examples{
# per cell activations of a five factor NMF
sc <- demo_single_cells()
res <- nmf_sc(sc, k = 5L, .verbose = FALSE)
dim(get_h(res))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
