% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_hotspot_membership}
\alias{get_hotspot_membership}
\alias{get_hotspot_membership.Hotspot}
\title{Get the hotspot gene membership table}
\usage{
get_hotspot_membership(x)

\method{get_hotspot_membership}{Hotspot}(x)
}
\arguments{
\item{x}{The object from which to retrieve the hotspot gene membership}
}
\description{
Get the hotspot gene membership table
}
\examples{
# the gene to module table hotspot clustered out
sc <- demo_single_cells()
hs <- hotspot_gene_cor_sc(sc, .verbose = FALSE)
hs <- generate_hotspot_membership(hs)
head(get_hotspot_membership(hs))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
