% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_hvg.sc_mapper}
\alias{get_hvg.sc_mapper}
\title{Get the indices of the HVG}
\usage{
\method{get_hvg}{sc_mapper}(x)
}
\arguments{
\item{x}{An \code{sc_mapper} object}
}
\value{
The gene indices (0-based for Rust) of the HVG
}
\description{
Get the indices of the HVG
}
