% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{get_ica_stability_res}
\alias{get_ica_stability_res}
\title{Get the ICA component data (stability, convergence, redundancy)}
\usage{
get_ica_stability_res(object)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}
}
\value{
data.table with the ICA parameter data (if found. Otherwise \code{NULL}.)
}
\description{
Getter function to extract the ICA component data in terms of stability,
convergence and maximum absolute correlation between the component loadings
(\code{max_abs_loading_cor}). If not
found will return \code{NULL}.
}
\examples{
# the stability table behind the ncomp choice
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- ica_processing(obj, .verbose = FALSE)
obj <- ica_evaluate_comp(
  obj,
  ica_type = "logcosh",
  ncomp_params = params_ica_ncomp(custom_seq = seq(2L, 20L, by = 2L)),
  .verbose = FALSE
)
head(get_ica_stability_res(obj))
}
