% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_index_cells}
\alias{get_index_cells}
\alias{get_index_cells.miloR}
\title{Get the index cells}
\usage{
get_index_cells(x)

\method{get_index_cells}{miloR}(x)
}
\arguments{
\item{x}{An object from which get the index cells.}
}
\value{
The indices of the cells in the neighbourhood.
}
\description{
Get the index cells
}
\examples{
# the cells the neighbourhoods were centred on
sc <- demo_single_cells(
  syn_data_params = params_sc_synthetic_data(
    n_cells = 500L,
    n_genes = 50L,
    n_samples = 6L,
    sample_bias = "even"
  )
)
milo <- get_miloR_abundances_sc(
  sc,
  sample_id_col = "sample_id",
  miloR_params = params_sc_miloR(k_refine = 10L),
  .verbose = FALSE
)
head(get_index_cells(milo))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
