% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_inflection_point}
\alias{get_inflection_point}
\title{Identify the inflection point for elbow-like data}
\usage{
get_inflection_point(x, y, span = 0.5)
}
\arguments{
\item{x, y}{The x and y values.}

\item{span}{The span parameter for the loess function.}
}
\value{
A list containing:
\itemize{
\item inflection_idx - Index of the inflection point
\item gradient_change - Absolute change in the first derivative
}
}
\description{
This function will identify the index of the inflection point of an arbitrary
series \code{x ~ y} via the biggest increase in the first derivative. Useful to
identify key points in elbow plots.
}
\examples{
# elbow of a decaying curve that flattens out
x <- 1:20
y <- c(exp(-x[1:10] / 2), rep(0.005, 10))
get_inflection_point(x, y)$inflection_idx
}
