% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_kmeans_clusters}
\alias{get_kmeans_clusters}
\alias{get_kmeans_clusters.SingleCellFastClusters}
\title{Get k-means cluster assignments from a fast cluster result}
\usage{
get_kmeans_clusters(x)

\method{get_kmeans_clusters}{SingleCellFastClusters}(x)
}
\arguments{
\item{x}{\code{SingleCellFastClusters} object.}
}
\description{
Get k-means cluster assignments from a fast cluster result
}
\examples{
# the centroid each cell was assigned to
sc <- demo_single_cells()
res <- fast_cluster_sc(
  sc,
  resolutions = 1.0,
  n_centroids = 30L,
  return_kmeans = TRUE,
  .verbose = FALSE
)
head(get_kmeans_clusters(res))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
