% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_others.R,
%   R/classes_single_cell_subset.R
\name{get_knn_dist}
\alias{get_knn_dist}
\alias{get_knn_dist.MetaCells}
\alias{get_knn_dist.ScCache}
\alias{get_knn_dist.SingleCells}
\alias{get_knn_dist.SingleCellNearestNeighbour}
\alias{get_knn_dist.SingleCellsSubset}
\title{Get the KNN distance}
\usage{
get_knn_dist(x, ...)

## S7 method for class <bixverse::MetaCells>
get_knn_dist(x, ...)

\method{get_knn_dist}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
get_knn_dist(x, ...)

\method{get_knn_dist}{SingleCellNearestNeighbour}(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_knn_dist(x, ...)
}
\arguments{
\item{x}{An object to get the kNN distances from.}

\item{...}{Other parameters.}
}
\description{
Getter for an integer matrix of samples x distances. Useful in combination
with \code{\link[=get_knn_mat]{get_knn_mat()}}.
}
\examples{
# cells x neighbour distances
sc <- demo_single_cells()
dim(get_knn_dist(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
