% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_others.R,
%   R/classes_single_cell_subset.R
\name{get_knn_mat}
\alias{get_knn_mat}
\alias{get_knn_mat.MetaCells}
\alias{get_knn_mat.ScCache}
\alias{get_knn_mat.SingleCells}
\alias{get_knn_mat.SingleCellNearestNeighbour}
\alias{get_knn_mat.SingleCellsSubset}
\title{Get the KNN matrix}
\usage{
get_knn_mat(x, ...)

## S7 method for class <bixverse::MetaCells>
get_knn_mat(x, ...)

\method{get_knn_mat}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
get_knn_mat(x, ...)

\method{get_knn_mat}{SingleCellNearestNeighbour}(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_knn_mat(x, ...)
}
\arguments{
\item{x}{An object to get the kNN matrix from.}

\item{...}{Other parameters.}
}
\description{
Getter for an integer matrix of samples x neighbours.
}
\examples{
# cells x neighbours
sc <- demo_single_cells()
dim(get_knn_mat(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
