% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_knn_mat.sc_cache}
\alias{get_knn_mat.sc_cache}
\title{Get the KNN matrix}
\usage{
\method{get_knn_mat}{sc_cache}(x)
}
\arguments{
\item{x}{An \code{sc_cache} object}
}
\value{
The KNN matrix (0-indexed).
}
\description{
Get the KNN matrix
}
