% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{get_knn_obj}
\alias{get_knn_obj}
\alias{get_knn_obj.MetaCells}
\alias{get_knn_obj.ScCache}
\alias{get_knn_obj.SingleCells}
\alias{get_knn_obj.SingleCellsSubset}
\title{Get the KNN object}
\usage{
get_knn_obj(x, ...)

## S7 method for class <bixverse::MetaCells>
get_knn_obj(x, ...)

\method{get_knn_obj}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
get_knn_obj(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_knn_obj(x, ...)
}
\arguments{
\item{x}{An object to get the KNN class from.}

\item{...}{Other parameters.}
}
\value{
The \code{SingleCellNearestNeighbour} object.
}
\description{
Returns the \code{SingleCellNearestNeighbour} from the object. This function is
used for the single cell-related classes and methods.
}
\examples{
# the cached kNN, cells x neighbours
sc <- demo_single_cells()
dim(get_knn_mat(get_knn_obj(sc)))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
