% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_aggregations.R
\name{get_meta_cells_sc}
\alias{get_meta_cells_sc}
\title{Generate meta cells and return the matrices}
\usage{
get_meta_cells_sc(
  object,
  sc_meta_cell_params = params_sc_metacells(),
  regenerate_knn = FALSE,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  target_size = 10000,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{sc_meta_cell_params}{List. Output of \code{\link[=params_sc_metacells]{params_sc_metacells()}}.
A list with the following items:
\itemize{
\item max_shared - Maximum number of allowed shared neighbours for
the meta cell to be considered.
\item target_no_metacells - Number of target meta cells you wish to reach.
\item max_iter - Maximum number of iterations you want to use for the
algorithm.
\item k - Number of neighbours for the kNN search. Only relevant if you
set regenerate_knn to \code{TRUE}.
\item n_trees - Integer. Number of trees to use for the annoy algorithm.
Only relevant if you set regenerate_knn to \code{TRUE}.
\item search_budget - Integer. Search budget per tree for the annoy
algorithm. Only relevant if you set regenerate_knn to \code{TRUE}.
\item knn_algorithm - String. Which of the two implemented kNN algorithms
to use. Defaults to \code{"annoy"}. Only relevant if you set regenerate_knn to
\code{TRUE}.
}}

\item{regenerate_knn}{Boolean. Shall a kNN graph be regenerated.}

\item{embd_to_use}{String. The embedding to use. Atm, the only option is
\code{"pca"}. Only relevant if you set regenerate_knn to \code{TRUE}.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used. Only relevant if you set regenerate_knn to
\code{TRUE}.}

\item{target_size}{Numeric. The library target size to normalise the meta
cells to.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\description{
This function implements the meta cell aggregation from Morabito, et al.
The generation of metacells is a useful approach for subsequent application
of for example correlation-based methods to identify co-regulated genes.
}
\references{
Morabito, et al. Cell Rep Methods, 2023
}
