% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_add.R
\name{get_obs_data}
\alias{get_obs_data}
\alias{get_obs_data.sc_proportion_res}
\alias{get_obs_data.scrublet_res}
\alias{get_obs_data.boost_res}
\title{Get the ready obs data from various sub method}
\usage{
get_obs_data(x, ...)

\method{get_obs_data}{sc_proportion_res}(x, ...)

\method{get_obs_data}{scrublet_res}(x, ...)

\method{get_obs_data}{boost_res}(x, ...)
}
\arguments{
\item{x}{An object to set gene mapping for}

\item{...}{Other parameters}
}
\value{
Returns a data.table with a cell_idx column for the cells included
in the analysis and additional columns to be added to the obs table.
}
\description{
Helper method that creates data.tables with cell indices which were used
in the given analysis + the values that are to be added to the obs table
in the DuckDB.
}
