% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_similarity.R
\name{get_ontology_ancestors}
\alias{get_ontology_ancestors}
\title{Return ancestor terms from an ontology}
\usage{
get_ontology_ancestors(parent_child_dt)
}
\arguments{
\item{parent_child_dt}{data.table. The data.table with column parent and
child.}
}
\value{
A named list with ancestor terms as values
}
\description{
this function will return all ancestor terms based on a provided
data.table with parent-child terms
}
