% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R, R/classes_lda.R,
%   R/classes_single_cell_others.R
\name{get_params}
\alias{get_params}
\alias{get_params.LdaResult}
\alias{get_params.LdaKSweepResult}
\alias{get_params.Hotspot}
\alias{get_params.miloR}
\alias{get_params.ScenicGrn}
\alias{get_params.NmfResult}
\alias{get_params.StabilisedNmfResult}
\alias{get_params.ConsensusNmfResult}
\alias{get_params.DialogueResult}
\alias{get_params.ScNebula}
\title{Get the parameters that were used.}
\usage{
get_params(object, to_json = FALSE, pretty_json = FALSE)

get_params.LdaResult(object, ...)

get_params.LdaKSweepResult(object, ...)

get_params.Hotspot(object, to_json = FALSE, pretty_json = FALSE)

get_params.miloR(object, to_json = FALSE, pretty_json = FALSE)

get_params.ScenicGrn(object, to_json = FALSE, pretty_json = FALSE)

get_params.NmfResult(object, to_json = FALSE, pretty_json = FALSE)

get_params.StabilisedNmfResult(object, to_json = FALSE, pretty_json = FALSE)

get_params.ConsensusNmfResult(object, to_json = FALSE, pretty_json = FALSE)

get_params.DialogueResult(object, ...)

get_params.ScNebula(object, to_json = FALSE, pretty_json = FALSE)
}
\arguments{
\item{object}{A class within bixverse that inherits from
\code{\link[=BixverseBaseClass]{BixverseBaseClass()}} or defined S3 classes.}

\item{to_json}{Shall the params be returned as a JSON string.}

\item{pretty_json}{Shall the params be returned as a pretty JSON string.}

\item{...}{Unused, present so the S3 methods sharing this page match the
generic.}
}
\value{
Depending on parameters either the R list or a (pretty) JSON string.
}
\description{
Extracts parameters from the \code{BixverseBaseClass} class (or
child classes) and has options to return (pretty) JSONs. This generic
also gets inherited by other classes and can be used to extract parameters.
Also, can dispatch to specific methods for certain S3 classes.
}
\examples{
# parameters stored in a freshly created class
object <- SimilarityNetworkFusion(snf_params = params_snf(k = 3L))
names(get_params(object))
get_params(object, to_json = TRUE, pretty_json = TRUE)
}
