% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_pca_factors.sc_cache}
\alias{get_pca_factors.sc_cache}
\title{Get the PCA factors}
\usage{
\method{get_pca_factors}{sc_cache}(x)
}
\arguments{
\item{x}{An \code{sc_cache} object}
}
\value{
The PCA factor matrix.
}
\description{
Get the PCA factors
}
