% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{get_pca_loadings}
\alias{get_pca_loadings}
\alias{get_pca_loadings.MetaCells}
\alias{get_pca_loadings.ScCache}
\alias{get_pca_loadings.SingleCells}
\alias{get_pca_loadings.SingleCellsSubset}
\title{Get the PCA loadings}
\usage{
get_pca_loadings(x, ...)

## S7 method for class <bixverse::MetaCells>
get_pca_loadings(x, ...)

\method{get_pca_loadings}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
get_pca_loadings(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_pca_loadings(x, ...)
}
\arguments{
\item{x}{An object to get PCA loadings from.}

\item{...}{Other parameters.}
}
\value{
The PCA feature loadings from the object (if found).
}
\description{
Returns the PCA loadings (feature-based scores). This function is used for
the single cell-related classes and methods.
}
\examples{
# HVGs x PCs
sc <- demo_single_cells()
dim(get_pca_loadings(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
