% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{get_pca_singular_val}
\alias{get_pca_singular_val}
\alias{get_pca_singular_val.MetaCells}
\alias{get_pca_singular_val.ScCache}
\alias{get_pca_singular_val.SingleCells}
\alias{get_pca_singular_val.SingleCellsSubset}
\title{Get the PCA singular values}
\usage{
get_pca_singular_val(x, ...)

## S7 method for class <bixverse::MetaCells>
get_pca_singular_val(x, ...)

\method{get_pca_singular_val}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
get_pca_singular_val(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_pca_singular_val(x, ...)
}
\arguments{
\item{x}{An object to get PCA singular values from.}

\item{...}{Other parameters.}
}
\value{
The PCA singular values from the object (if found).
}
\description{
Returns the PCA singular values (can be useful to assess cumulative variance
explained). This function is used for the single cell-related classes and
methods.
}
\examples{
# singular values, largest first
sc <- demo_single_cells()
head(get_pca_singular_val(sc), 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
