% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R
\name{get_per_sample_graph}
\alias{get_per_sample_graph}
\title{Get the per-sample graph from a MetaSpot}
\usage{
get_per_sample_graph(object, exp_id)
}
\arguments{
\item{object}{A \code{\link[=MetaSpot]{MetaSpot()}}.}

\item{exp_id}{String. The experiment identifier.}
}
\value{
A \code{dgCMatrix} adjacency at metaspot resolution.
}
\description{
Get the per-sample graph from a MetaSpot
}
