% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R, R/classes_spatial_spot_subset.R
\name{get_per_sample_image_features}
\alias{get_per_sample_image_features}
\alias{get_per_sample_image_features.SpCache}
\alias{get_per_sample_image_features,bixverse::SpatialSpot-method}
\alias{get_per_sample_image_features,bixverse::SpatialSpotSubset-method}
\title{Get the per-sample histology image features}
\usage{
get_per_sample_image_features(x, exp_id, resolution, ...)

\method{get_per_sample_image_features}{SpCache}(x, exp_id, resolution, ...)

## S7 method for class <bixverse::SpatialSpot>
get_per_sample_image_features(x, exp_id, resolution, ...)

## S7 method for class <bixverse::SpatialSpotSubset>
get_per_sample_image_features(x, exp_id, resolution, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{resolution}{String. The image resolution the features were cut from,
one of \code{c("lowres", "hires", "cytassist", "fullres")}.}

\item{...}{Other parameters.}
}
\description{
Get the per-sample histology image features
}
