% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R, R/classes_spatial_spot_subset.R
\name{get_per_sample_pca}
\alias{get_per_sample_pca}
\alias{get_per_sample_pca.SpCache}
\alias{get_per_sample_pca,bixverse::SpatialSpot-method}
\alias{get_per_sample_pca,bixverse::SpatialSpotSubset-method}
\title{Get the per-sample PCA}
\usage{
get_per_sample_pca(x, exp_id, ...)

\method{get_per_sample_pca}{SpCache}(x, exp_id, ...)

## S7 method for class <bixverse::SpatialSpot>
get_per_sample_pca(x, exp_id, ...)

## S7 method for class <bixverse::SpatialSpotSubset>
get_per_sample_pca(x, exp_id, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{...}{Other parameters.}
}
\description{
Get the per-sample PCA
}
