% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_rbh_res}
\alias{get_rbh_res}
\title{Get the RBH results}
\usage{
get_rbh_res(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=RbhGraph]{RbhGraph()}}.}
}
\value{
The data.table with the RBH result if found, otherwise NULL.
}
\description{
Pulls out the RBH results if you ran
\code{\link[=generate_rbh_graph]{generate_rbh_graph()}}
}
\examples{
# reciprocal best hits between modules of two data sets
set.seed(123)
modules <- data.table::data.table(
  origin = rep(c("set_a", "set_b"), each = 20),
  module = rep(c("m1", "m2", "m3", "m4"), each = 10),
  gene = unlist(replicate(4, sample(letters, 10), simplify = FALSE))
)
object <- RbhGraph(
  modules,
  rbh_type = "set",
  dataset_col = "origin",
  module_col = "module",
  value_col = "gene"
)
object <- generate_rbh_graph(object, minimum_similarity = 0)
head(get_rbh_res(object))
}
