% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R, R/classes_single_cell_others.R
\name{get_results}
\alias{get_results}
\alias{get_results.DialogueResult}
\title{Get the final results from the class}
\usage{
get_results(object)

get_results.DialogueResult(object, ...)
}
\arguments{
\item{object}{The underlying \code{\link[=BixverseBaseClass]{BixverseBaseClass()}}
class. The class functionality is usually inherited by other S7 classes in
\code{bixverse}.}

\item{...}{Unused, present so the S3 methods sharing this page match the
generic.}
}
\value{
Returns the final results if any have been stored in the class.
}
\description{
Get the final results from \code{BixverseBaseClass} class (or child
classes).
}
\examples{
# communities found after a network diffusion
set.seed(42)
g <- igraph::sample_pa(15, directed = FALSE)
edges <- data.table::setDT(igraph::as_data_frame(g))[, `:=`(
  from = sprintf("node_\%i", from),
  to = sprintf("node_\%i", to)
)]
object <- NetworkDiffusions(edges, weighted = FALSE, directed = FALSE)
object <- diffuse_seed_nodes(object, c(node_1 = 1, node_3 = 1), "max")
object <- permute_seed_nodes(object, perm_iters = 100L, .verbose = FALSE)
object <- community_detection(
  object,
  community_params = params_community_detection(
    min_seed_nodes = 0L,
    min_nodes = 2L
  )
)
head(get_results(object))
}
