% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R,
%   R/classes_single_cell_groups.R
\name{get_rust_count_cell_f_path}
\alias{get_rust_count_cell_f_path}
\title{Get the file path to the counts in cell format}
\usage{
get_rust_count_cell_f_path(object)

get_rust_count_cell_f_path(object)
}
\arguments{
\item{object}{\code{SingleCellsGrps} class.}
}
\value{
The path to the \code{counts_cells.bin}

The path to the \code{counts_cells.bin}
}
\description{
Get the file path to the counts in cell format

Get the file path to the counts in cell format
}
\keyword{internal}
