% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_rust_gene_indices}
\alias{get_rust_gene_indices}
\title{Get the Rust gene indices based on gene IDs}
\usage{
get_rust_gene_indices(x, gene_ids)
}
\arguments{
\item{x}{An object to set gene mapping for}

\item{gene_ids}{String vector. The names of the genes you wish to get the
Rust index identifiers for.}
}
\description{
Get the Rust gene indices based on gene IDs
}
