% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_rust_gene_indices.sc_mapper}
\alias{get_rust_gene_indices.sc_mapper}
\title{Get updated gene indices}
\usage{
\method{get_rust_gene_indices}{sc_mapper}(x, gene_ids)
}
\arguments{
\item{x}{An \code{sc_mapper} object}

\item{gene_ids}{String vector. The names of the genes you wish to get the
Rust index identifiers for.}
}
\value{
Updated gene indices
}
\description{
Get updated gene indices
}
