% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R
\name{get_sample}
\alias{get_sample}
\title{Get a single SpatialSample by exp_id}
\usage{
get_sample(object, exp_id)
}
\arguments{
\item{object}{A \code{\link[=SpatialSpot]{SpatialSpot()}}.}

\item{exp_id}{String. The experiment identifier.}
}
\value{
A \code{\link[=new_spatial_sample]{new_spatial_sample()}}.
}
\description{
Get a single SpatialSample by exp_id
}
