% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R
\name{get_sample_ids}
\alias{get_sample_ids}
\title{Get the exp_ids present in the samples list}
\usage{
get_sample_ids(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialSpot]{SpatialSpot()}} or \code{\link[=MetaSpot]{MetaSpot()}}.}
}
\value{
Character vector of \code{exp_id} values.
}
\description{
Get the exp_ids present in the samples list
}
