% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R
\name{get_samples}
\alias{get_samples}
\title{Get the samples list from a spatial object}
\usage{
get_samples(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialSpot]{SpatialSpot()}} or \code{\link[=MetaSpot]{MetaSpot()}}.}
}
\value{
Named list of \code{\link[=new_spatial_sample]{new_spatial_sample()}} objects keyed
by \code{exp_id}.
}
\description{
Get the samples list from a spatial object
}
