% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{get_sc_available_features}
\alias{get_sc_available_features}
\title{Returns the available features for single cell applications}
\usage{
get_sc_available_features(object)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} (or potentially other) class.}
}
\value{
A data.table with available features.
}
\description{
Returns a data.table with available features in the obs table and in the
count matrices.
}
\examples{
# what can be queried from the obs table and the counts
sc <- demo_single_cells(prepped = FALSE)
head(get_sc_available_features(sc), 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
