% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_sc_cache}
\alias{get_sc_cache}
\title{Getter the memory-stored data from the class}
\usage{
get_sc_cache(object)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}
}
\value{
Returns the sc_cache class from the object with all the
memory-stored data.
}
\description{
Getter the memory-stored data from the class
}
