% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{get_sc_counts}
\alias{get_sc_counts}
\title{Getter the counts}
\usage{
get_sc_counts(
  object,
  assay = c("raw", "norm"),
  return_format = c("cell", "gene"),
  cell_indices = NULL,
  gene_indices = NULL,
  use_cells_to_keep = TRUE,
  modality = c("rna", "adt"),
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells}, \code{SingleCellsMultiModal} class.}

\item{assay}{String. Which slot to return. One of \code{c("raw", "norm")}.
Defaults to \code{"raw"}.}

\item{return_format}{String. One of \code{c("cell", "gene")}. Return data in
cell-centric compressed format (CSR) or gene-centric compressed format (CSC).
Defaults to \code{"cell"}. Not relevant for \code{MetaCells}.}

\item{cell_indices}{Optional cell indices.}

\item{gene_indices}{Optional gene indices.}

\item{use_cells_to_keep}{Boolean. Shall cells to keep be found in the class,
shall the counts be reduced to these. Not relevant for \code{MetaCells}.}

\item{modality}{String. The modality to return. One of \code{c("rna", "adt")}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The counts table
}
\description{
Getter the counts
}
\examples{
# raw counts for the first ten genes, cell-centric (CSR)
sc <- demo_single_cells(prepped = FALSE)
counts <- get_sc_counts(sc, gene_indices = 1:10, .verbose = FALSE)
dim(counts)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
