% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_sc_duckdb}
\alias{get_sc_duckdb}
\title{Getter for the single cell DuckDB connection}
\usage{
get_sc_duckdb(object)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}
}
\value{
The DuckDB connector
}
\description{
Getter for the single cell DuckDB connection
}
