% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_sc_map}
\alias{get_sc_map}
\title{Getter for the different maps in the object}
\usage{
get_sc_map(object)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}
}
\value{
Returns the maps within the class.
}
\description{
Getter for the different maps in the object
}
