% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{get_sc_obs}
\alias{get_sc_obs}
\title{Getter the obs table}
\usage{
get_sc_obs(object, indices = NULL, cols = NULL, filtered = FALSE)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells}, \code{SingleCellsMultiModal} class.}

\item{indices}{Optional integer vector. The integer positions of the cells
to return.}

\item{cols}{Optional string vector. The columns from the obs table to return.}

\item{filtered}{Boolean. Whether to return all cells or filtered to \code{to_keep}
cells. Not relevant for \code{MetaCells}.}
}
\value{
The obs table
}
\description{
Getter the obs table
}
\examples{
# the obs table, restricted to a few columns
sc <- demo_single_cells(prepped = FALSE)
head(get_sc_obs(sc, cols = c("cell_id", "cell_grp", "lib_size")), 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
