% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R,
%   R/classes_single_cell_groups.R
\name{get_sc_rust_ptr}
\alias{get_sc_rust_ptr}
\title{Getter for the single cell Rust pointer}
\usage{
get_sc_rust_ptr(object)

get_sc_rust_ptr(object)
}
\arguments{
\item{object}{\code{SingleCellsGrps} class.}
}
\value{
The Rust structure

The Rust structure
}
\description{
Getter for the single cell Rust pointer

Getter for the single cell Rust pointer
}
\keyword{internal}
