% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R,
%   R/classes_single_cell_others.R
\name{get_scores}
\alias{get_scores}
\alias{get_scores.ScDblFinderRes}
\alias{get_scores.ScTypeResults}
\alias{get_scores.ScTypeCellResults}
\title{Get scores}
\usage{
get_scores(x, ...)

\method{get_scores}{ScDblFinderRes}(x, ..., score_type = c("weighted", "cxds_scores"))

\method{get_scores}{ScTypeResults}(x, ...)

\method{get_scores}{ScTypeCellResults}(x, ...)
}
\arguments{
\item{x}{An object to get scores from.}

\item{...}{Additional arguments passed to methods.}

\item{score_type}{Either \code{"weighted"} or \code{"cxds_scores"}.}
}
\value{
The score matrix or data.table held by the object, depending on the
class. Methods exist for \code{ScTypeResults}, \code{ScTypeCellResults} and
\code{ScDblFinderRes}.

A numeric matrix of cells x cell types.

A numeric vector with the winning score per cell.
}
\description{
Get scores
}
\examples{
# the ScType score matrix of cell types by cluster
sc <- demo_single_cells()
markers <- data.table::data.table(
  cell_type = rep(sprintf("cell_type_\%i", 1:3), each = 10),
  gene_id = sprintf("gene_\%02d", 1:30)
)
cell_markers <- prepare_cell_markers(obj = sc, marker_df = markers)
res <- calc_sc_type_scores(sc, cell_marker_list = cell_markers,
                           .verbose = FALSE)
dim(get_scores(res))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
