% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_aggregations.R
\name{get_seacells_sc}
\alias{get_seacells_sc}
\title{Generate meta cells based on SEACells and return a \code{meta_cells} object}
\usage{
get_seacells_sc(
  object,
  seacell_params = params_sc_seacells(),
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  target_size = 1e+05,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{seacell_params}{List. Output of \code{\link[=params_sc_seacells]{params_sc_seacells()}}.
A list with the following items:
\itemize{
\item n_sea_cells - Number of SEA cells to detect.
\item max_fw_iters - Maximum iterations for the Franke-Wolfe algorithm per
matrix update.
\item convergence_epsilon - Convergence threshold. Algorithm stops when
RSS change < epsilon * RSS(0).
\item max_iter - Maximum iterations to run SEACells for.
\item min_iter - Minimum iterations to run SEACells for.
\item greedy_threshold - Maximum number of cells before defaulting to rapid
random selection of archetypes.
\item graph_building - Graph building method.
\item k - Number of neighbours for the kNN algorithm.
\item knn_method - String. Which kNN algorithm to use. One of
\code{c("annoy", "hnsw", "nndescent")}. Defaults to \code{"annoy"}.
\item n_trees - Integer. Number of trees to use for the annoy algorithm.
\item search_budget - Integer. Search budget during querying for the annoy
algorithm.
\item nn_max_iter - Integer. Maximum iterations for NN Descent. Only
relevant if you use \code{"nndescent"}.
\item rho - Numeric. Sampling rate for NN Descent. Only relevant if you
use \code{"nndescent"}.
\item delta - Numeric. Early termination criterion for NN Descent. Only
relevant if you use \code{"nndescent"}.
}}

\item{embd_to_use}{String. The embedding to use. Atm, the only option is
\code{"pca"}. Only relevant if you set regenerate_knn to \code{TRUE}.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used. Only relevant if you set regenerate_knn to
\code{TRUE}.}

\item{target_size}{Numeric. The library target size to normalise the meta
cells to.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A \code{\link[=meta_cells]{meta_cells()}} with the data generated by this meta cell
aggregation method.
}
\description{
This function implements the meta cell aggregation from Persad et al., and
returns the resuling SEACells. Compared to other algorithms, a kernel
archetype analysis is used to identify the metacells. For details, please
refer to the publication.
}
\references{
Morabito, et al. Cell Rep Methods, 2023
}
