% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{get_semantic_similarities}
\alias{get_semantic_similarities}
\title{Get the ontology term similarities}
\usage{
get_semantic_similarities(object)
}
\arguments{
\item{object}{\verb{ontology class}. See \code{\link[=ontology]{ontology()}}.}
}
\value{
Returns the semantic similarity data.table from the class
}
\description{
Get the ontology term similarities
}
