% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{get_seurat_counts}
\alias{get_seurat_counts}
\title{Pull the raw counts out of a Seurat object}
\usage{
get_seurat_counts(seurat_obj)
}
\arguments{
\item{seurat_obj}{\code{Seurat} class. The class to extract the counts from.}
}
\value{
The raw counts as a \code{dgRMatrix} of cells x genes.
}
\description{
Pull the raw counts out of a Seurat object
}
\keyword{internal}
