% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{get_seurat_counts_to_list}
\alias{get_seurat_counts_to_list}
\title{Transform Seurat raw counts into a List}
\usage{
get_seurat_counts_to_list(seurat_obj)
}
\arguments{
\item{seurat_obj}{\code{Seurat} class. The class from which to extract the counts
from and transform to a list.}
}
\value{
A list with the following elements
\itemize{
\item indptr - Index pointers of the sparse data.
\item indices - Indices of the data.
\item data - The underlying data.
\item format - String that defines if the data is CSR or CSC.
\item nrow - The number of rows.
\item ncol - The number of columns.
}
}
\description{
Transform Seurat raw counts into a List
}
