% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{get_sim_matrix}
\alias{get_sim_matrix}
\title{Get the similarity matrix}
\usage{
get_sim_matrix(object, as_data_table = FALSE, .verbose = TRUE)
}
\arguments{
\item{object}{\verb{OntologySim class}. See \code{\link[=OntologySim]{OntologySim()}}.}

\item{as_data_table}{Boolean. Shall the data be returned as a long
data.table.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Returns the semantic similarity data.table from the class
}
\description{
Get the similarity matrix
}
\examples{
# the Wang similarity matrix back out of the class
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f"),
  type = c("part_of", "part_of", "part_of", "is_a", "is_a")
)
onto_obj <- OntologySim(onto, .verbose = FALSE)
onto_obj <- calculate_wang_sim_onto(
  onto_obj,
  weights = c(part_of = 0.8, is_a = 0.6),
  .verbose = FALSE
)
round(get_sim_matrix(onto_obj, .verbose = FALSE), 3)
}
