% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_methods.R
\name{get_similarity_matrix}
\alias{get_similarity_matrix}
\title{Calculate the Resnik and Lin semantic similarity}
\usage{
get_similarity_matrix(
  terms,
  ancestor_list,
  ic_list,
  similarity_type = c("resnik", "lin"),
  return_matrix = TRUE,
  max_ic = 1
)
}
\arguments{
\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}

\item{ancestor_list}{R list with names being the term and the elements in the list the names
of the ancestors.}

\item{ic_list}{R list with the names being the term and the elements the information content
of this given term. Needs to be a single float!}

\item{similarity_type}{String. Can take values of \code{c("resnik", "lin")}.}

\item{return_matrix}{Boolean. Return the similarity as a matrix or vector, Default = TRUE}

\item{max_ic}{Double. The maximum information content observed in the data. This will return
normalised Resnik (i.e., scaled between 0 and 1). If set to 1, it will return the Resnik distance. Default = 1}
}
\description{
This function calculates the Resnik or Lin similarity for a given ontology.
}
