% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_snf_adjcacency_mat}
\alias{get_snf_adjcacency_mat}
\title{Get an individual affinity matrix}
\usage{
get_snf_adjcacency_mat(object, name)
}
\arguments{
\item{object}{The underlying class \code{\link[=SimilarityNetworkFusion]{SimilarityNetworkFusion()}}.}

\item{name}{String. The name of the individual data modality affinity
matrix to return.}
}
\value{
Returns adjcacency matrix if found.
}
\description{
Get an individual affinity matrix
}
\examples{
# affinity matrix of a single modality
set.seed(42)
continuous <- matrix(rnorm(120), nrow = 12, ncol = 10)
rownames(continuous) <- sprintf("sample_\%02i", 1:12)
colnames(continuous) <- sprintf("feature_\%i", 1:10)
object <- SimilarityNetworkFusion(
  data = continuous,
  data_name = "continuous",
  snf_params = params_snf(k = 3L)
)
dim(get_snf_adjcacency_mat(object, name = "continuous"))
}
