% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{get_snn_graph}
\alias{get_snn_graph}
\alias{get_snn_graph.MetaCells}
\alias{get_snn_graph.ScCache}
\alias{get_snn_graph.SingleCells}
\alias{get_snn_graph.SingleCellsSubset}
\title{Get the sNN graph}
\usage{
get_snn_graph(x, ...)

## S7 method for class <bixverse::MetaCells>
get_snn_graph(x, ...)

\method{get_snn_graph}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
get_snn_graph(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_snn_graph(x, ...)
}
\arguments{
\item{x}{An object to get the sNN graph from.}

\item{...}{Other parameters.}
}
\value{
The igraph that has the shared nearest neighbours.
}
\description{
Returns the shared nearest neighbour graph from the object. This function is
used for the single cell-related classes and methods.
}
\examples{
# the sNN graph the clustering methods run on
sc <- demo_single_cells()
igraph::vcount(get_snn_graph(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
