% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_snn_graph.sc_cache}
\alias{get_snn_graph.sc_cache}
\title{Get the sNN graph}
\usage{
\method{get_snn_graph}{sc_cache}(x)
}
\arguments{
\item{x}{An \code{sc_cache} object}
}
\value{
The sNN igraph.
}
\description{
Get the sNN graph
}
