% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R
\name{get_sp_cache}
\alias{get_sp_cache}
\title{Get the SpCache from a spatial object}
\usage{
get_sp_cache(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialSpot]{SpatialSpot()}}.}
}
\value{
The \code{\link[=new_sp_cache]{new_sp_cache()}} S3 object.
}
\description{
Get the SpCache from a spatial object
}
