% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_aggregations.R
\name{get_supercells_sc}
\alias{get_supercells_sc}
\title{Generate SuperCells and return a \code{meta_cells} object}
\usage{
get_supercells_sc(
  object,
  sc_supercell_params = params_sc_supercell(),
  regenerate_knn = FALSE,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  cells_to_use = NULL,
  target_size = 1e+05,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{sc_supercell_params}{List. Output of \code{\link[=params_sc_supercell]{params_sc_supercell()}}.
A list with the following items:
\itemize{
\item walk_length - Integer. Walk length for the Walktrap algorithm.
\item graining_factor - Numeric. Graining level of data (proportion of
number of single cells in the initial dataset to the number of metacells in
the final dataset)
\item linkage_dist - String. Which type of distance metric to use for the
linkage. Defaults to \code{"complete"}.
\item k - Number of neighbours for the kNN search. Only relevant if you
set regenerate_knn to \code{TRUE}.
\item knn_method - String. Which kNN algorithm to use. One of
\code{c("annoy", "hnsw", "nndescent")}. Defaults to \code{"annoy"}. Only relevant if
you set regenerate_knn to \code{TRUE}.
\item ann_dist - String. Distance metric for the approximate neighbour
search. One of \code{c("cosine", "euclidean")}. Defaults to \code{"cosine"}. Only
relevant if you set regenerate_knn to \code{TRUE}.
\item n_trees - Integer. Number of trees to use for the annoy algorithm.
Only relevant if you set regenerate_knn to \code{TRUE}.
\item search_budget - Integer. Search budget per tree for the annoy
algorithm. Only relevant if you set regenerate_knn to \code{TRUE}.
\item nn_max_iter - Integer. Maximum iterations for NN Descent. Only
relevant if you set regenerate_knn to \code{TRUE} and use
\code{"nndescent"}.
\item rho - Numeric. Sampling rate for NN Descent. Only relevant if you
set regenerate_knn to \code{TRUE} and use \code{"nndescent"}.
\item delta - Numeric. Early termination criterion for NN Descent. Only
relevant if you set regenerate_knn to \code{TRUE} and use \code{"nndescent"}.
}}

\item{regenerate_knn}{Boolean. Shall a kNN graph be regenerated.}

\item{embd_to_use}{String. The embedding to use. Only relevant if you set
regenerate_knn to \code{TRUE}.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used. Only relevant if you set regenerate_knn to
\code{TRUE}.}

\item{cells_to_use}{Optional string. Names of the cells to use for the
generation of the meta-cells. If provided, this function will regenerate the
kNN graph no matter what.}

\item{target_size}{Numeric. The library target size to normalise the meta
cells to.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A \code{\link[=meta_cells]{meta_cells()}} with the data generated by this meta cell
aggregation method.
}
\description{
This function implements the meta cell aggregation from Bilous, et al.
The core idea is to use the walktrap community detection on the kNN graph.
Briefly, either the provided kNN graph in the object will be used or you
can regenerate that one (see parameters). The walk trap algorithm, a
community detection algorithm that uses short random walks, is applied to
the kNN graph and the number of cuts (i.e., final resulting communities) is
defined by the \code{graining_factor}.
}
\details{
The function will be executed on the cells that are defined by
\code{\link[=set_cells_to_keep]{set_cells_to_keep()}}, if cells_to_use is set to \code{NULL}. This is
the default behaviour of the function.
}
\references{
Bilous, et al. BMC Bioinform., 2022
}
