% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_references.R
\name{get_symphony_hvg_names}
\alias{get_symphony_hvg_names}
\title{Getter for the HVG gene names of a Symphony reference}
\usage{
get_symphony_hvg_names(object)
}
\arguments{
\item{object}{\code{SymphonyReference} class.}
}
\value{
Character vector of HVG gene names in reference loading order.
}
\description{
Getter for the HVG gene names of a Symphony reference
}
\examples{
# the HVG names in reference loading order
ref <- demo_single_cells(
  syn_data_params = params_sc_synthetic_data(
    n_cells = 500L,
    n_genes = 50L,
    n_batches = 2L
  )
)
symphony_ref <- build_symphony_ref(
  ref,
  batch_column = "batch_index",
  hvg = get_hvg(ref) + 1L,
  harmony_params = params_sc_harmony(k = 10L),
  no_pcs = 10L,
  label_columns = "cell_grp",
  .verbose = FALSE
)
head(get_symphony_hvg_names(symphony_ref))

unlink(ref@dir_data, recursive = TRUE, force = TRUE)
}
