% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_tf_to_gene}
\alias{get_tf_to_gene}
\alias{get_tf_to_gene.ScenicGrn}
\title{Extract the TF to gene data from the ScenicGrn object}
\usage{
get_tf_to_gene(x)

\method{get_tf_to_gene}{ScenicGrn}(x)
}
\arguments{
\item{x}{\code{ScenicGrn} object from which to extract the TF to gene data.table.}
}
\value{
data.table with TF to gene information
}
\description{
Extract the TF to gene data from the ScenicGrn object
}
\examples{
# the TF to gene links that survived the importance filter
sc <- demo_single_cells()
grn <- scenic_grn_sc(
  sc,
  tf_ids = sprintf("gene_\%02d", 1:5),
  scenic_params = params_scenic(
    min_counts = 1L,
    learner_params = list(n_trees = 20L)
  ),
  .verbose = FALSE
)
grn <- identify_tf_to_genes(
  grn,
  method = "top_k",
  k_tfs = 3L,
  .verbose = FALSE
)
head(get_tf_to_gene(grn))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
