% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_updated_gene_indices}
\alias{get_updated_gene_indices}
\title{Set gene mapping for sc_mapper object}
\usage{
get_updated_gene_indices(x, gene_indices)
}
\arguments{
\item{x}{An object to get the updated gene indices from}

\item{gene_indices}{Integers. The old gene indices}
}
\description{
Set gene mapping for sc_mapper object
}
