% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_updated_gene_indices.sc_mapper}
\alias{get_updated_gene_indices.sc_mapper}
\title{Get updated gene indices}
\usage{
\method{get_updated_gene_indices}{sc_mapper}(x, gene_indices)
}
\arguments{
\item{x}{An \code{sc_mapper} object}

\item{gene_indices}{Integer vector. The gene indices to map.}
}
\value{
Updated gene indices
}
\description{
Get updated gene indices
}
