% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{get_used_cells}
\alias{get_used_cells}
\title{Extract the cells used of a given object}
\usage{
get_used_cells(x)
}
\arguments{
\item{x}{The object from which to extract the used cells.}
}
\value{
If found, the cell indices of the used cells.
}
\description{
Extract the cells used of a given object
}
