% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_meta_cell.R
\name{head,bixverse::MetaCells-method}
\alias{head,bixverse::MetaCells-method}
\title{head Method for MetaCells object}
\usage{
## S7 method for class <bixverse::MetaCells>
head(x, n = 6L, ...)
}
\arguments{
\item{x}{An object of class \code{MetaCells}.}

\item{n}{Integer. Number of rows to return. Defaults to \code{6L}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A data.table with the first \code{n} rows of the obs table.
}
\description{
Returns the first \code{n} rows of the obs table from a \code{MetaCells} object.
}
\keyword{internal}
