% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{head.SingleCells}
\alias{head.SingleCells}
\title{head Method for SingleCells object}
\arguments{
\item{x}{An object of class \code{SingleCells}.}

\item{n}{Integer. Number of rows to return. Defaults to \code{6L}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A data.table with the first \code{n} rows of the obs table.
}
\description{
Returns the first \code{n} rows of the obs table from a \code{SingleCells} object.
}
\keyword{internal}
