% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_groups.R
\name{head.SingleCellsSubset}
\alias{head.SingleCellsSubset}
\title{head Method for SingleCellsSubset object}
\usage{
## S7 method for class <bixverse::SingleCellsSubset>
head(x, n = 6L, ...)
}
\arguments{
\item{x}{An object of class \code{SingleCellsSubset}.}

\item{n}{Integer. Number of rows to return. Defaults to \code{6L}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A data.table with the first \code{n} rows of the obs table.
}
\description{
Returns the first \code{n} rows of the obs table from a \code{SingleCellsSubset} object.
}
\keyword{internal}
