% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{hedges_g_dge_list}
\alias{hedges_g_dge_list}
\title{Calculate the effect}
\usage{
hedges_g_dge_list(meta_info, main_contrast, dge_list, ..., .verbose = TRUE)
}
\arguments{
\item{meta_info}{data.table. The meta information about the experiment in
which the contrast info can be found.}

\item{main_contrast}{String. Which column contains the main groups you want
to calculate the Hedge's G effect for. Every permutation of the groups
will be tested.}

\item{dge_list}{DGEList, see \code{\link[edgeR:DGEList]{edgeR::DGEList()}}.}

\item{...}{Additional parameters to forward to \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr()}}.}

\item{.verbose}{Controls verbosity of the function.}
}
\value{
A data.table with the effect sizes and standard errors based on the
Hedge's G effect size for all found permutations of the groups.
}
\description{
Calculate the effect
}
