% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{hypergeom_test_list}
\alias{hypergeom_test_list}
\title{Run hypergeometric tests over a list of target gene sets.}
\usage{
hypergeom_test_list(target_gene_lists, gene_sets, gene_universe)
}
\description{
Run hypergeometric tests over a list of target gene sets.
}
