% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{ica_general_params}
\alias{ica_general_params}
\title{Wrapper function for standard ICA parameters}
\usage{
ica_general_params(maxit = 200L, alpha = 1, max_tol = 1e-04, verbose = FALSE)
}
\arguments{
\item{maxit}{Integer. Maximum number of iterations for ICA.}

\item{alpha}{Float. The alpha parameter for the logcosh version of ICA.
Should be between 1 to 2.}

\item{max_tol}{Numeric. Should be \verb{0 < max_tol < 1}. Maximum tolerance of the
algorithm.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the parameters for usage in subsequent functions.
}
\description{
Wrapper function for standard ICA parameters
}
