% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{ica_optimal_ncomp}
\alias{ica_optimal_ncomp}
\title{Identify stability inflection point}
\usage{
ica_optimal_ncomp(object, span = 0.2, show_plot = TRUE, .verbose = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}. You need to apply
\code{\link[=ica_evaluate_comp]{ica_evaluate_comp()}} before running this function.}

\item{span}{Float. A value between 0.1 and 1.0 for the loess function
defining the span. Defaults to \code{0.2}.}

\item{show_plot}{Boolean. Shall a plot be shown of the loess function fitted.
Defaults to \code{TRUE}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
\code{BulkCoExp} with optimal ncomp based on the inflection point method.
}
\description{
This function can be used after having run \code{\link[=ica_evaluate_comp]{ica_evaluate_comp()}}.
It will calculate the inflection point, based on the first derivative of a
loess function fitted \code{ncomp ~ combined_score} with the combined score being
a product of the median stability, \code{1 - max_abs_loading_cor} and proportion
of convergence
and add these info to the object. Should the loess function raise a warning
(e.g., singularity), the class will be returned as is and manual
determination of optimal ncomp is warranted. Additionally, you have the
option to plot the loess function for additional control over the span
parameter (defaults to \code{TRUE}).
}
\examples{
# inflection point of the combined score
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- ica_processing(obj, .verbose = FALSE)
obj <- ica_evaluate_comp(
  obj,
  ica_type = "logcosh",
  ncomp_params = params_ica_ncomp(custom_seq = seq(2L, 20L, by = 2L)),
  .verbose = FALSE
)
obj <- ica_optimal_ncomp(
  obj, span = 0.4, show_plot = FALSE, .verbose = FALSE
)
obj@params$ica_stability_assessment$optimal_ncomp
}
