% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{knn_graph_label_propagation}
\alias{knn_graph_label_propagation}
\title{kNN-based graph label propagation}
\usage{
knn_graph_label_propagation(
  from,
  to,
  labels,
  weights = NULL,
  label_prop_params = params_label_propagation()
)
}
\arguments{
\item{from}{Integer vector. Source node indices for each edge.}

\item{to}{Integer vector. Target node indices for each edge. Must be the
same length as \code{from}.}

\item{labels}{Character vector. The labels with \code{NA} indicating unlabelled.}

\item{weights}{Optional numeric vector. The weights between the nodes. Must
be the same length as \code{from}. If not provided, all edges are treated as
unweighted.}

\item{label_prop_params}{List. Named list of label propagation parameters as
generated by \code{\link[=params_label_propagation]{params_label_propagation()}}. If not provided, defaults will be
used.}
}
\value{
A list with the following elements:
\itemize{
\item \code{assignment_probs} - Matrix with the assignment probabilities.
\item \code{final_labels} - Final labels in the graph.
}
}
\description{
In case of a kNN graph with a subset of nodes having labels, this function
can be used to propagate the labels through the graph. This function uses
the label spreading version of the algorithm.
}
\examples{
# spread two labels through a five node kNN graph
from <- as.integer(c(1, 1, 2, 3, 4))
to <- as.integer(c(2, 3, 3, 4, 5))
labels <- c("A", NA, "B", NA, NA)
res <- knn_graph_label_propagation(from = from, to = to, labels = labels)
res$final_labels
}
