% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{limma_contrasts}
\alias{limma_contrasts}
\title{Create all limma contrasts from}
\usage{
limma_contrasts(limma_fit, contrast_list)
}
\arguments{
\item{limma_fit}{The fitted limma model, i.e., output of \code{\link[limma:lmFit]{limma::lmFit()}}.}

\item{contrast_list}{String vector. The list of contrast of interest.}
}
\value{
The Limma contrasts for further usage.
}
\description{
Create all limma contrasts from
}
