% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_go_human_data}
\alias{load_go_human_data}
\title{Get the Gene Ontology data human}
\usage{
load_go_human_data()
}
\value{
A list containing:
\itemize{
\item go_info - data.table. The gene ontology identifier, name and namespace
can be found in this one
\item gene_ontology - data.table. The relationships between different
gene ontology terms.
\item go_to_genes - data.table. The gene ontology term to gene (ensembl id)
relationships.
}
}
\description{
This function loads in gene ontology data stored in the package. This is
for humans only.
}
\examples{
# the human gene ontology tables shipped with the package
go_data <- load_go_human_data()
names(go_data)
head(go_data$go_info)
}
