% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_meta_cell.R
\name{merge.MetaCells}
\alias{merge.MetaCells}
\title{Merging of MetaCell objects}
\usage{
\method{merge}{MetaCells}(x, feature_space = c("union"))
}
\arguments{
\item{x}{List. Named list of \code{MetaCells} objects to merge}

\item{feature_space}{String. How do deal with different set of features
across samples, can be either using the "intersect" or the "union".
(Default: "union")}
}
\description{
A named list of \code{MetaCells} objects can be provided. \code{MetaCells}
objects will be merged by concatenating the counts and obs table.
The meta_cell_id is renamed using the names of the list.
}
