% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{meta_cells}
\alias{meta_cells}
\title{bixverse meta cell class (deprecated)}
\usage{
meta_cells(meta_cell_data, var_data, meta_cell_method)
}
\arguments{
\item{meta_cell_data}{Named list. Output of meta-cell generation Rust
functions. Will contain the aggregated raw and normalised counts, plus
additional information on the origin of the meta cells.}

\item{var_data}{data.table with the variable/feature informations.}

\item{meta_cell_method}{String describing the origin of the metacell.}
}
\value{
Returns a \code{\link[=MetaCells]{MetaCells()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=MetaCells]{MetaCells()}}.
}
\keyword{internal}
