% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{multilevel_error}
\alias{multilevel_error}
\title{Helper function to calculate the multi-level error}
\usage{
multilevel_error(pval, sample_size)
}
\arguments{
\item{pval}{Numerical vector. The p-values.}

\item{sample_size}{Integer. The sample size.}
}
\value{
Returns the log2error
}
\description{
Helper function to calculate the multi-level error
}
