% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{network_diffusions}
\alias{network_diffusions}
\title{Network diffusion class (deprecated)}
\usage{
network_diffusions(edge_data_frame, weighted, directed)
}
\arguments{
\item{edge_data_frame}{data.table that contains the edge information. It is
expected to have the columns 'from' and 'to'.}

\item{weighted}{Boolean. Is the graph weighted. If set to TRUE, the
\code{edge_data_frame} needs to have a weight column.}

\item{directed}{Boolean. Shall the graph be stored as directed.}
}
\value{
Returns a \code{\link[=NetworkDiffusions]{NetworkDiffusions()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=NetworkDiffusions]{NetworkDiffusions()}}.
}
\keyword{internal}
