% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{new_adt_counts}
\alias{new_adt_counts}
\title{Generates a new \code{ADTCounts} class}
\usage{
new_adt_counts(raw_counts, cell_names)
}
\arguments{
\item{raw_counts}{Numeric matrix. The raw ADT counts.}

\item{cell_names}{String. The barcodes to keep (from the transcriptomics
for example.)}
}
\value{
\code{ADTCounts} that contains the raw and normalised ADT counts.
}
\description{
Generates a new \code{ADTCounts} class
}
