% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{new_adt_counts_clr}
\alias{new_adt_counts_clr}
\title{Generates a new \code{ADTCounts} class}
\usage{
new_adt_counts_clr(
  raw_counts,
  cell_info,
  seurat_clr = FALSE,
  clean_clr_counts = TRUE,
  percentile = 0.01
)
}
\arguments{
\item{raw_counts}{Numeric matrix. The raw ADT counts.}

\item{cell_info}{Named integer vector. Output of \code{\link[=get_cell_info]{get_cell_info()}}. Defines
as elements the cell indices (R-based) and as names the barcodes.}

\item{seurat_clr}{Boolean. Shall a Seurat-style CLR be applied.}

\item{clean_clr_counts}{Boolean. Shall the per-protein 1st percentile be
removed from the CLR normalised counts.}

\item{percentile}{Numeric. The percentile to remove to reduce background
effects.}
}
\value{
\code{ADTCounts} that contains the raw and normalised ADT counts.
}
\description{
This function generates a new \code{ADTCounts} class which uses CLR normalisation
under the hood. You have the choice between Seurat-style CLR (no negative
values) and normal CLR (allows negative values).
}
\examples{
# CLR normalisation of synthetic ADT counts
adt <- generate_single_cell_test_data_adt()
cell_info <- stats::setNames(
  seq_len(nrow(adt$counts)),
  rownames(adt$counts)
)
new_adt_counts_clr(adt$counts, cell_info = cell_info)
}
