% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{new_sc_cache}
\alias{new_sc_cache}
\title{Helper function to hold relevant cached data}
\usage{
new_sc_cache()
}
\value{
Generates an empty version of the \code{sc_cache} class.
}
\description{
Helper class that contains various data that is held in memory and not on
disk.
}
