% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{new_sc_hotspot_res}
\alias{new_sc_hotspot_res}
\title{Helper function to generate HotSpot data}
\usage{
new_sc_hotspot_res(hotspot_res, used_genes, used_cells)
}
\arguments{
\item{hotspot_res}{Named list with the following items:
\itemize{
\item z - Pairwise gene-gene matrix with the Z-scores.
\item cor - Pairwise gene-gene matrix with the local correlation
coefficients.
}}

\item{used_genes}{Character vector. The used genes to generate the matrices.}

\item{used_cells}{Character vector. The used cells to generate the matrices.}
}
\value{
Generates the \code{sc_hotspot} class.
}
\description{
Wrapper class that stores the HotSpot
}
