% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{new_sc_knn}
\alias{new_sc_knn}
\title{Helper function to generate kNN data with distances}
\usage{
new_sc_knn(knn_data, used_cells)
}
\arguments{
\item{knn_data}{Named list with the following items:
\itemize{
\item indices - Integer matrix containing the indices of the nearest
neighbours (0-indexed).
\item dist - Numerical matrix containing the distances to the nearest
neighbours.
\item dist_metric - String. Distance metric used.
}}

\item{used_cells}{Character vector. The cells used to generate the kNN graph
with the distances.}
}
\value{
Generates the \code{SingleCellNearestNeighbour} class.
}
\description{
Wrapper class that stores kNN data for subsequent usage in various other
functions, methods and helpers.
}
\examples{
# rebuild the kNN wrapper from an existing graph
sc <- demo_single_cells()
knn <- generate_knn_sc(sc, .validate_index = FALSE, .verbose = FALSE)
new_sc_knn(
  knn_data = list(
    indices = get_knn_mat(knn),
    dist = get_knn_dist(knn),
    dist_metric = "euclidean"
  ),
  used_cells = get_cell_names(sc)
)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
