% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{new_sc_list}
\alias{new_sc_list}
\title{Generate an ScListRes instance}
\usage{
new_sc_list(res, cell_indices)
}
\arguments{
\item{res}{Named list. The results from a single cell analysis in list form.}

\item{cell_indices}{Integer. The cells that were included in the analysis.}
}
\value{
\code{ScListRes} class for subsequant usage.
}
\description{
Generate an ScListRes instance
}
\keyword{internal}
