% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{new_sc_mapping}
\alias{new_sc_mapping}
\title{Helper function to construct relevant maps}
\usage{
new_sc_mapping()
}
\value{
Generates an empty version of the \code{sc_mapper} class.
}
\description{
Helper class that contains various mappings and makes it easier to use
getters/setters with
}
