% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{new_sc_matrix}
\alias{new_sc_matrix}
\title{Generate an ScMatrixRes instance}
\usage{
new_sc_matrix(res, cell_indices)
}
\arguments{
\item{res}{Numeric matrix. Of shape samples x features.}

\item{cell_indices}{Integer. The cells that were included in the analysis.}
}
\value{
\code{ScMatrixRes} class for subsequant usage.
}
\description{
Generate an ScMatrixRes instance
}
\keyword{internal}
