% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nightly_feature}
\alias{nightly_feature}
\title{Helper functions to mark something unstable/nightly}
\usage{
nightly_feature()
}
\value{
Throws a warning.
}
\description{
Helper functions to mark something unstable/nightly
}
