% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{Helper to generate one-hot encodings}
\usage{
one_hot_encode(labels)
}
\arguments{
\item{labels}{String vector. \code{NA}s signify unlabelled data.}
}
\value{
One-hot encoded matrix.
}
\description{
Helper function that generate one-hot encoding with the option
of unlabelled data. (Denoted as \code{NA} in the labels vector.)
}
