% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ontology}
\alias{ontology}
\title{Ontology class (deprecated)}
\usage{
ontology(parent_child_dt, .verbose = TRUE)
}
\arguments{
\item{parent_child_dt}{A data.table containing the ontological information
as parent-child relationships. Must contain the \code{c("parent", "child")}
columns.}

\item{.verbose}{Boolean. Controls the verbosity of the class.}
}
\value{
Returns a \code{\link[=OntologySim]{OntologySim()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=OntologySim]{OntologySim()}}.
}
\keyword{internal}
