% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{OntologySim}
\alias{OntologySim}
\title{OntologySim class}
\usage{
OntologySim(parent_child_dt, .verbose = TRUE)
}
\arguments{
\item{parent_child_dt}{A data.table that contains the ontological information
in terms of parent child relationships. Need to contain the
\code{c("parent", "child")} columns.}

\item{.verbose}{Boolean. Controls the verbosity of the class}
}
\value{
Returns the class for subsequent usage.
}
\description{
This class is used to store any ontology and apply different methods to it.
Currently implemented are different types of term similarities, i.e., based
on semantic similarities and the Wang similarity.
}
\section{Properties}{

\describe{
\item{edge_dt}{data.table. Contains the parent-child relationships. (For
Wang similarity also the relationship type.)}
\item{outputs}{List. Contains various intermediary results used for some
methods.}
\item{sim_mat}{List. Contains the potentially calculated similarity
matrix in form of an R6 class. Getters to access the data are provided.}
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{Final results stored in the class.}
}
}

\examples{
# wrap a toy parent-child ontology into the class
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f"),
  type = c("part_of", "part_of", "part_of", "is_a", "is_a")
)
OntologySim(onto, .verbose = FALSE)
}
