% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{ot_harmonic_score}
\alias{ot_harmonic_score}
\title{Calculates a harmonic sum normalised between 0 to 1.}
\usage{
ot_harmonic_score(x)
}
\arguments{
\item{x}{Numeric vector. Needs to be between 0 and 1.}
}
\value{
Harmonic, normalised sum of the provided scores.
}
\description{
The function takes in a vector of scores between 0 and 1 and calculates a
harmonic sum, based on the approach OpenTargets takes to do their gene -
disease evidence scores, see:
https://platform-docs.opentargets.org/associations.
}
\examples{
# harmonic sum over four evidence scores
ot_harmonic_score(c(1, 0.8, 0.5, 0.1))
}
