% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_community_detection}
\alias{params_community_detection}
\title{Wrapper function to generate community detection parameters}
\usage{
params_community_detection(
  max_nodes = 300L,
  min_nodes = 10L,
  min_seed_nodes = 2L,
  initial_res = 0.5,
  threshold_type = c("prop_based", "pval_based"),
  network_threshold = 0.5,
  pval_threshold = 0.1
)
}
\arguments{
\item{max_nodes}{Integer. Maximum number of nodes in a given community.
Defaults to \code{300L}.}

\item{min_nodes}{Integer. Minimum number of nodes in a given community.
Defaults to \code{10L}.}

\item{min_seed_nodes}{Integer. Minimum number of seed nodes within a
community. Defaults to \code{2L}.}

\item{initial_res}{Numeric. Initial resolution parameter to start with.
Defaults to \code{0.5}.}

\item{threshold_type}{String. You can chose to include a certain proportion
of the network with the highest diffusion scores, or use p-values based on
permutations. One of \code{c("prop_based", "pval_based")}. Defaults to
\code{"prop_based"}.}

\item{network_threshold}{Numeric. The proportion of the network to include.
Used if \code{threshold_type = "prop_based"}. Defaults to \code{0.5}.}

\item{pval_threshold}{Numeric. The maximum p-value for nodes to be included.
Used if \code{threshold_type = "pval_based"}. Defaults to \code{0.1}.}
}
\value{
A named list with the following elements:
\itemize{
\item max_nodes - Integer. Maximum number of nodes in a given community.
Defaults to \code{300L}.
\item min_nodes - Integer. Minimum number of nodes in a given community.
Defaults to \code{10L}.
\item min_seed_nodes - Integer. Minimum number of seed nodes within a
community. Defaults to \code{2L}.
\item initial_res - Numeric. Initial resolution parameter to start with.
Defaults to \code{0.5}.
\item threshold_type - String. You can chose to include a certain proportion
of the network with the highest diffusion scores, or use p-values based on
permutations. One of \code{c("prop_based", "pval_based")}. Defaults to
\code{"prop_based"}.
\item network_threshold - Numeric. The proportion of the network to include.
Used if \code{threshold_type = "prop_based"}. Defaults to \code{0.5}.
\item pval_threshold - Numeric. The maximum p-value for nodes to be
included. Used if \code{threshold_type = "pval_based"}. Defaults to \code{0.1}.
}
}
\description{
Wrapper function to generate community detection parameters
}
