% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_cor_graph}
\alias{params_cor_graph}
\title{Wrapper function for graph generation}
\usage{
params_cor_graph(
  epsilon = 2,
  min_cor = 0.2,
  fdr_threshold = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{epsilon}{Numeric. Defines the epsilon parameter for the radial basis
function. Defaults to \code{2.0}.}

\item{min_cor}{Numeric. Minimum absolute correlation that needs to be
observed in either data set. Only relevant for differential correlation-based
graphs. Defaults to \code{0.2}.}

\item{fdr_threshold}{Numeric. Maximum FDR for the differential correlation
p-value. Defaults to \code{0.05}.}

\item{verbose}{Boolean. Controls verbosity of the graph generation function.
Defaults to \code{TRUE}.}
}
\value{
A named list with the following elements:
\itemize{
\item epsilon - Numeric. Defines the epsilon parameter for the radial basis
function. Defaults to \code{2.0}.
\item min_cor - Numeric. Minimum absolute correlation that needs to be
observed in either data set. Only relevant for differential
correlation-based graphs. Defaults to \code{0.2}.
\item fdr_threshold - Numeric. Maximum FDR for the differential correlation
p-value. Defaults to \code{0.05}.
\item verbose - Boolean. Controls verbosity of the graph generation
function. Defaults to \code{TRUE}.
}
}
\description{
Wrapper function for graph generation
}
