% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_coremo}
\alias{params_coremo}
\title{Wrapper function to generate CoReMo parameters}
\usage{
params_coremo(
  epsilon = 2,
  k_min = 2L,
  k_max = 150L,
  min_size = NULL,
  junk_module_threshold = 0.05,
  rbf_func = c("gaussian", "inverse_quadratic", "bump"),
  cor_method = c("spearman", "pearson")
)
}
\arguments{
\item{epsilon}{Numeric. Epsilon parameter for the chosen RBF function, see
\code{rbf_func}. The higher, the more aggressively low correlations will be
shrunk. Defaults to \code{2.0}.}

\item{k_min}{Integer. Minimum and maximum number of cuts to use for the
hierarchical clustering. Defaults to \code{2L}.}

\item{k_max}{Integer. Minimum and maximum number of cuts to use for the
hierarchical clustering. Defaults to \code{150L}.}

\item{min_size}{Integer or \code{NULL}. Minimum size of the clusters. Smaller
clusters will be combined together. Defaults to \code{NULL}.}

\item{junk_module_threshold}{Numeric. Threshold for the minimum correlation
to be observed in a module. Defaults to \code{0.05}.}

\item{rbf_func}{String. Type of RBF you wish to apply to down-weigh weak
correlations. One of \code{c("gaussian", "inverse_quadratic", "bump")}. Defaults
to \code{"gaussian"}.}

\item{cor_method}{String. The type of correlation to use. One of
\code{c("spearman", "pearson")}. Defaults to \code{"spearman"}.}
}
\value{
A named list with the following elements:
\itemize{
\item epsilon - Numeric. Epsilon parameter for the chosen RBF function, see
\code{rbf_func}. The higher, the more aggressively low correlations will be
shrunk. Defaults to \code{2.0}.
\item k_min - Integer. Minimum and maximum number of cuts to use for the
hierarchical clustering. Defaults to \code{2L}.
\item k_max - Integer. Minimum and maximum number of cuts to use for the
hierarchical clustering. Defaults to \code{150L}.
\item min_size - Integer or \code{NULL}. Minimum size of the clusters. Smaller
clusters will be combined together. Defaults to \code{NULL}.
\item junk_module_threshold - Numeric. Threshold for the minimum correlation
to be observed in a module. Defaults to \code{0.05}.
\item rbf_func - String. Type of RBF you wish to apply to down-weigh weak
correlations. One of \code{c("gaussian", "inverse_quadratic", "bump")}. Defaults
to \code{"gaussian"}.
\item cor_method - String. The type of correlation to use. One of
\code{c("spearman", "pearson")}. Defaults to \code{"spearman"}.
}
}
\description{
Wrapper function to generate CoReMo parameters
}
