% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_dgrdl}
\alias{params_dgrdl}
\title{Wrapper function to generate DGRDL parameters}
\usage{
params_dgrdl(
  sparsity = 5L,
  dict_size = 5L,
  alpha = 1,
  beta = 1,
  max_iter = 20L,
  k_neighbours = 5L,
  admm_iter = 5L,
  rho = 1
)
}
\arguments{
\item{sparsity}{Integer. Sparsity constraint (max non-zero coefficients per
signal) Defaults to \code{5L}.}

\item{dict_size}{Integer. Dictionary size Defaults to \code{5L}.}

\item{alpha}{Numeric. Sample context regularisation weight. Defaults to
\code{1.0}.}

\item{beta}{Numeric. Feature effect regularisation weight. Defaults to \code{1.0}.}

\item{max_iter}{Integer. Maximum number of iterations for the main algorithm.
Defaults to \code{20L}.}

\item{k_neighbours}{Integer. Number of neighbours in the KNN graph. Defaults
to \code{5L}.}

\item{admm_iter}{Integer. ADMM iterations for sparse coding. Defaults to
\code{5L}.}

\item{rho}{Numeric. ADMM step size. Defaults to \code{1.0}.}
}
\value{
A named list with the following elements:
\itemize{
\item sparsity - Integer. Sparsity constraint (max non-zero coefficients per
signal) Defaults to \code{5L}.
\item dict_size - Integer. Dictionary size Defaults to \code{5L}.
\item alpha - Numeric. Sample context regularisation weight. Defaults to
\code{1.0}.
\item beta - Numeric. Feature effect regularisation weight. Defaults to
\code{1.0}.
\item max_iter - Integer. Maximum number of iterations for the main
algorithm. Defaults to \code{20L}.
\item k_neighbours - Integer. Number of neighbours in the KNN graph.
Defaults to \code{5L}.
\item admm_iter - Integer. ADMM iterations for sparse coding. Defaults to
\code{5L}.
\item rho - Numeric. ADMM step size. Defaults to \code{1.0}.
}
}
\description{
Wrapper function to generate DGRDL parameters
}
