% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_fast_cluster_default}
\alias{params_fast_cluster_default}
\title{Helper function to generate default parameters for the fast clustering for
the doublet detection methods}
\usage{
params_fast_cluster_default()
}
\value{
A named list with the following elements:
\itemize{
\item km_type - String. The type of k-means clustering. One of
\code{c("minibatch", "standard")}. Defaults to \code{"minibatch"}.
\item n_centroids - Integer or \code{NULL}. The number of centroids to use.
\code{NULL} uses \code{sqrt(N_cells) * 4} centroids. Defaults to \code{NULL}.
\item kmeans_iters - Integer. Number of maximum k-means iterations. Defaults
to \code{100L}.
\item batch_size - Integer. Max batch size, capped at \code{N_cells / 2}
depending on the data set. Defaults to \code{4098L}.
}
}
\description{
Helper function to generate default parameters for the fast clustering for
the doublet detection methods
}
