% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_graph_resolution}
\alias{params_graph_resolution}
\title{Wrapper function to generate resolution parameters for Leiden or Louvain
clustering.}
\usage{
params_graph_resolution(min_res = 0.1, max_res = 10, number_res = 15L)
}
\arguments{
\item{min_res}{Numeric. Minimum resolution to test. Defaults to \code{0.1}.}

\item{max_res}{Numeric. Maximum resolution to test. Defaults to \code{10.0}.}

\item{number_res}{Integer. Number of resolutions to test between the
\code{max_res} and \code{min_res.} Defaults to \code{15L}.}
}
\value{
A named list with the following elements:
\itemize{
\item min_res - Numeric. Minimum resolution to test. Defaults to \code{0.1}.
\item max_res - Numeric. Maximum resolution to test. Defaults to \code{10.0}.
\item number_res - Integer. Number of resolutions to test between the
\code{max_res} and \code{min_res.} Defaults to \code{15L}.
}
}
\description{
Wrapper function to generate resolution parameters for Leiden or Louvain
clustering.
}
