% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_gsea}
\alias{params_gsea}
\title{Wrapper function to generate GSEA parameters}
\usage{
params_gsea(
  min_size = 5L,
  max_size = 500L,
  gsea_param = 1,
  sample_size = 101L,
  eps = 1e-50
)
}
\arguments{
\item{min_size}{Integer. Minimum number of genes per gene set. Defaults to
\code{5L}.}

\item{max_size}{Integer. Maximum number of genes per gene set. Defaults to
\code{500L}.}

\item{gsea_param}{Numeric. GSEA parameter. Defaults to \code{1.0}.}

\item{sample_size}{Integer. Number of samples to iterate through for the
multi-level implementation of fgsea. Defaults to \code{101L}.}

\item{eps}{Numeric. Boundary for calculating the p-value. Used for the multi-
level implementation of fgsea. Defaults to \code{1e-50}.}
}
\value{
A named list with the following elements:
\itemize{
\item min_size - Integer. Minimum number of genes per gene set. Defaults to
\code{5L}.
\item max_size - Integer. Maximum number of genes per gene set. Defaults to
\code{500L}.
\item gsea_param - Numeric. GSEA parameter. Defaults to \code{1.0}.
\item sample_size - Integer. Number of samples to iterate through for the
multi-level implementation of fgsea. Defaults to \code{101L}.
\item eps - Numeric. Boundary for calculating the p-value. Used for the
multi- level implementation of fgsea. Defaults to \code{1e-50}.
}
}
\description{
Wrapper function to generate GSEA parameters
}
