% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_gsva}
\alias{params_gsva}
\title{Wrapper function to generate GSVA parameters}
\usage{
params_gsva(
  tau = 1,
  min_size = 5L,
  max_size = 500L,
  max_diff = TRUE,
  abs_rank = FALSE
)
}
\arguments{
\item{tau}{Numeric. Tau parameter, usual recommendation is to use \code{1.0} here.
Larger values emphasise the tails more. Defaults to \code{1.0}.}

\item{min_size}{Integer. Minimum number of genes per gene set. Defaults to
\code{5L}.}

\item{max_size}{Integer. Maximum number of genes per gene set. Defaults to
\code{500L}.}

\item{max_diff}{Boolean. Scoring mode for GSVA, if \code{TRUE} = difference; if
\code{FALSE} = larger absolute value. Defaults to \code{TRUE}.}

\item{abs_rank}{Boolean. If \code{TRUE} = pos - neg, \code{FALSE} = pos + neg. Defaults
to \code{FALSE}.}
}
\value{
A named list with the following elements:
\itemize{
\item tau - Numeric. Tau parameter, usual recommendation is to use \code{1.0}
here. Larger values emphasise the tails more. Defaults to \code{1.0}.
\item min_size - Integer. Minimum number of genes per gene set. Defaults to
\code{5L}.
\item max_size - Integer. Maximum number of genes per gene set. Defaults to
\code{500L}.
\item max_diff - Boolean. Scoring mode for GSVA, if \code{TRUE} = difference; if
\code{FALSE} = larger absolute value. Defaults to \code{TRUE}.
\item abs_rank - Boolean. If \code{TRUE} = pos - neg, \code{FALSE} = pos + neg.
Defaults to \code{FALSE}.
}
}
\description{
Wrapper function to generate GSVA parameters
}
