% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_ica_general}
\alias{params_ica_general}
\title{Wrapper function for standard ICA parameters}
\usage{
params_ica_general(maxit = 200L, alpha = 1, max_tol = 1e-04, verbose = FALSE)
}
\arguments{
\item{maxit}{Integer. Maximum number of iterations for ICA. Defaults to
\code{200L}.}

\item{alpha}{Numeric. The alpha parameter for the logcosh version of ICA.
Should be between 1 to 2. Defaults to \code{1.0}.}

\item{max_tol}{Numeric. Should be \verb{0 < max_tol < 1}. Maximum tolerance of the
algorithm. Defaults to \code{1e-04}.}

\item{verbose}{Boolean. Controls verbosity of the function. Defaults to
\code{FALSE}.}
}
\value{
A named list with the following elements:
\itemize{
\item maxit - Integer. Maximum number of iterations for ICA. Defaults to
\code{200L}.
\item alpha - Numeric. The alpha parameter for the logcosh version of ICA.
Should be between 1 to 2. Defaults to \code{1.0}.
\item max_tol - Numeric. Should be \verb{0 < max_tol < 1}. Maximum tolerance of
the algorithm. Defaults to \code{1e-04}.
\item verbose - Boolean. Controls verbosity of the function. Defaults to
\code{FALSE}.
}
}
\description{
Wrapper function for standard ICA parameters
}
