% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_ica_ncomp}
\alias{params_ica_ncomp}
\title{Wrapper function for ICA ncomp iterations}
\usage{
params_ica_ncomp(max_no_comp = 75L, steps = 5L, custom_seq = NULL)
}
\arguments{
\item{max_no_comp}{Integer. Maximum number of ncomp to test. Defaults to
\code{75L}.}

\item{steps}{Integer. In which steps to move from 5 onwards. Defaults to
\code{5L}.}

\item{custom_seq}{Integer vector or \code{NULL}. If you wish to provide a custom
version of no_comp to iterate through. If NULL, you will iterate through
\verb{c(2, 3, 4, 5, 5 + step, ... max_no_comp - step, max_no_comp)} Defaults to
\code{NULL}.}
}
\value{
A named list with the following elements:
\itemize{
\item max_no_comp - Integer. Maximum number of ncomp to test. Defaults to
\code{75L}.
\item steps - Integer. In which steps to move from 5 onwards. Defaults to
\code{5L}.
\item custom_seq - Integer vector or \code{NULL}. If you wish to provide a custom
version of no_comp to iterate through. If NULL, you will iterate through
\verb{c(2, 3, 4, 5, 5 + step, ... max_no_comp - step, max_no_comp)} Defaults to
\code{NULL}.
}
}
\description{
Wrapper function to provide parameters through which ncomps to
iterate through.
}
