% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_kmeans_defaults}
\alias{params_kmeans_defaults}
\title{K-mean parameter defaults.}
\usage{
params_kmeans_defaults()
}
\value{
A named list with the following elements:
\itemize{
\item k_means_iter - Integer. The number of iterations to use for the
clustering. Defaults to \code{30L}.
\item k_means_init - String. The initialisation. One of \code{c("parallel", "random")}. Defaults to \code{"parallel"}.
\item gemm - Boolean or \code{NULL}. Controls which CPU implementation is used by
the method. GEMM is faster with large dimensionality. Defaults to \code{FALSE}.
\item hamerly - Boolean or \code{NULL}. Shall a faster exact method be used
leveraging the triangle inequality. Faster on large data sets with large
numbers of centroids. Defaults to \code{TRUE}.
}
}
\description{
Helper function to generate defaults for the k-mean clustering
were more control is needed.
}
