% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_norm_doublet_detection_defaults}
\alias{params_norm_doublet_detection_defaults}
\title{Helper function to generate normalisation defaults for doublet detection.}
\usage{
params_norm_doublet_detection_defaults()
}
\value{
A list with the following parameters for normalisation specifically
designed for doublet detection methods:
\itemize{
\item log_transform - Boolean. Shall the counts be log-normalised.
Defaults to \code{TRUE}.
\item mean_center - Boolean. Shall mean centreing be applied. Defaults
to \code{FALSE}.
\item normalise_variance - Boolean. Shall the variance be normalised.
Defaults to \code{FALSE}.
\item target_size - Target library size. Defaults to \code{1e6}
}
}
\description{
Helper function to generate normalisation defaults for doublet detection.
}
