% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_norm_doublets_defaults}
\alias{params_norm_doublets_defaults}
\title{Helper function to generate normalisation defaults for doublet detection.}
\usage{
params_norm_doublets_defaults()
}
\value{
A named list with the following elements:
\itemize{
\item log_transform - Boolean. Shall the counts be log-normalised. Defaults
to \code{TRUE}.
\item mean_center - Boolean. Shall mean centring be applied. Defaults to
\code{FALSE}.
\item normalise_variance - Boolean. Shall the variance be normalised.
Defaults to \code{FALSE}.
\item target_size - Numeric. Target library size. Defaults to \code{1000000.0}.
}
}
\description{
Helper function to generate normalisation defaults for doublet detection.
}
