% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_sc_fastmnn}
\alias{params_sc_fastmnn}
\title{Wrapper function for the fastMNN parameters}
\usage{
params_sc_fastmnn(
  ndist = 3,
  cos_norm = TRUE,
  no_pcs = 30L,
  sparse_svd = TRUE,
  knn = list(k = 20L),
  pca = params_sc_pca()
)
}
\arguments{
\item{ndist}{Numeric. Number of median distances for the tricube kernel
bandwidth. Defaults to \code{3.0}.}

\item{cos_norm}{Boolean. Apply cosine normalisation before computing
distances. Defaults to \code{TRUE}.}

\item{no_pcs}{Integer. Number of PCs to use for MNN calculations. Defaults to
\code{30L}.}

\item{sparse_svd}{Boolean. Shall the sparse SVD be used. Defaults to \code{TRUE}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{extract_knn}, \code{m}, \code{ef_construction},
\code{ef_search}, \code{n_list} and \code{n_probe}. See \code{\link[=params_knn_defaults]{params_knn_defaults()}} for the
available elements. Defaults to \code{list(k = 20L)}.}

\item{pca}{List. Parameters to feed through to the optional recalculation of
the PCA, see \code{\link[=params_sc_pca]{params_sc_pca()}}. See \code{\link[=params_sc_pca]{params_sc_pca()}} for the available
elements. Defaults to \code{params_sc_pca()}.}
}
\value{
A named list with the following elements:
\itemize{
\item ndist - Numeric. Number of median distances for the tricube kernel
bandwidth. Defaults to \code{3.0}.
\item cos_norm - Boolean. Apply cosine normalisation before computing
distances. Defaults to \code{TRUE}.
\item no_pcs - Integer. Number of PCs to use for MNN calculations. Defaults
to \code{30L}.
\item sparse_svd - Boolean. Shall the sparse SVD be used. Defaults to
\code{TRUE}.
\item The elements of \code{\link[=params_knn_defaults]{params_knn_defaults()}}, overridden by \code{knn}, spliced
in at this position.
\item The elements of \code{\link[=params_sc_pca]{params_sc_pca()}}, overridden by \code{pca}, spliced in at
this position.
}
}
\description{
Wrapper function for the fastMNN parameters
}
