% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_metacells}
\alias{params_sc_metacells}
\title{Wrapper function for parameters for meta cell generation}
\usage{
params_sc_metacells(
  max_shared = 15L,
  target_no_metacells = 1000L,
  max_iter = 5000L,
  knn = list()
)
}
\arguments{
\item{max_shared}{Integer. Maximum number of allowed shared neighbours for
the meta cell to be considered. Defaults to \code{15L}.}

\item{target_no_metacells}{Integer. Target number of meta-cells to generate.
Defaults to \code{1000L}.}

\item{max_iter}{Integer. Maximum number of iterations for the algorithm.
Defaults to \code{5000L}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{nn_max_iter},
\code{rho}, \code{delta}. Note: This function uses \code{k = 25L} and \code{ann_dist = "cosine"}
as defaults.}
}
\value{
A list with the metacell parameters.
}
\description{
Wrapper function for parameters for meta cell generation
}
