% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_sc_min_quality}
\alias{params_sc_min_quality}
\title{Wrapper function to generate QC metric params for single cell}
\usage{
params_sc_min_quality(
  min_unique_genes = 100L,
  min_lib_size = 250L,
  min_cells = 10L,
  target_size = 10000
)
}
\arguments{
\item{min_unique_genes}{Integer. Minimum number of unique genes per cell/spot
to be included. Defaults to \code{100L}.}

\item{min_lib_size}{Integer. Minimum library size per cell/spot to be
included. Defaults to \code{250L}.}

\item{min_cells}{Integer. Minimum number of cells a gene has to be expressed
to be included. Defaults to \code{10L}.}

\item{target_size}{Numeric. The target size for the normalisation. Defaults
to \code{10000.0}.}
}
\value{
A named list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of unique genes per
cell/spot to be included. Defaults to \code{100L}.
\item min_lib_size - Integer. Minimum library size per cell/spot to be
included. Defaults to \code{250L}.
\item min_cells - Integer. Minimum number of cells a gene has to be
expressed to be included. Defaults to \code{10L}.
\item target_size - Numeric. The target size for the normalisation. Defaults
to \code{10000.0}.
}
}
\description{
Wrapper function to generate QC metric params for single cell
}
