% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_sc_mtx_io}
\alias{params_sc_mtx_io}
\title{Wrapper function to provide data for mtx-based loading}
\usage{
params_sc_mtx_io(path_mtx, path_obs, path_var, cells_as_rows, has_hdr)
}
\arguments{
\item{path_mtx}{Any. Path to the .mtx file Required.}

\item{path_obs}{Any. Path to the file containing cell/barcode info. Required.}

\item{path_var}{Any. Path to the file containing gene/variable info.
Required.}

\item{cells_as_rows}{Boolean. Do cells represent the rows or columns.
Required.}

\item{has_hdr}{Boolean. Do the plain text files have headers. Required.}
}
\value{
A named list with the following elements:
\itemize{
\item path_mtx - Any. Path to the .mtx file Required.
\item path_obs - Any. Path to the file containing cell/barcode info.
Required.
\item path_var - Any. Path to the file containing gene/variable info.
Required.
\item cells_as_rows - Boolean. Do cells represent the rows or columns.
Required.
\item has_hdr - Boolean. Do the plain text files have headers. Required.
}
}
\description{
Wrapper function to provide data for mtx-based loading
}
