% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_sc_seacells}
\alias{params_sc_seacells}
\title{Wrapper function for the SEACells parameters}
\usage{
params_sc_seacells(
  n_sea_cells,
  max_fw_iters = 50L,
  convergence_epsilon = 0.001,
  max_iter = 100L,
  min_iter = 10L,
  greedy_threshold = 20000L,
  graph_building = "union",
  pruning = TRUE,
  pruning_threshold = 1e-07,
  n_landmarks = NULL,
  knn = list()
)
}
\arguments{
\item{n_sea_cells}{Integer. Number of SEA cells to detect. Required.}

\item{max_fw_iters}{Integer. Maximum iterations for the Franke-Wolfe
algorithm. Defaults to \code{50L}.}

\item{convergence_epsilon}{Numeric. Convergence threshold. Algorithm stops
when RSS change < epsilon * RSS(0). Defaults to \code{0.001}.}

\item{max_iter}{Integer. Maximum iterations to run SEACells for. Defaults to
\code{100L}.}

\item{min_iter}{Integer. Minimum iterations to run SEACells for. Defaults to
\code{10L}.}

\item{greedy_threshold}{Integer. Maximum number of cells before defaulting to
rapid random selection of archetypes. Defaults to \code{20000L}.}

\item{graph_building}{String. Graph building method. Defaults to \code{"union"}.}

\item{pruning}{Boolean. Shall tiny values be pruned during Franke-Wolfe
updates. Defaults to \code{TRUE}.}

\item{pruning_threshold}{Numeric. If \code{pruning = TRUE} values below which
threshold shall be pruned. Defaults to \code{1e-07}.}

\item{n_landmarks}{Integer or \code{NULL}. If provided, it will use the Nystroem
extension during the archetype finding. Useful for larger data sets. Defaults
to \code{NULL}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{extract_knn}, \code{m}, \code{ef_construction},
\code{ef_search}, \code{n_list} and \code{n_probe}. See \code{\link[=params_knn_defaults]{params_knn_defaults()}} for the
available elements. Defaults to \code{list()}.}
}
\value{
A named list with the following elements:
\itemize{
\item n_sea_cells - Integer. Number of SEA cells to detect. Required.
\item max_fw_iters - Integer. Maximum iterations for the Franke-Wolfe
algorithm. Defaults to \code{50L}.
\item convergence_epsilon - Numeric. Convergence threshold. Algorithm stops
when RSS change < epsilon * RSS(0). Defaults to \code{0.001}.
\item max_iter - Integer. Maximum iterations to run SEACells for. Defaults
to \code{100L}.
\item min_iter - Integer. Minimum iterations to run SEACells for. Defaults
to \code{10L}.
\item greedy_threshold - Integer. Maximum number of cells before defaulting
to rapid random selection of archetypes. Defaults to \code{20000L}.
\item graph_building - String. Graph building method. Defaults to \code{"union"}.
\item pruning - Boolean. Shall tiny values be pruned during Franke-Wolfe
updates. Defaults to \code{TRUE}.
\item pruning_threshold - Numeric. If \code{pruning = TRUE} values below which
threshold shall be pruned. Defaults to \code{1e-07}.
\item n_landmarks - Integer or \code{NULL}. If provided, it will use the Nystroem
extension during the archetype finding. Useful for larger data sets.
Defaults to \code{NULL}.
\item The elements of \code{\link[=params_knn_defaults]{params_knn_defaults()}}, overridden by \code{knn}, spliced
in at this position.
}
}
\description{
Wrapper function for the SEACells parameters
}
