% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_sc_synthetic_data}
\alias{params_sc_synthetic_data}
\title{Default parameters for generation of synthetic single cell data (RNA)}
\usage{
params_sc_synthetic_data(
  n_cells = 1000L,
  n_genes = 100L,
  n_batches = 1L,
  marker_genes = list(cell_type_1 = list(marker_genes = 0:9), cell_type_2 =
    list(marker_genes = 10:19), cell_type_3 = list(marker_genes = 20:29)),
  batch_effect_strength = c("strong", "medium", "weak"),
  n_samples = NULL,
  sample_bias = NULL
)
}
\arguments{
\item{n_cells}{Integer. Number of cells. Defaults to \code{1000L}.}

\item{n_genes}{Integer. Number of genes. Defaults to \code{100L}.}

\item{n_batches}{Integer. Number of batches. Defaults to \code{1L}.}

\item{marker_genes}{Any. A nested list that indicates which gene indices are
markers for which cell. Defaults to \code{list(cell_type_1 = list(marker_genes = 0:9), cell_type_2 = list(marker_genes = 10:19), cell_type_3 = list(marker_genes = 20:29))}.}

\item{batch_effect_strength}{String. The strength of the batch effect to add.
One of \code{c("strong", "medium", "weak")}. Defaults to \code{"strong"}.}

\item{n_samples}{Integer or \code{NULL}. Shall sample membership be added to the
synthetic data. If you want sample information you need to provide
\code{n_samples} and \code{sample_bias}. Defaults to \code{NULL}.}

\item{sample_bias}{Any. One of \code{c("even", "slightly_uneven", "very_uneven")}
Defaults to \code{NULL}.}
}
\value{
A named list with the following elements:
\itemize{
\item n_cells - Integer. Number of cells. Defaults to \code{1000L}.
\item n_genes - Integer. Number of genes. Defaults to \code{100L}.
\item marker_genes - Any. A nested list that indicates which gene indices
are markers for which cell. Defaults to \code{list(cell_type_1 = list(marker_genes = 0:9), cell_type_2 = list(marker_genes = 10:19), cell_type_3 = list(marker_genes = 20:29))}.
\item n_batches - Integer. Number of batches. Defaults to \code{1L}.
\item batch_effect_strength - String. The strength of the batch effect to
add. One of \code{c("strong", "medium", "weak")}. Defaults to \code{"strong"}.
\item n_samples - Integer or \code{NULL}. Shall sample membership be added to the
synthetic data. If you want sample information you need to provide
\code{n_samples} and \code{sample_bias}. Defaults to \code{NULL}.
\item sample_bias - Any. One of \code{c("even", "slightly_uneven", "very_uneven")} Defaults to \code{NULL}.
}
}
\description{
For the generation of synthetic single cell data mostly for
testing or showcasing purposes. The default configurations generates 1000
cells x 100 genes with genes 1:10 being cell markers for cell type 1, genes
11:20 for cell type 2 and genes 21:30 for cell type.
}
