% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_sc_synthetic_data_adt}
\alias{params_sc_synthetic_data_adt}
\title{Default parameters for generation of synthetic single cell data (ADT)}
\usage{
params_sc_synthetic_data_adt(
  n_cells = 1000L,
  n_proteins = 15L,
  n_batches = 1L,
  marker_genes = list(cell_type_1 = list(marker_genes = 0:2), cell_type_2 =
    list(marker_genes = 3:5), cell_type_3 = list(marker_genes = 6:8)),
  isotype_controls = 12:14,
  batch_effect_strength = c("strong", "medium", "weak")
)
}
\arguments{
\item{n_cells}{Integer. Number of cells. Defaults to \code{1000L}.}

\item{n_proteins}{Integer. Number of proteins Defaults to \code{15L}.}

\item{n_batches}{Integer. Number of batches. Defaults to \code{1L}.}

\item{marker_genes}{Any. A nested list that indicates which gene indices are
markers for which cell. Defaults to \code{list(cell_type_1 = list(marker_genes = 0:2), cell_type_2 = list(marker_genes = 3:5), cell_type_3 = list(marker_genes = 6:8))}.}

\item{isotype_controls}{Integer vector. The columns that defines the isotype
controls. (0-indexed!) Defaults to \code{12:14}.}

\item{batch_effect_strength}{String. The strength of the batch effect to add.
One of \code{c("strong", "medium", "weak")}. Defaults to \code{"strong"}.}
}
\value{
A named list with the following elements:
\itemize{
\item n_cells - Integer. Number of cells. Defaults to \code{1000L}.
\item n_proteins - Integer. Number of proteins Defaults to \code{15L}.
\item marker_genes - Any. A nested list that indicates which gene indices
are markers for which cell. Defaults to \code{list(cell_type_1 = list(marker_genes = 0:2), cell_type_2 = list(marker_genes = 3:5), cell_type_3 = list(marker_genes = 6:8))}.
\item n_batches - Integer. Number of batches. Defaults to \code{1L}.
\item isotype_controls - Integer vector. The columns that defines the
isotype controls. (0-indexed!) Defaults to \code{12:14}.
\item batch_effect_strength - String. The strength of the batch effect to
add. One of \code{c("strong", "medium", "weak")}. Defaults to \code{"strong"}.
}
}
\description{
For the generation of synthetic single cell data mostly for
testing or showcasing purposes. In this case, ADT counts to test multi-modal
integration. The default configurations generates 1000 cells x 15 proteins
with probes 1:3 being cell markers for cell type 1, genes 4:6 for cell type 2
and genes 7:9 for cell type. Columns 13:15 represents isotype controls.
}
