% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_scenic_gradient_boosting_defaults}
\alias{params_scenic_gradient_boosting_defaults}
\title{Default parameters for the SCENIC GradientBoosting (GRNBoost2) regression
learner}
\usage{
params_scenic_gradient_boosting_defaults()
}
\value{
A named list with the following elements:
\itemize{
\item n_trees_max - Any. TODO Defaults to \code{1000L}.
\item learning_rate - Any. TODO Defaults to \code{0.01}.
\item max_depth - Any. TODO Defaults to \code{3L}.
\item min_samples_leaf - Any. TODO Defaults to \code{50L}.
\item early_stop_window - Any. TODO Defaults to \code{25L}.
\item subsample_rate - Any. TODO Defaults to \code{0.9}.
\item n_features_split - Any. TODO Defaults to \code{0L}.
}
}
\description{
Default parameters for the SCENIC GradientBoosting (GRNBoost2) regression
learner
}
