% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_scenic_random_forest_defaults}
\alias{params_scenic_random_forest_defaults}
\title{Default parameters for the SCENIC RandomForest regression learner}
\usage{
params_scenic_random_forest_defaults()
}
\value{
A named list with the following elements:
\itemize{
\item n_trees - Any. TODO Defaults to \code{250L}.
\item min_samples_leaf - Any. TODO Defaults to \code{50L}.
\item n_features_split - Any. TODO Defaults to \code{0L}.
\item subsample_rate - Any. TODO Defaults to \code{0.632}.
\item bootstrap - Any. TODO Defaults to \code{FALSE}.
\item max_depth - Any. TODO Defaults to \code{8L}.
\item subsample_frac - Any. TODO Defaults to \code{NULL}.
}
}
\description{
Default parameters for the SCENIC RandomForest regression learner
}
