% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_snf}
\alias{params_snf}
\title{Wrapper function to generate SNF parameters}
\usage{
params_snf(
  k = 20L,
  t = 20L,
  mu = 0.5,
  alpha = 1,
  normalise = TRUE,
  distance_metric = c("euclidean", "manhattan", "canberra", "cosine")
)
}
\arguments{
\item{k}{Integer. Number of neighbours to consider. Defaults to \code{20L}.}

\item{t}{Integer. Number of iterations for the SNF algorithm. Defaults to
\code{20L}.}

\item{mu}{Numeric. Normalisation factor for the Gaussian kernel width.
Defaults to \code{0.5}.}

\item{alpha}{Numeric. Normalisation parameter controlling the fusion
strength. Defaults to \code{1.0}.}

\item{normalise}{Boolean. Shall continuous values be Z-scored. Defaults to
\code{TRUE}.}

\item{distance_metric}{String. Which distance metric to use for the
continuous calculations. In case of pure categorical, Hamming will be used,
for mixed data types Gower distance is used. One of \code{c("euclidean", "manhattan", "canberra", "cosine")}. Defaults to \code{"euclidean"}.}
}
\value{
A named list with the following elements:
\itemize{
\item k - Integer. Number of neighbours to consider. Defaults to \code{20L}.
\item t - Integer. Number of iterations for the SNF algorithm. Defaults to
\code{20L}.
\item mu - Numeric. Normalisation factor for the Gaussian kernel width.
Defaults to \code{0.5}.
\item alpha - Numeric. Normalisation parameter controlling the fusion
strength. Defaults to \code{1.0}.
\item distance_metric - String. Which distance metric to use for the
continuous calculations. In case of pure categorical, Hamming will be used,
for mixed data types Gower distance is used. One of \code{c("euclidean", "manhattan", "canberra", "cosine")}. Defaults to \code{"euclidean"}.
\item normalise - Boolean. Shall continuous values be Z-scored. Defaults to
\code{TRUE}.
}
}
\description{
Wrapper function to generate SNF parameters
}
