% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_ssgsea}
\alias{params_ssgsea}
\title{Wrapper function to generate ssGSEA parameters}
\usage{
params_ssgsea(alpha = 0.25, min_size = 5L, max_size = 500L, normalise = TRUE)
}
\arguments{
\item{alpha}{Numeric. The exponent defining the weight of the tail in the
random walk performed by ssGSEA. Defaults to \code{0.25}.}

\item{min_size}{Integer. Minimum number of genes per gene set. Defaults to
\code{5L}.}

\item{max_size}{Integer. Maximum number of genes per gene set. Defaults to
\code{500L}.}

\item{normalise}{Boolean. Shall the scores be normalised. Defaults to \code{TRUE}.}
}
\value{
A named list with the following elements:
\itemize{
\item alpha - Numeric. The exponent defining the weight of the tail in the
random walk performed by ssGSEA. Defaults to \code{0.25}.
\item min_size - Integer. Minimum number of genes per gene set. Defaults to
\code{5L}.
\item max_size - Integer. Maximum number of genes per gene set. Defaults to
\code{500L}.
\item normalise - Boolean. Shall the scores be normalised. Defaults to
\code{TRUE}.
}
}
\description{
Wrapper function to generate ssGSEA parameters
}
