% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_symphony_map}
\alias{params_symphony_map}
\title{Default parameters for Symphony query mapping}
\usage{
params_symphony_map(sigma = 0.1, lambda = 1)
}
\arguments{
\item{sigma}{Numeric. Soft-clustering fuzziness for query -> reference
centroid assignment. Symphony R default is 0.1. Defaults to \code{0.1}.}

\item{lambda}{Numeric. Ridge penalty on batch coefficients. Symphony R
hardcodes 1.0. Defaults to \code{1.0}.}
}
\value{
A named list with the following elements:
\itemize{
\item sigma - Numeric. Soft-clustering fuzziness for query -> reference
centroid assignment. Symphony R default is 0.1. Defaults to \code{0.1}.
\item lambda - Numeric. Ridge penalty on batch coefficients. Symphony R
hardcodes 1.0. Defaults to \code{1.0}.
}
}
\description{
Default parameters for Symphony query mapping
}
