% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{permute_seed_nodes}
\alias{permute_seed_nodes}
\title{Generate permuation scores for the diffusion}
\usage{
permute_seed_nodes(
  object,
  perm_iters = 1000L,
  random_seed = 10101L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{NetworkDiffusions} object. The underlying class
\code{\link[=NetworkDiffusions]{NetworkDiffusions()}}.}

\item{perm_iters}{Integer. Number of permutations to test for. Defaults to
\code{1000L}.}

\item{random_seed}{Integer. Random seed for determinism.}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
The class with added diffusion score based on a single set of seed
genes. Additionally, the seed genes are stored in the class.
}
\description{
This function generate node-degree adjusted permutations of a given diffusion
score and adds Z-scores to the object. The function will automatically
determine if the original diffusion was a single or tied diffusion and
construct permutations accordingly.
}
\examples{
# 100 node-degree adjusted permutations of a single diffusion
set.seed(42)
g <- igraph::sample_pa(15, directed = FALSE)
edges <- data.table::setDT(igraph::as_data_frame(g))[, `:=`(
  from = sprintf("node_\%i", from),
  to = sprintf("node_\%i", to)
)]
object <- NetworkDiffusions(edges, weighted = FALSE, directed = FALSE)
object <- diffuse_seed_nodes(object, c(node_1 = 1, node_3 = 1), "max")
object <- permute_seed_nodes(object, perm_iters = 100L, .verbose = FALSE)
head(get_diffusion_perms(object))
}
