% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{plot.CellQc}
\alias{plot.CellQc}
\title{Plot per-cell QC violin plots from a CellQc object}
\usage{
\method{plot}{CellQc}(x, ...)
}
\arguments{
\item{x}{A \code{CellQc} object.}

\item{...}{Ignored.}
}
\value{
A named list of ggplot objects, one per metric.
}
\description{
Plot per-cell QC violin plots from a CellQc object
}
\keyword{internal}
