% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_spatial_others.R
\name{plot.SpNhoodRes}
\alias{plot.SpNhoodRes}
\title{Plot the neighbourhood enrichment Z-scores}
\usage{
\method{plot}{SpNhoodRes}(x, ...)
}
\arguments{
\item{x}{An \code{SpNhoodRes} object.}

\item{...}{Further arguments (currently unused).}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Heatmap of the K x K Z-score matrix. Red is more co-occurrence than the
permuted null gives, blue is less.
}
\keyword{internal}
