% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bulk.R
\name{plot_hvgs}
\alias{plot_hvgs}
\title{Plot the highly variable genes}
\usage{
plot_hvgs(object, bins = 50L)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}

\item{bins}{Integer. Number of bins to plot.}
}
\description{
Plots the median-absolute deviation of the genes and applied thresholds.
Expects that \code{\link[=preprocess_bulk_coexp]{preprocess_bulk_coexp()}} was run and will throw an
error otherwise.
}
\examples{
# MAD distribution with the selected genes highlighted
syn <- synthetic_bulk_cor_matrix()
mat <- log1p(t(syn$counts))
meta <- data.table::data.table(sample_id = rownames(mat))
object <- BulkCoExp(raw_data = mat, meta_data = meta)
object <- preprocess_bulk_coexp(object, hvg = 200L, .verbose = FALSE)
plot_hvgs(object)
}
