% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{plot_ica_ncomp_params}
\alias{plot_ica_ncomp_params}
\title{Plot various parameters with no comp}
\usage{
plot_ica_ncomp_params(object)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}. You need to apply
\code{\link[=ica_evaluate_comp]{ica_evaluate_comp()}} before running this function.}
}
\value{
The plot with no comp ~ vs. various parameters.
}
\description{
Helper function to plot various parameters with the no of components. You
have:
\itemize{
\item Median stability - The median stability of the clusters at this
no of components given random initialisations.
\item \% Converged - The percentage of ICA runs that converged at this
no of components.
\item IC redundancy - \code{1 - max_abs_loading_cor}, one minus the largest
absolute correlation between two component loadings at this level.
\item Combined score - The product of the three other scores.
}
If found, the function will also add the optimal number of components based
on \code{\link[=ica_optimal_ncomp]{ica_optimal_ncomp()}} (if a loess function could be fitted).
}
\examples{
# stability, convergence and orthogonality against ncomp
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- ica_processing(obj, .verbose = FALSE)
obj <- ica_evaluate_comp(
  obj,
  ica_type = "logcosh",
  ncomp_params = params_ica_ncomp(custom_seq = seq(2L, 20L, by = 2L)),
  .verbose = FALSE
)
plot_ica_ncomp_params(obj)
}
