% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{plot_ica_stability_individual}
\alias{plot_ica_stability_individual}
\title{Plot the stability of the ICA components}
\usage{
plot_ica_stability_individual(object)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}. You need to apply
\code{\link[=ica_evaluate_comp]{ica_evaluate_comp()}} before running this function.}
}
\value{
A ggplot with the per-component stability profiles over the tested
number of components.
}
\description{
Helper function to plot the individual stability profiles over the tested
ncomps.
}
\examples{
# per-component stability profiles across the tested ncomps
mat <- t(synthetic_signal_matrix()$mat)
obj <- BulkCoExp(mat, data.table::data.table(sample_id = rownames(mat)))
obj <- preprocess_bulk_coexp(obj, hvg = 0.3, .verbose = FALSE)
obj <- ica_processing(obj, .verbose = FALSE)
obj <- ica_evaluate_comp(
  obj,
  ica_type = "logcosh",
  ncomp_params = params_ica_ncomp(custom_seq = seq(2L, 20L, by = 2L)),
  .verbose = FALSE
)
plot_ica_stability_individual(obj)
}
