% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{plot_ica_stability_summarised}
\alias{plot_ica_stability_summarised}
\title{Plot the maximum and 75 \%ile of the stability.}
\usage{
plot_ica_stability_summarised(object)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. You need to apply
\code{\link[=ica_evaluate_comp]{ica_evaluate_comp()}} before running this function.}
}
\description{
Helper function to plot the individual maximum and 75th percentile in terms
of stability over the tested ncomp to identify inflection points in the
stability profiles.
}
