% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bulk.R
\name{plot_preprocessing_outliers}
\alias{plot_preprocessing_outliers}
\title{Helper plot function for identification of outliers}
\usage{
plot_preprocessing_outliers(samples, group_col, min_perc, max_perc)
}
\arguments{
\item{samples}{data.table with sample information with perc_detected_genes
and a column specifying the cohort.}

\item{group_col}{String specifying the column with cohort information}

\item{min_perc}{Numeric. Lower cutoff to identify outliers.}

\item{max_perc}{Numeric. Upper cutoff to identify outliers}
}
\value{
ggplot object, i.e., beeswarm plot with outlier indication
}
\description{
Beeswarm plot of the percentage of detected genes per sample, with the
outlier cutoffs drawn in. Used in the bulk DGE pre-processing report.
}
