% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{plot_synthetic_GEX_HT}
\alias{plot_synthetic_GEX_HT}
\title{Plot the synthetic gene expression}
\usage{
plot_synthetic_GEX_HT(synthetic_GEX)
}
\arguments{
\item{synthetic_GEX}{List. Output from Hephaestus::create_synthetic_signal_matrix().
It will throw an error if this is not an output from the function.}
}
\value{
A plotted heatmap showing the DEG.
}
\description{
Plot the synthetic gene expression
}
\examples{
\dontrun{

synthetic_GEX <- create_synthetic_signal_matrix()

plot_synthetic_GEX_HT(synthetic_GEX)
}
}
