% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bulk.R
\name{plot_voom_normalization}
\alias{plot_voom_normalization}
\title{Helper plot function for Voom normalisation}
\usage{
plot_voom_normalization(norm_counts)
}
\arguments{
\item{norm_counts}{Numeric matrix. The voom log2-CPM values, genes x
samples.}
}
\value{
ggplot object, i.e., voom normalisation plot.
}
\description{
Helper plot function for Voom normalisation
}
\examples{
# mean variance trend after voom
syn <- synthetic_bulk_cor_matrix()
grp <- rep(c("case", "control"), each = 50)
norm_counts <- rs_cpm(syn$counts, lib_size = NULL, log = TRUE,
  prior_count = 0.5)
plot_voom_normalization(norm_counts)
}
