% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{pre_process_sim_onto}
\alias{pre_process_sim_onto}
\title{Pre-process data for subsequent ontology similarity}
\usage{
pre_process_sim_onto(object, .verbose = TRUE)
}
\arguments{
\item{object}{\verb{OntologySim class}. See \code{\link[=OntologySim]{OntologySim()}}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added pre-processed data for semantic similarities to
the properties.
}
\description{
This function calculates needed information for semantic
similiary calculations
}
\examples{
# ancestors, descendants and information content added to the class
onto <- data.table::data.table(
  parent = c("a", "b", "b", "b", "c"),
  child = c("b", "c", "d", "e", "f"),
  type = c("part_of", "part_of", "part_of", "is_a", "is_a")
)
onto_obj <- OntologySim(onto, .verbose = FALSE)
onto_obj <- pre_process_sim_onto(onto_obj, .verbose = FALSE)
names(S7::prop(onto_obj, "outputs"))
}
