% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{prep_limma_contrasts}
\alias{prep_limma_contrasts}
\title{Create all limma contrasts from a provided string}
\usage{
prep_limma_contrasts(limma_fit, contrast_list)
}
\arguments{
\item{limma_fit}{The fitted limma model, i.e., output of \code{\link[limma:lmFit]{limma::lmFit()}}.}

\item{contrast_list}{String vector. The strings need to have the form of
\code{"contrast1-contrast2"}.}
}
\value{
The Limma contrasts for further usage.
}
\description{
Create all limma contrasts from a provided string
}
