% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{prep_stats_pathways}
\alias{prep_stats_pathways}
\title{Helper function to prepare data for GSEA}
\usage{
prep_stats_pathways(stats, pathways, min_size = 5L, max_size = 500L)
}
\arguments{
\item{stats}{Named numeric vector. The gene level statistic.}

\item{pathways}{List. A named list with each element containing the genes for
this pathway.}

\item{min_size}{Integer. Minimum size of the gene sets. Defaults to \code{5L}.}

\item{max_size}{Integer. Maximum size of the gene sets. Defaults to \code{500L}.}
}
\value{
A list with the following elements:
\itemize{
\item stats - The sorted gene level statistics
\item pathways - The prepared pathways for further usage in the functions
\item pathway_sizes - The sizes of the pathways
}
}
\description{
Helper function to prepare data for GSEA
}
