% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_batch_corr.R
\name{print.BatchLisiScores}
\alias{print.BatchLisiScores}
\title{Print method for BatchLisiScores}
\usage{
\method{print}{BatchLisiScores}(x, ...)
}
\arguments{
\item{x}{A \code{BatchLisiScores} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for BatchLisiScores
}
