% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_batch_corr.R
\name{print.BatchSilhouetteScores}
\alias{print.BatchSilhouetteScores}
\title{Print method for BatchSilhouetteScores}
\usage{
\method{print}{BatchSilhouetteScores}(x, ...)
}
\arguments{
\item{x}{A \code{BatchSilhouetteScores} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for BatchSilhouetteScores
}
