% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{print.BoostRes}
\alias{print.BoostRes}
\title{Print a BoostRes object}
\usage{
\method{print}{BoostRes}(x, ...)
}
\arguments{
\item{x}{A \code{BoostRes} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print a BoostRes object
}
\examples{
# doublet calls from the boosted classifier
sc <- demo_single_cells(prepped = FALSE)
res <- doublet_detection_boost_sc(
  sc,
  boost_params = params_boost(
    hvg = list(min_gene_var_pctl = 0.0),
    pca = list(no_pcs = 10L),
    n_iters = 5L
  ),
  .verbose = FALSE
)
print(res)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
