% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_coexp.R
\name{print.bulk_coexp}
\alias{print.bulk_coexp}
\title{print Method for bulk_coexp object}
\arguments{
\item{x}{An object of class \code{bulk_coexp}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a bulk_coexp object.
}
