% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{print.gene_ontology_data}
\alias{print.gene_ontology_data}
\title{print Method for gene_ontology_data object}
\arguments{
\item{x}{An object of class \code{gene_ontology_data}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a gene_ontology_data object.
}
