% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{print.GeneOntologyElim}
\alias{print.GeneOntologyElim}
\title{print Method for GeneOntologyElim object}
\arguments{
\item{x}{An object of class \code{GeneOntologyElim}, see
\code{\link[=GeneOntologyElim]{GeneOntologyElim()}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a GeneOntologyElim object.
}
