% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_batch_corr.R
\name{print.KbetScores}
\alias{print.KbetScores}
\title{Print method for KbetScores}
\usage{
\method{print}{KbetScores}(x, ...)
}
\arguments{
\item{x}{A \code{KbetScores} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for KbetScores
}
\keyword{internal}
