% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{print.Ontology}
\alias{print.Ontology}
\title{print Method for Ontology object}
\arguments{
\item{x}{An object of class \code{Ontology}, see \code{\link[=Ontology]{Ontology()}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Ontology object.
}
