% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{print.OntologySim}
\alias{print.OntologySim}
\title{print Method for OntologySim object}
\arguments{
\item{x}{An object of class \code{OntologySim}, see \code{\link[=OntologySim]{OntologySim()}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a OntologySim object.
}
