% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{print.single_cell_exp}
\alias{print.single_cell_exp}
\title{print Method for single_cell_exp object}
\arguments{
\item{x}{An object of class \code{single_cell_exp}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a single_cell_exp object.
}
