% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{print.SingleCells}
\alias{print.SingleCells}
\title{print Method for SingleCells object}
\arguments{
\item{x}{An object of class \code{SingleCells}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a SingleCells object.
}
\keyword{internal}
