% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_groups.R
\name{print.SingleCellsGrps}
\alias{print.SingleCellsGrps}
\title{print Method for SingleCellsGrps object}
\usage{
## S7 method for class <bixverse::SingleCellsGrps>
print(x, ...)
}
\arguments{
\item{x}{An object of class \code{SingleCellsGrps}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a SingleCellsGrps object.
}
\keyword{internal}
