% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{print.SingleCellsMultiModal}
\alias{print.SingleCellsMultiModal}
\title{print Method for SingleCellsMultiModal object}
\usage{
## S7 method for class <bixverse::SingleCellsMultiModal>
print(x, ...)
}
\arguments{
\item{x}{An object of class \code{SingleCellsMultiModal}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a SingleCellsMultiModal object.
}
\keyword{internal}
