% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_subset.R
\name{print.SingleCellsSubset}
\alias{print.SingleCellsSubset}
\title{print Method for SingleCellsSubset object}
\usage{
## S7 method for class <bixverse::SingleCellsSubset>
print(x, ...)
}
\arguments{
\item{x}{An object of class \code{SingleCellsSubset}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
print Method for SingleCellsSubset object
}
\keyword{internal}
