% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_spatial_spot.R
\name{print.SpCache}
\alias{print.SpCache}
\title{Print method for SpCache objects}
\usage{
\method{print}{SpCache}(x, ...)
}
\arguments{
\item{x}{A \code{SpCache} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for SpCache objects
}
