% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_symphony.R
\name{print.SymphonyReference}
\alias{print.SymphonyReference}
\title{print Method for SymphonyReference object}
\usage{
## S7 method for class <bixverse::SymphonyReference>
print(x, ...)
}
\arguments{
\item{x}{An object of class \code{SymphonyReference}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a SymphonyReference object.
}
\keyword{internal}
