% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cis_target.R
\name{process_cistarget_res}
\alias{process_cistarget_res}
\title{Helper to process CisTarget results}
\usage{
process_cistarget_res(cs_ls, gs_name, represented_motifs, represented_genes)
}
\arguments{
\item{cs_ls}{List. The result list from the Rust wrapper.}

\item{gs_name}{String. Name of the tested gene set.}

\item{represented_motifs}{Character vector. The represented motifs in the
rankings.}

\item{represented_genes}{Character vector. The represented genes in the
rankings.}
}
\value{
A data.table with the results if there were any significant motifs.
}
\description{
Helper to process CisTarget results
}
