% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_go_data}
\alias{process_go_data}
\title{Process Gene Ontology data into the right format}
\usage{
process_go_data(go_info, go_genes, go_relationships)
}
\arguments{
\item{go_info}{data.table. Contains \code{go_id}, \code{go_name} and \code{namespace.}}

\item{go_genes}{data.table. Contains \code{go_id} and corresponding \code{ensembl_id}.}

\item{go_relationships}{data.table. Contains \code{parent}, \code{child} and
\code{relationship}}
}
\value{
data.table ready for usage in \code{\link[=GeneOntologyElim]{GeneOntologyElim()}}.
}
\description{
Helper function that takes in different files containing gene
ontology data and puts them together for various gene set enrichment methods
using the ontological information
}
