% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rbh_graph}
\alias{rbh_graph}
\title{Reciprocal best hit graph (deprecated)}
\usage{
rbh_graph(
  module_results,
  rbh_type = c("cor", "set"),
  dataset_col = NULL,
  module_col = NULL,
  value_col = NULL
)
}
\arguments{
\item{module_results}{data.table or list containing the gene modules.}

\item{rbh_type}{String. One of \code{c("cor", "set")}.}

\item{dataset_col}{The column indicating the data set/method origin.}

\item{module_col}{The column storing the module names.}

\item{value_col}{The column storing the genes within each module.}
}
\value{
Returns a \code{\link[=RbhGraph]{RbhGraph()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=RbhGraph]{RbhGraph()}}.
}
\keyword{internal}
