% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cis_target.R
\name{read_motif_annotation_file}
\alias{read_motif_annotation_file}
\title{Read in the motif annotation file}
\usage{
read_motif_annotation_file(annot_file)
}
\arguments{
\item{annot_file}{String. Path to the motif2tf file that you downloaded.}
}
\value{
data.table with the motif to transcription factor information.
}
\description{
This function loads in the motif2tf information that you can get from
\verb{https://resources.aertslab.org/cistarget/motif2tf/}.
The function will generate a data.table that can be subsequently used.
}
