% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cis_target.R
\name{read_motif_ranking}
\alias{read_motif_ranking}
\title{Read in the motif rankings and transform them into a matrix}
\usage{
read_motif_ranking(ranking_file)
}
\arguments{
\item{ranking_file}{String. The file path to the .feather file}
}
\value{
An integer matrix that has been transposed for easier use in the
underlying Rust code.
}
\description{
This function loads in the .feather files with the motif to target gene
rankings. These can be found here:
\verb{https://resources.aertslab.org/cistarget/databases/}
}
\examples{
\dontrun{
# transposed motif rankings from the downloaded feather file
paths <- download_cistarget_hg38()
rankings <- read_motif_ranking(paths$rankings)
dim(rankings)
}
}
