% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{remove_knn}
\alias{remove_knn}
\alias{remove_knn.MetaCells}
\alias{remove_knn.ScCache}
\alias{remove_knn.SingleCells}
\alias{remove_knn.SingleCellsSubset}
\title{Remove the KNN data}
\usage{
remove_knn(x, ...)

## S7 method for class <bixverse::MetaCells>
remove_knn(x, ...)

\method{remove_knn}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
remove_knn(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
remove_knn(x, ...)
}
\arguments{
\item{x}{An object from which to remove the kNN data.}

\item{...}{Other parameters.}
}
\description{
Remove the KNN data
}
\examples{
# drop the cached kNN, for example before rebuilding it
sc <- demo_single_cells()
sc <- remove_knn(sc)
is.null(get_knn_obj(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
