% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R
\name{remove_residual_fit}
\alias{remove_residual_fit}
\alias{remove_residual_fit.MetaCells}
\alias{remove_residual_fit.ScCache}
\alias{remove_residual_fit.SingleCells}
\title{Remove the fitted residual model}
\usage{
remove_residual_fit(x, ...)

## S7 method for class <bixverse::MetaCells>
remove_residual_fit(x, ...)

\method{remove_residual_fit}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
remove_residual_fit(x, ...)
}
\arguments{
\item{x}{An object from which to remove the fitted model.}

\item{...}{Other parameters.}
}
\value{
The object with the fitted model dropped.
}
\description{
Remove the fitted residual model
}
\examples{
# drop the fitted model again
sc <- demo_single_cells()
sc <- fit_residuals_sc(sc, method = "analytic_pearson", .verbose = FALSE)
sc <- remove_residual_fit(sc)
is.null(get_residual_fit(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
