% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples from object}
\usage{
remove_samples(object, samples_to_remove, ...)
}
\arguments{
\item{object}{The underlying object, either \code{\link[=BulkCoExp]{BulkCoExp()}} or
\code{\link[=BulkDge]{BulkDge()}}.}

\item{samples_to_remove}{Character vector. The sample identifiers to remove.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
Returns the object with the samples removed. This will regenerated
the object from the start and remove any data in it.
}
\description{
This function allows to remove certain samples from the object
}
\examples{
# drop the first two samples and rebuild the object
syn <- synthetic_bulk_cor_matrix()
meta <- data.table::data.table(sample_id = colnames(syn$counts))
object <- BulkDge(raw_counts = syn$counts, meta_data = meta)
object <- remove_samples(object, c("sample_1", "sample_2"))
dim(S7::prop(object, "raw_counts"))
}
