% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{remove_snn_graph}
\alias{remove_snn_graph}
\alias{remove_snn_graph.MetaCells}
\alias{remove_snn_graph.ScCache}
\alias{remove_snn_graph.SingleCells}
\alias{remove_snn_graph.SingleCellsSubset}
\title{Remove the sNN graph}
\usage{
remove_snn_graph(x, ...)

## S7 method for class <bixverse::MetaCells>
remove_snn_graph(x, ...)

\method{remove_snn_graph}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
remove_snn_graph(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
remove_snn_graph(x, ...)
}
\arguments{
\item{x}{An object from which to remove the sNN graph}

\item{...}{Other parameters.}
}
\description{
Remove the sNN graph
}
\examples{
# drop the cached sNN graph
sc <- demo_single_cells()
sc <- remove_snn_graph(sc)
is.null(get_snn_graph(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
