% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell_lr.R
\name{resolve_weights}
\alias{resolve_weights}
\title{Resolve prioritisation weights from the scenario or validate a user vector}
\usage{
resolve_weights(weights, scenario, has_condition_de)
}
\arguments{
\item{weights}{description}
}
\description{
Resolve prioritisation weights from the scenario or validate a user vector
}
\keyword{internal}
