% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{robust_scale}
\alias{robust_scale}
\title{Robust scaler.}
\usage{
robust_scale(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
x, robustly scaled.
}
\description{
Robust scaling, i.e., removes the median and scales data based on the
interquartile range (IQR). Useful if outliers are expected. NAs will be
ignored.
}
\examples{
# median-centred, IQR-scaled vector
set.seed(123)
head(robust_scale(rnorm(10)))
}
