% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_batch_lisi}
\alias{rs_batch_lisi}
\title{Calculate batch LISI scores}
\usage{
rs_batch_lisi(knn_mat, batch_vector, verbose)
}
\arguments{
\item{knn_mat}{Integer matrix. The rows represent the cells and the
columns the neighbour indices.}

\item{batch_vector}{Integer vector. The integers indicate to which
batch a given cell belongs.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following items
\itemize{
\item per_cell - Per-cell LISI scores
\item mean_lisi - Mean LISI
\item median_lisi - Median LISI
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Computes the Local Inverse Simpson's Index on batch labels using the
kNN graph. Measures the effective number of batches in each cell's
neighbourhood. Under perfect mixing LISI equals the number of batches,
under no mixing LISI equals 1.
}
