% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_bbknn}
\alias{rs_bbknn}
\title{BBKNN implementation in Rust}
\usage{
rs_bbknn(embd, batch_labels, bbknn_params, seed, verbose)
}
\arguments{
\item{embd}{Numerical matrix. The embedding matrix to use to generate the
BBKNN parameters. Usually PCA. Rows represent cells.}

\item{batch_labels}{Integer vector. These represent to which batch a given
cell belongs (0-indexed!).}

\item{bbknn_params}{List. Contains all of the BBKNN parameters.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with
\itemize{
\item distances - Sparse list representation of the kNN distances.
\item connectivities - Sparse list representation of the
connectivities.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function implements the BBKNN algorithm from Polański, et al.
}
\references{
Polański, et al., Bioinformatics, 2020
}
\keyword{internal}
