% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_calc_es}
\alias{rs_calc_es}
\title{Calculates the traditional GSEA enrichment score}
\usage{
rs_calc_es(stats, pathway_r)
}
\arguments{
\item{stats}{Named numerical vector. Needs to be sorted. The gene level
statistics.}

\item{pathway_r}{String vector. The genes in the pathway.}
}
\value{
The enrichment score
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
