% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_calc_gsea_stat_cumulative_batch}
\alias{rs_calc_gsea_stat_cumulative_batch}
\title{Helper function to generate fgsea simple-based permutations}
\usage{
rs_calc_gsea_stat_cumulative_batch(
  stats,
  pathway_scores,
  pathway_sizes,
  iters,
  gsea_param,
  return_add_stats,
  seed
)
}
\arguments{
\item{stats}{Numeric vector. The gene level statistic. Needs to be
sorted in descending nature.}

\item{pathway_scores}{Numeric vector. The enrichment scores for the
pathways}

\item{pathway_sizes}{Integer vector. The sizes of the pathways.}

\item{iters}{Integer. Number of permutations.}

\item{gsea_param}{Float. The Gene Set Enrichment parameter.}

\item{return_add_stats}{Boolean. Returns additional statistics
necessary for the multi-level calculations.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
List with the following elements
\itemize{
\item es Enrichment scores for the gene sets
\item nes Normalised enrichment scores for the gene sets
\item pvals The calculated p-values.
\item n_more_extreme Number of times the enrichment score was
bigger or smaller than the permutation (pending sign).
\item size Pathway size.
}

If \code{return_add_stats} is set to \code{TRUE}, there are additional elements in
the list:
\itemize{
\item le_zero Number of times the permutation was less than zero.
\item ge_zero Number of times the permutation was greater than zero.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
