% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_calc_gsea_stat_traditional_batch}
\alias{rs_calc_gsea_stat_traditional_batch}
\title{Helper function to generate traditional GSEA-based permutations}
\usage{
rs_calc_gsea_stat_traditional_batch(
  stats,
  pathway_scores,
  pathway_sizes,
  iters,
  seed
)
}
\arguments{
\item{stats}{Numeric vector. The gene level statistic. Needs to be
sorted in descending nature.}

\item{pathway_scores}{Numeric vector. The enrichment scores for the
pathways}

\item{pathway_sizes}{Integer vector. The sizes of the pathways.}

\item{iters}{Integer. Number of permutations.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
List with the following elements
\itemize{
\item es Enrichment scores for the gene sets
\item nes Normalised enrichment scores for the gene sets
\item pvals The calculated p-values.
\item n_more_extreme Number of times the enrichment score was
bigger or smaller than the permutation (pending sign).
\item size Pathway size.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
