% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_calc_gsea_stats}
\alias{rs_calc_gsea_stats}
\title{Rust implementation of the fgsea::calcGseaStat() function}
\usage{
rs_calc_gsea_stats(
  stats,
  gs_idx,
  gsea_param,
  return_leading_edge,
  return_all_extremes
)
}
\arguments{
\item{stats}{Numeric vector. The gene level statistic. Needs to be
sorted in descending nature.}

\item{gs_idx}{Integer vector. The 1-based indices of the gene set genes.}

\item{gsea_param}{Float. The GSEA parameter. Usually defaults to 1.0.}

\item{return_leading_edge}{Boolean. Return the leading edge indices.}

\item{return_all_extremes}{Boolean. Shall the extreme values be returned
for plotting.}
}
\value{
List with the following elements
\itemize{
\item es Enrichment score for that gene set
\item leading_edge 1-based indices of the leading edge genes.
\item top Top values of the curve.
\item bottom Bottom values of the curve.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
