% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_calculate_dge_mann_whitney}
\alias{rs_calculate_dge_mann_whitney}
\title{Calculate DGEs between cells based on Mann Whitney stats}
\usage{
rs_calculate_dge_mann_whitney(
  f_path,
  cell_indices_1,
  cell_indices_2,
  min_prop,
  alternative,
  verbose
)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{cell_indices_1}{Integer. Index positions (0-indexed) of the cells
of group 1.}

\item{cell_indices_2}{Integer. Index positions (0-indexed) of the cells
of group 2.}

\item{min_prop}{Numeric. Minimum proportion of expression in at least one
of the two groups to be tested.}

\item{alternative}{String. One of \code{c("twosided", "greater", "less")}. Null
hypothesis.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following elements
\itemize{
\item lfc - Log fold changes between the two groups.
\item prop1 - Proportion of cells expressing the gene in group 1.
\item prop2 - Proportion of cells expressing the gene in group 2.
\item z_scores - Z-scores based on the Mann Whitney statistic.
\item p_values - P-values of the Mann Whitney statistic.
\item fdr - False discovery rate after BH adjustment
\item genes_to_keep - Boolean indicating which genes were tested.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The function will take two sets of cell indices and calculate the
differential gene expression based on the Mann Whitney test between the
two groups.
}
\keyword{internal}
