% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cluster_stability}
\alias{rs_cluster_stability}
\title{Helper function to assess cluster stability}
\usage{
rs_cluster_stability(data)
}
\arguments{
\item{data}{Integer matrix. Assumes that each column represents a given
resampling/bootstrap and the rows represent the features, while each integer
indicates cluster membership.}
}
\value{
A list containing, one entry per feature:
\itemize{
\item mean_jaccard - mean Jaccard similarity of the feature's cluster
across all pairs of bootstraps/resamplings.
\item std_jaccard - the (population) standard deviation of these Jaccard
similarities.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
