% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_compute_cluster_expr_stats}
\alias{rs_compute_cluster_expr_stats}
\title{Compute cluster statistics for NicheNet prioritisation}
\usage{
rs_compute_cluster_expr_stats(f_path_gene, gene_indices, clusters)
}
\arguments{
\item{f_path_gene}{Path to the \code{counts_genes.bin} file.}

\item{gene_indices}{Integer vector. 0-indexed(!) positions of the genes
to include.}

\item{clusters}{List. A list that contains within the cell indices of the
clusters of interest (0-indexed).}
}
\value{
A list with two matrices
\itemize{
\item mean - The average expression. Shape genes x clusters.
\item frac - The fraction of cells expressing. Shape genes x clusters.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Helper function to pull out average expression and fraction of cells for
genes of interest.
}
\keyword{internal}
