% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_constrained_page_rank_list}
\alias{rs_constrained_page_rank_list}
\title{Calculate a constrained page-rank score over a list.}
\usage{
rs_constrained_page_rank_list(
  personalisation_list,
  node_names,
  node_types,
  from,
  to,
  weights,
  edge_type,
  sink_nodes,
  sink_edges
)
}
\arguments{
\item{personalisation_list}{List. The list with the personalisation vectors.
The sum must equal to 1, otherwise the function panics!}

\item{node_names}{String vector. Name of the graph nodes.}

\item{node_types}{String vector. The node types.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{weights}{Numerical vector. The edge weights from the edge list.}

\item{edge_type}{String vector. The edge types.}

\item{sink_nodes}{Optional string vector. Should these node types be seen as
sinks, i.e., the reset occurs when this node is reached.}

\item{sink_edges}{Optional string vector. Shall an automatic reset occur
when this edge type is traversed.}
}
\value{
A list of the personalised (constrained) page rank values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function can be used to get constrained personalised page-rank scores
akin to Ruiz, et al. You can provide optionally \code{sink_nodes} (node types
that will force a reset) and/or \code{sink_edges} (edge types that will force a
reset). This version can take in a list of personalisation vectors and
returns a list as result.
}
\references{
Ruiz, et al., Nat Commun, 2021
}
