% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor}
\alias{rs_cor}
\title{Calculate the column wise correlations.}
\usage{
rs_cor(x, spearman)
}
\arguments{
\item{x}{R matrix with doubles.}

\item{spearman}{Shall the Spearman correlation be calculated instead of
Pearson.}
}
\value{
The correlation matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the pairwise correlations of the columns.
}
