% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor_upper_triangle}
\alias{rs_cor_upper_triangle}
\title{Calculate the column wise correlations and returns the upper triangle}
\usage{
rs_cor_upper_triangle(x, spearman, shift)
}
\arguments{
\item{x}{R matrix with doubles.}

\item{spearman}{Shall the Spearman correlation be calculated instead of
Pearson.}

\item{shift}{Boolean. If \code{TRUE}, the diagonal is excluded, otherwise it is
included.}
}
\value{
The upper triangle of the correlation matrix iterating through the
rows, with or without the diagonal depending on \code{shift}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the correlation matrix of the columns, but returns the upper
triangle only as a flat vector.
}
