% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cos}
\alias{rs_cos}
\title{Calculate the column wise cosine similarities}
\usage{
rs_cos(x)
}
\arguments{
\item{x}{R matrix with doubles.}
}
\value{
The cosine similarity matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the pairwise cosine similarity matrix of the columns.
}
