% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_count_zeroes}
\alias{rs_count_zeroes}
\title{Helper to get zero stats from a given matrix}
\usage{
rs_count_zeroes(x)
}
\arguments{
\item{x}{Numeric matrix. The matrix for which to calculate the total zeroes,
column and row zeroes.}
}
\value{
A list with:
\itemize{
\item total_zeroes - Total zeroes in the matrix.
\item row_zeroes - Vector with number of zeroes per row.
\item col_zeroes - Vector with number of zeroes per column.
}
}
\description{
Calculates in a single matrix pass the total number of zeroes, the row
zeroes and column zeroes.
}
