% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cov2cor}
\alias{rs_cov2cor}
\title{Calculates the correlation matrix from the co-variance matrix}
\usage{
rs_cov2cor(x)
}
\arguments{
\item{x}{R matrix with doubles that is the co-variance matrix}
}
\value{
The correlation matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the correlation matrix from a co-variance matrix.
}
