% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_covariance}
\alias{rs_covariance}
\title{Calculate the column-wise co-variance.}
\usage{
rs_covariance(x)
}
\arguments{
\item{x}{R matrix with doubles.}
}
\value{
The co-variance matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the pairwise co-variance of the columns.
}
