% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_critval}
\alias{rs_critval}
\title{Calculate the critical value}
\usage{
rs_critval(values, iters, alpha, seed)
}
\arguments{
\item{values}{Numeric vector. The full data set for which to calculate the
critical value.}

\item{iters}{Integer. Size of the bootstrap sample.}

\item{alpha}{Float. The alpha value. For example, 0.001 would return the
value exceeded by roughly 0.1 percent of the bootstrap sample.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
The critical value for the given parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the critical value for a given set based on a
bootstrap sample (with replacement) and a given alpha value. The sample is
sorted in decreasing order and the value at the upper \code{alpha} tail is
returned.
}
\keyword{internal}
