% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_critval_mat}
\alias{rs_critval_mat}
\title{Calculate the critical value}
\usage{
rs_critval_mat(mat, iters, alpha, seed)
}
\arguments{
\item{mat}{Numeric matrix. The symmetric matrix with all of the values.}

\item{iters}{Integer. Size of the bootstrap sample.}

\item{alpha}{Float. The alpha value. For example, 0.001 would return the
value exceeded by roughly 0.1 percent of the bootstrap sample.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
The critical value for the given parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Same as \code{\link[=rs_critval]{rs_critval()}}, but the values are taken from the upper
triangle (diagonal excluded) of a symmetric matrix.
}
\keyword{internal}
