% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_dense_to_upper_triangle}
\alias{rs_dense_to_upper_triangle}
\title{Generate a vector-based representation of the upper triangle of a matrix}
\usage{
rs_dense_to_upper_triangle(x, shift)
}
\arguments{
\item{x}{Numeric matrix. The symmetric matrix to flatten.}

\item{shift}{Boolean. If \code{TRUE}, the diagonal is excluded, otherwise it is
included.}
}
\value{
Numeric vector with the upper triangle values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function generates a vector from the upper triangle of a given
symmetric matrix, iterating through the rows. You have the option to
remove the diagonal with setting \code{shift = TRUE}.
}
