% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_differential_cor}
\alias{rs_differential_cor}
\title{Calculate the column wise differential correlation between two sets of data.}
\usage{
rs_differential_cor(x_a, x_b, spearman)
}
\arguments{
\item{x_a}{Numeric matrix a, samples x features.}

\item{x_b}{Numeric matrix b, samples x features. Needs the same number of
columns as \code{x_a}.}

\item{spearman}{Boolean. Shall the Spearman correlation be calculated
instead of Pearson.}
}
\value{
A list containing, one entry per upper-triangle feature pair:
\itemize{
\item r_a - The correlation coefficients in the upper triangle of
matrix a.
\item r_b - The correlation coefficients in the upper triangle of
matrix b.
\item z_score - The z-scores of the difference in correlation
coefficients.
\item p_val - The z-scores transformed to two-sided p-values.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the differential correlation based on the Fisher
method. For speed purposes, the function will only calculate the
differential correlation on the upper triangle of the two correlation
matrices.
}
