% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_extract_counts_plots}
\alias{rs_extract_counts_plots}
\title{Helper to extract single cell counts as a dense vector for plotting}
\usage{
rs_extract_counts_plots(f_path, cell_indices, gene_index, norm, scale, clip)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep.}

\item{gene_index}{Integer. Gene index position to return (0-indexed!).}

\item{norm}{Boolean. Shall normalised counts be returned.}

\item{scale}{Boolean. Shall the normalised counts be z-scored across the
selected cells.}

\item{clip}{Optional float. Clips the Z-scores to \verb{[-clip, clip]}. Only
used if \code{scale = TRUE}.}
}
\value{
Numerical vector with one expression value per cell in
\code{cell_indices}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Extract dense counts for this given gene.
}
\keyword{internal}
