% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_extract_several_genes_plots}
\alias{rs_extract_several_genes_plots}
\title{Helper to extract single cell counts for several genes}
\usage{
rs_extract_several_genes_plots(f_path, cell_indices, gene_indices, scale, clip)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep.}

\item{gene_indices}{Integer vector. Gene index positions to return
(0-indexed!).}

\item{scale}{Boolean. Shall the normalised counts be z-scored per gene
across the selected cells.}

\item{clip}{Optional float. Clips the Z-scores to \verb{[-clip, clip]}. Only
used if \code{scale = TRUE}.}
}
\value{
A list of numerical vectors, one per gene in \code{gene_indices}, each
with one normalised value per cell in \code{cell_indices}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Extract the normalised single cell counts of several genes at once.
}
\keyword{internal}
