% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_fast_auc}
\alias{rs_fast_auc}
\title{Fast AUC calculation}
\usage{
rs_fast_auc(pos_scores, neg_scores, iters, seed)
}
\arguments{
\item{pos_scores}{The scores of your hits.}

\item{neg_scores}{The scores of your non-hits.}

\item{iters}{Number of iterations to run the function for.
Recommended size: 10000L.}

\item{seed}{Seed.}
}
\value{
The approximate AUC.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates rapidly AUCs based on an approximation: the
fraction of \code{iters} random (positive, negative) pairs, drawn with
replacement, where the positive score is larger.
}
