% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_fdr_adjustment}
\alias{rs_fdr_adjustment}
\title{Calculate a BH-based FDR}
\usage{
rs_fdr_adjustment(pvals)
}
\arguments{
\item{pvals}{Numeric vector. The p-values you wish to adjust.}
}
\value{
The Benjamini-Hochberg adjusted p-values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Rust implementation that will be faster if you have a terrifying amount of
p-values to adjust.
}
