% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gaussian_affinity_kernel}
\alias{rs_gaussian_affinity_kernel}
\title{Apply a Gaussian affinity kernel to a distance metric}
\usage{
rs_gaussian_affinity_kernel(x, bandwidth)
}
\arguments{
\item{x}{The distance metric. Should be positive values only!}

\item{bandwidth}{The bandwidth of the kernel. Smaller values will yield
smaller affinities.}
}
\description{
Applies a Gaussian kernel to a vector of distances.
}
