% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_geom_elim_fgsea}
\alias{rs_geom_elim_fgsea}
\title{Run fgsea simple method for gene ontology with elimination method}
\usage{
rs_geom_elim_fgsea(
  stats,
  levels,
  go_obj,
  gsea_param,
  elim_threshold,
  min_size,
  max_size,
  iters,
  seed,
  debug
)
}
\description{
Run fgsea simple method for gene ontology with elimination method
}
