% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_get_metacells}
\alias{rs_get_metacells}
\title{Generate meta cells (hdWGCNA method)}
\usage{
rs_get_metacells(
  f_path,
  knn_mat,
  embd,
  meta_cell_params,
  target_size,
  seed,
  verbose
)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{knn_mat}{Optional integer matrix. The kNN matrix you wish to use
for the generation of the meta cells. This function expects 0-indices!}

\item{embd}{Optional numerical matrix. The embedding matrix (for example
PCA embedding) you wish to use for the generation of the kNN graph that
is used subsequently for aggregation of the meta cells.}

\item{meta_cell_params}{A list containing the meta cell parameters.}

\item{target_size}{Numeric. Target library size for re-normalisation of
the meta cells. Typically \code{1e4}.}

\item{seed}{Integer. For reproducibility purposes.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item assignments - A list containing assignment information with elements:
assignments (vector), metacells (list), unassigned (vector), n_metacells,
n_cells, n_unassigned
\item aggregated - A list with indptr, indices, raw_counts, norm_counts,
nrow, ncol in sparse format.
}
}
\description{
This function implements the approach from Morabito, et al.
to generate meta cells. You can provide an already pre-computed kNN matrix
or an embedding to regenerate the kNN matrix with specified parameters in
the meta_cell_params. If \code{knn_mat} is provided, this one will be used. You
need to at least provide \code{knn_mat} or \code{embd}!
}
