% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gse_geom_elim}
\alias{rs_gse_geom_elim}
\title{Run hypergeometric enrichment over the gene ontology}
\usage{
rs_gse_geom_elim(
  target_genes,
  levels,
  go_obj,
  gene_universe_length,
  min_genes,
  elim_threshold,
  min_overlap,
  fdr_threshold
)
}
\arguments{
\item{target_genes}{A character vector representing the target gene set.}

\item{levels}{A character vector representing the levels to iterate through.
The order will be the one the iterations are happening in.}

\item{go_obj}{The \code{GeneOntologyElim} S7 class. See
\code{\link[=GeneOntologyElim]{GeneOntologyElim()}}.}

\item{gene_universe_length}{The length of the gene universe.}

\item{min_genes}{number of minimum genes for the gene ontology term to be
tested.}

\item{elim_threshold}{p-value below which the elimination procedure shall be
applied to the ancestors.}

\item{min_overlap}{Optional minimum overlap threshold.}

\item{fdr_threshold}{Optional fdr threshold.}
}
\value{
A list containing (for the terms passing the thresholds):
\itemize{
\item go_ids - The gene ontology identifier.
\item pvals - The calculated p-values.
\item fdr - The calculated FDRs.
\item odds_ratios - The calculated odds ratios.
\item hits - The size of the overlap.
\item gene_set_lengths - The length of the gene sets.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function implements a Rust version of the gene ontology enrichment with
elimination: the starting point are the leaves of the ontology and
hypergeometric tests will first be conducted there. Should the
hypergeometric test p-value be below a certain threshold, the genes of that
gene ontology term will be removed from all ancestors.
}
