% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gsea_traditional}
\alias{rs_gsea_traditional}
\title{Traditional Gene Set Enrichment analysis (in Rust)}
\usage{
rs_gsea_traditional(stats, iters, pathway_list, seed)
}
\arguments{
\item{stats}{Named numerical vector. Needs to be sorted. The gene level statistics.}

\item{iters}{Integer. Number of permutations to test for.}

\item{pathway_list}{List. A named list with each element containing the genes for this
pathway.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
List with the following elements
\itemize{
\item es Enrichment scores for the gene sets
\item nes Normalised enrichment scores for the gene sets
\item pvals The calculated p-values.
}
}
\description{
This function serves as an internal control. It implements the
gene set enrichment analysis in the traditional way without the convex approximations
used in fgsea implementations.
}
