% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_h5ad_data}
\alias{rs_h5ad_data}
\title{Load in h5ad data via Rust}
\usage{
rs_h5ad_data(f_path, cs_type, nrows, ncols, cell_quality, slot, verbose)
}
\arguments{
\item{f_path}{File path. The path to the h5ad file.}

\item{cs_type}{String. One of \code{c("csr", "csc")}. How the data is stored in
the file. Other values raise an error.}

\item{nrows}{Integer. Number of rows in the file.}

\item{ncols}{Integer. Number of columns in the file.}

\item{cell_quality}{List. Specifying the cell quality. Please refer
to \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}.}

\item{slot}{String. In which slot the raw data can be found. One of
\code{c("X", "raw", "layers.counts")}. Unknown strings default to \code{"X"}.}

\item{verbose}{Boolean. Controls verbosity of the function}
}
\value{
A list with the CSR data (cells x genes) of the cells and genes
passing \code{cell_quality}:
\itemize{
\item data - The counts of the sparse matrix.
\item indices - The 0-based gene indices of the sparse matrix.
\item indptr - The index pointers of the sparse matrix.
\item no_genes - No of genes in the sparse matrix (i.e., ncol).
\item no_cells - No of cells in the sparse matrix (i.e., nrow).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Loads in h5ad data within Rust and automatically converts the data into
CSR with cells x genes.
}
