% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hamming_dist}
\alias{rs_hamming_dist}
\title{Calculates the Hamming distance between categorical columns}
\usage{
rs_hamming_dist(x)
}
\arguments{
\item{x}{Integer matrix. The integers represent the factor data.}
}
\value{
The Hamming distance matrix between the columns, values in
\verb{[0, 1]}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the pairwise Hamming distance between the columns, i.e. the
fraction of rows in which two columns differ.
}
