% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hedges_g}
\alias{rs_hedges_g}
\title{Calculate the Hedge's G effect}
\usage{
rs_hedges_g(mat_a, mat_b, small_sample_correction)
}
\arguments{
\item{mat_a}{The matrix of samples and features in grp A for which to
calculate the Hedge's G effect.}

\item{mat_b}{The matrix of samples and features in grp B for which to
calculate the Hedge's G effect.}

\item{small_sample_correction}{Shall the small sample correction be applied.}
}
\value{
A list with the following items:
\itemize{
\item effect_sizes - Hedge's G effect size per feature.
\item standard_errors - Standard error of the effect size per feature.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the Hedge's G effect for two sets of matrices. The function
assumes that rows = samples and columns = features.
}
