% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hypergeom_test}
\alias{rs_hypergeom_test}
\title{Run a single hypergeometric test.}
\usage{
rs_hypergeom_test(
  target_genes,
  gene_sets,
  gene_universe,
  min_overlap,
  fdr_threshold
)
}
\arguments{
\item{target_genes}{String vector. Represents the target gene set.}

\item{gene_sets}{List. Contains the strings that represent the gene sets to
test against.}

\item{gene_universe}{String vector. The features representing the gene
universe from which the target genes and gene sets are sampled from.}

\item{min_overlap}{Optional integer. Shall a filter be applied on the
minimum of overlapping genes.}

\item{fdr_threshold}{Optional float. Shall a filter be applied for the
maximum tolerated FDR.}
}
\value{
A list containing (only for the gene sets passing the filters):
\itemize{
\item pvals - The p-values from the hypergeometric test
\item odds_ratios - The calculated odds ratios
\item hits - The size of the overlap
\item gene_set_lengths - The length of the gene sets.
\item fdr - The FDR calculated across all gene sets (before filtering).
\item to_keep - 1-based indices of the gene sets that passed the
(optional) thresholds.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Given a set of target genes, this is a Rust implementation of an
hypergeometric test testing for overenrichment of the target genes in the
gene sets.
}
