% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hypergeom_test_list}
\alias{rs_hypergeom_test_list}
\title{Run a hypergeometric test over a list of target genes}
\usage{
rs_hypergeom_test_list(
  target_genes_list,
  gene_sets,
  gene_universe,
  min_overlap,
  fdr_threshold
)
}
\arguments{
\item{target_genes_list}{A list of character vectors, each representing a
target gene set.}

\item{gene_sets}{A list of strings that represent the gene sets to test
against.}

\item{gene_universe}{A character vector representing the gene universe from
which the target genes and gene sets are sampled from.}

\item{min_overlap}{Optional integer. Shall a filter be applied on the
minimum of overlapping genes.}

\item{fdr_threshold}{Optional float. Shall a filter be applied for the
maximum tolerated FDR.}
}
\value{
A list containing (results of all target sets concatenated, only
for the tests passing the filters):
\itemize{
\item pvals - The p-values from the hypergeometric test.
\item fdr - The FDRs for each target set calculated across all gene sets.
\item odds_ratios - The calculated odds ratios
\item hits - The size of the overlap between the target gene set and
individual gene sets.
\item gene_set_lengths - The length of the gene sets.
\item to_keep - 1-based indices (into \code{gene_sets}) of the tests that
passed.
\item tests_passed - How many tests passed the filter criteria for that
target set.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Given a list of target gene sets, this function will test for each of the
individual target gene sets the hypergeometric enrichment against the
specified gene sets.
}
