% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_knn_mat_to_edge_list}
\alias{rs_knn_mat_to_edge_list}
\title{Flatten kNN matrix to edge list}
\usage{
rs_knn_mat_to_edge_list(knn_mat, one_index)
}
\arguments{
\item{knn_mat}{Integer matrix. Rows represent the samples and the columns
the 0-based indices of the k-nearest neighbours.}

\item{one_index}{Boolean. Shall 1-based indices be returned.}
}
\value{
A flat vector representing the edge list, alternating from and
to, i.e. \code{c(from_1, to_1, from_2, to_2, ...)}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Helper function to leverage Rust to transform a kNN matrix into a flat
edge list.
}
