% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mad_outlier}
\alias{rs_mad_outlier}
\title{Calculate MAD outlier detection in Rust.}
\usage{
rs_mad_outlier(x, threshold, direction)
}
\arguments{
\item{x}{Numerical vector to test.}

\item{threshold}{Numeric. Number of (unscaled) MADs from the median in
either direction that is acceptable.}

\item{direction}{String. One of \code{c("below", "above", "twosided")}. Shall
the outlier direction be done for values below the threshold, above the
threshold or in both directions. Unknown strings default to twosided tests.}
}
\value{
A list with the following items:
\itemize{
\item outlier - Boolean vector if element is an outlier
\item threshold - Applied margin, i.e. \code{threshold * MAD}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Should you provide an empty vector, the function will return an empty
boolean and a threshold of 0.
}
