% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_pca_residuals}
\alias{rs_mc_pca_residuals}
\title{Calculates PCA on Pearson residuals for meta cells}
\usage{
rs_mc_pca_residuals(
  sparse_data,
  residual_fit,
  no_pcs,
  pca_params,
  gene_indices,
  seed,
  verbose
)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{cs_type}. Shape is (metacells, genes). Pass
raw counts.}

\item{residual_fit}{List. A fit from \code{\link[=rs_mc_fit_residuals]{rs_mc_fit_residuals()}}.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{pca_params}{Named list. Contains the parameters to use for this PCA
run.}

\item{gene_indices}{Integer vector. The gene indices to use. (0-indexed!)}

\item{seed}{Integer. Random seed for the randomised SVD.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following items
\itemize{
\item scores - The samples projected on the PCA space.
\item loadings - The loadings of the features for the PCA.
\item singular_values - The singular values for the PCA.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
In-memory version of \code{\link[=rs_sc_pca_residuals]{rs_sc_pca_residuals()}}. As there, \code{clr} and
\code{normalise_variance} must both be \code{FALSE}.
}
\keyword{internal}
