% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_residual_variance}
\alias{rs_mc_residual_variance}
\title{Residual variance and variable features for meta cells}
\usage{
rs_mc_residual_variance(sparse_data, residual_fit, n_hvg, verbose)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{cs_type}. Shape is (metacells, genes). Pass
raw counts.}

\item{residual_fit}{List. A fit from \code{\link[=rs_mc_fit_residuals]{rs_mc_fit_residuals()}}.}

\item{n_hvg}{Integer. Variable features to take from each group.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following items
\itemize{
\item genes - The gene indices the variances are indexed by. (0-indexed!)
\item variance - Matrix of residual variance, genes by groups.
\item hvg - The selected gene indices, ascending. (0-indexed!)
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
In-memory version of \code{\link[=rs_sc_residual_variance]{rs_sc_residual_variance()}}.
}
\keyword{internal}
