% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_vision}
\alias{rs_mc_vision}
\title{Calculate VISION pathway scores in Rust (for meta cells)}
\usage{
rs_mc_vision(sparse_data, gs_list, verbose)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{cs_type}. Shape is (metacells, genes) and the
data need to be the \strong{normalised} counts.}

\item{gs_list}{Nested list. Each sublist contains the (0-indexed!) positive
and negative gene indices of that specific gene set.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A matrix of meta cells x vision scores per gene set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The function will take in a list of gene sets that contains lists of \code{"pos"}
and \code{"neg"} gene indices (0-indexed). You don't have to provide the \code{"neg"},
but it can be useful to classify the delta of two stats (EMT, Th1; Th2) etc.
This version works on MetaCell counts which are stored in memory directly.
}
\references{
DeTomaso, et al., Nat. Commun., 2019
}
