% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_meld_sc}
\alias{rs_meld_sc}
\title{Run MELD}
\usage{
rs_meld_sc(embd, knn_data, meld_params, labels, n_labels, seed, verbose)
}
\arguments{
\item{embd}{Numeric matrix. The original embedding that was used to generate
the kNN graph.}

\item{knn_data}{Optional named list. This contains pre-computed kNN data
(\code{indices} (0-indexed), \code{dist}, \code{dist_metric} and \code{k}). The user has to
ensure consistency! If provided, this will be used.}

\item{meld_params}{Named list. Contains the parameters to use for MELD.}

\item{labels}{Integer. The group label per cell (1-indexed!), in
\verb{1..n_labels}.}

\item{n_labels}{Integer. Number of labels represented in the data.}

\item{seed}{Integer. For reproducibility.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following items
\itemize{
\item raw_scores - Matrix of cells x labels with the raw MELD scores.
\item norm_scores - Same shape. Negative values were clamped to 0 and the
rows L1 normalised. This yields probability-like values.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This implements a Rust-based version of the MELD algorithm, see Burkhardt,
et al. Nat. Biotechnol., 2021.
}
\keyword{internal}
