% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_onto_sim_wang_mat}
\alias{rs_onto_sim_wang_mat}
\title{Calculate the Wang similarity matrix for an ontology}
\usage{
rs_onto_sim_wang_mat(parents, children, w, flat_matrix)
}
\arguments{
\item{parents}{String vector. The names of the parents.}

\item{children}{String vector. The names of the children. The length of
\code{parents} needs to be equal to \code{children}.}

\item{w}{Numerics. The weights between the parents and children. Need
to be values between 0 and 1.}

\item{flat_matrix}{Boolean. Shall only the upper triangle (row-wise,
diagonal excluded) be returned.}
}
\value{
A list with:
\itemize{
\item sim_mat - the Wang similarity matrix (flat or as matrix.)
\item names - the row and column names for the calculated matrix.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the Wang similarity matrix for a given ontology.
}
