% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_onto_similarity}
\alias{rs_onto_similarity}
\title{Calculate the semantic similarity in an ontology}
\usage{
rs_onto_similarity(terms, ancestor_list, ic_list, similarity_type)
}
\arguments{
\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}

\item{ancestor_list}{R list with names being the term and the elements in the list the names
of the ancestors.}

\item{ic_list}{R list with the names being the term and the elements the information content
of this given term. Needs to be a single float!}

\item{similarity_type}{String. Need to be one of \code{c("resnik", "lin")}.}
}
\value{
A list with:
\itemize{
\item terms - The supplied iterated terms.
\item similarities - Numeric vector of the similarities.
}
}
\description{
This function calculates the Resnik or Lin similarity for a given ontology.
}
