% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_onto_similarity_both}
\alias{rs_onto_similarity_both}
\title{Calculate the Resnik and Lin semantic similarity}
\usage{
rs_onto_similarity_both(terms, ancestor_list, ic_list, max_ic)
}
\arguments{
\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}

\item{ancestor_list}{R list with names being the term and the elements in the list the names
of the ancestors.}

\item{ic_list}{R list with the names being the term and the elements the information content
of this given term. Needs to be a single float!}

\item{max_ic}{Double. The maximum information content observed in the data. This will return
normalised Resnik (i.e., scaled between 0 and 1). If set to 1, it will return the Resnik distance.}
}
\value{
A list with:
\itemize{
\item terms - The supplied iterated terms.
\item resnik_sim - The Resnic similarities.
\item lin_sim - The Lin similarities.
}
}
\description{
This function calculates the Resnik and Lin similarity for a given ontology in
one call.
}
