% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_page_rank}
\alias{rs_page_rank}
\title{Rust version of calculating the personalised page rank}
\usage{
rs_page_rank(
  node_names,
  from,
  to,
  weights,
  personalised,
  undirected,
  damping_factor,
  tol
)
}
\arguments{
\item{node_names}{String vector. Name of the graph nodes.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{weights}{Optional weight vector. If NULL, defaults to 1.0 as weight
for all edges.}

\item{personalised}{Numerical vector. The reset values. They must sum to 1
and be of same length of \code{node_names}!}

\item{undirected}{Boolean. Is this an undirected graph.}

\item{damping_factor}{Numeric in \verb{[0, 1]}. Probability of continuing the
walk.}

\item{tol}{Numeric > 0. Convergence threshold on the L1 change between
iterations.}
}
\value{
The personalised page rank values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Personalised page rank with at most 1000 iterations.
}
