% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_page_rank_parallel}
\alias{rs_page_rank_parallel}
\title{Calculate massively parallelised personalised page rank scores}
\usage{
rs_page_rank_parallel(
  node_names,
  from,
  to,
  weights,
  diffusion_scores,
  undirected,
  damping_factor,
  tol
)
}
\arguments{
\item{node_names}{String vector. Name of the graph nodes.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{weights}{Optional weight vector. If NULL, defaults to 1.0 as weight
for all edges.}

\item{diffusion_scores}{List. The personalised vectors for the page rank
reset values. Each element must sum to 1 and be of same length of
\code{node_names}!}

\item{undirected}{Boolean. Is this an undirected graph.}

\item{damping_factor}{Numeric in \verb{[0, 1]}. Probability of continuing the
walk.}

\item{tol}{Numeric > 0. Convergence threshold on the L1 change between
iterations.}
}
\value{
A matrix of the scores with each row representing an element in the
\code{diffusion_scores} list (in order), and each column representing the value
of the personalised page rank diffusion for this node.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Helper function to calculate in parallel on the same network the
personalised page rank as fast as possible. Can be used for permutation
type approaches.
}
