% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_page_rank_permutations}
\alias{rs_page_rank_permutations}
\title{Helper function to calculate permutation-based page rank scores}
\usage{
rs_page_rank_permutations(node_names, from, to, diffusion_scores, undirected)
}
\arguments{
\item{node_names}{String vector. Name of the graph nodes.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{diffusion_scores}{List. The personalised vectors for the page rank reset
values. Each element must sum to 1 and be of same length of \code{node_names}!}

\item{undirected}{Boolean. Is this an undirected graph.}
}
\value{
A list containing:
\itemize{
\item means - The mean personalised page-rank scores based on the permutations.
\item sd - The standard deviation of the personalised page-rank scores based on
permutations.
}
}
\description{
Helper function to calculate permutation based page-rank scores.
}
