% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_page_rank_permutations_tied}
\alias{rs_page_rank_permutations_tied}
\title{Helper function to calculate permutation-based tied page rank scores}
\usage{
rs_page_rank_permutations_tied(
  node_names,
  from,
  to,
  diffusion_scores_1,
  diffusion_scores_2,
  summarisation_fun,
  undirected
)
}
\arguments{
\item{node_names}{String vector. Name of the graph nodes.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{diffusion_scores_1}{List. The first set of personalised vectors for
the page rank reset values. Each element must sum to 1 and be of same length
of \code{node_names}!}

\item{diffusion_scores_2}{List. The second set of personalised vectors for
the page rank reset values. Each element must sum to 1 and be of same length
of \code{node_names}!}

\item{summarisation_fun}{String. One of \code{c("min", "max", "avg")}. Which type
of summarisation function to use to calculate the tied diffusion.}

\item{undirected}{Boolean. Is this an undirected graph.}
}
\value{
A list containing:
\itemize{
\item means - The mean personalised page-rank scores based on the permutations.
\item sd - The standard deviation of the personalised page-rank scores based on
permutations.
}
}
\description{
Helper function to calculate permutation based tied page-rank
scores.
}
