% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_pointwise_mutual_info}
\alias{rs_pointwise_mutual_info}
\title{Calculates the point wise mutual information}
\usage{
rs_pointwise_mutual_info(x, normalise)
}
\arguments{
\item{x}{Logical matrix. The columns represent features and the rows
represent samples}

\item{normalise}{Shall the normalised pointwise mutual information be
returned.}
}
\value{
The (normalised) pointwise mutual information matrix between the
columns. Pairs without co-occurrence are \code{-Inf}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the pointwise mutual information (can be also normalised) across
all columns of the data. This can be used to identify (dis)similar samples
based on Boolean characteristics.
}
