% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_prcomp}
\alias{rs_prcomp}
\title{Rust implementation of prcomp}
\usage{
rs_prcomp(x, scale, top_pcs)
}
\arguments{
\item{x}{Numeric matrix. Rows = samples, columns = features.}

\item{scale}{Boolean. Shall the columns be variance normalised. (Mean
centring will automatically occur.)}

\item{top_pcs}{Optional integer. Only return the top PCs (under the hood
all of them will be calculated).}
}
\value{
A list with:
\itemize{
\item scores - The product of x (centred and potentially scaled) with v.
\item v - v matrix of the SVD.
\item s - Eigenvalues of the SVD.
\item scaled - Boolean. Was the matrix scaled.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Runs the singular value decomposition over the matrix x. Assumes that
samples = rows, and columns = features.
}
