% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_prepare_gsva_gs}
\alias{rs_prepare_gsva_gs}
\title{Prepare a pathway list for GSVA}
\usage{
rs_prepare_gsva_gs(feature_names, pathway_list, min_size, max_size)
}
\arguments{
\item{feature_names}{String vector. The feature names of the matrix (should
be the rows).}

\item{pathway_list}{List. A list containing the pathways and the respective
genes}

\item{min_size, max_size}{Integer. The minimum and maximum size respectively.}
}
\value{
Returns a named list with the sorted, 0-based indices of the
pathways whose size (after matching to \code{feature_names}) is within
\verb{[min_size, max_size]}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
