% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_random_svd}
\alias{rs_random_svd}
\title{Run randomised SVD over a matrix}
\usage{
rs_random_svd(x, scale, rank, seed, oversampling, n_power_iter)
}
\arguments{
\item{x}{Numeric matrix. Rows = samples, columns = features.}

\item{scale}{Boolean. Shall the columns be variance normalised. (Mean
centring will automatically occur.)}

\item{rank}{Integer. The rank to use.}

\item{seed}{Integer. Random seed for reproducibility.}

\item{oversampling}{Optional integer. Defaults to \code{10L} if \code{NULL}.}

\item{n_power_iter}{Optional integer. Number of power iterations (each with
a QR decomposition). Defaults to \code{2L} if \code{NULL}.}
}
\value{
A list with:
\itemize{
\item scores - u matrix of the SVD multiplied by the singular values.
\item v - v matrix of the SVD.
\item s - Singular values of the SVD.
\item scaled - Boolean. Was the matrix scaled.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Runs a randomised singular value decomposition over a matrix. This
implementation is faster than the full SVD on large data sets, with slight
loss in precision.
}
