% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_rank_matrix_col}
\alias{rs_rank_matrix_col}
\title{Gene rank matrix}
\usage{
rs_rank_matrix_col(exp)
}
\arguments{
\item{exp}{Numerical matrix. The expression matrix (rows = genes, columns =
samples).}
}
\value{
Returns a matrix of ranks with the same shape as \code{exp}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Ranks the columns of a matrix with the tie method.
}
