% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_rbf_function}
\alias{rs_rbf_function}
\title{Apply a Radial Basis Function}
\usage{
rs_rbf_function(x, epsilon, rbf_type)
}
\arguments{
\item{x}{Numeric vector. The distances you wish to apply the RBF onto.}

\item{epsilon}{Float. Epsilon parameter for the RBF.}

\item{rbf_type}{String. Needs to be from
\code{c("gaussian", "bump", "inverse_quadratic")}. Other values raise an error.}
}
\value{
The affinities after the Kernel was applied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Applies a radial basis function (RBF) to a given distance vector. Has the
option to apply a Gaussian, Bump or Inverse Quadratic RBF.
}
