% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_rbf_iterate_epsilons}
\alias{rs_rbf_iterate_epsilons}
\title{Helper to identify the right epsilon parameter}
\usage{
rs_rbf_iterate_epsilons(dist, epsilon_vec, original_dim, shift, rbf_type)
}
\arguments{
\item{dist}{Numeric vector. The distances you wish to apply the RBF function
to.}

\item{epsilon_vec}{Numeric vector. The epsilons you wish to use/test.}

\item{original_dim}{Integer. The original dimensions of the symmetric
distance matrix.}

\item{shift}{Boolean. Was the matrix shifted up (\code{FALSE} = diagonal
included; \code{TRUE} = diagonal not included).}

\item{rbf_type}{String. One of \code{c("gaussian", "bump", "inverse_quadratic")}
for the currently implemented RBF function. Unknown strings default to
Gaussian.}
}
\value{
A matrix with rows representing the features and columns the
epsilons tested. Values are the summed affinity to other features.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function will take a distance vector from the upper triangle of a
symmetric distance matrix and apply the desired RBF with the supplied
epsilon from epsilon vec. Subsequently, the column sums will be measured to
identify the total similarity of each feature with other features. This data
can be used to see if the data follows scale-free topology for example to
identify the right epsilon parameter with the given RBF.
}
\keyword{internal}
