% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_rbh_cor}
\alias{rs_rbh_cor}
\title{Generate reciprocal best hits based on correlations}
\usage{
rs_rbh_cor(module_matrices, k_best, spearman, min_similarity)
}
\arguments{
\item{module_matrices}{A named list of matrices with row and column names.
Rows represent features and columns the modules you wish to calculate the
correlations for. Only features shared between two matrices are used.}

\item{k_best}{Integer. Number of best neighbours to consider. If set to
\code{1L}, this behaves as the traditional reciprocal best hit. If you set this
to \code{3L} you consider edges if the modules is in the top 3 best modules
by similarity for each other.}

\item{spearman}{Boolean. Shall Spearman correlation be used.}

\item{min_similarity}{Numeric. Only hits with an absolute correlation
strictly above this are returned.}
}
\value{
A list containing:
\itemize{
\item origin - The name of the origin of the gene modules.
\item target - The name of the target of the gene modules.
\item comparisons - Integer vector indicating how many RBH hits were
identified in this comparison
\item origin_modules - Names of the gene modules from the origin.
\item target_modules - Names of the gene modules from the target.
\item similarity - The absolute correlations between the two respective
gene modules.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function takes list of (named) matrices which represent for example
matrix factorisation results you wish to identify (k-th) reciprocal best
hits (RBH) for. The rows need to represent the features and the columns the
parts you wish to calculate the RBH for.
}
