% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_rbh_sets}
\alias{rs_rbh_sets}
\title{Generate reciprocal best hits based on set similarities}
\usage{
rs_rbh_sets(module_list, k_best, overlap_coefficient, min_similarity)
}
\arguments{
\item{module_list}{A nested named list. The outer list should contain the
origin of the gene modules, the inner list the names of the gene modules and
the respective genes in them.}

\item{k_best}{Integer. Number of best neighbours to consider. If set to
\code{1L}, this behaves as the traditional reciprocal best hit. If you set this
to \code{3L} you consider edges if the modules is in the top 3 best modules
by similarity for each other.}

\item{overlap_coefficient}{Boolean. Shall the overlap coefficient instead
of the Jaccard similarity be used.}

\item{min_similarity}{Numeric. Only hits with a similarity strictly above
this are returned.}
}
\value{
A list containing:
\itemize{
\item origin - The name of the origin of the gene modules.
\item target - The name of the target of the gene modules.
\item comparisons - Integer vector indicating how many RBH hits were
identified in this comparison
\item origin_modules - Names of the gene modules from the origin.
\item target_modules - Names of the gene modules from the target.
\item similarity - The similarities between the two respective gene
modules.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function takes a nested list that contains gene modules/sets derived
from various methods and identifies (k-th) reciprocal best hits between
gene modules/sets across every pair of origins.
}
