% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_doublet_detection}
\alias{rs_sc_doublet_detection}
\title{Detect Doublets via BoostClassifier (in Rust)}
\usage{
rs_sc_doublet_detection(
  f_path_gene,
  f_path_cell,
  cells_to_keep,
  boost_params,
  seed,
  verbose,
  streaming
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file.}

\item{cells_to_keep}{Integer vector. The indices (0-indexed!) of the cells
to include in this analysis.}

\item{boost_params}{List. Parameter list, see
\code{\link[=params_boost]{params_boost()}}.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{streaming}{Boolean. Shall the data be streamed for the HVG
calculations.}
}
\value{
A list with
\itemize{
\item doublet - Boolean vector indicating which observed cells are
predicted as doublets (TRUE = doublet, FALSE = singlet).
\item doublet_score - Numerical vector with the doublet score per cell,
averaged across iterations.
\item voting_avg - Voting average per cell across the iterations.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Implementation of BoostClassifier in Rust.
}
\keyword{internal}
