% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_get_gene_set_perc}
\alias{rs_sc_get_gene_set_perc}
\title{Calculate the percentage of gene sets in the cells}
\usage{
rs_sc_get_gene_set_perc(
  f_path_cell,
  gene_set_idx,
  cell_indices,
  streaming,
  verbose
)
}
\arguments{
\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file.}

\item{gene_set_idx}{Named list with integer(!) positions (0-indexed!) as
elements of the genes of interest.}

\item{cell_indices}{Integer. The indices of the cells for which to calculate
the proportions. (0-indexed!)}

\item{streaming}{Boolean. Shall the data be worked on in chunks.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list, named like \code{gene_set_idx}, with one numerical vector per
gene set holding the percentage of counts per cell.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function allows to calculate for example the proportion of
mitochondrial genes, or ribosomal genes in the cells for QC purposes.
}
\keyword{internal}
