% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_knn}
\alias{rs_sc_knn}
\title{Generates the kNN graph}
\usage{
rs_sc_knn(embd, knn_params, validate_index, verbose, seed)
}
\arguments{
\item{embd}{Numerical matrix. The embedding matrix to use to generate the
kNN graph.}

\item{knn_params}{List. The kNN parameters defined by
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.}

\item{validate_index}{Boolean. If you want to validate the index via
an exhaustive search in a subset of cells.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{seed}{Integer. Seed for reproducibility purposes.}
}
\value{
An integer matrix of N x k with the neighbour indices
(0-indexed!), N being the number of cells and k the number of neighbours.
Rows the search left short are padded by repeating their last neighbour.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function is a wrapper over the Rust-based generation of the approximate
nearest neighbours.
}
