% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_otsu_method}
\alias{rs_sc_otsu_method}
\title{Run Otsu's method}
\usage{
rs_sc_otsu_method(scores, bins)
}
\arguments{
\item{scores}{Numeric vector. The vector for which to identify the
threshold.}

\item{bins}{Integer. Number of bins to use for the histogram building.}
}
\value{
The threshold based on Otsu's method
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Maximises between-class variance of the observed score distribution to
find the optimal binary split. Robust to both bimodal and skewed
distributions.
}
\keyword{internal}
