% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_scenic_grn}
\alias{rs_scenic_grn}
\title{SCENIC: Generating gene-regulatory networks}
\usage{
rs_scenic_grn(
  f_path_genes,
  cell_indices,
  gene_indices,
  tf_indices,
  scenic_params,
  seed,
  verbose
)
}
\arguments{
\item{f_path_genes}{Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer vector. 0-indexed(!) positions of cells to
include in the analysis}

\item{gene_indices}{Integer vector. 0-indexed(!) positions of the genes
to include.}

\item{tf_indices}{Integer vector. 0-indexed(!) positions of the TF
predictor variables to use in the generation of the regression learners.}

\item{scenic_params}{Named list. Contains all of the parameters needed for
SCENIC.}

\item{seed}{Integer. Controls reproducibility of the function.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A gene x TF importance matrix
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
