% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_sim_list}
\alias{rs_set_sim_list}
\title{Set similarities}
\usage{
rs_set_sim_list(string, string_list, similarity_index)
}
\arguments{
\item{string}{The String vector against which to calculate the set similarities.}

\item{string_list}{The list of character vectors for which to calculate the set similarities.}

\item{similarity_index}{Shall the similarity index instead of the Jaccard similarity be calculated.}
}
\description{
This function calculates the Jaccard or similarity index between a given
string vector and a list of other string vectors.
}
