% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_similarity}
\alias{rs_set_similarity}
\title{Set similarities}
\usage{
rs_set_similarity(s_1, s_2, overlap_coefficient)
}
\arguments{
\item{s_1}{Character vector. The first set.}

\item{s_2}{Character vector. The second set.}

\item{overlap_coefficient}{Boolean. Use the overlap coefficient instead of
the Jaccard similarity.}
}
\value{
The Jaccard similarity or overlap coefficient.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the Jaccard similarity or overlap coefficient
between two string vectors.
}
