% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_similarity_list2}
\alias{rs_set_similarity_list2}
\title{Set similarities over two list}
\usage{
rs_set_similarity_list2(s_1_list, s_2_list, overlap_coefficient)
}
\arguments{
\item{s_1_list}{R list. The first list of string elements you want to
compare against.}

\item{s_2_list}{R list. The second list of string elements you want to
compare against.}

\item{overlap_coefficient}{Boolean. Use the overlap coefficient instead of
the Jaccard similarity.}
}
\value{
A matrix of the similarities between the elements. The rows
represent \code{s_1_list} and the columns \code{s_2_list}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the Jaccard similarity or overlap coefficient
between all elements of two lists.
}
