% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_seurat_cca}
\alias{rs_seurat_cca}
\title{Seurat CCA batch correction in Rust}
\usage{
rs_seurat_cca(
  f_path_gene,
  f_path_cell,
  cell_indices,
  gene_indices,
  batch_indices,
  precomputed_pca,
  cca_params,
  verbose,
  seed
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file. Used if
you wish to use the PFlogPF transformation during the optional PCA step.}

\item{cell_indices}{Integer. The cell indices to use. (0-indexed!)}

\item{gene_indices}{Integer. The gene indices to use. (0-indexed!) Ideally
these are batch-aware highly variable genes.}

\item{batch_indices}{Integer vector. These represent to which batch a given
cell belongs. Need to be 0-indexed and contiguous.}

\item{precomputed_pca}{Optional PCA matrix. If you want to provide a
pre-computed matrix.}

\item{cca_params}{List. Contains all of the Seurat CCA parameters.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{seed}{Integer. Seed for reproducibility purposes.}
}
\value{
Numerical matrix, cells x dimensions, with the batch-corrected
embedding.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function implements the canonical correlation analysis (CCA) anchor
integration from Stuart, et al. Anchors are identified in a per-pair
L2-normalised canonical correlation embedding, scored via shared
neighbours, filtered in gene space and then used to apply a kernel-weighted
correction on the union PCA embedding.
}
\references{
Stuart, et al., Cell, 2019
}
\keyword{internal}
