% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_snf}
\alias{rs_snf}
\title{Similarity network fusion}
\usage{
rs_snf(aff_mat_list, k, t, alpha)
}
\arguments{
\item{aff_mat_list}{A list of numerical matrices. The affinity matrices to
fuse together.}

\item{k}{Integer. Number of neighbours to consider.}

\item{t}{Integer. Number of iterations for the algorithm.}

\item{alpha}{Float. Normalisation parameter controlling the fusion strength.}
}
\value{
The final affinity matrix after the fusion.
}
\description{
This function iteratively fuses the affinity matrices together.
}
