% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_snf_affinity_cat}
\alias{rs_snf_affinity_cat}
\title{Calculate the SNF affinity matrix for categorical values}
\usage{
rs_snf_affinity_cat(data, k, mu)
}
\arguments{
\item{data}{Integer matrix. Needs to be oriented features x samples! The
integers represent the factor values of the categories.}

\item{k}{Integer. Number of neighbours to consider.}

\item{mu}{Float. Normalisation factor for the Gaussian kernel width.}
}
\value{
The affinity matrix based on categorical values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
