% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_snf_affinity_mixed}
\alias{rs_snf_affinity_mixed}
\title{Calculate the SNF affinity matrix for mixed values}
\usage{
rs_snf_affinity_mixed(data, is_cat, k, mu)
}
\arguments{
\item{data}{Numerical matrix. Needs to be oriented features x samples! This
function will calculate the Gower distance under the hood for the affinity
calculation.}

\item{is_cat}{Boolean vector. Which of the features are categorical. Needs
to be of \code{nrow(data)}.}

\item{k}{Integer. Number of neighbours to consider.}

\item{mu}{Float. Normalisation factor for the Gaussian kernel width.}
}
\value{
The affinity matrix based on mixed values.
}
\description{
Calculate the SNF affinity matrix for mixed values
}
