% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sparse_dict_dgrdl}
\alias{rs_sparse_dict_dgrdl}
\title{Generate a sparse dictionary with DGRDL}
\usage{
rs_sparse_dict_dgrdl(x, dgrdl_params, seed, verbose)
}
\arguments{
\item{x}{Numerical matrix. Rows = samples, columns = features.}

\item{dgrdl_params}{A list with the parameters for the algorithm. Missing
items fall back to defaults. Expects the following items.
\itemize{
\item sparsity - Integer. Sparsity constraint (max non-zero coefficients
per signal).
\item dict_size - Integer. Size of the dictionary.
\item alpha - Float. Sample context regularisation weight. The higher
the stronger the regularisation.
\item beta - Float. Feature context regularisation weight. The higher
the stronger the regularisation.
\item max_iter - Integer. Maximum iteration for the algorithm.
\item k_neighbours - Integer. Number of k neighbours for the sample and
feature Laplacian matrix for the regularisation.
\item admm_iter - Integer. Number of iterations for using alternating
direction method of multipliers (ADMM).
\item rho - Float. ADMM step size.
}}

\item{seed}{Integer. Seed for the initialisation of the algorithm.}

\item{verbose}{Boolean. Controls the verbosity of the function and reports
timing of individual steps.}
}
\value{
A list with the following elements:
\itemize{
\item dictionary - The dictionary of samples x dict_size.
\item coefficients - The feature loadings of size dict_size x features.
\item feature_laplacian - The kNN graph Laplacian of the features as a
CSR list with \code{data}, \code{indptr}, \code{indices}, \code{nrow}, \code{ncol} and \code{cs_type}.
\item sample_laplacian - The kNN graph Laplacian of the samples, same
format.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This is the Rust implementation of dual graph regularised dictionary
learning in the implementation of Pan, et al., Cell Systems, 2022.
}
