% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_spectral_clustering}
\alias{rs_spectral_clustering}
\title{Rust implementation of spectral clustering}
\usage{
rs_spectral_clustering(
  data,
  distance_type,
  epsilon,
  k_neighbours,
  n_clusters,
  max_iters,
  seed
)
}
\arguments{
\item{data}{Numerical matrix. The data to cluster. Rows = samples, columns =
features.}

\item{distance_type}{String. One of
\code{c("euclidean", "manhattan", "canberra", "cosine", "correlation")}. Unknown
strings raise an error.}

\item{epsilon}{Numerical. The epsilon parameter for the Gaussian Radial
Basis function}

\item{k_neighbours}{Integer. Number of neighbours to consider in the kNN
graph generation}

\item{n_clusters}{Integer. Number of clusters to identify}

\item{max_iters}{Integer. Number of iterations for k-means clustering}

\item{seed}{Integer. Seed for reproducibility}
}
\value{
A vector with the 1-based cluster membership of the samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This version can take in data as is and will calculate the distance matrix
internally and convert it into similarities with a Gaussian RBF.
}
