% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_spectral_clustering_sim}
\alias{rs_spectral_clustering_sim}
\title{Rust implementation of spectral clustering}
\usage{
rs_spectral_clustering_sim(
  similarities,
  k_neighbours,
  n_clusters,
  max_iters,
  seed
)
}
\arguments{
\item{similarities}{Numerical matrix representing the similarities. Needs
to be symmetric!}

\item{k_neighbours}{Integer. Number of neighbours to consider in the kNN
graph generation}

\item{n_clusters}{Integer. Number of clusters to identify}

\item{max_iters}{Integer. Number of iterations for k-means clustering}

\item{seed}{Integer. Seed for reproducibility}
}
\value{
A vector with the 1-based cluster membership of the samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Spectral clustering on a pre-calculated similarity matrix.
}
