% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_ssgsea}
\alias{rs_ssgsea}
\title{Rust version of the ssGSEA algorithm}
\usage{
rs_ssgsea(exp, gs_list, alpha, normalise, timings)
}
\arguments{
\item{exp}{Numerical matrix. The expression matrix with rows = genes, and
columns = samples}

\item{gs_list}{List. A list containing the 0-based indices of the pathway
genes. See \code{\link[=rs_prepare_gsva_gs]{rs_prepare_gsva_gs()}}.}

\item{alpha}{Float. The alpha parameter to adjust the weights.}

\item{normalise}{Boolean. Shall the scores be normalised.}

\item{timings}{Boolean. Prints timings from the algorithm.}
}
\value{
Returns a matrix of gene set ES scores x samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Rust-based implementation of the popular single sample GSEA algorithm. Has
further performance optimisations compared to the original implementation.
}
