% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_synthetic_sc_data_csc}
\alias{rs_synthetic_sc_data_csc}
\title{Generate synthetic single cell data (Seurat type)}
\usage{
rs_synthetic_sc_data_csc(n_genes, n_cells, min_genes, max_genes, max_exp, seed)
}
\arguments{
\item{n_genes}{Integer. Number of genes you wish to have in the synthetic
data.}

\item{n_cells}{Integer. Number of cells you wish to have in the synthetic
data.}

\item{min_genes}{Integer. Minimum number of genes expressed per cell.}

\item{max_genes}{Integer. Maximum number of genes expressed per cell.}

\item{max_exp}{Upper bound in terms of expression. Expression values will be
sampled from \code{1:max_exp}.}

\item{seed}{Integer. Seed for reproducibility purposes.}
}
\value{
The list with the synthetic data with the following items:
\itemize{
\item data - The synthetic counts
\item col_ptrs - The column pointers
\item row_indices - The row indices
\item nrow - Number of rows (cells)
\item ncol - Number of cols (genes)
}
}
\description{
This function generates pseudo data to test single cell
functions in form of the Seurat version, with cells = columns and genes =
rows. The data is CSC type.
}
