% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_upper_triangle_to_dense}
\alias{rs_upper_triangle_to_dense}
\title{Reconstruct a matrix from a flattened upper triangle vector}
\usage{
rs_upper_triangle_to_dense(data, shift, n)
}
\arguments{
\item{data}{Numeric vector. The vector of for example correlation
coefficients that you want to use to go back to a dense matrix.}

\item{shift}{Boolean. If \code{TRUE}, \code{data} excludes the diagonal and the
diagonal is set to \code{1}. If \code{FALSE}, \code{data} includes the diagonal.}

\item{n}{Integer. Original dimension (i.e., ncol/nrow) of the matrix to be
reconstructed.}
}
\value{
The dense R matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function takes a flattened vector of the upper triangle from a
symmetric matrix (think correlation matrix) and reconstructs the full
dense matrix for you.
}
