% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_whiten_matrix}
\alias{rs_whiten_matrix}
\title{Whiten a matrix}
\usage{
rs_whiten_matrix(x)
}
\arguments{
\item{x}{The matrix to whiten. The whitening will happen over the columns.}
}
\value{
The whitened matrix (will be transposed compared to x).
}
\description{
Whitens the matrix for subsequent usage. This is a need pre-
processing step for ICA.
}
