% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{safe_plot_capture}
\alias{safe_plot_capture}
\title{Helper function to capture R-derived plots}
\usage{
safe_plot_capture(plot)
}
\arguments{
\item{plot}{The R plot to capture.}
}
\value{
The recorded plot if it works, otherwise \code{NULL}.
}
\description{
Helper function to capture R-derived plots
}
