% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_sc.R
\name{sc_pipeline}
\alias{sc_pipeline}
\title{Construct an empty single cell pipeline}
\usage{
sc_pipeline()
}
\value{
An empty \code{ScPipeline} object.
}
\description{
Linear container of \code{ScStep}s. Append steps with \code{\link{\%>>\%}} and execute with
\code{\link[=apply_pipeline]{apply_pipeline()}}. Pipelines are inert until applied; steps can be
inspected via \code{pipeline$steps}.
}
\examples{
# an empty container, filled with `\%>>\%`
sc_pipeline() \%>>\%
  step_hvg_sc(hvg_no = 30L) \%>>\%
  step_pca_sc(no_pcs = 10L)
}
