% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc_cache_state.R
\name{SC_STAMP_ATTR}
\alias{SC_STAMP_ATTR}
\title{Attribute name under which every stamp is stored}
\usage{
SC_STAMP_ATTR
}
\description{
Attribute name under which every stamp is stored
}
\keyword{internal}
