% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{score_clusters}
\alias{score_clusters}
\alias{score_clusters.ScTypeResults}
\title{Score clusters based on ScType}
\usage{
score_clusters(x, cluster_labels)

\method{score_clusters}{ScTypeResults}(x, cluster_labels)
}
\arguments{
\item{x}{\code{ScTypeResults} object.}

\item{cluster_labels}{Integer vector. Cluster assignment, of length of the
scored cells.}
}
\value{
A \code{data.table} with cluster_id, cell_type, scores and n_cells.
}
\description{
Score clusters based on ScType
}
\examples{
# aggregate the per cell ScType scores onto a clustering
sc <- demo_single_cells()
markers <- data.table::data.table(
  cell_type = rep(sprintf("cell_type_\%i", 1:3), each = 10),
  gene_id = sprintf("gene_\%02d", 1:30)
)
cell_markers <- prepare_cell_markers(obj = sc, marker_df = markers)
scores <- calc_sc_type_scores(
  sc,
  cell_marker_list = cell_markers,
  .verbose = FALSE
)
clusters <- fast_cluster_sc(
  sc,
  resolutions = 1.0,
  n_centroids = 30L,
  .verbose = FALSE
)
head(score_clusters(scores, cluster_labels = get_data(clusters)$res_1))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
