% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_user_option}
\alias{select_user_option}
\title{Helper function for user option selection}
\usage{
select_user_option(options)
}
\arguments{
\item{options}{Character vector. The options the user has to choose from.}
}
\value{
The chosen option
}
\description{
Helper function for user option selection
}
