% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_add_embedding}
\alias{set_add_embedding}
\alias{set_add_embedding.sc_cache}
\title{Add additional embeddings to the class}
\usage{
set_add_embedding(x, embd, name)

\method{set_add_embedding}{sc_cache}(x, embd, name)
}
\arguments{
\item{x}{An object to add the singular values for.}

\item{embd}{Numerical matrix representing the additional embedding.}

\item{name}{String. Name of the embedding.}
}
\description{
Add additional embeddings to the class
}
