% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R
\name{set_cell_mapping}
\alias{set_cell_mapping}
\alias{set_cell_mapping.ScMap}
\alias{set_cell_mapping.SingleCells}
\title{Set cell mapping}
\usage{
set_cell_mapping(x, cell_map)

\method{set_cell_mapping}{ScMap}(x, cell_map)

## S7 method for class <bixverse::SingleCells>
set_cell_mapping(x, cell_map)
}
\arguments{
\item{x}{An object to set cell mapping for}

\item{cell_map}{Named integer indicating indices and names of the cells}
}
\description{
Set a cell mapping for a given object. This is used for the
single cell-related classes with streaming from disk.
}
\examples{
# the mapping is normally written during ingestion
sc <- demo_single_cells(prepped = FALSE)
cells <- get_cell_names(sc)
sc <- set_cell_mapping(sc, stats::setNames(seq_along(cells), cells))
head(get_cell_names(sc), 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
