% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_cell_mapping.sc_mapper}
\alias{set_cell_mapping.sc_mapper}
\title{Set cell mapping method for sc_mapper}
\usage{
\method{set_cell_mapping}{sc_mapper}(x, cell_map)
}
\arguments{
\item{x}{An \code{sc_mapper} object}

\item{cell_map}{Named integer vector for cell mapping}
}
\value{
Updated \code{sc_mapper} object with cell mapping set
}
\description{
Set cell mapping method for sc_mapper
}
