% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R
\name{set_cells_to_keep}
\alias{set_cells_to_keep}
\alias{set_cells_to_keep.ScMap}
\alias{set_cells_to_keep.SingleCells}
\title{Set cells to keep}
\usage{
set_cells_to_keep(x, cells_to_keep)

\method{set_cells_to_keep}{ScMap}(x, cells_to_keep)

## S7 method for class <bixverse::SingleCells>
set_cells_to_keep(x, cells_to_keep)
}
\arguments{
\item{x}{An object to set cells to keep for}

\item{cells_to_keep}{String or integer. The names or indices of the cells
to keep in downstream analysis.}
}
\description{
Set the cells to keep. This is used for the single cell-related
classes with streaming from disk and tells subsequent (Rust) methods which
cells to include.
}
\examples{
# restrict everything downstream to the first 100 cells
sc <- demo_single_cells(prepped = FALSE)
sc <- set_cells_to_keep(sc, get_cell_names(sc)[1:100])
length(get_cells_to_keep(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
