% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_column_index_map}
\alias{set_column_index_map}
\title{Set column index map for sc_mapper object}
\usage{
set_column_index_map(x, gene_mask)
}
\arguments{
\item{x}{An object to set column index mapping for}

\item{gene_mask}{Boolean vector indicating which genes are being kept}
}
\description{
Set column index map for sc_mapper object
}
