% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_column_index_map.sc_mapper}
\alias{set_column_index_map.sc_mapper}
\title{Set column index map for sc_mapper using gene mask}
\usage{
\method{set_column_index_map}{sc_mapper}(x, gene_mask)
}
\arguments{
\item{x}{An \code{sc_mapper} object}

\item{gene_mask}{Logical vector indicating which genes to include}
}
\value{
Updated \code{sc_mapper} object with column index mapping set
}
\description{
Set column index map for sc_mapper using gene mask
}
