% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_column_index_map.single_cell_exp}
\alias{set_column_index_map.single_cell_exp}
\title{Set column index method for \code{single_cell_exp}}
\description{
Set column index method for \code{single_cell_exp}
}
