% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{set_embedding}
\alias{set_embedding}
\alias{set_embedding.MetaCells}
\alias{set_embedding.ScCache}
\alias{set_embedding.SingleCells}
\alias{set_embedding.SingleCellsSubset}
\title{Add additional embeddings to the class}
\usage{
set_embedding(x, embd, name, ...)

## S7 method for class <bixverse::MetaCells>
set_embedding(x, embd, name, ...)

\method{set_embedding}{ScCache}(x, embd, name, ...)

## S7 method for class <bixverse::SingleCells>
set_embedding(x, embd, name, ...)

## S7 method for class <bixverse::SingleCellsSubset>
set_embedding(x, embd, name, ...)
}
\arguments{
\item{x}{An object to add the singular values for.}

\item{embd}{Numerical matrix representing the additional embedding.}

\item{name}{String. Name of the embedding.}

\item{...}{Other parameters.}
}
\description{
Add additional embeddings to the class
}
\examples{
# the first two PCs stored as an embedding in their own right
sc <- demo_single_cells()
sc <- set_embedding(sc, get_pca_factors(sc)[, 1:2], name = "pca_2d")
get_available_embeddings(sc)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
