% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R
\name{set_gene_mapping}
\alias{set_gene_mapping}
\alias{set_gene_mapping.ScMap}
\alias{set_gene_mapping.SingleCells}
\title{Set gene mapping}
\usage{
set_gene_mapping(x, gene_map)

\method{set_gene_mapping}{ScMap}(x, gene_map)

## S7 method for class <bixverse::SingleCells>
set_gene_mapping(x, gene_map)
}
\arguments{
\item{x}{An object to set gene mapping for}

\item{gene_map}{Named integer indicating indices and names of the genes}
}
\description{
Set a gene mapping for a given object. This is used for the
single cell-related classes with streaming from disk.
}
\examples{
# the mapping is normally written during ingestion
sc <- demo_single_cells(prepped = FALSE)
genes <- get_gene_names(sc)
sc <- set_gene_mapping(sc, stats::setNames(seq_along(genes), genes))
head(get_gene_names(sc), 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
