% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_gene_mapping.sc_mapper}
\alias{set_gene_mapping.sc_mapper}
\title{Set gene mapping method for sc_mapper}
\usage{
\method{set_gene_mapping}{sc_mapper}(x, gene_map)
}
\arguments{
\item{x}{An \code{sc_mapper} object}

\item{gene_map}{Named integer vector for gene mapping}
}
\value{
Updated \code{sc_mapper} object with gene mapping set
}
\description{
Set gene mapping method for sc_mapper
}
