% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_gene_mapping.single_cell_exp}
\alias{set_gene_mapping.single_cell_exp}
\title{Set gene mapping method for \code{single_cell_exp}}
\description{
Set gene mapping method for \code{single_cell_exp}
}
