% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_add.R
\name{set_hotspot_membership}
\alias{set_hotspot_membership}
\alias{set_hotspot_membership.sc_hotspot}
\title{Identify hotspot gene clusters}
\usage{
set_hotspot_membership(x, fdr_threshold = 0.05, min_size = 10L)

\method{set_hotspot_membership}{sc_hotspot}(x, fdr_threshold = 0.05, min_size = 10L)
}
\arguments{
\item{x}{An object to generate the hotspot gene clusters for.}

\item{fdr_threshold}{Numeric. The maximum FDR for a given gene-gene local
correlation to be included.}

\item{min_size}{Integer. Minimum cluster size.}
}
\description{
Identify hotspot gene clusters
}
