% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{set_hvg}
\alias{set_hvg}
\alias{set_hvg.MetaCells}
\alias{set_hvg.ScMap}
\alias{set_hvg.SingleCells}
\alias{set_hvg.SingleCellsSubset}
\title{Set the HVG genes}
\usage{
set_hvg(x, hvg)

## S7 method for class <bixverse::MetaCells>
set_hvg(x, hvg)

\method{set_hvg}{ScMap}(x, hvg)

## S7 method for class <bixverse::SingleCells>
set_hvg(x, hvg)

## S7 method for class <bixverse::SingleCellsSubset>
set_hvg(x, hvg)
}
\arguments{
\item{x}{An object to set the HVGs for}

\item{hvg}{String or integer. The names or indices of the highly variable
genes.}
}
\description{
Stores within the class the index positions of the HVG. This is used for
the single cell-related classes and methods.
}
\examples{
# HVGs picked by hand instead of by find_hvg_sc()
sc <- demo_single_cells(prepped = FALSE)
sc <- set_hvg(sc, get_gene_names(sc)[1:20])
head(get_hvg(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
