% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_hvg.sc_mapper}
\alias{set_hvg.sc_mapper}
\title{Set the HVG for sc_mapper}
\usage{
\method{set_hvg}{sc_mapper}(x, hvg)
}
\arguments{
\item{x}{An \code{sc_mapper} object}

\item{hvg}{String or integer. The names or indices of the highly variable
genes.}
}
\value{
Updated \code{sc_mapper} object with column index mapping set
}
\description{
Set the HVG for sc_mapper
}
