% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{set_knn}
\alias{set_knn}
\alias{set_knn.MetaCells}
\alias{set_knn.ScCache}
\alias{set_knn.SingleCells}
\alias{set_knn.SingleCellsSubset}
\title{Set/add KNN}
\usage{
set_knn(x, knn, ...)

## S7 method for class <bixverse::MetaCells>
set_knn(x, knn, ...)

\method{set_knn}{ScCache}(x, knn, ...)

## S7 method for class <bixverse::SingleCells>
set_knn(x, knn, ...)

## S7 method for class <bixverse::SingleCellsSubset>
set_knn(x, knn, ...)
}
\arguments{
\item{x}{An object to add the KNN data to}

\item{knn}{\code{SingleCellNearestNeighbour} class to add to the classes.}

\item{...}{Other parameters.}
}
\description{
Set/add KNN
}
\examples{
# a kNN built outside the object put into the cache
sc <- demo_single_cells()
knn <- generate_knn_sc(sc, .validate_index = FALSE, .verbose = FALSE)
sc <- set_knn(sc, knn)
dim(get_knn_mat(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
