% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_knn.sc_cache}
\alias{set_knn.sc_cache}
\title{Set KNN for sc_mapper}
\usage{
\method{set_knn}{sc_cache}(x, knn_mat)
}
\arguments{
\item{x}{An \code{sc_cache} object}

\item{knn_mat}{Numerical matrix. The matrix with the PCA factors.}
}
\value{
Updated \code{sc_cache} object with added PCA loadings.
}
\description{
Set KNN for sc_mapper
}
