% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{set_pca_factors}
\alias{set_pca_factors}
\alias{set_pca_factors.MetaCells}
\alias{set_pca_factors.ScCache}
\alias{set_pca_factors.SingleCells}
\alias{set_pca_factors.SingleCellsSubset}
\title{Set/add PCA factors}
\usage{
set_pca_factors(x, pca_factor, ...)

## S7 method for class <bixverse::MetaCells>
set_pca_factors(x, pca_factor, ...)

\method{set_pca_factors}{ScCache}(x, pca_factor, ...)

## S7 method for class <bixverse::SingleCells>
set_pca_factors(x, pca_factor, ...)

## S7 method for class <bixverse::SingleCellsSubset>
set_pca_factors(x, pca_factor, ...)
}
\arguments{
\item{x}{An object to add the PCA factors for.}

\item{pca_factor}{Numerical matrix. The matrix with the PCA factors.}

\item{...}{Other parameters.}
}
\description{
Set/add PCA factors
}
\examples{
# an externally computed embedding pushed into the cache
sc <- demo_single_cells(prepped = FALSE)
sc <- set_pca_factors(sc, matrix(stats::rnorm(dim(sc)[1] * 2), ncol = 2))
dim(get_pca_factors(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
