% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_pca_factors.sc_cache}
\alias{set_pca_factors.sc_cache}
\title{Set PCA factors for sc_cache}
\usage{
\method{set_pca_factors}{sc_cache}(x, pca_factor)
}
\arguments{
\item{x}{An \code{sc_cache} object}

\item{pca_factor}{Numerical matrix. The matrix with the PCA factors.}
}
\value{
Updated \code{sc_cache} object with added PCA factors.
}
\description{
Set PCA factors for sc_cache
}
