% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_pca_factors.single_cell_exp}
\alias{set_pca_factors.single_cell_exp}
\title{Set PCA factors method for \code{single_cell_exp}}
\description{
Set PCA factors method for \code{single_cell_exp}
}
