% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{set_pca_loadings}
\alias{set_pca_loadings}
\alias{set_pca_loadings.MetaCells}
\alias{set_pca_loadings.ScCache}
\alias{set_pca_loadings.SingleCells}
\alias{set_pca_loadings.SingleCellsSubset}
\title{Set/add PCA loadings}
\usage{
set_pca_loadings(x, pca_loading, ...)

## S7 method for class <bixverse::MetaCells>
set_pca_loadings(x, pca_loading, ...)

\method{set_pca_loadings}{ScCache}(x, pca_loading, ...)

## S7 method for class <bixverse::SingleCells>
set_pca_loadings(x, pca_loading, ...)

## S7 method for class <bixverse::SingleCellsSubset>
set_pca_loadings(x, pca_loading, ...)
}
\arguments{
\item{x}{An object to add the PCA loadings for.}

\item{pca_loading}{Numerical matrix. The Matrix with the PCA loadings.}

\item{...}{Other parameters.}
}
\description{
Set/add PCA loadings
}
\examples{
# only the first two loading vectors kept
sc <- demo_single_cells()
sc <- set_pca_loadings(sc, get_pca_loadings(sc)[, 1:2])
dim(get_pca_loadings(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
