% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{set_pca_singular_vals}
\alias{set_pca_singular_vals}
\alias{set_pca_singular_vals.MetaCells}
\alias{set_pca_singular_vals.ScCache}
\alias{set_pca_singular_vals.SingleCells}
\alias{set_pca_singular_vals.SingleCellsSubset}
\title{Set/add PCA singular values}
\usage{
set_pca_singular_vals(x, singular_vals, ...)

## S7 method for class <bixverse::MetaCells>
set_pca_singular_vals(x, singular_vals, ...)

\method{set_pca_singular_vals}{ScCache}(x, singular_vals, ...)

## S7 method for class <bixverse::SingleCells>
set_pca_singular_vals(x, singular_vals, ...)

## S7 method for class <bixverse::SingleCellsSubset>
set_pca_singular_vals(x, singular_vals, ...)
}
\arguments{
\item{x}{An object to add the singular values for.}

\item{singular_vals}{Numerical vector. The singular values.}

\item{...}{Other parameters.}
}
\description{
Set/add PCA singular values
}
\examples{
# singular values from a decomposition done elsewhere
sc <- demo_single_cells(prepped = FALSE)
sc <- set_pca_singular_vals(sc, c(4.1, 2.3))
get_pca_singular_val(sc)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
