% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R, R/classes_spatial_spot_subset.R
\name{set_per_sample_image_features}
\alias{set_per_sample_image_features}
\alias{set_per_sample_image_features.SpCache}
\alias{set_per_sample_image_features,bixverse::SpatialSpot-method}
\alias{set_per_sample_image_features,bixverse::SpatialSpotSubset-method}
\title{Set the per-sample histology image features}
\usage{
set_per_sample_image_features(x, exp_id, resolution, features, ...)

\method{set_per_sample_image_features}{SpCache}(x, exp_id, resolution, features, ...)

## S7 method for class <bixverse::SpatialSpot>
set_per_sample_image_features(x, exp_id, resolution, features, ...)

## S7 method for class <bixverse::SpatialSpotSubset>
set_per_sample_image_features(x, exp_id, resolution, features, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{resolution}{String. The image resolution the features were cut from,
one of \code{c("lowres", "hires", "cytassist", "fullres")}.}

\item{features}{The image feature result, see
\code{\link[=new_sp_image_features_res]{new_sp_image_features_res()}}.}

\item{...}{Other parameters.}
}
\description{
Set the per-sample histology image features
}
\keyword{internal}
