% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R, R/classes_spatial_spot_subset.R
\name{set_per_sample_morans_i}
\alias{set_per_sample_morans_i}
\alias{set_per_sample_morans_i.SpCache}
\alias{set_per_sample_morans_i,bixverse::SpatialSpot-method}
\alias{set_per_sample_morans_i,bixverse::SpatialSpotSubset-method}
\title{Set the per-sample Moran's I results}
\usage{
set_per_sample_morans_i(x, exp_id, morans_i, ...)

\method{set_per_sample_morans_i}{SpCache}(x, exp_id, morans_i, ...)

## S7 method for class <bixverse::SpatialSpot>
set_per_sample_morans_i(x, exp_id, morans_i, ...)

## S7 method for class <bixverse::SpatialSpotSubset>
set_per_sample_morans_i(x, exp_id, morans_i, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{morans_i}{\code{data.table} of Moran's I results for this sample.}

\item{...}{Other parameters.}
}
\description{
Set the per-sample Moran's I results
}
\keyword{internal}
