% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R, R/classes_spatial_spot_subset.R
\name{set_per_sample_nhood_enrichment}
\alias{set_per_sample_nhood_enrichment}
\alias{set_per_sample_nhood_enrichment.SpCache}
\alias{set_per_sample_nhood_enrichment,bixverse::SpatialSpot-method}
\alias{set_per_sample_nhood_enrichment,bixverse::SpatialSpotSubset-method}
\title{Set the per-sample neighbourhood enrichment results}
\usage{
set_per_sample_nhood_enrichment(x, exp_id, label_col, result, ...)

\method{set_per_sample_nhood_enrichment}{SpCache}(x, exp_id, label_col, result, ...)

## S7 method for class <bixverse::SpatialSpot>
set_per_sample_nhood_enrichment(x, exp_id, label_col, result, ...)

## S7 method for class <bixverse::SpatialSpotSubset>
set_per_sample_nhood_enrichment(x, exp_id, label_col, result, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{label_col}{String. The obs column name the enrichment was
computed against.}

\item{result}{Named list / \code{data.table} with the enrichment result.}

\item{...}{Other parameters.}
}
\description{
Set the per-sample neighbourhood enrichment results
}
\keyword{internal}
