% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R, R/classes_spatial_spot_subset.R
\name{set_per_sample_pca}
\alias{set_per_sample_pca}
\alias{set_per_sample_pca.SpCache}
\alias{set_per_sample_pca,bixverse::SpatialSpot-method}
\alias{set_per_sample_pca,bixverse::SpatialSpotSubset-method}
\title{Set the per-sample PCA}
\usage{
set_per_sample_pca(x, exp_id, pca, ...)

\method{set_per_sample_pca}{SpCache}(x, exp_id, pca, ...)

## S7 method for class <bixverse::SpatialSpot>
set_per_sample_pca(x, exp_id, pca, ...)

## S7 method for class <bixverse::SpatialSpotSubset>
set_per_sample_pca(x, exp_id, pca, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{pca}{Named list with \code{factors}, \code{loadings} and \code{singular_vals}.}

\item{...}{Other parameters.}
}
\description{
Set the per-sample PCA
}
\keyword{internal}
