% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R
\name{set_residual_fit}
\alias{set_residual_fit}
\alias{set_residual_fit.MetaCells}
\alias{set_residual_fit.ScCache}
\alias{set_residual_fit.SingleCells}
\title{Set/add the fitted residual model}
\usage{
set_residual_fit(x, residual_fit, ...)

## S7 method for class <bixverse::MetaCells>
set_residual_fit(x, residual_fit, ...)

\method{set_residual_fit}{ScCache}(x, residual_fit, ...)

## S7 method for class <bixverse::SingleCells>
set_residual_fit(x, residual_fit, ...)
}
\arguments{
\item{x}{An object to add the fitted model to.}

\item{residual_fit}{\code{ScResidualFit} class, as returned by
\code{\link[=fit_residuals_sc]{fit_residuals_sc()}}.}

\item{...}{Other parameters.}
}
\value{
The object with the fitted model attached.
}
\description{
Set/add the fitted residual model
}
\examples{
# the fitted model taken out and put back
sc <- demo_single_cells()
sc <- fit_residuals_sc(sc, method = "analytic_pearson", .verbose = FALSE)
fit <- get_residual_fit(sc)
sc <- set_residual_fit(remove_residual_fit(sc), fit)
get_residual_fit(sc)$method

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
