% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{set_sc_new_obs_col}
\alias{set_sc_new_obs_col}
\title{Add a new column to the obs table}
\usage{
set_sc_new_obs_col(object, col_name, new_data)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{col_name}{String. The name of the column to add.}

\item{new_data}{Atomic vector. The data to add to the column. Needs to be
of same length as \code{\link[=get_cells_to_keep]{get_cells_to_keep()}} and have the same order!}
}
\value{
The class with updated obs table in the DuckDB
}
\description{
Add a new column to the obs table
}
\examples{
# one new column, in the order of the cells that passed quality control
sc <- demo_single_cells(prepped = FALSE)
sc <- set_sc_new_obs_col(
  sc,
  col_name = "arm",
  new_data = rep(c("ctrl", "treated"), length.out = dim(sc)[1])
)
table(get_sc_obs(sc)$arm)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
