% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{set_sc_new_obs_col_multiple}
\alias{set_sc_new_obs_col_multiple}
\title{Add multiple new columns to the obs table}
\usage{
set_sc_new_obs_col_multiple(object, new_data)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{new_data}{Named list. The names will be the column names and the
elements will be added to the obs table. Needs to be of same length as
\code{\link[=get_cells_to_keep]{get_cells_to_keep()}} and have the same order!}
}
\value{
The class with updated obs table in the DuckDB
}
\description{
Add multiple new columns to the obs table
}
\examples{
# several columns in one go
sc <- demo_single_cells(prepped = FALSE)
n_cells <- dim(sc)[1]
sc <- set_sc_new_obs_col_multiple(
  sc,
  new_data = list(
    arm = rep(c("ctrl", "treated"), length.out = n_cells),
    donor = rep(sprintf("donor_\%i", 1:4), length.out = n_cells)
  )
)
head(get_sc_obs(sc)[, c("arm", "donor")], 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
