% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{set_snn_graph}
\alias{set_snn_graph}
\alias{set_snn_graph.MetaCells}
\alias{set_snn_graph.ScCache}
\alias{set_snn_graph.SingleCells}
\alias{set_snn_graph.SingleCellsSubset}
\title{Set/add KNN}
\usage{
set_snn_graph(x, snn_graph, ...)

## S7 method for class <bixverse::MetaCells>
set_snn_graph(x, snn_graph, ...)

\method{set_snn_graph}{ScCache}(x, snn_graph, ...)

## S7 method for class <bixverse::SingleCells>
set_snn_graph(x, snn_graph, ...)

## S7 method for class <bixverse::SingleCellsSubset>
set_snn_graph(x, snn_graph, ...)
}
\arguments{
\item{x}{An object to add the KNN data to.}

\item{snn_graph}{Igraph. The sNN graph for subsequent clustering.}

\item{...}{Other parameters.}
}
\description{
Set/add KNN
}
\examples{
# the sNN graph taken out and put back
sc <- demo_single_cells()
snn <- get_snn_graph(sc)
sc <- set_snn_graph(remove_snn_graph(sc), snn)
igraph::vcount(get_snn_graph(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\keyword{internal}
