% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_snn_graph.sc_cache}
\alias{set_snn_graph.sc_cache}
\title{Set the sNN graph for sc_mapper}
\usage{
\method{set_snn_graph}{sc_cache}(x, snn_graph)
}
\arguments{
\item{x}{An \code{sc_cache} object}

\item{snn_graph}{Igraph. The sNN graph for subsequent clustering.}
}
\value{
Updated \code{sc_cache} object with added PCA loadings.
}
\description{
Set the sNN graph for sc_mapper
}
