% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{setnames_sc}
\alias{setnames_sc}
\title{Rename columns in the obs or var table}
\usage{
setnames_sc(object, table = c("obs", "var"), old, new)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} (or potentially other) class.}

\item{table}{String. One of \code{c("obs", "var")}. In which of the tables to
rename the columns.}

\item{old}{Character vector. The old column names.}

\item{new}{Character vector. The new column names.}
}
\value{
Invisible self
}
\description{
Renames the columns in the obs or var table of single cell-related classes.
}
\examples{
# rename a column in the obs table
sc <- demo_single_cells(prepped = FALSE)
sc <- setnames_sc(sc, table = "obs", old = "cell_grp", new = "cell_type")
head(get_sc_obs(sc)$cell_type, 3)

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
