% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{single_cell_duckdb_con}
\alias{single_cell_duckdb_con}
\title{Class for storing single cell experimental data in DuckDB (nightly!)}
\description{
This class wraps up the DB connection and methods to interact with the
observation/cell metadata and the feature/gene metadata.
}
\section{Super class}{
\code{\link[bixverse:single_cell_duckdb_base]{bixverse::single_cell_duckdb_base}} -> \code{single_cell_duckdb_con}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-single_cell_duckdb_con-populate_obs_from_h5}{\code{single_cell_duckdb_con$populate_obs_from_h5()}}
\item \href{#method-single_cell_duckdb_con-populate_vars_from_h5}{\code{single_cell_duckdb_con$populate_vars_from_h5()}}
\item \href{#method-single_cell_duckdb_con-clone}{\code{single_cell_duckdb_con$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="get_obs_table"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-get_obs_table'><code>bixverse::single_cell_duckdb_base$get_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="get_vars_table"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-get_vars_table'><code>bixverse::single_cell_duckdb_base$get_vars_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="initialize"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-initialize'><code>bixverse::single_cell_duckdb_base$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_obs_from_h5"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_obs_from_h5}{}}}
\subsection{Method \code{populate_obs_from_h5()}}{
This function populates the obs table from an h5 file (if found).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_obs_from_h5(h5_path, .verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h5_path}}{String. Path to the h5 file from which to load in the
observation table.}

\item{\code{.verbose}}{Boolean. Controls verbosity of the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns invisible self. As a side effect, it will load in the
obs data from the h5ad file into the DuckDB.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_vars_from_h5"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_vars_from_h5}{}}}
\subsection{Method \code{populate_vars_from_h5()}}{
This populates the vars table from an h5 file (if found.)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_vars_from_h5(h5_path, .verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h5_path}}{String. Path to the h5 file from which to load in the
observation table.}

\item{\code{.verbose}}{Boolean. Controls verbosity of the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns invisible self. As a side effect, it will load in the
obs data from the h5ad file into the DuckDB.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-clone"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
