% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{single_cell_exp}
\alias{single_cell_exp}
\title{bixverse single cell class (deprecated)}
\usage{
single_cell_exp(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which the experimental files will
be stored.}
}
\value{
Returns a \code{\link[=SingleCells]{SingleCells()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=SingleCells]{SingleCells()}}.
}
\keyword{internal}
