% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{SingleCells}
\alias{SingleCells}
\title{bixverse SingleCells class}
\usage{
SingleCells(dir_data)
}
\arguments{
\item{dir_data}{String. This is the directory in which the experimental files
will be}
}
\value{
Returns the \code{SingleCells} class for further operations.
}
\description{
This is the \code{bixverse}-based SingleCells class. Under the hood it uses a
DuckDB for obs and vars storing, and a Rust-based binarised file format to
store the raw and normalised counts. In both cases, the idea is not to hold
any data that is not needed at a given point of time in memory, but leverage
speedy on-disk computations and streaming engines powered by Rust and DuckDB
to run the analysis. This version is specifically designed for single cell
RNAseq. If you want to use multi-modal data, please refer to
\code{\link[=SingleCellsMultiModal]{SingleCellsMultiModal()}} - this class can store multiple layers of 'omics.
}
\section{Properties}{

\describe{
\item{db_connection}{This contains an R6 class with DuckDB pointers and
wrappers to interact with the table-like data for this experiment.}
\item{count_connection}{This contains an R6-like environment that points
to Rust functions that can work on the counts more specifically.}
\item{dir_data}{Path to the directory in which the data will be saved on
disk.}
\item{sc_cache}{Class with cached data. Contains less memory-heavy objects
such as embeddings, kNN information or sNN graphs.}
\item{sc_map}{Class containing various mapping information such as HVG
indices, cells to keep, etc.}
\item{dims}{Dimensions of the original data.}
}
}

\examples{
# demo_single_cells() wraps the construction and the ingestion
sc <- demo_single_cells(prepped = FALSE)
sc

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
