% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{snf}
\alias{snf}
\title{Similarity network fusion (deprecated)}
\usage{
snf(data = NULL, data_name = NULL, snf_params = params_snf())
}
\arguments{
\item{data}{Optional data to transform into adjacency data. Can be a
data.table (categorical/mixed) or a matrix (continuous).}

\item{data_name}{Optional string. Name of the data modality.}

\item{snf_params}{List. The SNF parameters, see \code{\link[=params_snf]{params_snf()}}.}
}
\value{
Returns a \code{\link[=SimilarityNetworkFusion]{SimilarityNetworkFusion()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This constructor has been renamed to \code{\link[=SimilarityNetworkFusion]{SimilarityNetworkFusion()}}.
}
\keyword{internal}
