% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{snf_process_aff_cat_mixed}
\alias{snf_process_aff_cat_mixed}
\title{Helper function to process categorical or mixed data for SNF}
\usage{
snf_process_aff_cat_mixed(data, k, mu)
}
\arguments{
\item{data}{data.table of structure samples x features. The function will
assume that the first column represents the sample identifiers.}

\item{k}{Integer. Number of neighbours to consider.}

\item{mu}{Float. Normalisation factor for the Gaussian kernel width.}
}
\value{
The affinity matrix based on categorical and/or mixed values.
}
\description{
Helper function to process categorical or mixed data for SNF
}
