% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{snf_process_aff_continuous}
\alias{snf_process_aff_continuous}
\title{Helper function to process continuous data for SNF}
\usage{
snf_process_aff_continuous(data, k, mu, distance_metric, normalise)
}
\arguments{
\item{data}{Numerical matrix. Is of type samples x features and is named.}

\item{k}{Integer. Number of neighbours to consider.}

\item{mu}{Float. Normalisation factor for the Gaussian kernel width.}

\item{distance_metric}{String. One of
\code{c("euclidean", "manhattan", "canberra", "cosine")}. Which distance metric
to use for the continuous calculations.}

\item{normalise}{Boolean. Shall continuous values be Z-scored.}
}
\value{
The affinity matrix based on continuous values.
}
\description{
Helper function to process continuous data for SNF
}
