% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sparse_mat_to_list}
\alias{sparse_mat_to_list}
\title{Helper function to transform R sparse matrices to list}
\usage{
sparse_mat_to_list(sparse_mat)
}
\arguments{
\item{sparse_mat}{Sparse matrix. The matrix to transform into a list.}
}
\value{
A list with the following elements
\itemize{
\item indptr - Index pointers of the sparse data.
\item indices - Indices of the data.
\item data - The underlying data.
\item format - String that defines if the data is CSR or CSC.
\item nrow - The number of rows.
\item ncol - The number of columns.
}
}
\description{
Helper function to transform R sparse matrices to list
}
