% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_spatial_spot_db.R
\name{SpatialSpotDuckDB}
\alias{SpatialSpotDuckDB}
\title{Class for storing SpatialSpot experimental data in DuckDB}
\description{
R6 subclass of \code{\link[=SingleCellDuckDB]{SingleCellDuckDB()}} that adds spatial-
specific helpers. The obs table is expected to carry an \code{exp_id}
column identifying which slide/section each spot belongs to.
}
\keyword{internal}
\section{Super classes}{
\code{\link[bixverse:SingleCellDuckDBBase]{SingleCellDuckDBBase}} -> \code{\link[bixverse:SingleCellDuckDB]{SingleCellDuckDB}} -> \code{SpatialSpotDuckDB}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SpatialSpotDuckDB-get_exp_ids}{\code{SpatialSpotDuckDB$get_exp_ids()}}
    \item \href{#method-SpatialSpotDuckDB-get_spot_indices_for_exp}{\code{SpatialSpotDuckDB$get_spot_indices_for_exp()}}
    \item \href{#method-SpatialSpotDuckDB-create_metaspot_assignments_table}{\code{SpatialSpotDuckDB$create_metaspot_assignments_table()}}
    \item \href{#method-SpatialSpotDuckDB-get_metaspot_assignments}{\code{SpatialSpotDuckDB$get_metaspot_assignments()}}
    \item \href{#method-SpatialSpotDuckDB-clone}{\code{SpatialSpotDuckDB$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="add_data_obs"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-add_data_obs'><code>SingleCellDuckDBBase$add_data_obs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="add_data_var"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-add_data_var'><code>SingleCellDuckDBBase$add_data_var()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="drop_columns"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-drop_columns'><code>SingleCellDuckDBBase$drop_columns()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="filter_var_table"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-filter_var_table'><code>SingleCellDuckDBBase$filter_var_table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="get_cells_to_keep"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-get_cells_to_keep'><code>SingleCellDuckDBBase$get_cells_to_keep()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="get_obs_cols"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-get_obs_cols'><code>SingleCellDuckDBBase$get_obs_cols()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="get_obs_index_map"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-get_obs_index_map'><code>SingleCellDuckDBBase$get_obs_index_map()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="get_obs_table"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-get_obs_table'><code>SingleCellDuckDBBase$get_obs_table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="get_var_index_map"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-get_var_index_map'><code>SingleCellDuckDBBase$get_var_index_map()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="get_vars_adt_table"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-get_vars_adt_table'><code>SingleCellDuckDBBase$get_vars_adt_table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="get_vars_table"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-get_vars_table'><code>SingleCellDuckDBBase$get_vars_table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="initialize"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-initialize'><code>SingleCellDuckDBBase$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="join_data_obs"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-join_data_obs'><code>SingleCellDuckDBBase$join_data_obs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="rename_column"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-rename_column'><code>SingleCellDuckDBBase$rename_column()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="set_cells_to_keep"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-set_cells_to_keep'><code>SingleCellDuckDBBase$set_cells_to_keep()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDBBase" data-id="set_to_keep_column"><a href='../../bixverse/html/SingleCellDuckDBBase.html#method-SingleCellDuckDBBase-set_to_keep_column'><code>SingleCellDuckDBBase$set_to_keep_column()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_data.table"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_data.table'><code>SingleCellDuckDB$populate_obs_from_data.table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_h5ad"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_h5ad'><code>SingleCellDuckDB$populate_obs_from_h5ad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_multi_duckdb"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_multi_duckdb'><code>SingleCellDuckDB$populate_obs_from_multi_duckdb()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_multi_h5ad"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_multi_h5ad'><code>SingleCellDuckDB$populate_obs_from_multi_h5ad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_multi_plain_text"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_multi_plain_text'><code>SingleCellDuckDB$populate_obs_from_multi_plain_text()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_multi_tenx_h5"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_multi_tenx_h5'><code>SingleCellDuckDB$populate_obs_from_multi_tenx_h5()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_plain_text"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_plain_text'><code>SingleCellDuckDB$populate_obs_from_plain_text()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_obs_from_tenx_h5"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_obs_from_tenx_h5'><code>SingleCellDuckDB$populate_obs_from_tenx_h5()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_var_adt_from_data.table"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_var_adt_from_data.table'><code>SingleCellDuckDB$populate_var_adt_from_data.table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_var_from_data.table"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_var_from_data.table'><code>SingleCellDuckDB$populate_var_from_data.table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_var_from_plain_text"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_var_from_plain_text'><code>SingleCellDuckDB$populate_var_from_plain_text()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_var_minimal"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_var_minimal'><code>SingleCellDuckDB$populate_var_minimal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_vars_from_duckdb_reordered"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_vars_from_duckdb_reordered'><code>SingleCellDuckDB$populate_vars_from_duckdb_reordered()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_vars_from_h5ad"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_vars_from_h5ad'><code>SingleCellDuckDB$populate_vars_from_h5ad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bixverse" data-topic="SingleCellDuckDB" data-id="populate_vars_from_tenx_h5"><a href='../../bixverse/html/SingleCellDuckDB.html#method-SingleCellDuckDB-populate_vars_from_tenx_h5'><code>SingleCellDuckDB$populate_vars_from_tenx_h5()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialSpotDuckDB-get_exp_ids"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialSpotDuckDB-get_exp_ids}{}}}
\subsection{\code{SpatialSpotDuckDB$get_exp_ids()}}{
  Returns the distinct experiment identifiers from the obs table.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpatialSpotDuckDB$get_exp_ids()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character vector of \code{exp_id} values.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialSpotDuckDB-get_spot_indices_for_exp"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialSpotDuckDB-get_spot_indices_for_exp}{}}}
\subsection{\code{SpatialSpotDuckDB$get_spot_indices_for_exp()}}{
  Returns the spot indices for a given experiment, expressed in
the original (0-based) obs space - i.e. positions in the
original obs table NOT renumbered after filtering. This matches
the convention used by the Rust streaming reader.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpatialSpotDuckDB$get_spot_indices_for_exp(exp_id, filtered = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exp_id}}{String. The experiment identifier.}
      \item{\code{filtered}}{Boolean. Honour the \code{to_keep} filter. Defaults
to \code{TRUE}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer vector of 0-based original obs indices.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialSpotDuckDB-create_metaspot_assignments_table"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialSpotDuckDB-create_metaspot_assignments_table}{}}}
\subsection{\code{SpatialSpotDuckDB$create_metaspot_assignments_table()}}{
  Create the \code{metaspot_assignments} sidecar table that links
metaspots in this aggregated SpatialSpot back to the original
spots in the parent. Composite primary key on
\verb{(metaspot_idx, original_spot_idx)}. Used by
\verb{generate_superspot_*} when emitting a new on-disk SpatialSpot.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpatialSpotDuckDB$create_metaspot_assignments_table(assignments)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{assignments}}{A \code{data.frame}/\code{data.table} with columns
\code{metaspot_idx} (integer), \code{original_spot_idx} (integer) and
\code{exp_id} (character).}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Invisibly self.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialSpotDuckDB-get_metaspot_assignments"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialSpotDuckDB-get_metaspot_assignments}{}}}
\subsection{\code{SpatialSpotDuckDB$get_metaspot_assignments()}}{
  Retrieve metaspot assignment rows. Optional filters on
\code{metaspot_idx}, \code{original_spot_idx} or \code{exp_id}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpatialSpotDuckDB$get_metaspot_assignments(
  metaspot_idx = NULL,
  original_spot_idx = NULL,
  exp_id = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{metaspot_idx}}{Optional integer vector to filter on.}
      \item{\code{original_spot_idx}}{Optional integer vector to filter on.}
      \item{\code{exp_id}}{Optional string to filter on.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{data.table} with \code{metaspot_idx}, \code{original_spot_idx},
\code{exp_id} columns.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialSpotDuckDB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialSpotDuckDB-clone}{}}}
\subsection{\code{SpatialSpotDuckDB$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpatialSpotDuckDB$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
