% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_sc.R
\name{step_clusters}
\alias{step_clusters}
\title{Pipeline step: graph-based clustering}
\usage{
step_clusters(
  cluster_algorithm = c("leiden", "louvain"),
  res = 1,
  name = "leiden_clustering",
  modality = c("rna", "adt", "wnn"),
  seed = 42L
)
}
\arguments{
\item{cluster_algorithm}{String. One of \code{c("leiden", "louvain")}.}

\item{res}{Numeric. The resolution parameter for \code{\link[igraph:cluster_leiden]{igraph::cluster_leiden()}}
or \code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}.}

\item{name}{String. The name to add to the obs table in the DuckDB.}

\item{modality}{String. On which modality to run the UMAP. One of
\code{c("rna", "adt", "wnn")}. The two latter options are only available for
multi-modal versions with the added data.}

\item{seed}{Integer. For reproducibility.}
}
\value{
An \code{ScStep}.
}
\description{
Wraps \code{\link[=find_clusters_sc]{find_clusters_sc()}} as an \code{ScStep}.
}
