% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_sc.R
\name{step_neighbours}
\alias{step_neighbours}
\title{Pipeline step: nearest neighbours}
\usage{
step_neighbours(
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  modality = c("rna", "adt"),
  neighbours_params = params_sc_neighbours(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{embd_to_use}{String. The embedding to use. Whichever you chose, it
needs to be part of the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{modality}{String. One of \code{c("rna", "adt")}. You can only use \code{"adt"}
on \code{SingleCellsMultiModal} class.}

\item{neighbours_params}{List. Output of \code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
A list with the following items:
\itemize{
\item full_snn - Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours.
\item pruning - Numeric. Weights below this threshold will be set to 0 in
the generation of the sNN graph.
\item snn_similarity - String. One of \code{c("rank", "jaccard")}. Defines how
the weight from the SNN graph is calculated. For details, please see
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{seed}{Integer. For reproducibility.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
An \code{ScStep}.
}
\description{
Wraps \code{\link[=find_neighbours_sc]{find_neighbours_sc()}} as an \code{ScStep}.
}
