% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{stream_h5_cellranger}
\alias{stream_h5_cellranger}
\title{Stream in CellRanger h5 files to \code{single_cell_exp}}
\usage{
stream_h5_cellranger(
  object,
  h5_path,
  sc_qc_param = params_sc_min_quality(),
  max_genes_in_memory = 2000L,
  cell_batch_size = 100000L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{h5_path}{File path to the CellRanger h5 file.}

\item{sc_qc_param}{List. Output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}. A
list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of genes to be detected
in the cell to be included.
\item min_lib_size - Integer. Minimum library size in the cell to be
included.
\item min_cells - Integer. Minimum number of cells a gene needs to be
detected to be included.
\item target_size - Float. Target size to normalise to. Defaults to \code{1e5}.
}}

\item{max_genes_in_memory}{Integer. How many genes shall be held in memory
at a given point. Defaults to \code{2000L}.}

\item{cell_batch_size}{Integer. How big are the batch sizes for the cells
in the transformation from the cell-based to gene-based format. Defaults to
\code{100000L}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
It will populate the files on disk and return the class with updated
shape information.
}
\description{
This function loads CellRanger h5 files (v2 or v3 formats) via streaming
with automatic format detection. It loads cell and gene identifiers into
the DuckDB of the \code{single_cell_exp} class and the counts into a
Rust-binarised format for rapid access. During the reading in of the counts,
the log CPM transformation will occur automatically. This function is
specifically designed to deal with larger amounts of data and is slower than
\code{\link[=load_h5_cellranger]{load_h5_cellranger()}}.
}
