% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{synthetic_cPCA_data}
\alias{synthetic_cPCA_data}
\title{Generates synthetic data for cPCA exploration.}
\usage{
synthetic_cPCA_data(seed = 10101L)
}
\arguments{
\item{seed}{Integer. Initial random seed for generation of the synthetic
data. Default: 10101L.}
}
\value{
A list with the following elements:
\itemize{
\item target - The target matrix.
\item background - The background matrix.
\item target_labels - The target labels
}
}
\description{
Generates three elements: target matrix, background matrix and labels for the
target matrix.
}
