% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{testSNFParams}
\alias{testSNFParams}
\title{Test SNF params}
\usage{
testSNFParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
Boolean. \code{TRUE} if the check was successful, otherwise
\code{FALSE}.
}
\description{
Test SNF params
}
\keyword{internal}
