% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{top_genes_perc_sc}
\alias{top_genes_perc_sc}
\title{Calculate the proportions of reads for the Top N genes}
\usage{
top_genes_perc_sc(
  object,
  top_n_vals = c(25L, 50L, 100L),
  streaming = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{top_n_vals}{Integer. The Top N thresholds to test.}

\item{streaming}{Optional Boolean. Shall the data be streamed in. Useful for
larger data sets where you wish to avoid loading in the whole data. If
\code{NULL}, will automatically detect.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
It will add the columns based on the names in the \code{gene_set_list} to
the obs table.
}
\description{
This is a helper function that calculates proportions of reads to the Top N
genes by expression in a given cell. High values here can indicate low
complexity, quality cells. The values will be automatically added to the
obs table.
}
\examples{
# share of a cell's reads taken by its top 5 and top 10 genes
sc <- demo_single_cells(prepped = FALSE)
sc <- top_genes_perc_sc(sc, top_n_vals = c(5L, 10L), .verbose = FALSE)
head(unlist(sc[["top_5_genes_percentage"]]))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
