% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{tree_cut_iter}
\alias{tree_cut_iter}
\title{Coremo: Iterate over k for gene module detection.}
\usage{
tree_cut_iter(
  tree,
  cor_mat,
  dist_mat,
  k_min = 1L,
  k_max = 100L,
  min_size = NULL,
  cor_method = c("spearman", "pearson"),
  seed = 42L
)
}
\arguments{
\item{tree}{hclust object. The hierarchical clustering of the correlation
matrix (or the distance thereof).}

\item{cor_mat}{Numerical matrix. Correlation matrix.}

\item{dist_mat}{Numerical matrix. Distance matrix.}

\item{k_min, k_max}{Integer. The minimum and maximum number of cuts.}

\item{min_size}{Integer. Optional minimum size of resulting modules.}

\item{cor_method}{String. Method for the correlation function. One of
\code{c("pearson", "spearman")}.}

\item{seed}{Integer. For reproducibility purposes.}
}
\value{
a data.table with stats (median size of the clusters, median weighted
R^2, and median R^2) on the varying levels of k.
}
\description{
This function uses a tree (output of \code{\link[stats:hclust]{stats::hclust()}}) and cuts it according
across all values from k_min to k_max and returns the quality at each
individual cut.
}
