% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_triangular_mat.R
\name{upper_triangle_diffcor_mat}
\alias{upper_triangle_diffcor_mat}
\title{Class for symmetric differential correlation matrices}
\description{
The class allows to store the upper triangular matrices of of differential
correlation results in an memory-efficient form and return a data.table or
dense R matrix for a given parameter if need be.
}
\section{Super class}{
\code{\link[bixverse:upper_triangular_cor_mat]{bixverse::upper_triangular_cor_mat}} -> \code{upper_triangular_diffcor_mat}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-upper_triangular_diffcor_mat-new}{\code{upper_triangle_diffcor_mat$new()}}
\item \href{#method-upper_triangular_diffcor_mat-get_data_table}{\code{upper_triangle_diffcor_mat$get_data_table()}}
\item \href{#method-upper_triangular_diffcor_mat-get_cor_matrix}{\code{upper_triangle_diffcor_mat$get_cor_matrix()}}
\item \href{#method-upper_triangular_diffcor_mat-clone}{\code{upper_triangle_diffcor_mat$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="upper_triangular_cor_mat" data-id="get_cor_vector"><a href='../../bixverse/html/upper_triangular_cor_mat.html#method-upper_triangular_cor_mat-get_cor_vector'><code>bixverse::upper_triangular_cor_mat$get_cor_vector()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="upper_triangular_cor_mat" data-id="print"><a href='../../bixverse/html/upper_triangular_cor_mat.html#method-upper_triangular_cor_mat-print'><code>bixverse::upper_triangular_cor_mat$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upper_triangular_diffcor_mat-new"></a>}}
\if{latex}{\out{\hypertarget{method-upper_triangular_diffcor_mat-new}{}}}
\subsection{Method \code{new()}}{
Initialises the R6 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{upper_triangle_diffcor_mat$new(diff_cor_res, features)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{diff_cor_res}}{A list of differential correlation results.}

\item{\code{features}}{String vector. The features of the correlation matrices.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the initialised class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upper_triangular_diffcor_mat-get_data_table"></a>}}
\if{latex}{\out{\hypertarget{method-upper_triangular_diffcor_mat-get_data_table}{}}}
\subsection{Method \code{get_data_table()}}{
Returns the data in form of a data.table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{upper_triangle_diffcor_mat$get_data_table(factor = FALSE, .verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor}}{Boolean. Shall the string columns be transformed into
factors. Reduces size of the object; however, takes longer to generate.}

\item{\code{.verbose}}{Boolean. Controls verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table with three columns:
\itemize{
\item feature_a: The name of the first feature in the correlation matrix.
\item feature_b: The name of the second feature in the correlation
matrix.
\item cor_a: The correlation coefficients for data set a between the
features.
\item cor_b: The correlation coefficients for data set b between the
features.
\item z_score: The differential correlation z-score.
\item p_val: The p-value of the differential correlation.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upper_triangular_diffcor_mat-get_cor_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-upper_triangular_diffcor_mat-get_cor_matrix}{}}}
\subsection{Method \code{get_cor_matrix()}}{
Return the full correlation matrix of either sample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{upper_triangle_diffcor_mat$get_cor_matrix(
  to_ret = c("cor_a", "cor_b"),
  .verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{to_ret}}{String. Option of \verb{("cor_a", "cor_b")}, pending on which
correlation matrix you want to retrieve.}

\item{\code{.verbose}}{Boolean. Controls verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the specified correlation matrix as a dense R matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upper_triangular_diffcor_mat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-upper_triangular_diffcor_mat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{upper_triangle_diffcor_mat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
