% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{upper_triangle_to_sparse}
\alias{upper_triangle_to_sparse}
\title{Transform an upper triangle-stored matrix to a sparse one}
\usage{
upper_triangle_to_sparse(upper_triangle_vals, shift, n, type = c("csc", "csr"))
}
\arguments{
\item{upper_triangle_vals}{Numerical vector. The values of the upper triangle
stored in a row major format.}

\item{shift}{Boolean. Did you exclude the diagonal.}

\item{n}{Integer. Number of columns and rows of the symmetric matrix.}

\item{type}{String. One of \code{c("csc", "csr")}. Which type of of sparse matrix
to return.}
}
\value{
The sparse matrix.
}
\description{
Transform an upper triangle-stored matrix to a sparse one
}
\examples{
# 3 x 3 symmetric matrix from its off-diagonal upper triangle
mat <- upper_triangle_to_sparse(c(0.5, 0.2, 0.8), shift = TRUE, n = 3L)
dim(mat)
}
