% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{vision_w_autocor_sc}
\alias{vision_w_autocor_sc}
\title{Calculate VISION scores (with auto-correlation scores)}
\usage{
vision_w_autocor_sc(
  object,
  gs_list,
  embd_to_use,
  no_embd_to_use = NULL,
  use_knn = TRUE,
  vision_params = params_sc_vision(),
  streaming = NULL,
  random_seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} (or potentially other) class.}

\item{gs_list}{Named nested list. Every element must itself be a list with
at least a \code{"pos"} element holding the gene identifiers, and optionally a
\code{"neg"} one. A bare character vector is not accepted. The gene identifiers
need to be part of the variables of the object.}

\item{embd_to_use}{String. The embedding to use. Whichever you chose, it
needs to be part of the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{use_knn}{Boolean. Shall the internal kNN be used. If set to yes, you
need to ensure consistency.}

\item{vision_params}{List with vision parameters, see
\code{\link[=params_sc_vision]{params_sc_vision()}} with the following elements:
\itemize{
\item n_perm - Integer. Number of random permutations
\item n_cluster - Integer. Number of random clusters to generate to
associate each set with.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{streaming}{Optional Boolean. Shall the data be streamed in. Useful for
larger data sets where you wish to avoid loading in the whole data. If
\code{NULL}, will automatically detect. Ignored when applied to \code{MetaCells}.}

\item{random_seed}{Integer. The random seed.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
A list with the following elements:
\itemize{
\item vision_matrix - Matrix of cells x signatures with the VISION
pathway scores as values.
\item auto_cor_dt - data.table with the auto-correlation results per
gene set, i.e., \code{auto_cor} (1 - Gaery's C), \code{p_val} and \code{fdr}.
}
}
\description{
Calculates VISION-type scores for pathways based on DeTomaso, et al.
Compared to other score types, you can also calculate delta-type scores
between positive and negative gene indices, think epithelial vs mesenchymal
gene signature, etc. Additionally, this function also calculates the auto-
correlation values, answering the question if a given signature shows non-
random enrichment on the kNN graph. The kNN graph (and distance measures)
will be generated on-the-fly based on the embedding you wish to use.
}
\examples{
# scores plus whether they sit non-randomly on the kNN graph
sc <- demo_single_cells()
gs_list <- list(
  programme_a = list(pos = get_gene_names(sc)[1:10]),
  programme_b = list(pos = get_gene_names(sc)[21:30])
)
res <- vision_w_autocor_sc(
  sc,
  gs_list = gs_list,
  embd_to_use = "pca",
  vision_params = params_sc_vision(n_perm = 50L, n_cluster = 3L),
  .verbose = FALSE
)
res$auto_cor_dt

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
\references{
DeTomaso, et al., Nat. Commun., 2019
}
