% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{write_h5ad_sc}
\alias{write_h5ad_sc}
\title{Helper function to write data to h5ad format}
\usage{
write_h5ad_sc(f_path, counts, obs, var, overwrite = TRUE, .verbose = TRUE)
}
\arguments{
\item{f_path}{String. The filepath to which to save the data}

\item{counts}{Sparse matrix. Needs to be of class \code{dgRMatrix} or
\code{dgCMatrix}.}

\item{obs}{data.table. The observations. Needs to have
\code{nrow(obs) == nrow(counts)}.}

\item{var}{data.table. The variable data. Needs to have
\code{ncol(var) == ncol(counts)}.}

\item{overwrite}{Boolean. Shall any found h5ad file be overwritten.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Returns invisible
}
\description{
This is a helper to write synthetic data to h5ad file. This
version will write the data into the common compressed sparse data format.
}
\examples{
# round trip synthetic counts through a sparse h5ad
data <- generate_single_cell_test_data(
  syn_data_params = params_sc_synthetic_data(n_cells = 200L, n_genes = 40L)
)
f_path <- tempfile(fileext = ".h5ad")
write_h5ad_sc(f_path, data$counts, data$obs, data$var, .verbose = FALSE)
get_h5ad_dimensions(f_path)$dims

unlink(f_path)
}
