% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{add_adt_counts_sc}
\alias{add_adt_counts_sc}
\title{Add ADT counts to \code{SingleCellsMultiModal}}
\usage{
add_adt_counts_sc(object, adt_counts)
}
\arguments{
\item{object}{\code{SingleCellsMultiModal} class.}

\item{adt_counts}{Numeric matrix. Of shape samples x features with the ADT
raw counts.}
}
\value{
Returns a \code{SingleCellsMultiModal} with the ADT data added.
}
\description{
This method allows you to add the ADT counts to a \code{SingleCellsMultiModal}.
This assumes you already have ingested transcriptomics data and you know
which cells to keep. The method will subset the ADT counts into these cells
and add a \code{"var_adt"} to the DuckDB and the \code{ADTCounts} to the class, too.
A CLR transformation will be automatically applied to the counts and
normalised counts will be added, too.
}
