% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{add_new_metadata}
\alias{add_new_metadata}
\title{Replace the meta data}
\usage{
add_new_metadata(object, new_metadata, ...)
}
\arguments{
\item{object}{The class}

\item{new_metadata}{data.table. The new meta data you wish to add.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
The object with updated metadata.
}
\description{
This function will replace the meta data within the given object
}
