% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{add_sc_new_obs}
\alias{add_sc_new_obs}
\title{Add an obs table derived from a method to the SingleCells.}
\usage{
add_sc_new_obs(object, obs_data)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{obs_data}{data.table. A data.table you generated with \code{\link[=get_obs_data]{get_obs_data()}}
on some sub class.}
}
\value{
The class with updated obs table in the DuckDB
}
\description{
Add an obs table derived from a method to the SingleCells.
}
