% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{add_snf_data_modality}
\alias{add_snf_data_modality}
\title{Add a data modality for SNF generation}
\usage{
add_snf_data_modality(object, data, data_name, params = NULL)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=snf]{snf()}}.}

\item{data}{matrix or data.table. The data to transform into adjacency data
and add to the class. Any data supplied will be assumed to be samples x
features. The provided data type can be a data.table (for categorical and/or
mixed types) or a matrix (for continous types). If you provide a data.table,
the function will assume the first column are the sample identifiers. Please
ensure that setting.}

\item{data_name}{String. The data modality name.}

\item{params}{Optional List. If you wish to overwite the already set up
parameters for SNF, see \code{\link[=params_snf]{params_snf()}}. If \code{NULL}, the settings
from within the object will be used. If not NULL, the new parameters will
be used for this modality specifically and only for this modality!}
}
\value{
The class with added adjacency matrix for this data.
}
\description{
This function will add a modality as an affinity matrix for subsequent fusion
to the object. To note: all of the modalities need to have the same number
of rows, i.e., samples!
}
