% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{all_limma_contrasts}
\alias{all_limma_contrasts}
\title{Create all limma contrasts}
\usage{
all_limma_contrasts(limma_fit)
}
\arguments{
\item{limma_fit}{The fitted limma model, i.e., output of \code{\link[limma:lmFit]{limma::lmFit()}}.}
}
\value{
The Limma contrasts for further usage.
}
\description{
Create all limma contrasts
}
