% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_h5.R
\name{AnnDataParser}
\alias{AnnDataParser}
\title{Class for Anndata}
\description{
This class helps dealing with h5ad objects from Python AnnData. You have
several options will allow for retrieval of the underlying data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnDataParser-new}{\code{AnnDataParser$new()}}
\item \href{#method-AnnDataParser-get_obs_table}{\code{AnnDataParser$get_obs_table()}}
\item \href{#method-AnnDataParser-get_var_info}{\code{AnnDataParser$get_var_info()}}
\item \href{#method-AnnDataParser-get_raw_counts}{\code{AnnDataParser$get_raw_counts()}}
\item \href{#method-AnnDataParser-get_key_data}{\code{AnnDataParser$get_key_data()}}
\item \href{#method-AnnDataParser-clone}{\code{AnnDataParser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataParser-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataParser-new}{}}}
\subsection{Method \code{new()}}{
Initialises the Anndata Parser.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataParser$new(h5_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h5_path}}{String. Path to the h5 file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the initialised class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataParser-get_obs_table"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataParser-get_obs_table}{}}}
\subsection{Method \code{get_obs_table()}}{
Returns the observation table with all the data from the
h5ad file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataParser$get_obs_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table. The found observations are returned. The pandas index
will be named \code{sample_id}. Remaining columns (if found) will be returned
as factors due to the way the data is stored in h5.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataParser-get_var_info"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataParser-get_var_info}{}}}
\subsection{Method \code{get_var_info()}}{
Returns the variable table with all the data from the
h5ad file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataParser$get_var_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table. The found observations are returned. The pandas index
will be named \code{var_id}. Remaining columns (if found) will be returned as
factors due to the way the data is stored in h5.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataParser-get_raw_counts"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataParser-get_raw_counts}{}}}
\subsection{Method \code{get_raw_counts()}}{
Returns the counts that are stored in \code{X} slot of the
anndata object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataParser$get_raw_counts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the count matrix with samples = columns and rows =
features.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataParser-get_key_data"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataParser-get_key_data}{}}}
\subsection{Method \code{get_key_data()}}{
Wrapper function that returns a list of the stored count
data and the metadata found in the h5ad file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataParser$get_key_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List with following elements:
\itemize{
\item metadata - metadata from the respective h5ad file
\item var_info - metadata on the variables from the respective h5ad
file.
\item counts - counts that were found in the h5ad file.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataParser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataParser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataParser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
