% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertLdaParams}
\alias{assertLdaParams}
\title{Assert LDA parameters}
\usage{
assertLdaParams(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to check/assert}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages.}
}
\value{
Invisibly returns the checked object if the assertion is
successful.
}
\description{
Checkmate extension for asserting the LDA solver parameters,
see \code{\link[=params_lda]{params_lda()}}.
}
\keyword{internal}
