% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertSctypeCellParams}
\alias{assertSctypeCellParams}
\title{Assert per-cell ScType parameters}
\usage{
assertSctypeCellParams(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The list to assert.}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in checkmate.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns \code{x} if the assertion is successful.
}
\description{
Checkmate extension for asserting the per-cell ScType parameters
as returned by \code{\link[=params_sctype_cells]{params_sctype_cells()}}.
}
\keyword{internal}
