% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{assertSpatialSample}
\alias{assertSpatialSample}
\title{Assert that an object is a valid SpatialSample}
\usage{
assertSpatialSample(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is
successful.
}
\description{
Checkmate assertion for the \code{SpatialSample} S3 class built by
\code{\link[=new_spatial_sample]{new_spatial_sample()}}.
}
\keyword{internal}
