% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_analysis.R
\name{aucell_sc}
\alias{aucell_sc}
\title{Calculate AUC scores (akin to AUCell)}
\usage{
aucell_sc(
  object,
  gs_list,
  auc_type = c("wilcox", "auroc"),
  streaming = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{gs_list}{Named list. The elements have the gene identifiers of the
respective gene sets.}

\item{auc_type}{String. Which type of AUC to calculate. Choice of
\code{c("wilcox", "auroc")}.}

\item{streaming}{Boolean. Shall the cell data be streamed in. Useful for
larger data sets.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
data.table with the DGE results from the test.
}
\description{
Calculates an AUC-type score akin to AUCell across the gene sets. You have
the options to calculate the AUC. Two options here: calculate this
with proper AUROC calculations (useful for marker gene expression, use the
\code{"auroc"} version) or based on the Mann-Whitney statistic (useful for pathway
activity measurs, use the \code{"wilcox"}). Data can be streamed in chunks of 50k
cells per or loaded in in one go.
}
\details{
The function will be executed on the cells that are defined by
\code{\link[=set_cells_to_keep]{set_cells_to_keep()}}. If you wish to apply it to a different set
of cells, you will have to change that parameter.
}
