% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{batch_correction_bulk_dge}
\alias{batch_correction_bulk_dge}
\title{Run a linear batch correction}
\usage{
batch_correction_bulk_dge(
  object,
  contrast_column,
  batch_col,
  scale_genes = FALSE,
  no_hvg_genes = 2500L
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_dge]{bulk_dge()}}.}

\item{contrast_column}{String. The contrast column in which the groupings
are stored. Needs to be found in the meta_data within the properties.}

\item{batch_col}{String. The column in which the batch effect groups can
be found.}

\item{scale_genes}{Boolean. Shall the log(cpm) counts be scaled prior the PCA
calculation. Defaults to \code{FALSE}.}

\item{no_hvg_genes}{Integer. Number of highly variable genes to include.
Defaults to 2500.}
}
\value{
Returns the class with additional data added to the outputs.
}
\description{
Runs a linear batch correction over the data regressing out batch effects
and adding \code{normalised_counts_corrected} to the object. Should these counts
be found by \code{\link[=calculate_dge_hedges]{calculate_dge_hedges()}}, they will be used for
calculations of effect sizes based on Hedge's G.
}
