% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bixverse-package.R
\docType{package}
\name{bixverse-package}
\alias{bixverse}
\alias{bixverse-package}
\title{bixverse: Rust-accelerated standard bioinformatics, computational biology workflows}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Your favourite methods used in bioinformatics and computational biology made VERY fast via Rust. Contains various gene set enrichment methods, graph- based methods, co-expression module detection and much more. Additionally, has now a single cell analysis suite, enabling million cell scale analysis on local compute.
}
\section{Options}{

\describe{
\item{\code{bixverse.cache_check}}{Controls how strictly the single cell
classes police stale cached artefacts (PCA, embeddings, kNN, sNN) against
the current cell set. Unset, the default, is two tier: the getters warn,
while \code{\link[=assert_sc_state]{assert_sc_state()}} errors inside the functions that hand cached
indices to Rust. \code{"error"} promotes the getter warnings to errors too,
\code{"warn"} spells out the default getter behaviour, and \code{"none"} disables
both tiers. See \code{\link[=get_sc_cache_status]{get_sc_cache_status()}}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://gregorlueg.github.io/bixverse}
}

}
\author{
\strong{Maintainer}: Gregor Lueg \email{gregorlueg@me.com}

Authors:
\itemize{
  \item Gregor Lueg \email{gregorlueg@me.com}
}

Other contributors:
\itemize{
  \item Liesbeth Francois \email{francois.liesbeth@gmail.com} [contributor]
  \item Andrew Skelton \email{Ajskelton@proton.me} [contributor]
  \item Grant Neilson \email{grant.neilson5@gmail.com} [contributor]
}

}
\keyword{internal}
