% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bixverse-package.R
\docType{package}
\name{bixverse-package}
\alias{bixverse}
\alias{bixverse-package}
\title{bixverse: Rust-accelerated standard bioinformatics, computational biology workflows}
\description{
Your favourite methods used in bioinformatics and computational biology made VERY fast via Rust. Contains various gene set enrichment methods, graph- based methods, co-expression module detection and much more. Additionally, has now single cell functionality, enabling million cell analysis on local compute.
}
\author{
\strong{Maintainer}: Gregor Lueg \email{gregorlueg@me.com}

Other contributors:
\itemize{
  \item Liesbeth Francois \email{francois.liesbeth@gmail.com} [contributor]
  \item Andrew Skelton \email{Ajskelton@proton.me} [contributor]
  \item Grant Neilson \email{grant.neilson5@gmail.com} [contributor]
}

}
\keyword{internal}
