% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_coexp.R
\name{bulk_coexp}
\alias{bulk_coexp}
\title{Bulk RNAseq co-expression modules}
\usage{
bulk_coexp(raw_data, meta_data)
}
\arguments{
\item{raw_data}{The raw count matrix. Rows = samples, columns = features.}

\item{meta_data}{Metadata information on the samples. It expects to have a
column sample_id and case_control column.}
}
\value{
Returns the \code{bulk_coexp} class for further operations.
}
\description{
Class for applying various co-expression module detection methods on top of
bulk RNAseq data.
}
\section{Properties}{

\describe{
\item{raw_data}{A numerical matrix of the provided raw data.}
\item{meta_data}{A data.table with the meta-information about the samples.}
\item{processed_data}{A list in which various types of processed data will
be stored.}
\item{outputs}{A list in which key outputs will be stored.}
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{A data.table that will contain the final results.}
}
}

