% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_classes.R
\name{bulk_coexp}
\alias{bulk_coexp}
\title{Bulk RNAseq co-expression modules}
\usage{
bulk_coexp(raw_data, meta_data)
}
\arguments{
\item{raw_data}{The raw count matrix. Rows = samples, columns = features.}

\item{meta_data}{Metadata information on the samples. It expects to have a
column sample_id and case_control column.}
}
\value{
Returns the \code{bulk_coexp} class for further operations.
}
\description{
Class for applying various co-expression module detection methods on top of
bulk RNAseq data.
}
