% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_coexp_classes.R
\name{bulk_coexp}
\alias{bulk_coexp}
\title{Bulk RNAseq co-expression modules}
\usage{
bulk_coexp(raw_data, meta_data)
}
\arguments{
\item{raw_data}{...}

\item{meta_data}{...}
}
\value{
Returns the \code{bulk_coexp} class for further operations.
}
\description{
...
}
