% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{bulk_dge}
\alias{bulk_dge}
\title{Bulk RNAseq differential gene expression class}
\usage{
bulk_dge(raw_counts, meta_data, variable_info = NULL, .verbose = TRUE, ...)
}
\arguments{
\item{raw_counts}{matrix. The raw count matrix. Rows = genes, columns =
samples. Note: this is different from the \code{\link[=bulk_coexp]{bulk_coexp()}} class!}

\item{meta_data}{data.table. Metadata information on the samples. It expects
to have a column sample_id and case_control column.}

\item{variable_info}{data.table. Metadata information on the features. This
is an optional table.}

\item{.verbose}{Boolean. Controls verbosity of the function.}

\item{...}{Parameters to forward to \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr()}}.}
}
\value{
Returns the \code{bulk_coexp} class for further operations.
}
\description{
Class for coordinating differential gene expression analyses with subsequent
GSE in a structured format. The class will automatically generated a filtered
count matrix in which lowly expressed genes are removed.
}
\section{Properties}{

\describe{
\item{raw_counts}{A numerical matrix of the provided raw data.}
\item{meta_data}{A data.table with the meta-information about the samples.}
\item{filtered_counts}{A numerical matrix with the filtered counts by
minimum expression.}
\item{variable_info}{An optional data.table containing the variable info.}
\item{outputs}{A list in which key outputs will be stored.}
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{A list in which final results will be stored.}
}
}

