% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{bulk_dge_from_h5ad}
\alias{bulk_dge_from_h5ad}
\title{Wrapper function to generate bulk_dge object from h5ad}
\usage{
bulk_dge_from_h5ad(h5_path, .verbose = TRUE)
}
\arguments{
\item{h5_path}{String. Path to the h5ad object.}

\item{.verbose}{Controls verbosity of the function}
}
\value{
\code{bulk_dge} object.
}
\description{
This is a helper function that can be used to create a \code{bulk_dge} object
(see \code{\link[=bulk_dge]{bulk_dge()}}) directly from h5ad objects.
}
