% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mc_processing.R
\name{calc_diffusion_coordinates}
\alias{calc_diffusion_coordinates}
\title{Calculate diffusion coordinates}
\usage{
calc_diffusion_coordinates(
  object,
  knn_data,
  n_dcs = 10L,
  k_density = 150L,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{MetaCells} class.}

\item{knn_data}{\code{SingleCellNearestNeighbour} class. Contains the kNN graph
from the original cells.}

\item{n_dcs}{Integer. Number of diffusion coordinates to use. Defaults to
\code{10L}.}

\item{k_density}{Integer. The k-th neighbour to use for the density region
estimation. Defaults to \code{150L}.}

\item{seed}{Integer. Seed for reproducibility}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
The class with the diffusion map coordinates, density distance and
region attached.
}
\description{
To leverage the quality metrics from Persad, et al., we need the diffusion
coordinates to then calculate if a cell is a dense or sparse region of the
manifold, its compactness and separation to other meta cells. To do so,
generate a diffusion map on the original data based on the approach of
SEACells and add the data to the object, see Persad, et al.
}
\references{
Persad, et al. Nat Biotechnol, 2023
}
