% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{calc_gsea_traditional}
\alias{calc_gsea_traditional}
\title{Bixverse implementation of the traditional GSEA algorithm}
\usage{
calc_gsea_traditional(
  stats,
  pathways,
  nperm = 2000L,
  gsea_params = params_gsea(),
  seed = 123L
)
}
\arguments{
\item{stats}{Named numeric vector. The gene level statistic.}

\item{pathways}{List. A named list with each element containing the genes for
this pathway.}

\item{nperm}{Integer. Number of permutation tests. Defaults to \code{2000L}.}

\item{gsea_params}{List. The GSEA parameters, see \code{\link[=params_gsea]{params_gsea()}}
wrapper function. This function generates a list containing:
\itemize{
\item min_size - Integer. Minimum size for the gene sets.
\item max_size - Integer. Maximum size for the gene sets.
\item gsea_param - Float. The GSEA parameter. Defaults to \code{1.0}.
\item sample_size - Integer. Number of samples to iterate through for the
multi-level implementation of fgsea.
\item eps - Float. Boundary for calculating the p-value. Used for the multi-
level implementation of fgsea.
}}

\item{seed}{Random seed for reproducibility.}
}
\value{
A data.table with the results from the GSEA with the following
columns:
\itemize{
\item es - Float. The enrichment score for this pathway.
\item nes - Float. The normalised enrichment score for this pathway.
\item pvals - Float. The p-value for this pathway.
\item n_more_extreme - Integer. Number of permutation that had more extreme
enrichment scores than the actual.
\item size - Integer. The size of the pathway.
\item pathway_name - Character. The name of the pathway.
\item leading_edge - List of character vectors with the leading edge genes.
\item fdr - Float. The adjusted pval.
}
}
\description{
Rust-based version of the traditional permutation-based GSEA algorithm.
}
