% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{calc_gsva}
\alias{calc_gsva}
\title{Bixverse implementation of GSVA}
\usage{
calc_gsva(
  exp,
  pathways,
  kernel = c("gaussian", "poisson", "none"),
  gaussian = deprecated(),
  gsva_params = params_gsva(),
  .verbose = FALSE
)
}
\arguments{
\item{exp}{Numerical matrix. Rows represents the features, columns the
features/genes.}

\item{pathways}{List. A named list with each element containing the genes for
this pathway.}

\item{kernel}{String. One of \code{c("gaussian", "poisson", "none")}. The kernel
to use.}

\item{gaussian}{Boolean. If set to \code{TRUE} the Gaussian kernel will be used,
if \code{FALSE} the Poisson will be used. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{gsva_params}{List. The GSVA parameters, see \code{\link[=params_gsva]{params_gsva()}}
wrapper function. This function generates a list containing:
\itemize{
\item tau - Float. The tau parameter of the algorithm. Large values will
emphasise the tails more. Defaults to \code{1.0}.
\item min_size - Integer. Minimum size for the gene sets.
\item max_size - Integer. Maximum size for the gene sets.
\item max_diff - Boolean. Influences the scoring. If \code{TRUE} the difference
will be used; if \code{FALSE}, the largest absolute value.
\item abs_rank - Boolean. If \code{TRUE} = pos-neg, \code{FALSE} = pos+neg for the
internal calculations.
}}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
A matrix of shape pathways (that passed the thresholds) x samples.
}
\description{
Implementation of the bixverse version of the gene set variation analysis
(GSVA), see Hänzelmann, et al.
}
\references{
see Hänzelmann, et al. Bmc Bioinformatics, 2013
}
