% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mc_processing.R
\name{calc_manifold_metrics}
\alias{calc_manifold_metrics}
\title{Calculate manifold metrics}
\usage{
calc_manifold_metrics(object)
}
\value{
The class with the compactness and separation scores added.
}
\description{
This function will calculate the compactness and separation of your metacells
on the manifold (defined by the diffusion map). You must have run
\code{\link[=calc_diffusion_coordinates]{calc_diffusion_coordinates()}} before calling this function. The idea is
that compactness indicates how tight the metacell spans the manifold, whereas
separation indicates how well the different metacells span the manifold.
}
\references{
Persad, et al. Nat Biotechnol, 2023
}
