% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mc_processing.R
\name{calc_meta_cell_purity}
\alias{calc_meta_cell_purity}
\title{Calculate meta cell purity}
\usage{
calc_meta_cell_purity(object, original_cell_type)
}
\arguments{
\item{object}{\code{meta_cells} class.}

\item{original_cell_type}{Character vector. The original cell type
annotations. The indices need to match with the original cell indices used
to generate the meta-cells! (1-indexed)}
}
\value{
The \code{meta_cells} with an added columns to the observation table
with the purity measures
}
\description{
A potential metric to see how well the meta cells are aggregated is their
cell type purity. This helper function helps to plot the meta-cell purity
based on annotated cell types. These can be also just unsupervised
memberships to graph-based clustering, etc.
}
