% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{calc_mitch}
\alias{calc_mitch}
\title{Calculate a mitch gene set enrichments on contrast}
\usage{
calc_mitch(contrast_mat, gene_set_list, min_size = 5L)
}
\arguments{
\item{contrast_mat}{Numerical matrix. The rows represent the gene statistic
per contrast and each column represents the contrast.}

\item{gene_set_list}{Named list. Contains the pathways you wish to test
against.}

\item{min_size}{Integer. Minimum size of the gene set to be included.}
}
\value{
A data.table with the Mitch enrichment results.
}
\description{
Rust-based version of the mitch multi-contrast enrichment, see Kaspi and
Ziemann. Takes in a matrix representing the contrasts you wish to test
against.
}
\references{
Kaspi and Ziemann, Bmc Genomics, 2020
}
