% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_annotations_reference.R
\name{calc_sc_type_scores}
\alias{calc_sc_type_scores}
\title{Calculate ScType scores per cell}
\usage{
calc_sc_type_scores(
  object,
  cell_marker_list,
  sensitivity = TRUE,
  weight_floor = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} or \code{SingleCellsMultiModal}}

\item{cell_marker_list}{A list, see \code{\link[=prepare_cell_markers]{prepare_cell_markers()}}.}

\item{sensitivity}{Boolean. Shall shared marker genes be downweighted (like
in the original reference). Defaults to \code{TRUE}.}

\item{weight_floor}{Optional numeric. A value between 0 to 1 and sets the
floor for the weights if \code{sensitivity = TRUE}. If not provided, defaults to
\code{0.1}.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
An \code{ScTypeResults} results class
}
\description{
Implements the approach from
}
