% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{calc_ssgsea}
\alias{calc_ssgsea}
\title{Bixverse implementation of ssGSEA}
\usage{
calc_ssgsea(exp, pathways, ssgsea_params = params_ssgsea(), .verbose = FALSE)
}
\arguments{
\item{exp}{Numerical matrix. Rows represents the features, columns the
features/genes.}

\item{pathways}{List. A named list with each element containing the genes for
this pathway.}

\item{ssgsea_params}{List. The GSVA parameters, see \code{\link[=params_ssgsea]{params_ssgsea()}}
wrapper function. This function generates a list containing:
\itemize{
\item alpha - Float. The exponent defining the weight of the tail in the
random walk performed by ssGSEA.
\item min_size - Integer. Minimum size for the gene sets.
\item max_size - Integer. Maximum size for the gene sets.
\item normalise - Boolean. Shall the scores be normalised.
}}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
A matrix of shape pathways (that passed the thresholds) x samples.
}
\description{
Implementation of the bixverse version of the single sample gene set
enrichment analysis (ssGSEA)
}
