% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{calculate_all_dges}
\alias{calculate_all_dges}
\title{Calculate all possible DGE variants}
\usage{
calculate_all_dges(
  object,
  contrast_column,
  filter_column = NULL,
  gene_filter = NULL,
  co_variates = NULL,
  ...,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_dge]{bulk_dge()}}.}

\item{contrast_column}{String. The contrast column in which the groupings
are stored. Needs to be found in the meta_data within the properties.}

\item{filter_column}{Optional String. If there is a column you wish to use as
sub groupings, this can be provided here. An example could be different
sampled tissues and you wish to run the DGE analyses within each tissue
separately.}

\item{gene_filter}{Optional String vector. You can provide a vector of
strings for genes you wish to filter for, e.g., protein-coding genes.}

\item{co_variates}{Optional String vector. Any co-variates you wish to
consider during the Limma Voom modelling.}

\item{...}{Additional parameters to forward to \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr()}}.}

\item{.verbose}{Controls verbosity of the function.}
}
\value{
Returns the class with additional data added to the outputs.
}
\description{
This class is a wrapper class around applying various differential gene
expression on the data. At a minimum you will need to provide a
\code{contrast_column} that can be found in the meta-data. Every permutation of
groups represented in that column will be tested against each other. Two
variants of differential gene expression analyses will be applied:
\enumerate{
\item A standard Limma Voom approach will be applied. For this specific
approach you can also provide co-variates that will be used in the model
fitting. The function will return the Limma Voom results for every
combination of contrasts found. For more details, please refer to
\code{\link[=run_limma_voom]{run_limma_voom()}}.
\item Secondly, the Hedge's effect size for each gene between all
combinations of the groups will be calculated. The effect sizes can
subsequently be used for e.g., meta-analyses across various studies of
interest.
}
}
