% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{calculate_critical_value}
\alias{calculate_critical_value}
\title{Calculates the critical value}
\usage{
calculate_critical_value(x, alpha, permutations = 100000L, seed = 10101L)
}
\arguments{
\item{x}{Numerical matrix or \verb{ontology class}, see \code{\link[=ontology]{ontology()}}.}

\item{alpha}{Float. The alpha value. For example, 0.001 would mean that the
critical value is smaller than 0.1 percentile of the random permutations.}

\item{permutations}{Number of random permutations.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
The critical value.
}
\description{
This function calculates the critical value for a given ontology
similarity matrix
}
