% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{calculate_dge_limma}
\alias{calculate_dge_limma}
\title{Calculates the Limma Voom DGE}
\usage{
calculate_dge_limma(
  object,
  contrast_column,
  contrast_list = NULL,
  filter_column = NULL,
  co_variates = NULL,
  limma_params = params_limma_voom(),
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=BulkDge]{BulkDge()}}.}

\item{contrast_column}{String. The contrast column in which the groupings
are stored. Needs to be found in the meta_data within the properties.}

\item{contrast_list}{Optional string vector. A vectors that contains the
contrast formatted as \code{"contrast1-contrast2"}. Default \code{NULL} will create
all possible contrast automatically.}

\item{filter_column}{Optional String. If there is a column you wish to use as
sub groupings, this can be provided here. An example could be different
sampled tissues and you wish to run the DGE analyses within each tissue
separately in the data.}

\item{co_variates}{Optional string vector. Any co-variates you wish to
consider during the Limma Voom modelling.}

\item{limma_params}{List. The limma parameters, see
\code{\link[=params_limma_voom]{params_limma_voom()}}.}

\item{.verbose}{Controls verbosity of the function.}
}
\value{
Returns the class with additional data added to the outputs.
}
\description{
This function will apply the Limma Voom DGE workflow. At a minimum you will
need to provide \code{contrast_column} that can be found in the meta-data. If you
do not provide a vector of contrasts that you wish to test for, every
permutation of groups represented in that column will be tested against each
other.
}
\examples{
# limma voom over every contrast in the case_control column
syn <- synthetic_bulk_cor_matrix()
meta <- data.table::data.table(
  sample_id = colnames(syn$counts),
  case_control = rep(c("case", "control"), each = 50)
)
object <- BulkDge(raw_counts = syn$counts, meta_data = meta)
object <- qc_bulk_dge(object, group_col = "case_control", .verbose = FALSE)
object <- normalise_bulk_dge(
  object,
  group_col = "case_control",
  .verbose = FALSE
)
object <- calculate_dge_limma(
  object,
  contrast_column = "case_control",
  .verbose = FALSE
)
head(get_dge_limma_voom(object))
}
