% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_graph_classes_methods.R
\name{calculate_diffusion_auc}
\alias{calculate_diffusion_auc}
\title{Calculate the AUROC for a diffusion score}
\usage{
calculate_diffusion_auc(network_diffusions, ...)

calculate_diffusion_auc(
 network_diffusions,
 hit_nodes,
 auc_iters = 10000L,
 random_aucs = 1000L,
 permutation_test = FALSE,
 seed = 42L
)
}
\arguments{
\item{network_diffusions}{The underlying \code{network_diffusions} class.}

\item{hit_nodes}{Which nodes would consist a hit to calculate the AUROC
against.}

\item{auc_iters}{How many iterations to run to approximate the AUROC.}

\item{random_aucs}{How many random AUROCs to calculate to estimate the
Z-score}

\item{seed}{Seed for reproducibility purposes.}
}
\value{
List with AUC and Z-score as the two named elements if permutations
test set to TRUE; otherwise just the AUC.
}
\description{
This is the generic function for calculating AUC/AUROCs for a set of genes
based on the diffusion vector of either single or tied diffusion.

This functions can take a given \code{network_diffusions} class and calculates an
AUC and generates a Z-score based on random permutation of \code{random_aucs} for
test for statistical significance if desired.
}
