% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{calculate_effect_size}
\alias{calculate_effect_size}
\title{Calculate the Hedge G effect between two matrices}
\usage{
calculate_effect_size(
  mat_a,
  mat_b,
  small_sample_correction = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{mat_a}{Numerical matrix. Contains the values for group a. Assumes that
rows = samples, and columns = features.}

\item{mat_b}{Numerical matrix. Contains the values for group b.}

\item{small_sample_correction}{Can be NULL (automatic determination if a
small sample size correction should be applied) or Boolean.}

\item{.verbose}{Boolean that controls verbosity of the function.}
}
\value{
x, robustly scaled.
}
\description{
This function takes two matrices in and calculate on a per column basis the
Hedge's G effect size and the standard error. These results can be
subsequently used for meta-analyses or other approaches.
}
