% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_batch_corr.R
\name{calculate_kbet_sc}
\alias{calculate_kbet_sc}
\title{Calculate kBET scores}
\usage{
calculate_kbet_sc(object, batch_column, threshold = 0.05, .verbose = TRUE)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{batch_column}{String. The column with the batch information in the
obs data of the class.}

\item{threshold}{Numeric. Number between 0 and 1. Below this threshold, the
test is considered significant. Defaults to \code{0.05}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
A list with the following elements
\itemize{
\item kbet_score - Number of significant tests over all cells. 0 indicates
perfect mixing, 1 indicates basically zero mixing between batches.
\item significant_tests - Boolean indicating for which cells the statistic
was below the threshold
\item chisquare_pvals - The p-values of the ChiSquare test.
}
}
\description{
This function calculates the k-nearest neighbour batch-effect test (kBET).
Briefly, the function leverages a Chi-Square statistic to calculate the
differences in batch proportions observed in the neighbourhood of a given
cell with the overall batch proportions. If the test is significant for that
cell it indicates poor mixing for that cell specifically.
Large number of positive tests indicate bad mixing overall. For more details,
please see Büttner et al.
}
\details{
Internally, the kBET scores will be calculated against the kNN graph that is
stored in the object. Should you be interested in calculating the kBET with
a different set of cells, you will have to define these via
\code{\link[=set_cells_to_keep]{set_cells_to_keep()}} and rerun the kNN graph generation, see
\code{\link[=find_neighbours_sc]{find_neighbours_sc()}}.
}
\references{
Büttner, et al., Nat. Methods, 2019
}
