% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_multimodal.R
\name{calculate_pca_adt_sc}
\alias{calculate_pca_adt_sc}
\title{Calculate the PCA on top of the normalised ADT counts}
\usage{
calculate_pca_adt_sc(
  object,
  no_pcs,
  features = NULL,
  randomised_svd = FALSE,
  seed = 42L
)
}
\arguments{
\item{object}{\code{SingleCellsMultiModal} class with ADT counts added.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{features}{Optional string vector. If you want to subset to a specific
set of ADT probes (for example to exclude isotypes).}

\item{randomised_svd}{Boolean. Shall randomised SVD be used. Faster, but
less precise.}

\item{seed}{Integer. Controls reproducibility. Only relevant if
\code{randomised_svd = TRUE}.}
}
\value{
The function will add the PCA factors, loadings and singular values
for the ADT data to the object.
}
\description{
This function will run PCA - via (randomised) SVD - on the normalised counts
and add the PCA results to the ScCache for the ADT counts.
}
