% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{calculate_pca_bulk_dge}
\alias{calculate_pca_bulk_dge}
\title{Calculate PCA on the expression.}
\usage{
calculate_pca_bulk_dge(
  object,
  scale = FALSE,
  pcs = 10L,
  hvg_genes = 2500L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_dge]{bulk_dge()}}.}

\item{scale}{Boolean. Shall the log(cpm) counts be scaled prior the PCA
calculation. Defaults to \code{FALSE}.}

\item{pcs}{Integer. Number of PCs to return and add to the outputs slot.}

\item{hvg_genes}{Integer. Number of highly variable genes to include.}

\item{.verbose}{Boolean.}
}
\value{
Returns the class with additional data added to the outputs.
}
\description{
Calculates the principal component on top of the filtered count matrix and
adds the information of the first 10 principal components to the outputs.
}
