% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{calculate_pca_sc}
\alias{calculate_pca_sc}
\title{Run PCA for single cell}
\usage{
calculate_pca_sc(
  object,
  no_pcs,
  pca_params = params_sc_pca(),
  sparse_svd = FALSE,
  hvg = NULL,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} (or potentially other) class.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{pca_params}{Named list. Controls the parameters to be used for the
PCA calculation which is single cell-specific, see \code{\link[=params_sc_pca]{params_sc_pca()}}}

\item{sparse_svd}{Boolean. Shall sparse solvers be used that do not do
scaling. If set to yes, in the case of \code{random_svd = FALSE}, Lanczos
iterations are used to solve the sparse SVD. With \code{random_svd = TRUE}, the
sparse initial matrix is multiplied with the random matrix, yielding a
much smaller dense matrix that does not increase the memory pressure
massively. Not used for \code{MetaCells}.}

\item{hvg}{Optional integer. If you want to provide your own HVG genes.
Otherwise, the function will default to what is found in
\code{\link[=get_hvg]{get_hvg()}}. Please provide 1-indexed genes here! If you provide
these, the internal HVG will be overwritten.}

\item{seed}{Integer. Controls reproducibility. Only relevant if
\code{randomised_svd = TRUE}.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
The function will add the PCA factors, loadings and singular values
to the object cache in memory.
}
\description{
This function will run PCA on the detected highly variable genes. You can
use randomised SVD for speed and there is an option for sparse SVD for very
large data sets to avoid memory pressure.
}
\examples{
# PCA on the highly variable genes
sc <- demo_single_cells(prepped = FALSE)
sc <- find_hvg_sc(sc, hvg_no = 30L, .verbose = FALSE)
sc <- calculate_pca_sc(sc, no_pcs = 10L, .verbose = FALSE)
dim(get_pca_factors(sc))

unlink(sc@dir_data, recursive = TRUE, force = TRUE)
}
