% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{calculate_pca_sc}
\alias{calculate_pca_sc}
\title{Run PCA for single cell}
\usage{
calculate_pca_sc(
  object,
  no_pcs,
  randomised_svd = TRUE,
  sparse_svd = FALSE,
  hvg = NULL,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{randomised_svd}{Boolean. Shall randomised SVD be used. Faster, but
less precise.}

\item{sparse_svd}{Boolean. Shall sparse solvers be used that do not do
scaling. If set to yes, in the case of \code{random_svd = FALSE}, Lanczos
iterations are used to solve the sparse SVD. With \code{random_svd = TRUE}, the
sparse initial matrix is multiplied with the random matrix, yielding a
much smaller dense matrix that does not increase the memory pressure
massively.}

\item{hvg}{Optional integer. If you want to provide your own HVG genes.
Otherwise, the function will default to what is found in
\code{\link[=get_hvg]{get_hvg()}}. Please provide 1-indexed genes here! If you provide
these, the internal HVG will be overwritten.}

\item{seed}{Integer. Controls reproducibility. Only relevant if
\code{randomised_svd = TRUE}.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
The function will add the PCA factors, loadings and singular values
to the object cache in memory.
}
\description{
This function will run PCA (option of full SVD and randomised SVD for now)
on the detected highly variable genes.
}
