% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{calculate_pca_sc}
\alias{calculate_pca_sc}
\title{Run PCA for single cell}
\usage{
calculate_pca_sc(
  object,
  no_pcs,
  randomised_svd = TRUE,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{randomised_svd}{Boolean. Shall randomised SVD be used. Faster, but
less precise.}

\item{seed}{Integer. Controls reproducibility. Only relevant if
\code{randomised_svd = TRUE}.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
The function will add the PCA factors, loadings and singular values
to the object cache in memory.
}
\description{
This function will run PCA (option of full SVD and randomised SVD for now)
on the detected highly variable genes.
}
\details{
The function will be executed on the cells that are defined by
\code{\link[=set_cells_to_keep]{set_cells_to_keep()}}. If you have not set this parameter, all
cells in the object will be initially used. The information which cells
were used for the calculation of the PCAs will be additionally stored.
}
