% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell_processing.R
\name{calculate_pca_sc}
\alias{calculate_pca_sc}
\title{Run PCA for single cell}
\usage{
calculate_pca_sc(
  object,
  no_pcs,
  randomised_svd = TRUE,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{randomised_svd}{Boolean. Shall randomised SVD be used. Faster, but
less precise.}

\item{seed}{Integer. Controls reproducibility. Only relevant if
\code{randomised_svd = TRUE}.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
The function will add the PCA factors and loadings to the object
cache in memory.
}
\description{
This function will run PCA (option of full SVD and randomised SVD for now)
on the detected highly variable genes.
}
