% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{calculate_rpkm}
\alias{calculate_rpkm}
\title{RPKM calculation}
\usage{
calculate_rpkm(counts, gene_lengths)
}
\arguments{
\item{counts}{Numeric matrix. Count matrix (gene x sample)}

\item{gene_lengths}{Named vector. Named vector with gene lengths.}
}
\value{
RPKM-normalised matrix.
}
\description{
RPKM calculation
}
