% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_methods.R
\name{calculate_semantic_sim}
\alias{calculate_semantic_sim}
\title{Calculate the Resnik and Lin semantic similarity}
\usage{
calculate_semantic_sim(terms, ancestor_list, ic_list, max_ic)
}
\arguments{
\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}

\item{ancestor_list}{List. Names being the term and the elements in the
list the names of the ancestors.}

\item{ic_list}{List. The names being the term and the elements the
information content of this given term. Needs to be a single float!}

\item{max_ic}{Double. The maximum information content observed in the data.
This value will be used to calculate the normalised Resnik similarity.}
}
\description{
This function calculates the semantic similarities based on
Resnik and Lin similarity for a given ontology.
}
