% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ontology.R
\name{calculate_semantic_sim_mat}
\alias{calculate_semantic_sim_mat}
\title{Calculate the Resnik or Lin semantic similarity matrix}
\usage{
calculate_semantic_sim_mat(similarity_type, ancestor_list, ic_list)
}
\arguments{
\item{similarity_type}{String. One of \code{c("resnik", "lin", "combined")}. The
Resnik similarity is automatically rescaled between 0 and 1 (dividing the
values by maximum information content observed).}

\item{ancestor_list}{List. Names being the term and the elements in the
list the names of the ancestors, see \code{\link[=get_ontology_ancestry]{get_ontology_ancestry()}}.}

\item{ic_list}{List. The names being the term and the elements the
information content of this given term. Needs to be a single float! See
\code{\link[=calculate_information_content]{calculate_information_content()}}.}
}
\value{
The symmetric similarity matrix for the specified data from the
ontology.
}
\description{
This function calculates the semantic similarities based on
Resnik or Lin similarity for a given ontology. Has also the option to
calculate a combined version of the two. Returns the full matrix of the data.
}
