% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_similarity.R
\name{calculate_semantic_sim_onto}
\alias{calculate_semantic_sim_onto}
\title{Calculate the Resnik and Lin semantic similarity for an ontology.}
\usage{
calculate_semantic_sim_onto(object, terms = NULL, .verbose = TRUE)
}
\arguments{
\item{object}{\verb{ontology class}. See \code{\link[=ontology]{ontology()}}.}

\item{terms}{Character vector or NULL. The function will calculate the
the semantic similarities for these terms. If NULL, the semantic similarity
for everything will be calculated.}

\item{.verbose}{}
}
\value{
The class with added semantic similarities to the properties.
}
\description{
This function calculates the semantic similarities based on
Resnik and Lin similarity for a given ontology. In this case, the ontology
is stored within an \code{\link[=ontology]{ontology()}} class.
}
