% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{calculate_semantic_sim_onto}
\alias{calculate_semantic_sim_onto}
\title{Calculate the Resnik or Lin semantic similarity for an ontology.}
\usage{
calculate_semantic_sim_onto(
  object,
  sim_type = c("resnik", "lin", "combined"),
  .verbose = TRUE
)
}
\arguments{
\item{object}{\verb{ontology class}. See \code{\link[=ontology]{ontology()}}.}

\item{sim_type}{String. One of \code{c("resnik", "lin", "combined")}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
The class with added semantic similarities to the properties.
}
\description{
This function calculates the specified semantic similarities for
the whole ontology and adds it to the class.
}
