% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{calculate_tpm}
\alias{calculate_tpm}
\title{TPM calculation}
\usage{
calculate_tpm(counts, gene_lengths)
}
\arguments{
\item{counts}{Numeric matrix. Count matrix (gene x sample)}

\item{gene_lengths}{Named vector. Named vector with gene lengths.}
}
\value{
TPM-normalised matrix.
}
\description{
TPM calculation
}
