% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{calculate_wang_sim}
\alias{calculate_wang_sim}
\title{Calculate the Wang similarity matrix}
\usage{
calculate_wang_sim(parent_child_dt, w = 0.8)
}
\arguments{
\item{parent_child_dt}{data.table. The data.table with column parent and
child.}

\item{w}{Float. The value assigned to the respective edge. Defaults to \code{0.8}.}
}
\value{
The symmetric Wang similarity matrix.
}
\description{
This function calculates the Wang similarity, based on the DAG
for a given ontology. The current implementation just allows for a single w
for the relationships.
}
