% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_add.R
\name{call_doublets_manual}
\alias{call_doublets_manual}
\title{Helper function to manually readjust Scrublet thresholds}
\usage{
call_doublets_manual(scrublet_res, threshold, .verbose = TRUE)
}
\arguments{
\item{scrublet_res}{\code{scrublet_res} result class.}

\item{threshold}{Numeric. The new threshold to use.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
\code{scrublet_res} class with updated doublet calls based on the new
threshold.
}
\description{
Can update the Scrublet thresholding after manual introspection of the
histogram.
}
