% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{change_gene_identifier}
\alias{change_gene_identifier}
\title{Change the primary gene identifier of bulk_dge}
\usage{
change_gene_identifier(object, alternative_gene_id, variable_info = NULL)
}
\description{
Changes the primary gene identifier in the bulk_dge class. To do so, you need
to either provide a \code{variable_info} data.table with the alternative gene
identifier you wish to use or it exists already in the object itself. If it
exists in the object, that variable_info will be used.
}
