% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkBlitzGseaParams}
\alias{checkBlitzGseaParams}
\title{Check blitzGSEA parameters}
\usage{
checkBlitzGseaParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the blitzGSEA parameters.
}
\keyword{internal}
