% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkBulkSparsityParams}
\alias{checkBulkSparsityParams}
\title{Check bulk sparsification parameters}
\usage{
checkBulkSparsityParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the bulk sparsification
parameters, see \code{\link[=params_bulk_sparsity]{params_bulk_sparsity()}}.
}
\keyword{internal}
