% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkCellsExist}
\alias{checkCellsExist}
\title{Check that the cell name exists in the object}
\usage{
checkCellsExist(x, cell_names)
}
\arguments{
\item{x}{The \code{single_cell_exp} object to check/assert.}

\item{cell_names}{String. The provided cell names.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking if the prodivided cell names
exist in the object.
}
