% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkCommunityParams}
\alias{checkCommunityParams}
\title{Check community detection parameters}
\usage{
checkCommunityParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the community detection
parameters for identifying genetically privileged communities.
}
