% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkEdgeRQlParams}
\alias{checkEdgeRQlParams}
\title{Check edgeR quasi-likelihood parameters}
\usage{
checkEdgeRQlParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the edgeR quasi-likelihood
parameters, see \code{\link[=params_edger_ql]{params_edger_ql()}}.
}
\keyword{internal}
