% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkGraphResParams}
\alias{checkGraphResParams}
\title{Check resolution graph parameters}
\usage{
checkGraphResParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the resolution parameters for
community detection with Leiden.
}
