% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkGSVAParams}
\alias{checkGSVAParams}
\title{Check GSVA parameters}
\usage{
checkGSVAParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the gene set variation analysis
(GSVA) parameters.
}
