% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkKMeansParams}
\alias{checkKMeansParams}
\title{Check k-means method parameters}
\usage{
checkKMeansParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the run parameters of the
k-means clustering methods.
}
\keyword{internal}
