% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkKnnParams}
\alias{checkKnnParams}
\title{Check kNN parameters}
\usage{
checkKnnParams(x, required_params = NULL)
}
\arguments{
\item{x}{The list to check/assert}

\item{required_params}{Character vector of required kNN parameter names}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking kNN parameters.
}
\keyword{internal}
