% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkLdaParams}
\alias{checkLdaParams}
\title{Check LDA parameters}
\usage{
checkLdaParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the LDA solver parameters,
see \code{\link[=params_lda]{params_lda()}}.
}
\keyword{internal}
