% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkLimmaVoomParams}
\alias{checkLimmaVoomParams}
\title{Check limma-voom parameters}
\usage{
checkLimmaVoomParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the limma-voom parameters,
see \code{\link[=params_limma_voom]{params_limma_voom()}}.
}
\keyword{internal}
