% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkModuleMembershipParams}
\alias{checkModuleMembershipParams}
\title{Check module membership parameters}
\usage{
checkModuleMembershipParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the module membership
parameters, see \code{\link[=params_module_membership]{params_module_membership()}}.
}
\keyword{internal}
