% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkNebulaParams}
\alias{checkNebulaParams}
\title{Check NEBULA parameters}
\usage{
checkNebulaParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the NEBULA parameters, see
\code{\link[=params_nebula]{params_nebula()}}.
}
\keyword{internal}
