% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkScGeneTrendParams}
\alias{checkScGeneTrendParams}
\title{Check gene trend parameters}
\usage{
checkScGeneTrendParams(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the gene trend parameters.
}
\keyword{internal}
