% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSctypeCellParams}
\alias{checkSctypeCellParams}
\title{Check per-cell ScType parameters}
\usage{
checkSctypeCellParams(x)
}
\arguments{
\item{x}{The list to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the per-cell ScType parameters
as returned by \code{\link[=params_sctype_cells]{params_sctype_cells()}}.
}
\keyword{internal}
