% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSingleSampleGSEAparams}
\alias{checkSingleSampleGSEAparams}
\title{Check ssGSEA parameters}
\usage{
checkSingleSampleGSEAparams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking single sample gene set
enrichment analysis parameters.
}
