% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_spatial_sample.R
\name{checkSpatialSample}
\alias{checkSpatialSample}
\title{Check that an object is a valid SpatialSample}
\usage{
checkSpatialSample(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message
string.
}
\description{
Checkmate extension for \code{\link[=SpatialSample]{SpatialSample()}}.
}
\keyword{internal}
