% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{checkSyntheticBulkParams}
\alias{checkSyntheticBulkParams}
\title{Check synthetic bulk RNAseq parameters}
\usage{
checkSyntheticBulkParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the synthetic bulk RNAseq
parameters, see \code{\link[=params_synthetic_bulk_rnaseq]{params_synthetic_bulk_rnaseq()}}.
}
\keyword{internal}
