% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_diff_class_methods.R
\name{community_detection}
\alias{community_detection}
\title{Identify privileged communities based on a given diffusion vector}
\usage{
community_detection(network_diffusions, ...)
}
\description{
This is the generic function for diffusing seed genes over a network.

...
}
