% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{community_detection}
\alias{community_detection}
\title{Identify privileged communities based on a given diffusion vector}
\usage{
community_detection(
  object,
  community_params = params_community_detection(),
  seed = 42L,
  .verbose = FALSE,
  .max_iters = 100L
)
}
\arguments{
\item{object}{The underlying class \code{\link[=network_diffusions]{network_diffusions()}}.}

\item{community_params}{List. Parameters for the community detection within
the reduced network, see \code{\link[=params_community_detection]{params_community_detection()}}. A list
with the following items:
\itemize{
\item max_nodes - Integer. Number of maximum nodes per community. Larger
communities will be recursively subclustered.
\item min_nodes - Integer. Minimum number of nodes per community.
\item min_seed_nodes - Integer. Minimum number of seed genes that have to
be found in a given community.
\item initial_res - Float. Initial resolution parameter for the Leiden
clustering.
\item threshold_type - String. One of \code{c("prop_based", "pval_based")}.
You can chose to include a certain proportion of the network (like in the
original paper) with the highest diffusion scores, or use p-values based
on permutations. Defaults to \code{"prop_based"}.
\item network_threshold - Float. The proportion of the network to
include. Used if \code{threshold_type = "prop_based"}.
\item pval_threshold - Float. The maximum p-value for nodes to be included.
Used if \code{threshold_type = "pval_based"}.
}}

\item{seed}{Random seed.}

\item{.verbose}{Controls the verbosity of the function.}

\item{.max_iters}{Controls how many iterations shall be tried for the
sub-clustering. To note, in each iteration of the sub-clustering, the
resolution parameter is increased by 0.05, to identify more granular
communities within the sub communities.}
}
\value{
The class with added diffusion community detection results (if any
could be identified with the provided parameters).
}
\description{
Detects privileged communities after a diffusion based on seed
nodes.
}
