% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cPCA.R
\name{contrastive_pca_processing}
\alias{contrastive_pca_processing}
\title{Prepare class for contrastive PCA}
\usage{
contrastive_pca_processing(
  object,
  background_matrix,
  scale = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{background_matrix}{Numeric matrix. The background matrix you wish to
remove. You should apply any data transformation to this matrix, too!}

\item{scale}{Boolean. Shall the data be scaled. Defaults to FALSE.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} with the needed data for contrastive PCA in the
properties of the class.
}
\description{
This function will prepare the \code{bulk_coexp} for subsequent usage of the
contrastive PCA functions. This is based on the work of Abid, et al.
}
\references{
Abid, et al., Nature Communications, 2018
}
