% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_methods_cPCA.R
\name{contrastive_pca_processing}
\alias{contrastive_pca_processing}
\title{Prepare class for contrastive PCA}
\usage{
contrastive_pca_processing(bulk_coexp, ...)

contrastive_pca_processing(
 bulk_coexp = bulk_coexp,
 background_matrix = background_matrix,
 scale = FALSE
 .verbose = TRUE
)
}
\arguments{
\item{bulk_coexp}{\code{bulk_coexp} class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{background_matrix}{Numeric matrix.}

\item{scale}{Boolean. Shall the data be scaled. Defaults to FALSE.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} with additional data in the slots
}
\description{
This is the generic function for doing the preprocessing for contrastive PCA.

This function will prepare the \code{bulk_coexp} for subsequent usage of the
contrastive PCA function. This is based on the work of Abid, et al.
}
\references{
Abid, et al., Nature Communications, 2018
}
