% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_check_epsilon}
\alias{cor_module_check_epsilon}
\title{Iterate through different epsilon parameters}
\usage{
cor_module_check_epsilon(
  object,
  epsilons = c(0.25, seq(from = 0.5, to = 5, by = 0.5)),
  .verbose = TRUE
)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. You need to run
\code{\link[=diffcor_module_processing]{diffcor_module_processing()}} before running this function.}

\item{epsilons}{Vector of floats. The different epsilon parameters you
would like to run.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added data to the properties for subsequent usage.
}
\description{
This functions iterates through a set of provided epsilons and checks for
each one to what extend the underlying affinity matrix will follow a power
law distribution.
}
