% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_coremo_clustering}
\alias{cor_module_coremo_clustering}
\title{Generates CoReMo-based gene modules.}
\usage{
cor_module_coremo_clustering(
  object,
  epsilon = 2,
  params_coremo = params_coremo(),
  seed = 42L,
  .verbose = FALSE
)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{epsilon}{...}

\item{params_coremo}{...}

\item{seed}{...}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added data to the properties for subsequent usage.
}
\description{
...
}
