% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{cor_module_coremo_clustering}
\alias{cor_module_coremo_clustering}
\title{Generates CoReMo-based gene modules}
\usage{
cor_module_coremo_clustering(
  object,
  coremo_params = params_coremo(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{coremo_params}{List. Parameters for the generation of the CoReMo
modules, see \code{\link[=params_coremo]{params_coremo()}}. Contains:
\itemize{
\item epsilon - Float. The epsilon parameter for the RBF. You can optimise
that one with \code{\link[=cor_module_check_epsilon]{cor_module_check_epsilon()}}. Defaults to \code{2}.
\item k_min - Integer. Minimum number of cuts. Defaults to \code{2L}.
\item k_max - Integer. Maximum number of cuts. Defaults to \code{150L}.
\item min_size - Optional integer. Minimum size of the clusters. If
provided, smaller clusters will be merged by eigengene similarity.
\item junk_module_threshold - Float. Minimum R2 median  value for a module
to not be considered a junk module. Defaults to \code{0.05}.
\item rbf_func - String. Type of RBF function to apply. Defaults to
\code{"gaussian"}.
\item cor_method - String. Type of correlation method to use for merging
the smaller cluster. Defaults to \code{"spearman"}.
}}

\item{seed}{Integer. Random seed for reproducibility purposes.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added data to the properties for subsequent usage.
}
\description{
This function creates gene modules, based on the framework from Srivastava
et al., 2018. Briefly, it applies an RBF function to the correlation matrix
to reduce the impact of weak correlations, leverages hierarchical clustering
for clustering the genes. It optimises the R2 (cor^2) within the clusters to
identify the optimal cut. Gene modules with low R2 are being considered as
the 'junk module'.
}
\references{
Srivastava, et al., Nat. Commun., 2018
}
