% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{cor_module_processing}
\alias{cor_module_processing}
\title{Prepare correlation-based module detection}
\usage{
cor_module_processing(bulk_coexp, ...)

...
}
\arguments{
\item{correlation_method}{String. Option of \code{c("pearson", "spearman")}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} with the needed data for subsequent identification of
correlation-based co-expression modules.
}
\description{
This is the generic function for running a single correlation across your
genes.

This function will generate necessarily data for correlation-based detection
of gene modules. In this case, this will be based on simple correlation
between the genes and the relevant data will be added to the \code{bulk_coexp}
class.
}
