% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{cor_module_processing}
\alias{cor_module_processing}
\title{Prepare correlation-based module detection}
\usage{
cor_module_processing(
  bulk_coexp,
  correlation_method = c("pearson", "spearman"),
  .verbose = TRUE
)
}
\arguments{
\item{correlation_method}{String. Option of \code{c("pearson", "spearman")}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. Ideally, you
should run \code{\link[=preprocess_bulk_coexp]{preprocess_bulk_coexp()}} before applying this class.

The class with added data to the properties for subsequent functions.
}
\description{
This function will calculate the correlation coefficients between the genes,
using the highly variable genes (if available, otherwise the function will
use the raw data). The data will be stored in a memory-efficient format as a
vector within the class.
}
