% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_dgrdl.R
\name{dgrdl_result}
\alias{dgrdl_result}
\title{Run DGRDL with the specified parameters}
\usage{
dgrdl_result(
  object,
  dgrdl_params = params_dgrdl(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. Ideally, you
should run \code{\link[=preprocess_bulk_coexp]{preprocess_bulk_coexp()}} before applying this function.}

\item{dgrdl_params}{List. Output of \code{\link[=params_dgrdl]{params_dgrdl()}}:
\itemize {
\item sparsity - Integer. Sparsity constraint (max non-zero coefficients
per signal)
\item dict size - Integer. The dictionary size.
\item alpha - Float. Sample context regularisation weight.
\item beta - Float. Feature effect regularisation weight.
\item max_iter - Integer. Maximum number of iterations for the main
algorithm.
\item k_neighbours - Integer. Number of neighbours for the KNN graph for the
feature and sample Laplacian.
\item admm_iter - Integer. ADMM iterations for sparse coding.
\item rho - Float. ADMM step size.
}}

\item{seed}{Integer. Seed for the initialisation of the dictionary.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\description{
Runs the DGRDL algorithm from Pan et al., with the specified hyperparamters.
To determine the hyperparameters, you can use
\code{\link[=dgrdl_grid_search]{dgrdl_grid_search()}}.
}
\references{
Pan et al., Cell Syst, 2022
}
