% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{diffuse_seed_nodes}
\alias{diffuse_seed_nodes}
\title{Diffuse seed genes over a network}
\usage{
diffuse_seed_nodes(
  object,
  diffusion_vector,
  summarisation = c("max", "mean", "harmonic_sum")
)
}
\arguments{
\item{object}{The underlying class \code{\link[=network_diffusions]{network_diffusions()}}.}

\item{diffusion_vector}{A named vector with values to use for the reset
parameter in the personalised page-rank diffusion. Names should represent
node names of the graph.}

\item{summarisation}{If there are duplicated names in the \code{diffusion_vector}
how to summarise these.}
}
\value{
The class with added diffusion score based on a single set of seed
genes. Additionally, the seed genes are stored in the class.
}
\description{
This function takes a diffusion vector and leverages personalised page-rank
diffusion to identify influential nodes. These can be used subsequently for
community detection or check AUROC values given a set of genes.
}
