% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_.R
\name{dim.SingleCells}
\alias{dim.SingleCells}
\title{dim Method for SingleCells object}
\usage{
## S7 method for class <bixverse::SingleCells>
dim(x)
}
\arguments{
\item{x}{An object of class \code{SingleCells}.}
}
\value{
An integer vector of length 2 with the number of cells and genes.
}
\description{
Returns the dimensions of a SingleCells object.
}
\keyword{internal}
