% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_plots_helpers.R
\name{.get_manifoldsr_knn}
\alias{.get_manifoldsr_knn}
\title{Helper function to generate manifoldsR nearest neighbours}
\usage{
.get_manifoldsr_knn(x)
}
\arguments{
\item{x}{\code{SingleCells} object from which to extract the kNN data}
}
\value{
The manifoldsR \code{NearestNeigbours} if possible
}
\description{
Helper function to generate manifoldsR nearest neighbours
}
\keyword{internal}
