% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{.tf_gene_cor_sc}
\alias{.tf_gene_cor_sc}
\title{Get TF gene correlations for SingleCells}
\usage{
.tf_gene_cor_sc(tf_to_gene, object)
}
\arguments{
\item{object}{}
}
\description{
Get TF gene correlations for SingleCells
}
