% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{doublet_detection_boost_sc}
\alias{doublet_detection_boost_sc}
\title{Doublet detection with boosted doublet classification}
\usage{
doublet_detection_boost_sc(
  object,
  boost_params = params_boost(),
  cells_to_use = NULL,
  group_by = NULL,
  seed = 42L,
  streaming = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{boost_params}{A list with the final scrublet parameters, see
\code{\link[=params_boost]{params_boost()}} for full details.}

\item{cells_to_use}{Optional string. Names of the cells to use for the
run of the boosted doublet detection. Useful when you wish to run doublet
detection on individual batches within your data. The object returned will be
specifically using these cells.}

\item{group_by}{Optional grouping variable. Useful if you want to run the
method on a per-sample basis.}

\item{seed}{Integer. Random seed.}

\item{streaming}{Optional Boolean. Shall the data be streamed in. Useful for
larger data sets where you wish to avoid loading in the whole data. If
\code{NULL}, will automatically detect.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
A \code{boost_res} class that has with the following items:
\itemize{
\item predicted_doublets - Boolean vector indicating which observed cells
predicted as doublets (TRUE = doublet, FALSE = singlet).
\item doublet_scores_obs - Numerical vector with the likelihood of being
a doublet for the observed cells.
\item voting_avg - Numerical vector with the average voting score.
}
}
\description{
This function implements the boosted doublet detection. It
generates through several iterations simulated doublets, generate kNN graphs,
runs Louvain clustering and assesses how often an observed cells clsuters
together with the simulated doublets.
}
