% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{doublet_detection_boost_sc}
\alias{doublet_detection_boost_sc}
\title{Doublet detection with boosted doublet classification}
\usage{
doublet_detection_boost_sc(
  object,
  boost_params = params_boost(),
  seed = 42L,
  streaming = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{boost_params}{A list with the final scrublet parameters, see
\code{\link[=params_boost]{params_boost()}} for full details.}

\item{seed}{Integer. Random seed.}

\item{streaming}{Boolean. Shall streaming be used during the HVG
calculations. Slower, but less memory usage.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A \code{boost_res} class that has with the following items:
\itemize{
\item predicted_doublets - Boolean vector indicating which observed cells
predicted as doublets (TRUE = doublet, FALSE = singlet).
\item doublet_scores_obs - Numerical vector with the likelihood of being
a doublet for the observed cells.
\item voting_avg - Numerical vector with the average voting score.
}
}
\description{
This function implements the boosted doublet detection. It
generates through several iterations simulated doublets, generate kNN graphs,
runs Louvain clustering and assesses how often an observed cells clsuters
together with the simulated doublets.
}
