% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_demuxlet_pbmc}
\alias{download_demuxlet_pbmc}
\title{Download PBMCs with demuxlet doublet information}
\usage{
download_demuxlet_pbmc()
}
\value{
String. The path to the extracted doublet detection data.
}
\description{
This function downloads a PBMC data set with demuxlet information to test
doublet detection methods.
}
