% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synthetic_sc.R
\name{download_pbmc3k}
\alias{download_pbmc3k}
\title{Download PBMC3K data from 10x Genomics}
\usage{
download_pbmc3k()
}
\value{
String. The path to the extracted PBMC3K data.
}
\description{
This function downloads the PBMC3K dataset from 10x Genomics and extracts
it to a temporary directory. It returns the path to the extracted data.
}
