% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_pbmc8k}
\alias{download_pbmc8k}
\title{Download PBMC8K data from Zenodo}
\usage{
download_pbmc8k(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean. If the download shall be quiet.}
}
\value{
String. The path to the extracted PBMC8K data.
}
\description{
This function downloads the PBMC8k dataset from 10x Genomics (uploaded on
Zenodo) and extracts it and returns the (temporary) paths.
}
