% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{download_pbmc_batches}
\alias{download_pbmc_batches}
\title{Download two different PBMC data sets for batch correction testing}
\usage{
download_pbmc_batches()
}
\value{
String. The path to the extracted doublet detection data.
}
\description{
This function downloads two different h5ad files for testing batch correction
methods.
}
