% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{f1_score_confusion_mat}
\alias{f1_score_confusion_mat}
\title{F1 scores on top of a confusion matrix}
\usage{
f1_score_confusion_mat(clusters_a, clusters_b)
}
\arguments{
\item{clusters_a}{String or factor. The clustering of algorithm 1.}

\item{clusters_b}{String or factor. The clustering of algorithm 2.}
}
\value{
Named vector with the F1 scores between the two clustering
algorithms.
}
\description{
Helper function to check for expected clustering vs actual clustering.
}
