% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{fast_ica_rust}
\alias{fast_ica_rust}
\title{Fast ICA via Rust}
\usage{
fast_ica_rust(
  X,
  K,
  n_icas,
  ica_fun = c("logcosh", "exp"),
  seed = NULL,
  ica_params = params_ica_general()
)
}
\arguments{
\item{X}{Numeric matrix. Processed data. Output of
\code{\link[=rs_prepare_whitening]{rs_prepare_whitening()}}.}

\item{K}{The K matrix. Whitening matrix. Output of
\code{\link[=rs_prepare_whitening]{rs_prepare_whitening()}}.}

\item{n_icas}{Integer. Number of independent components to recover.}

\item{ica_fun}{String, element of \code{c("logcosh", "exp")}.}

\item{seed}{Integer. Seed to ensure reproducible results.}

\item{ica_params}{List. The ICA parameters, see \code{\link[=params_ica_general]{params_ica_general()}}
wrapper function. This function generates a list containing:
\itemize{
\item maxit - Integer. Maximum number of iterations for ICA.
\item alpha - Float. The alpha parameter for the logcosh version of ICA.
Should be between 1 to 2.
\item max_tol - Maximum tolerance of the algorithm.
\item verbose - Controls verbosity of the function.
}}
}
\value{
A list containing:
\itemize{
\item w The mixing matrix w.
\item A ICA results matrix A.
\item S ICA results matrix S.
\item converged Boolean indicating if algorithm converged.
}
}
\description{
This functions is a wrapper over the Rust implementation of fastICA. It has
the same two options \code{c("logcosh", "exp")} to run ICA in parallel modus. This
function expected
}
