% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_batch_corr.R
\name{fast_mnn_sc}
\alias{fast_mnn_sc}
\title{Run fastMNN}
\usage{
fast_mnn_sc(
  object,
  batch_column,
  batch_hvg_genes,
  fastmnn_params = params_sc_fastmnn(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{batch_column}{String. The column with the batch information in the
obs data of the class.}

\item{batch_hvg_genes}{Integer vector. These are the highly variable genes,
identified by a batch-aware method. Please refer to
\code{\link[=find_hvg_batch_aware_sc]{find_hvg_batch_aware_sc()}} for more details.}

\item{fastmnn_params}{A list, please see \code{\link[=params_sc_fastmnn]{params_sc_fastmnn()}}. The
list has the following parameters:
\itemize{
\item k - Integer. Number of mutual nearest neighbours to identify.
Defaults to \code{20L}.
\item sigma - Numeric. Bandwidth of the Gaussian smoothing kernel (as
proportion of space radius). Defaults to \code{0.1}.
\item knn_method - String. One of \code{c("annoy", "hnsw")}. The method to use
for the approximate nearest neighbour search. Defaults to \code{"annoy"}.
\item dist_metric - String. One of \code{c("cosine", "euclidean")}. The
distance metric to be used for the approximate neighbour search. Defaults
to \code{"cosine"}.
\item annoy_n_trees - Integer. Number of trees for Annoy index. Defaults
to \code{100L}.
\item annoy_search_budget - Integer. Search budget per tree for Annoy.
Defaults to \code{100L}.
\item cos_norm - Logical. Apply cosine normalisation before computing
distances. Defaults to \code{TRUE}.
\item var_adj - Logical. Apply variance adjustment to avoid kissing
effects. Defaults to \code{TRUE}.
\item no_pcs - Integer. Number of PCs to use for MNN calculations.
Defaults to \code{30L}.
\item random_svd - Logical. Use randomised SVD. Defaults to \code{TRUE}.
}}

\item{seed}{Integer. Random seed.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
The object with the added fastMNN embeddings to the object.
}
\description{
This function implements the fast mutual nearest neighbour (MNN) from
Haghverdi, et al. This version works on the PCA embedding and generates
an embedding only and not fully corrected count matrix. The function will
iterate through the batches, identify the MNN and generate correction vectors
and generate a corrected embedding which is added to the function.
}
\references{
Haghverdi, et al., Nat Biotechnol, 2018
}
