% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{filter_similarities}
\alias{filter_similarities}
\title{Filter the calculated similarities}
\usage{
filter_similarities(
  object,
  alpha,
  permutations = 100000L,
  seed = 10101L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\verb{ontology class}. See \code{\link[=ontology]{ontology()}}.}

\item{alpha}{Float. The alpha value. For example, 0.001 would mean that the
critical value is smaller than 0.1 percentile of the random permutations.}

\item{permutations}{Number of random permutations.}

\item{seed}{Integer. For reproducibility purposes.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with filtered results added to the respective slot.
}
\description{
This function calculates the critical value, see
\code{\link[=calculate_critical_value]{calculate_critical_value()}} and filters subsequently all the
term pairs to the ones with a value ≥ critical value.
}
