% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{finalise_matrix}
\alias{finalise_matrix}
\title{Finalise the count matrix}
\usage{
finalise_matrix(matrix, object, return_format, cell_indices, gene_indices)
}
\arguments{
\item{matrix}{The sparse matrix to finalise}

\item{object}{\code{bixverse::single_cell_exp} class.}

\item{return_format}{String. One of \code{c("cell", "gene")}.}

\item{cell_indices}{Optional integer vector. The index positions of cells to
return.}

\item{gene_indices}{Optional integer vector. The index positions of genes to
return.}
}
\value{
The finalised matrix.
}
\description{
Finalise the count matrix
}
