% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_dge_diffabund.R
\name{find_all_markers_sc}
\alias{find_all_markers_sc}
\title{Find all markers}
\usage{
find_all_markers_sc(
  object,
  column_of_interest,
  method = "wilcox",
  alternative = c("greater", "less", "twosided"),
  min_prop = 0.05,
  downsampling = TRUE,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{column_of_interest}{String. The column you wish to use to identify
the markers between all combination. Needs to be in the obs table}

\item{method}{String. Which method to use for the calculations of the DGE.
At the moment the only option is \code{"wilcox"}, but the parameter is reserved
for future features.}

\item{alternative}{String. Test alternative. One of
\code{c("twosided", "greater", "less")}. This function will default to
\code{"greater"}, i.e., genes upregulated in the group.}

\item{min_prop}{Numeric. The minimum proportion of cells that need to express
the gene to be tested in any of the two groups.}

\item{downsampling}{Boolean. If the other group exceeds 100,000 cells, a
random subsample of 100,000 cells will be used.}

\item{seed}{Integer. Seed that is used for the downsampling.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
data.table with the DGE results from the test.
}
\description{
This function can be used to run differential gene expression for every
group of an unsupervised clustering method for example. You specify a column
and the function will start calculating differential gene expression of the
first cluster vs. everything else, second cluster vs. everything else, etc.
The function will automatically downsample everything else to a random set
of 100,000 cells if it should exceed that. This automatic downsampling can
be turned off however.
}
