% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell_processing.R
\name{find_clusters_sc}
\alias{find_clusters_sc}
\title{Graph-based clustering of cells on the sNN graph}
\usage{
find_clusters_sc(object, res = 1, name = "leiden_clustering")
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{res}{Numeric. The resolution parameter for \code{\link[igraph:cluster_leiden]{igraph::cluster_leiden()}}.}

\item{name}{String. The name to add to the obs table in the DuckDB.}
}
\value{
The object with added clustering in the obs table.
}
\description{
This function will apply Leiden clustering on the sNN graph with the
given resolution and add a column to the obs table.
}
