% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_batch_corr.R
\name{find_hvg_batch_aware_sc}
\alias{find_hvg_batch_aware_sc}
\title{Identify HVGs (batch aware)}
\usage{
find_hvg_batch_aware_sc(
  object,
  batch_column,
  hvg_no = 2000L,
  gene_comb_method = c("union", "average", "intersection"),
  hvg_params = params_sc_hvg(),
  streaming = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{batch_column}{String. The column name of the batch column in the obs
table.}

\item{hvg_no}{Integer. Number of highly variable genes to include. Defaults
to \code{2000L}.}

\item{gene_comb_method}{String. One of
\code{c("union", "average", "intersection")}. The method to combine the HVG across
the different batches. Defaults to \code{"union"}.}

\item{hvg_params}{List, see \code{\link[=params_sc_hvg]{params_sc_hvg()}}. This list contains
\itemize{
\item method - Which method to use. One of
\code{c("vst", "meanvarbin", "dispersion")}
\item loess_span - The span for the loess function to standardise the
variance
\item num_bin - Integer. Not yet implemented.
\item bin_method - String. One of \code{c("equal_width", "equal_freq")}. Not
implemented yet.
}}

\item{streaming}{Boolean. Shall the genes be streamed in. Useful for larger
data sets where you wish to avoid loading in the whole data. Defaults to
\code{FALSE}.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
This function will return a list with:
\itemize{
\item hvg_indices - The indices of the batch-aware HVG.
\item batch_hvg_data - data.table with the detailed information of the
variance per batch.
}
}
\description{
This is a helper function to identify highly variable genes in a batch-aware
manner. At the moment the implementation has only the VST-based version
(known as Seurat v3). The other methods will be implemented in the future.
This function will calculate the HVG per given experimental batch and you
can choose the way to combine them. The choices are union (of Top x HVG per
batch), based on the average variance per batch or only take genes that are
amongst the Top X HVG in all batches.
}
\details{
Internally, the batch-aware HVG will be calculated within the cells that are
defined by \code{\link[=set_cells_to_keep]{set_cells_to_keep()}}. If you have not set this
parameter, the function will use all cells that passed initial thresholding
during generation of the object.
}
