% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{find_hvg_sc}
\alias{find_hvg_sc}
\title{Identify HVGs}
\usage{
find_hvg_sc(
  object,
  hvg_no = 2000L,
  hvg_params = params_sc_hvg(),
  streaming = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} (or potentially other) class.}

\item{hvg_no}{Integer. Number of highly variable genes to include. Defaults
to \code{2000L}.}

\item{hvg_params}{List, see \code{\link[=params_sc_hvg]{params_sc_hvg()}}. This list contains
\itemize{
\item method - Which method to use. One of
\code{c("vst", "meanvarbin", "dispersion")}
\item loess_span - The span for the loess function to standardise the
variance
\item num_bin - Integer. Not yet implemented.
\item bin_method - String. One of \code{c("equal_width", "equal_freq")}. Not
implemented yet.
}}

\item{streaming}{Optional Boolean. Shall the data be streamed in. Useful for
larger data sets where you wish to avoid loading in the whole data. If
\code{NULL}, will automatically detect. Not used for \code{MetaCells}.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
It will add the mean, var, var_exp, var_std of each gene to the
the var table.
}
\description{
This is a helper function to identify highly variable genes for \code{SingleCells}
(using the Rust-based streaming of data) or \code{MetaCells}.
}
