% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_dge_diffabund.R
\name{find_markers_sc}
\alias{find_markers_sc}
\title{Calculate DGE between two cell groups}
\usage{
find_markers_sc(
  object,
  cells_1,
  cells_2,
  method = c("wilcox"),
  alternative = c("twosided", "greater", "less"),
  min_prop = 0.05,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} or \code{SingleCellsSubset} class.}

\item{cells_1}{String. The names of the cells in group 1. Need to be part
of the cell names in the object, see \code{\link[=get_cell_names]{get_cell_names()}}.}

\item{cells_2}{String. The names of the cells in group 2. Need to be part
of the cell names in the object, see \code{\link[=get_cell_names]{get_cell_names()}}.}

\item{method}{String. Which method to use for the calculations of the DGE.
At the moment the only option is \code{"wilcox"}, but the parameter is reserved
for future features.}

\item{alternative}{String. Test alternative. One of
\code{c("twosided", "greater", "less")}. Function will default to \code{"twosided"}.}

\item{min_prop}{Numeric. The minimum proportion of cells that need to express
the gene to be tested in any of the two groups.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
data.table with the DGE results from the test.
}
\description{
This function can be used to calculate differentially expressed genes
between two groups in the single cell data. At the moment, it has only
an implementation for the Wilcox-based rank statistic.
}
