% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{find_neighbours_sc}
\alias{find_neighbours_sc}
\title{Find the neighbours for single cell.}
\usage{
find_neighbours_sc(
  object,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  neighbours_params = params_sc_neighbours(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{embd_to_use}{String. The embedding to use. Whichever you chose, it
needs to be part of the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{neighbours_params}{List. Output of \code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
A list with the following items:
\itemize{
\item full_snn - Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours.
\item pruning - Numeric. Weights below this threshold will be set to 0 in
the generation of the sNN graph.
\item snn_similarity - String. One of \code{c("rank", "jaccard")}. Defines how
the weight from the SNN graph is calculated. For details, please see
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{seed}{Integer. For reproducibility.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
The object with added KNN matrix.
}
\description{
This function will generate the kNNs based on a given embedding. Four
different algorithms are implemented with different speed and accuracy to
approximate the nearest neighbours. \code{"hnsw"} implements a Hierarchical
Navigatable Small Worlds vector search that has slower index generation but
high precision. \code{"annoy"} is based on the Approximate Nearest Neighbours Oh
Yeah algorithm and is more rapid in terms of index generation, but querying
on large data sets can be slow. \code{"nndescent"} is a Rust-based implementation
of the PyNNDescent algorithm and is a good all-rounder and performs well
on very large data sets. \code{"exhaustive"} performs exact nearest neighbour
search. Subsequently, the kNN data will be used to generate an sNN igraph for
clustering methods.
}
