% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{find_neighbours_sc}
\alias{find_neighbours_sc}
\title{Find the neighbours for single cell.}
\usage{
find_neighbours_sc(
  object,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  neighbours_params = params_sc_neighbours(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} class.}

\item{embd_to_use}{String. The embedding to use. Whichever you chose, it
needs to be part of the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{neighbours_params}{List. Output of \code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
A list with the following items:
\itemize{
\item full_snn - Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours.
\item pruning - Numeric. Weights below this threshold will be set to 0 in
the generation of the sNN graph.
\item snn_similarity - String. One of \code{c("rank", "jaccard")}. Defines how
the weight from the SNN graph is calculated. For details, please see
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{seed}{Integer. For reproducibility.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
The object with added KNN matrix.
}
\description{
This function will generate the kNNs based on a given embedding. Available
algorithms are:
\itemize{
\item \code{hnsw} - Hierarchical Navigable Small World. A graph-based
approximate nearest neighbour search algorithm; works well on large data
sets. A benign race condition is leveraged during index build, making the
build non-deterministic. Bigger impact on smaller data sets.
\item \code{ivf} - Inverted file index. Uses first k-means clustering to
identify Voronoi cells and leverages these during querying. Works well
on large data sets with high dimensionality.
\item \code{nndescent} - Nearest neighbour descent. Similar to \code{PyNNDescent},
uses a first index to initialise the graph. Good all-rounder.
\item \code{annoy} - Approximate nearest neighbours Oh Yeah. Tree-based index,
used across different R single cell packages (Seurat, SCE). This version
is purely memory-based.
\item \code{exhaustive} - An exhaustive, flat index. On smaller data sets often
faster than the approximate nearest neighbour search algorithms.
}
Subsequently, the kNN graph will be additionally transformed into a shared
nearest neighbour graph for clustering methods.
}
