% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_methods.R
\name{find_rbh_communities}
\alias{find_rbh_communities}
\title{Find RBH communities}
\usage{
find_rbh_communities(rbh_graph)
}
\arguments{
\item{rbh_graph}{The \code{rbh_class} class, see \code{\link[=rbh_graph]{rbh_graph()}}.}
}
\value{
The class with added community detection results.
}
\description{
This is the generic function for identifying communities on top of an RBH
graph.
}
