% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_graph_classes_methods.R
\name{find_rbh_communities}
\alias{find_rbh_communities}
\title{Find RBH communities}
\usage{
find_rbh_communities(rbh_graph, ...)
}
\arguments{
\item{rbh_graph}{The \code{rbh_class} class.}
}
\value{
The class with added community detection results.
}
\description{
This is the generic function for identifying communities on top of an RBH
graph.

This function will generate an RBH graph based on set similarity between
gene modules. You have the option to use an overlap coefficient instead of
Jaccard similarity and to specify a
}
