% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{fix_meta_data_column}
\alias{fix_meta_data_column}
\title{Helper to fix meta-data columns to be R conform}
\usage{
fix_meta_data_column(object, col_names, ...)
}
\arguments{
\item{object}{The underlying object, either \code{\link[=BulkCoExp]{BulkCoExp()}} or
\code{\link[=BulkDge]{BulkDge()}}.}

\item{col_names}{Character vector. The columns to fix.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
Returns the object with the respective metadata columns updated.
}
\description{
This function will update the specified columns in the metadata of an
\code{\link[=BulkDge]{BulkDge()}} or \code{\link[=BulkCoExp]{BulkCoExp()}} to be conform with R standard
naming conventions. This is useful to do before running DGE methods as they
expect standardised names.
}
