% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{gene_ontology_data}
\alias{gene_ontology_data}
\title{Gene Ontology data}
\usage{
gene_ontology_data(go_data_dt, min_genes)
}
\arguments{
\item{go_data_dt}{A data.table that contains the gene ontology information.
This can be extract with for example \code{\link[=get_go_human_data]{get_go_human_data()}}.}

\item{min_genes}{data.frame. Meta-data information in form of a data.frame.}
}
\value{
Returns the class for subsequent usage.
}
\description{
This class is used to store the gene ontology information for usage in GSE
elimination methods.
}
\section{Properties}{

\describe{
\item{go_info}{data.table. Contains the gene ontology identifiers and
names.}
\item{go_to_genes}{List. Contains the genes within each gene ontology
term.}
\item{ancestry}{List. Contains the ancestors for each gene ontology term.}
\item{levels}{List. Which gene ontology terms sit at which level.}
\item{min_genes}{Integer, the minimum genes in the gene ontology term to
conduct the test.}
}
}

