% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_go_object.R
\name{gene_ontology_enrich_data}
\alias{gene_ontology_enrich_data}
\title{Constructor of the S7 class}
\usage{
gene_ontology_enrich_data(go_data_dt, min_genes)
}
\arguments{
\item{go_data_dt}{A data.table that contains the gene ontology information. This can for example
be produced with \code{biomind_to_go_data()}.}

\item{min_genes}{data.frame. Meta-data information in form of a data.frame.}
}
\value{
Returns the S7 object for further operations.
}
\description{
This class is used to store the gene ontology information for
\itemize{
\item go_info: data.table. Contains the gene ontology identifiers and names.
\item go_to_genes: List. Contains the genes within each gene ontology term.
\item ancestry: List. Contains the ancestors for each gene ontology term.
\item levels: List. Which gene ontology terms sit at which level.
\item min_genes: Integer, the minimum genes in the gene ontology term to conduct the test.
}
}
