% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_aggregations.R
\name{generate_bt_meta_cells_sc}
\alias{generate_bt_meta_cells_sc}
\title{Generate meta cells based on hdWGCNA and return a \code{MetaCells} object}
\usage{
generate_bt_meta_cells_sc(
  object,
  sc_meta_cell_params = params_sc_bt_metacells(),
  regenerate_knn = FALSE,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  cells_to_use = NULL,
  target_size = 1e+05,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} or \code{SingleCellsSubset} class.}

\item{sc_meta_cell_params}{List. Output of
\code{\link[=params_sc_bt_metacells]{params_sc_bt_metacells()}}.
A list with the following items:
\itemize{
\item max_shared - Maximum number of allowed shared neighbours for
the meta cell to be considered.
\item target_no_metacells - Number of target meta cells you wish to reach.
\item max_iter - Maximum number of iterations you want to use for the
algorithm.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{regenerate_knn}{Boolean. Shall a kNN graph be regenerated.}

\item{embd_to_use}{String. The embedding to use. Only relevant if you set
regenerate_knn to \code{TRUE}.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used. Only relevant if you set regenerate_knn to
\code{TRUE}.}

\item{cells_to_use}{Optional string. Names of the cells to use for the
generation of the meta-cells. If provided, this function will regenerate the
kNN graph no matter what.}

\item{target_size}{Numeric. The library target size to normalise the meta
cells to.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
A \code{\link[=MetaCells]{MetaCells()}} with the data generated by this meta cell
aggregation method.
}
\description{
This function implements the meta cell aggregation from Morabito, et al.
The generation of metacells is a useful approach for subsequent application
of for example correlation-based methods to identify co-regulated genes.
}
\references{
Morabito, et al. Cell Rep Methods, 2023
}
