% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{generate_knn_sc}
\alias{generate_knn_sc}
\title{Generate the KNN data with distances}
\usage{
generate_knn_sc(
  object,
  embd_to_use = "pca",
  cells_to_use = NULL,
  no_embd_to_use = NULL,
  neighbours_params = params_sc_neighbours(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{embd_to_use}{String. The embedding to use. Whichever you chose, it
needs to be part of the object.}

\item{cells_to_use}{String. Optional cell names to include in the generation
of the kNN graph. If \code{NULL} all (filtered) cells in the object will be used.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{neighbours_params}{List. Output of \code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
A list with the following items:
\itemize{
\item full_snn - Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours.
\item pruning - Numeric. Weights below this threshold will be set to 0 in
the generation of the sNN graph.
\item snn_similarity - String. One of \code{c("rank", "jaccard")}. Defines how
the weight from the SNN graph is calculated. For details, please see
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{seed}{Integer. For reproducibility.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
Initialised \code{sc_knn} with the kNN data.
}
\description{
This function will generate the kNNs based on a given embedding. Three
different algorithms are implemented with different speed and accuracy to
approximate the nearest neighbours. \code{"annoy"} is more
rapid and based on the \verb{Approximate Nearest Neigbours Oh Yeah} algorithm;
\code{"hnsw"} implements a \verb{Hierarchical Navigatable Small Worlds} vector
search that is slower, but more precise. Lastly, there is the option of
\code{"nndescent"}, a Rust-based implementation of the PyNNDescent algorithm. This
version skips the index generation and can be faster on smaller data sets.
This version of the function returns an \code{sc_knn} object that can be
used in other functions.
}
