% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{generate_knn_sc}
\alias{generate_knn_sc}
\title{Generate the KNN data with distances}
\usage{
generate_knn_sc(
  object,
  embd_to_use = "pca",
  cells_to_use = NULL,
  no_embd_to_use = NULL,
  neighbours_params = params_sc_neighbours(),
  seed = 42L,
  .validate_index = TRUE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{embd_to_use}{String. The embedding to use. Whichever you chose, it
needs to be part of the object.}

\item{cells_to_use}{String. Optional cell names to include in the generation
of the kNN graph. If \code{NULL} all (filtered) cells in the object will be used.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{neighbours_params}{List. Output of \code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
A list with the following items:
\itemize{
\item full_snn - Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours. Not used for this function.
\item pruning - Numeric. Weights below this threshold will be set to 0 in
the generation of the sNN graph. Not used for this function.
\item snn_similarity - String. One of \code{c("rank", "jaccard")}. Defines how
the weight from the SNN graph is calculated. For details, please see
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}. Not used for this function.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{seed}{Integer. For reproducibility.}

\item{.validate_index}{Boolean. Shall an exhaustive search against a subset
of cells be run to validate the approximate nearest neighbour index.}

\item{.verbose}{Boolean. Controls verbosity and returns run times.}
}
\value{
Initialised \code{sc_knn} with the kNN data.
}
\description{
This function will generate the kNNs based on a given embedding. Available
algorithms are:
\itemize{
\item \code{hnsw} - Hierarchical Navigable Small World. A graph-based
approximate nearest neighbour search algorithm; works well on large data
sets. A benign race condition is leveraged during index build, making the
build non-deterministic. Bigger impact on smaller data sets.
\item \code{ivf} - Inverted file index. Uses first k-means clustering to
identify Voronoi cells and leverages these during querying. Works well
on large data sets with high dimensionality.
\item \code{nndescent} - Nearest neighbour descent. Similar to \code{PyNNDescent},
uses a first index to initialise the graph. Good all-rounder.
\item \code{annoy} - Approximate nearest neighbours Oh Yeah. Tree-based index,
used across different R single cell packages (Seurat, SCE). This version
is purely memory-based.
\item \code{exhaustive} - An exhaustive, flat index. On smaller data sets often
faster than the approximate nearest neighbour search algorithms.
}
}
