% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{generate_personalisation_vec}
\alias{generate_personalisation_vec}
\title{Helper function to create personalisation vectors}
\usage{
generate_personalisation_vec(graph, node_weights)
}
\arguments{
\item{graph}{igraph. The graph for which to produce the personalisation
vector.}

\item{node_weights}{Named numeric. The names represent the nodes and the
values the strength of the reset.}
}
\value{
The personalisation vector for subsequent usage in page-rank
}
\description{
Helper function to create personalisation vectors
}
