% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_methods.R
\name{generate_rbh_graph}
\alias{generate_rbh_graph}
\title{Generate an RBH graph.}
\usage{
generate_rbh_graph(rbh_graph, ...)
}
\arguments{
\item{rbh_graph}{The \code{rbh_class} class.}

\item{minimum_similarity}{The minimum similarity to create an edge.}

\item{overlap_coefficient}{Shall the overlap coefficient be used instead of
Jaccard similarity.}

\item{.debug}{Debug flat that will create print messages from Rust.}
}
\description{
This is the generic function for generating an RBH graph.

This function will generate an RBH graph based on set similarity between
gene modules. You have the option to use an overlap coefficient instead of
Jaccard similarity and to specify a
}
