% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{generate_sc_knn}
\alias{generate_sc_knn}
\title{Generate a new SingleCellNearestNeighbour from data}
\usage{
generate_sc_knn(
  data,
  neighbours_params = params_sc_neighbours(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{data}{Numerical matrix. Samplex x features. The embedding matrix from
which to generate the kNN data.}

\item{neighbours_params}{List. Output of \code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.
A list with the following items:
\itemize{
\item full_snn - Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours. (Not used in this function.)
\item pruning - Numeric. Weights below this threshold will be set to 0 in
the generation of the sNN graph. (Not used in this function.)
\item snn_similarity - String. One of \code{c("rank", "jaccard")}. Defines how
the weight from the SNN graph is calculated. For details, please see
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}. (Not used in this function.)
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{seed}{Integer. Random seed for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
The \code{SingleCellNearestNeighbour} for downstream usage.
}
\description{
Generate a new SingleCellNearestNeighbour from data
}
