% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synthetic_sc.R
\name{generate_single_cell_test_data}
\alias{generate_single_cell_test_data}
\title{Single cell test data}
\usage{
generate_single_cell_test_data(
  syn_data_params = params_sc_synthetic_data(),
  seed = 42L
)
}
\arguments{
\item{syn_data_params}{List. Contains the parameters for the generation of
synthetic data, see: \code{\link[=params_sc_synthetic_data]{params_sc_synthetic_data()}}. Has the
following elements:
\itemize{
\item n_cells - Integer. Number of cells.
\item n_genes - Integer. Number of genes.
\item marker_genes - List. A nested list that indicates which gene indices
are markers for which cell.
\item n_batches - Integer. Number of batches.
}}

\item{seed}{Integer. The seed for the generation of the seed data.}
}
\value{
List with the following items
\itemize{
\item counts - dgRMatrix with cells x genes.
\item obs - data.table that contains the cell information.
\item var - data.table that contains the var information.
}
}
\description{
This function generates synthetic data for single cell test purposes. These
data can be used for testing functionality of various single cell functions.
}
