% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_single_cell_test_data}
\alias{generate_single_cell_test_data}
\title{Single cell test data}
\usage{
generate_single_cell_test_data(seed = 42L)
}
\arguments{
\item{seed}{Integer. The seed for the generation of the seed data.}
}
\value{
List with the following items
\itemize{
\item counts - dgRMatrix with cells x genes.
\item obs - data.table that contains the cell information.
\item var - data.table that contains the var information.
}
}
\description{
This function generates synthetic data for single cell test purposes. It has
hard-coded parameters and will generate a count matrix of 1000 cells x 100
genes, an obs table and a var table. There are three distinct cell types
that can be found in the data that each express between 2 to 8 marker genes
with higher expression compared to the background. The background genes
have some with higher expression and then less and less cells expressing
them.
}
