% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single_cell.R
\name{generate_single_cell_test_data_adt}
\alias{generate_single_cell_test_data_adt}
\title{Single cell test data (ADT)}
\usage{
generate_single_cell_test_data_adt(
  syn_data_params = params_sc_synthetic_data_adt(),
  seed = 42L
)
}
\arguments{
\item{syn_data_params}{List. Contains the parameters for the generation of
synthetic ADT data, see: \code{\link[=params_sc_synthetic_data_adt]{params_sc_synthetic_data_adt()}}. Has the
following elements:
\itemize{
\item n_cells - Integer. Number of cells.
\item n_proteins - Integer. Number of proteins.
\item marker_genes - List. A nested list that indicates which protein
indices are markers for which cell type.
\item n_batches - Integer. Number of batches.
\item isotype_controls - Integer vector. Column indices (0-based) of the
isotype controls.
\item batch_effect_strength - String. Indicates the strength of the batch
effect to add.
}}

\item{seed}{Integer. The seed for the generation of the synthetic data.}
}
\value{
List with the following items
\itemize{
\item counts - Numeric matrix with cells x proteins.
\item obs - data.table that contains the cell information.
\item var - data.table that contains the protein information.
}
}
\description{
This function generates synthetic ADT counts for single cell test purposes.
These data can be used for testing multi-modal functionality of various
single cell functions. Pairs cell-for-cell with
\code{\link[=generate_single_cell_test_data]{generate_single_cell_test_data()}} when generated with matching
\code{n_cells}, number of cell types and \code{n_batches}.
}
\keyword{internal}
