% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{get_adt_feature_info}
\alias{get_adt_feature_info}
\alias{get_adt_feature_info.ADTCounts}
\title{Get the ADT feature info}
\usage{
get_adt_feature_info(x)

\method{get_adt_feature_info}{ADTCounts}(x)
}
\arguments{
\item{x}{The object from which to get the ADT feature information}
}
\value{
A data.table with the ADT feature information
}
\description{
Returns the number of cells expressing a given ADT
}
