% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{get_adt_names}
\alias{get_adt_names}
\alias{get_adt_names.ADTCounts}
\title{Get the ADT feature names}
\usage{
get_adt_names(x)

\method{get_adt_names}{ADTCounts}(x)
}
\arguments{
\item{x}{An object to get the gene names from.}
}
\value{
The primary ADT feature identifiers stored in the class.
}
\description{
Get the main ADT feature names
}
