% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{get_adt_sample_info}
\alias{get_adt_sample_info}
\alias{get_adt_sample_info.ADTCounts}
\title{Get the ADT sample info}
\usage{
get_adt_sample_info(x)

\method{get_adt_sample_info}{ADTCounts}(x)
}
\arguments{
\item{x}{The object from which to get the ADT sample information}
}
\value{
A data.table with the ADT sample information
}
\description{
Returns sample information in terms of ADT capture, number of features, etc.
}
