% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R
\name{get_available_embeddings}
\alias{get_available_embeddings}
\alias{get_available_embeddings.MetaCells}
\alias{get_available_embeddings.ScCache}
\alias{get_available_embeddings.SingleCells}
\title{Get the available embeddings}
\usage{
get_available_embeddings(x, ...)

\method{get_available_embeddings}{ScCache}(x)
}
\arguments{
\item{x}{An object to get embedding from}

\item{...}{Other parameters.}
}
\value{
Get the names of the available embeddings.
}
\description{
Returns the available embedding as names from the class. This function is
used for the single cell-related classes and methods.
}
