% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{get_available_embeddings}
\alias{get_available_embeddings}
\alias{get_available_embeddings.ScCache}
\alias{get_available_embeddings.SingleCells}
\title{Get the available embeddings from the cache}
\usage{
get_available_embeddings(x)

\method{get_available_embeddings}{ScCache}(x)
}
\arguments{
\item{x}{An object to get embedding from}
}
\value{
Get the names of the available embeddings.
}
\description{
Returns the available embedding names from the cache.
}
