% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_available_embeddings}
\alias{get_available_embeddings}
\alias{get_available_embeddings.sc_cache}
\alias{get_available_embeddings.single_cell_exp}
\title{Get the available embeddings from the cache}
\usage{
get_available_embeddings(x)

\method{get_available_embeddings}{sc_cache}(x)
}
\arguments{
\item{x}{An object to get embedding from}
}
\description{
Returns the available embedding names from the cache.
}
