% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_c_pca.R
\name{get_c_pca_factors}
\alias{get_c_pca_factors}
\title{Get the contrastive PCA factors}
\usage{
get_c_pca_factors(object)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
The sample scores of the contrastive PCA run. If not found,
returns a warning and NULL.
}
\description{
Getter function for the feature factors of the contrastive PCA
}
