% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_c_pca.R
\name{get_c_pca_loadings}
\alias{get_c_pca_loadings}
\title{Get the contrastive PCA loadings}
\usage{
get_c_pca_loadings(object)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
The feature loadings of the contrastive PCA run. If not found,
returns a warning and NULL.
}
\description{
Getter function for the feature loadings of the contrastive PCA
}
