% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_cell_indices}
\alias{get_cell_indices}
\alias{get_cell_indices.sc_mapper}
\alias{get_cell_indices.single_cell_exp}
\title{Get the index position for a gene}
\usage{
get_cell_indices(x, cell_ids, rust_index)

\method{get_cell_indices}{sc_mapper}(x, cell_ids, rust_index)
}
\arguments{
\item{x}{An object to get the gene index from.}

\item{cell_ids}{String vector. The cell ids to search for.}

\item{rust_index}{Bool. Shall rust-based indexing be returned.}
}
\description{
Get the index position for a gene

Set the gene mapping for a \code{single_cell_exp} class.
}
