% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R
\name{get_cell_info}
\alias{get_cell_info}
\alias{get_cell_info.ScMap}
\alias{get_cell_info.SingleCells}
\title{Get the cell idx (R-based) and cell names}
\usage{
get_cell_info(x, filtered = TRUE)

\method{get_cell_info}{ScMap}(x, filtered = TRUE)
}
\arguments{
\item{x}{An object to get the cell info from}

\item{filtered}{Boolean. If \code{TRUE}, only the cells to keep will be returned.}
}
\value{
A named vector with elements -> cell_idx, names -> cell_names.
}
\description{
Returns the cell indices (R-based) and the cell names (usually barcodes)
from the object for further downstream usage.
}
