% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{get_cell_names}
\alias{get_cell_names}
\alias{get_cell_names.ScMap}
\alias{get_cell_names.SingleCells}
\title{Get the cell names}
\usage{
get_cell_names(x, filtered = FALSE)

\method{get_cell_names}{ScMap}(x, filtered = FALSE)
}
\arguments{
\item{x}{An object to get the cell names from.}

\item{filtered}{Boolean. Shall, if found only the cell names of the
\code{cells_to_keep} be returned (see \code{\link[=set_cells_to_keep]{set_cells_to_keep()}}. Defaults
to \code{FALSE}}
}
\description{
Get the cell names

Get the cell names from a \code{SingleCells}.
}
