% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_cell_names.sc_mapper}
\alias{get_cell_names.sc_mapper}
\title{Get the gene indices}
\usage{
\method{get_cell_names}{sc_mapper}(x)
}
\arguments{
\item{x}{An \code{sc_mapper} object}
}
\value{
The cell names
}
\description{
Get the gene indices
}
