% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{get_cell_ranger_params}
\alias{get_cell_ranger_params}
\title{Helper to generate cell ranger input parameters}
\usage{
get_cell_ranger_params(dir_data)
}
\arguments{
\item{dir_data}{String. The directory with the Cell Ranger outputs}
}
\value{
A list based on \code{\link[=params_sc_mtx_io]{params_sc_mtx_io()}}.
}
\description{
Helper to generate cell ranger input parameters
}
