% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R,
%   R/classes_single_cell_groups.R
\name{get_cells_to_keep}
\alias{get_cells_to_keep}
\alias{get_cells_to_keep.ScMap}
\alias{get_cells_to_keep.SingleCells}
\alias{get_cells_to_keep.SingleCellsGrps}
\title{Get the cells to keep}
\usage{
get_cells_to_keep(x)

\method{get_cells_to_keep}{ScMap}(x)

## S7 method for class <bixverse::SingleCells>
get_cells_to_keep(x)

## S7 method for class <bixverse::SingleCellsGrps>
get_cells_to_keep(x)
}
\arguments{
\item{x}{An object to get the gene index from.}
}
\description{
Get the cells to keep
}
