% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{get_cells_to_keep}
\alias{get_cells_to_keep}
\alias{get_cells_to_keep.ScMap}
\alias{get_cells_to_keep.SingleCells}
\title{Get the cells to keep}
\usage{
get_cells_to_keep(x)

\method{get_cells_to_keep}{ScMap}(x)
}
\arguments{
\item{x}{An object to get the gene index from.}
}
\description{
Get the cells to keep

Get the cells to keep from a \code{SingleCells}.
}
