% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_cells_to_keep}
\alias{get_cells_to_keep}
\alias{get_cells_to_keep.sc_mapper}
\alias{get_cells_to_keep.single_cell_exp}
\title{Get the cells to keep}
\usage{
get_cells_to_keep(x)

\method{get_cells_to_keep}{sc_mapper}(x)
}
\arguments{
\item{x}{An object to get the gene index from.}
}
\description{
Get the cells to keep

Get the cells to keep from a \code{single_cell_exp}.
}
