% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_cells_to_keep}
\alias{get_cells_to_keep}
\title{Get the cells to keep}
\usage{
get_cells_to_keep(x)
}
\arguments{
\item{x}{An object to get the gene index for.}
}
\description{
Get the cells to keep
}
