% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_centroids_sc}
\alias{get_centroids_sc}
\alias{get_centroids_sc.SingleCellFastClusters}
\title{Get k-means centroids from a fast cluster result}
\usage{
get_centroids_sc(x)

\method{get_centroids_sc}{SingleCellFastClusters}(x)
}
\arguments{
\item{x}{\code{SingleCellFastClusters} object.}
}
\description{
Get k-means centroids from a fast cluster result
}
