% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{get_cor_graph}
\alias{get_cor_graph}
\title{Get correlation-based graph}
\usage{
get_cor_graph(object, epsilon, .verbose)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{epsilon}{Float. The epsilon parameter for the RBF function, in this
case the bump function.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item graph - The igraph
\item params - A list that contains the parameters of the graph generation
and general graph information (node, edge numbers).
}
}
\description{
Helper function to get a correlation-based igraph from the class
}
