% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_methods_cor.R
\name{get_cor_graph}
\alias{get_cor_graph}
\title{Get correlation-based graph}
\usage{
get_cor_graph(object, kernel_bandwidth, min_affinity, .verbose)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{kernel_bandwidth}{Numerical. The bandwidth to use for the affinity
kernel}

\item{min_affinity}{Numerical. Minimum affinity needed to keep the edge.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item graph - The igraph
\item params - A list that contains the parameters of the graph generation
and general graph information (node, edge numbers).
}
}
\description{
Helper function to get a correlation-based igraph from the class
}
