% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R,
%   R/classes_single_cell_others.R, R/classes_spatial_others.R
\name{get_data}
\alias{get_data}
\alias{get_data.CellQc}
\alias{get_data.ScListRes}
\alias{get_data.ScMatrixRes}
\alias{get_data.ScrubletRes}
\alias{get_data.BoostRes}
\alias{get_data.ScDblFinderRes}
\alias{get_data.SingleCellFastClusters}
\alias{get_data.NmfResult}
\alias{get_data.SpImageFeatures}
\title{Get the ready obs data from various sub method}
\usage{
get_data(x, columns = NULL, ...)

\method{get_data}{CellQc}(x, ...)

\method{get_data}{ScListRes}(x, ...)

\method{get_data}{ScMatrixRes}(x, columns = NULL, ...)

\method{get_data}{ScrubletRes}(x, ...)

\method{get_data}{BoostRes}(x, ...)

\method{get_data}{ScDblFinderRes}(x, ...)

\method{get_data}{SingleCellFastClusters}(x, ...)

\method{get_data}{NmfResult}(x, ...)

\method{get_data}{SpImageFeatures}(x, columns = NULL, ...)
}
\arguments{
\item{x}{An object to get the data from.}

\item{columns}{Optional string. For some of the functions you can decide
to only extract specific columns.}

\item{...}{Other parameters}
}
\value{
Returns a data.table with a cell_idx column for the cells included
in the analysis and additional columns to be added to the obs table.
}
\description{
Helper method that creates data.tables with cell indices which were used
in the given analysis + the values that are to be added to the obs table
in the DuckDB.
}
