% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{get_dge_effect_sizes}
\alias{get_dge_effect_sizes}
\title{Return the effect size results}
\usage{
get_dge_effect_sizes(object)
}
\arguments{
\item{object}{\code{bulk_dge} class.}
}
\value{
Returns the effect size results. (If found.)
}
\description{
Getter function to extract the Effect size results from the
\code{\link[=bulk_dge]{bulk_dge()}} class.
}
