% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{get_dge_limma_voom}
\alias{get_dge_limma_voom}
\title{Return the Limma Voom results}
\usage{
get_dge_limma_voom(object)
}
\arguments{
\item{object}{\code{bulk_dge} class.}
}
\value{
Returns the Limma Voom results. (If found.)
}
\description{
Getter function to extract the Limma Voom results from the
\code{\link[=bulk_dge]{bulk_dge()}} class.
}
