% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{get_dge_list}
\alias{get_dge_list}
\title{Return the DGEList}
\usage{
get_dge_list(object)
}
\arguments{
\item{object}{\code{bulk_dge} class.}
}
\value{
Returns the DGEList stored in the class.
}
\description{
Getter function to extract the DGEList from the \code{\link[=bulk_dge]{bulk_dge()}} class.
}
