% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{get_dge_qc_plot}
\alias{get_dge_qc_plot}
\title{Return QC plots}
\usage{
get_dge_qc_plot(object, plot_choice = NULL)
}
\arguments{
\item{object}{\code{bulk_dge} class.}

\item{plot_choice}{Optional string or integer. Index or name of the plate.}
}
\value{
Returns the DGEList stored in the class.
}
\description{
Getter function to extract the QC plots from the \code{\link[=bulk_dge]{bulk_dge()}}
class. These are added when you run for example
\code{\link[=qc_bulk_dge]{qc_bulk_dge()}} and \code{\link[=normalise_bulk_dge]{normalise_bulk_dge()}}. You can
either leave the plot choice as \code{NULL} and provide input when prompted, or
you provide the name. The possible plots that might be in the class
\itemize{
\item p1_nb_genes_cohort Proportion of non-zero genes for the samples in the
respective cohorts (added after using \code{\link[=qc_bulk_dge]{qc_bulk_dge()}}).
\item p2_outliers An outlier plot based on the data from p1, added after
using \code{\link[=qc_bulk_dge]{qc_bulk_dge()}}.
\item p3_voom_normalization Initial Voom normalisation plot after filtering
lowly expressed genes. Added after using \code{\link[=normalise_bulk_dge]{normalise_bulk_dge()}}.
\item p4_boxplot_normalization Expression levels after normalisation. Added
after using \code{\link[=normalise_bulk_dge]{normalise_bulk_dge()}}.
\item p5_pca_case_control A PCA plot with the chosen case control category.
Added if \code{\link[=calculate_pca_bulk_dge]{calculate_pca_bulk_dge()}} is run.
\item p6_batch_correction_plot A PCA plot pre and post batch correction
with the case-control category overlayed. Added if
\code{\link[=batch_correction_bulk_dge]{batch_correction_bulk_dge()}} is run.
}
}
