% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{get_diffcor_graph}
\alias{get_diffcor_graph}
\title{Get differential correlation-based graph}
\usage{
get_diffcor_graph(object, min_cor = 0.2, fdr_threshold = 0.05, .verbose = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{min_cor}{Float. The minimum absolute correlation that needs to be
present in either data set.}

\item{fdr_threshold}{Float. The maximum FDR that is tolerated for the
generation of the graph.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item graph - The igraph
\item params - A list that contains the parameters of the graph generation
and general graph information (node, edge numbers).
}
}
\description{
Helper function to get a differential correlation-based igraph from the class
}
