% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_differential_abundance_res}
\alias{get_differential_abundance_res}
\alias{get_differential_abundance_res.sc_miloR}
\title{Get the differential abundance results}
\usage{
get_differential_abundance_res(x)

\method{get_differential_abundance_res}{sc_miloR}(x)
}
\arguments{
\item{x}{An object from which to get the differential abundance results
from.}
}
\value{
The differential abundance results stored in the object if found.
}
\description{
Get the differential abundance results
}
