% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_diffusion_perms}
\alias{get_diffusion_perms}
\title{Get the diffusion permutations}
\usage{
get_diffusion_perms(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=network_diffusions]{network_diffusions()}}.}
}
\value{
The diffusion Z scores if found. Otherwise \code{NULL}.
}
\description{
Returns the diffusion Z-scores if you ran
\code{\link[=permute_seed_nodes]{permute_seed_nodes()}}.
}
