% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_classes.R
\name{get_diffusion_vector}
\alias{get_diffusion_vector}
\title{Get the diffusion vector}
\usage{
get_diffusion_vector(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=network_diffusions]{network_diffusions()}}.}
}
\value{
The diffusion vector if found. If you did not run either diffusion
functions, it will return \code{NULL} and a warning.
}
\description{
Returns the diffusion vector if you ran \code{\link[=tied_diffusion]{tied_diffusion()}}
or \code{\link[=diffuse_seed_nodes]{diffuse_seed_nodes()}}.
}
