% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_groups.R
\name{get_embedding}
\alias{get_embedding}
\alias{get_embedding.MetaCells}
\alias{get_embedding.ScCache}
\alias{get_embedding.SingleCells}
\alias{get_embedding.SingleCellsSubset}
\title{Get the embedding}
\usage{
get_embedding(x, embd_name, ...)

## S7 method for class <bixverse::MetaCells>
get_embedding(x, embd_name, ...)

\method{get_embedding}{ScCache}(x, embd_name)

## S7 method for class <bixverse::SingleCells>
get_embedding(x, embd_name, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_embedding(x, embd_name, ...)
}
\arguments{
\item{x}{An object to get embedding from}

\item{embd_name}{String. The name of the embedding to return. The function
will throw an error if the embedding does not exist.}

\item{...}{Other parameters.}
}
\value{
Get the specified embeddings from the object (if found).
}
\description{
General wrapper function that can be used to pull out any embedding stored
in the class. This function is used for the single cell-related classes and
methods.
}
