% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_embedding}
\alias{get_embedding}
\alias{get_embedding.sc_cache}
\alias{get_embedding.single_cell_exp}
\title{Get the embedding from the cache}
\usage{
get_embedding(x, embd_name)

\method{get_embedding}{sc_cache}(x, embd_name)
}
\arguments{
\item{x}{An object to get embedding from}

\item{embd_name}{String. The name of the embedding to return. The function
will throw an error if the embedding does not exist.}
}
\description{
General wrapper function that can be used to pull out any embedding stored
in the \code{sc_cache}.
}
