% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{get_epsilon_res}
\alias{get_epsilon_res}
\title{Return the epsilon data}
\usage{
get_epsilon_res(object)
}
\arguments{
\item{object}{\code{bulk_coexp} class.}
}
\value{
Returns the epsilon data. (If found. Otherwise \code{NULL}).
}
\description{
Getter function to extract the \verb{epsilon param ~ power law goodness of fit}
data from the \code{\link[=bulk_coexp]{bulk_coexp()}} class.
}
