% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_feature_mat}
\alias{get_feature_mat}
\alias{get_feature_mat.ScDblFinderRes}
\title{Get the feature matrix used for the classifier}
\usage{
get_feature_mat(x)

\method{get_feature_mat}{ScDblFinderRes}(x, ...)
}
\arguments{
\item{x}{An object to get the feature matrix from. This will only include
the values of the observed cells.}
}
\description{
Get the feature matrix used for the classifier
}
