% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{get_fpkm_counts}
\alias{get_fpkm_counts}
\title{Return the FPKM-normalised counts}
\usage{
get_fpkm_counts(object)
}
\arguments{
\item{object}{\code{bulk_dge} class.}
}
\value{
Returns the FPKM-normalised counts. (If found.)
}
\description{
Getter function to extract the FPKM-normalised counts from the
\code{\link[=bulk_dge]{bulk_dge()}} class.
}
