% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_subset.R
\name{get_gene_indices}
\alias{get_gene_indices}
\alias{get_gene_indices.MetaCells}
\alias{get_gene_indices.ScMap}
\alias{get_gene_indices.SingleCells}
\alias{get_gene_indices.SingleCellsSubset}
\title{Get the index position for a gene}
\usage{
get_gene_indices(x, gene_ids, rust_index)

## S7 method for class <bixverse::MetaCells>
get_gene_indices(x, gene_ids, rust_index)

\method{get_gene_indices}{ScMap}(x, gene_ids, rust_index)

## S7 method for class <bixverse::SingleCells>
get_gene_indices(x, gene_ids, rust_index)

## S7 method for class <bixverse::SingleCellsSubset>
get_gene_indices(x, gene_ids, rust_index)
}
\arguments{
\item{x}{An object to get the gene index from.}

\item{gene_ids}{String vector. The gene ids to search for.}

\item{rust_index}{Bool. Shall Rust-based indexing be returned.}
}
\value{
The indices of the genes
}
\description{
Returns the index for a given gene based on the internal gene mapping. This
is used for the single cell-related classes and methods.
}
