% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell.R
\name{get_gene_indices}
\alias{get_gene_indices}
\alias{get_gene_indices.MetaCells}
\alias{get_gene_indices.ScMap}
\alias{get_gene_indices.SingleCells}
\title{Get the index position for a gene}
\usage{
get_gene_indices(x, gene_ids, rust_index)

\method{get_gene_indices}{ScMap}(x, gene_ids, rust_index)
}
\arguments{
\item{x}{An object to get the gene index from.}

\item{gene_ids}{String vector. The gene ids to search for.}

\item{rust_index}{Bool. Shall rust-based indexing be returned.}
}
\description{
Get the index position for a gene

Get the cell indices from a \code{MetaCells}.
}
