% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_gene_indices}
\alias{get_gene_indices}
\alias{get_gene_indices.sc_mapper}
\alias{get_gene_indices.single_cell_exp}
\title{Get the index position for a gene}
\usage{
get_gene_indices(x, gene_ids, rust_index)

\method{get_gene_indices}{sc_mapper}(x, gene_ids, rust_index)
}
\arguments{
\item{x}{An object to get the gene index from.}

\item{gene_ids}{String vector. The gene ids to search for.}

\item{rust_index}{Bool. Shall rust-based indexing be returned.}
}
\description{
Get the index position for a gene

Get the gene indices from a \code{single_cell_exp}.
}
