% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{get_gene_lengths}
\alias{get_gene_lengths}
\title{Get the gene lengths}
\usage{
get_gene_lengths(x, species = c("human", "mouse", "rat"), ...)
}
\arguments{
\item{x}{Object to extract gene lengths from. Can be a matrix or
bulk_coexp object.}

\item{species}{String. One of \code{c("human", "mouse", "rat")}.}

\item{...}{Additional parameters passed to methods.}
}
\description{
Get the gene lengths
}
