% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R,
%   R/classes_single_cell_subset.R
\name{get_gene_names}
\alias{get_gene_names}
\alias{get_gene_names.ScMap}
\alias{get_gene_names.SingleCells}
\alias{get_gene_names.SingleCellsSubset}
\title{Get the gene names}
\usage{
get_gene_names(x)

\method{get_gene_names}{ScMap}(x)

## S7 method for class <bixverse::SingleCells>
get_gene_names(x)

## S7 method for class <bixverse::SingleCellsSubset>
get_gene_names(x)
}
\arguments{
\item{x}{An object to get the gene names from.}
}
\value{
The primary gene identifiers stored in the class.
}
\description{
Get the main gene names (for example symbols or Ensembl identifiers).
}
