% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_gene_names_from_idx}
\alias{get_gene_names_from_idx}
\alias{get_gene_names_from_idx.sc_mapper}
\alias{get_gene_names_from_idx.single_cell_exp}
\title{Get the gene names based on the gene idx}
\usage{
get_gene_names_from_idx(x, gene_idx, rust_based = TRUE)

\method{get_gene_names_from_idx}{sc_mapper}(x, gene_idx, rust_based = TRUE)
}
\arguments{
\item{x}{An object to get the gene names from.}

\item{gene_idx}{Integer. The original gene indices for which to return
the gene names.}

\item{rust_based}{Boolean. Is it Rust-based, i.e., 0-index or R-based, i.e.,
1-indexed.}
}
\description{
Get the gene names based on the gene idx

Get the highly variable gene indices from a \code{single_cell_exp}.
}
