% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_go_data_human}
\alias{get_go_data_human}
\title{Wrapper function to load and process the gene ontology data.}
\usage{
get_go_data_human(.verbose = TRUE)
}
\arguments{
\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A data.table
}
\description{
This function loads in gene ontology data stored in the package and processes
it into the format for \code{\link[=gene_ontology_data]{gene_ontology_data()}}. Wraps
\code{\link[=load_go_human_data]{load_go_human_data()}} and \code{\link[=process_go_data]{process_go_data()}} into one.
}
