% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_dgrdl.R
\name{get_grid_search_res}
\alias{get_grid_search_res}
\title{Get the grid search results}
\usage{
get_grid_search_res(object)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
data.table with the grid search results (if found. Otherwise \code{NULL}.)
}
\description{
Getter function to extract the grid search results. If not found will return
\code{NULL}.
}
