% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_h}
\alias{get_h}
\alias{get_h.NmfResult}
\alias{get_h.StabilisedNmfResult}
\title{Get the H (cell activations) matrix}
\usage{
get_h(x)

\method{get_h}{NmfResult}(x)

\method{get_h}{StabilisedNmfResult}(x)
}
\arguments{
\item{x}{An object holding NMF results.}
}
\description{
Get the H (cell activations) matrix
}
