% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_h5ad_dimensions}
\alias{get_h5ad_dimensions}
\title{Helper function to get the dimensions and compressed sparse format}
\usage{
get_h5ad_dimensions(f_path)
}
\arguments{
\item{f_path}{File path to the \code{.h5ad} file.}
}
\value{
A list with the following elements:
\itemize{
\item dims - Dimensions of the stored data in the h5ad file.
\item type - Was the data stored in CSR (indptr = cells) or CSC (indptr =
genes).
}
}
\description{
Helper function to get the dimensions and compressed sparse format
}
