% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_hotspot_membership}
\alias{get_hotspot_membership}
\alias{get_hotspot_membership.sc_hotspot}
\title{Get the hotspot gene membership table}
\usage{
get_hotspot_membership(x)

\method{get_hotspot_membership}{sc_hotspot}(x)
}
\arguments{
\item{x}{The object from which to retrieve the hotspot gene membership}
}
\description{
Get the hotspot gene membership table
}
