% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell.R
\name{get_hvg}
\alias{get_hvg}
\alias{get_hvg.MetaCells}
\alias{get_hvg.ScMap}
\alias{get_hvg.SingleCells}
\title{Get the HVG}
\usage{
get_hvg(x)

\method{get_hvg}{ScMap}(x)
}
\arguments{
\item{x}{An object to get HVG from.}
}
\value{
Indices of the stored HVG genes.
}
\description{
Returns the HVG indices. Pending class type this are 1-based (for R) or
0-based for Rust.
}
