% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_hvg}
\alias{get_hvg}
\alias{get_hvg.ScMap}
\alias{get_hvg.SingleCells}
\title{Get the HVG}
\usage{
get_hvg(x)

\method{get_hvg}{ScMap}(x)
}
\arguments{
\item{x}{An object to get HVG from.}
}
\description{
Get the HVG

Get the highly variable gene indices from a \code{SingleCells}.
}
