% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{get_ica_stability_res}
\alias{get_ica_stability_res}
\title{Get the ICA component data (stability, convergence, nMI)}
\usage{
get_ica_stability_res(object)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
data.table with the ICA parameter data (if found. Otherwise \code{NULL}.)
}
\description{
Getter function to extract the ICA component data in terms of stability,
convergence and normalised mutual information between the components. If not
found will return \code{NULL}.
}
