% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R
\name{get_image}
\alias{get_image}
\title{Get the spatial image for an experiment}
\usage{
get_image(object, exp_id, resolution = c("lowres", "hires", "fullres"))
}
\arguments{
\item{object}{A \code{\link[=SpatialSpot]{SpatialSpot()}}.}

\item{exp_id}{String. The experiment identifier.}

\item{resolution}{String. One of \code{c("lowres", "hires", "fullres")}.}
}
\value{
A numeric array as returned by \code{\link[png:readPNG]{png::readPNG()}} or
\code{\link[jpeg:readJPEG]{jpeg::readJPEG()}}.
}
\description{
Get the spatial image for an experiment
}
