% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_index_cells}
\alias{get_index_cells}
\alias{get_index_cells.sc_miloR}
\title{Get the index cells}
\usage{
get_index_cells(x)

\method{get_index_cells}{sc_miloR}(x)
}
\arguments{
\item{x}{An object from which get the index cells.}
}
\value{
The indices of the cells in the neighbourhood.
}
\description{
Get the index cells
}
