% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{get_inflection_point}
\alias{get_inflection_point}
\title{Coremo: Identify the inflection point}
\usage{
get_inflection_point(x, y, span = 0.25)
}
\arguments{
\item{x, y}{The k and R2_weighted_median values.}

\item{span}{The span parameter for the loess function.}
}
\value{
Returns the inflection point.
}
\description{
This function will identify the optimal cut based on a loess-function fitted
to \code{k ~ R2_weighted_median} via the inflection point.
}
