% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell_lr.R
\name{get_influence}
\alias{get_influence}
\alias{get_influence.LigandTargetInfluence}
\title{Get the ligand-target influence matrix}
\usage{
get_influence(x)

\method{get_influence}{LigandTargetInfluence}(x)
}
\arguments{
\item{x}{An object holding ligand-target influence results.}
}
\description{
Get the ligand-target influence matrix
}
