% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_kmeans_clusters}
\alias{get_kmeans_clusters}
\alias{get_kmeans_clusters.SingleCellFastClusters}
\title{Get k-means cluster assignments from a fast cluster result}
\usage{
get_kmeans_clusters(x)

\method{get_kmeans_clusters}{SingleCellFastClusters}(x)
}
\arguments{
\item{x}{\code{SingleCellFastClusters} object.}
}
\description{
Get k-means cluster assignments from a fast cluster result
}
