% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell.R, R/classes_single_cell_others.R
\name{get_knn_dist}
\alias{get_knn_dist}
\alias{get_knn_dist.MetaCells}
\alias{get_knn_dist.ScCache}
\alias{get_knn_dist.SingleCells}
\alias{get_knn_dist.SingleCellNearestNeighbour}
\title{Get the KNN distance}
\usage{
get_knn_dist(x)

\method{get_knn_dist}{ScCache}(x)

\method{get_knn_dist}{SingleCellNearestNeighbour}(x)
}
\arguments{
\item{x}{An object to get the kNN distances from.}
}
\description{
Getter for an integer matrix of samples x distances. Useful in combination
with \code{\link[=get_knn_mat]{get_knn_mat()}}.
}
