% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell_others.R
\name{get_knn_dist}
\alias{get_knn_dist}
\alias{get_knn_dist.sc_knn}
\title{Get the KNN distance measures}
\usage{
get_knn_dist(x)

\method{get_knn_dist}{sc_knn}(x)
}
\arguments{
\item{x}{An object to get the kNN distances from.}
}
\description{
Get the KNN distance measures
}
