% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R,
%   R/classes_single_cell_others.R
\name{get_knn_mat}
\alias{get_knn_mat}
\alias{get_knn_mat.ScCache}
\alias{get_knn_mat.SingleCells}
\alias{get_knn_mat.SingleCellNearestNeighbour}
\title{Get the KNN matrix}
\usage{
get_knn_mat(x)

\method{get_knn_mat}{ScCache}(x)

\method{get_knn_mat}{SingleCellNearestNeighbour}(x)
}
\arguments{
\item{x}{An object to get the kNN matrix from.}
}
\description{
Get the KNN matrix

Get the KNN matrix from a \code{SingleCells}.
}
