% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{get_knn_obj}
\alias{get_knn_obj}
\alias{get_knn_obj.ScCache}
\alias{get_knn_obj.SingleCells}
\title{Get the KNN object}
\usage{
get_knn_obj(x)

\method{get_knn_obj}{ScCache}(x)
}
\arguments{
\item{x}{An object to get the KNN class from.}
}
\value{
The \code{SingleCellNearestNeighbour} object.
}
\description{
Get the KNN object

Get the KNN object from a \code{SingleCells}.
}
