% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R,
%   R/classes_single_cell_subset.R
\name{get_magic}
\alias{get_magic}
\alias{get_magic.ScCache}
\alias{get_magic.SingleCells}
\alias{get_magic.SingleCellsSubset}
\title{Get the MAGIC imputed layer}
\usage{
get_magic(x, ...)

\method{get_magic}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
get_magic(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_magic(x, ...)
}
\arguments{
\item{x}{An object to get the imputed layer from.}

\item{...}{Other parameters.}
}
\value{
The \code{ScMagic} object, or \code{NULL} when nothing was imputed.
}
\description{
Returns the \code{ScMagic} layer written by \code{\link[=run_magic_sc]{run_magic_sc()}}. This
function is used for the single cell-related classes and methods.
}
