% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{get_meta_cell_matrices}
\alias{get_meta_cell_matrices}
\title{Create sparse dgRMatrix matrices for raw and norm counts}
\usage{
get_meta_cell_matrices(meta_cell_data)
}
\arguments{
\item{meta_cell_data}{Named list. This contains the indptr, indices and data
for both raw counts and norm counts.}
}
\value{
A list of two items
\itemize{
\item raw - Sparse matrix representing the raw counts.
\item norm - Sparse matrix representing the norm counts.
}
}
\description{
Create sparse dgRMatrix matrices for raw and norm counts
}
