% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{get_metadata}
\alias{get_metadata}
\title{Return the metadata}
\usage{
get_metadata(object, ...)
}
\arguments{
\item{object}{The underlying object, either \code{\link[=BulkCoExp]{BulkCoExp()}} or
\code{\link[=BulkDge]{BulkDge()}}.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
Returns the metadata stored in the class.
}
\description{
Getter function to extract the metadata from the \code{\link[=BulkCoExp]{BulkCoExp()}} or
\code{\link[=BulkDge]{BulkDge()}}.
}
