% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_dge_diffabund.R
\name{get_miloR_abundances_sc}
\alias{get_miloR_abundances_sc}
\title{Generate an miloR abundance object for differential abundance testing}
\usage{
get_miloR_abundances_sc(
  object,
  sample_id_col,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  miloR_params = params_sc_miloR(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{sample_id_col}{Character. The column in the obs table representing
the sample identifier to count.}

\item{embd_to_use}{Character. The embedding to use for the refinement
procedure. Please use the same here as you used to generate the neighbours!
Defaults to \code{"pca"}.}

\item{no_embd_to_use}{Optional integer. If you only want to use a subset of
the embedding.}

\item{miloR_params}{A list, please see \code{\link[=params_sc_miloR]{params_sc_miloR()}}. The
list has the following parameters:
\itemize{
\item prop - Numeric. Proportion of cells to sample as neighbourhood
indices. Must be in (0,1).
\item k_refine - Integer. Number of neighbours to use for refinement.
\item refinement_strategy - String. Strategy for refining sampled indices.
One of \code{c("approximate", "bruteforce", "index")}.
\item index_type - String. Type of kNN index to use. One of
\code{c("annoy", "hnsw")}.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults. Note: \code{knn_method} cannot be
\code{"exhaustive"} for MiloR as it basically boils down to \code{"bruteforce"}.
}}

\item{seed}{Integer. Seed for reproducibility}

\item{.verbose}{Boolean. Controls verbosity of the method.}
}
\description{
This function implements the miloR differential abundance testing on top
of the kNN graph. The general idea of the approach is to use the kNN graph
generated from the single cell data, generate representative neighbourhoods
and calculate differential abundances within these neighbourhoods. For
further details on the method, please refer to Dann, et al. This function
will take a \code{single_cell_exp} class, run the neighbourhood detection,
count the occurrences of a sample and return a \code{sc_miloR} class for
subsequent differential abundance testing and further annotations.
}
\references{
Dann, et al., Nat Biotechnol, 2022
}
