% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_dge_diffabund.R
\name{get_miloR_abundances_sc}
\alias{get_miloR_abundances_sc}
\title{Generate an miloR abundance object for differential abundance testing}
\usage{
get_miloR_abundances_sc(
  object,
  sample_id_col,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  miloR_params = params_sc_miloR(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{sample_id_col}{Character. The column in the obs table representing
the sample identifier to count.}

\item{embd_to_use}{Character. The embedding to use for the refinement
procedure. Please use the same here as you used to generate the neighbours!
Defaults to \code{"pca"}.}

\item{no_embd_to_use}{Optional integer. If you only want to use a subset of
the embedding.}

\item{miloR_params}{A list, please see \code{\link[=params_sc_milor]{params_sc_milor()}}. The
list has the following parameters:
\itemize{
\item prop - Numeric. Proportion of cells to sample as neighbourhood
indices. Must be in (0,1). Defaults to \code{0.2}.
\item k_refine - Integer. Number of neighbours to use for refinement.
Defaults to \code{20L}.
\item refinement_strategy - String. Strategy for refining sampled indices.
One of \code{c("approximate", "bruteforce", "index")}. Defaults to
\code{"index"}.
\item index_type - String. Type of kNN index to use. One of
\code{c("annoy", "hnsw")}. Defaults to \code{"annoy"}.
\item k - Integer. Number of neighbours to consider. Defaults to \code{15L}.
\item knn_method - String. One of \code{c("annoy", "hnsw")}. The method to use
for the approximate nearest neighbour search. Defaults to \code{"annoy"}. Note:
\code{"nndescent"} is not supported for MiloR!
\item ann_dist - String. One of \code{c("cosine", "euclidean")}. The distance
metric to be used for the approximate neighbour search. Defaults to
\code{"euclidean"}.
\item search_budget - Integer. Search budget per tree for Annoy. Defaults
to \code{100L}.
\item n_trees - Integer. Number of trees to generate for Annoy. Defaults
to \code{100L}.
\item nn_max_iter - Integer. Maximum iterations for NNDescent. Defaults to
\code{15L}.
\item rho - Numeric. Sampling rate for NNDescent. Defaults to \code{1.0}.
\item delta - Numeric. Early termination criterion for NNDescent. Defaults
to \code{0.001}.
}}

\item{seed}{Integer. Seed for reproducibility}

\item{.verbose}{Boolean. Controls verbosity of the method.}
}
\description{
This function implements the miloR differential abundance testing on top
of the kNN graph. The general idea of the approach is to use the kNN graph
generated from the single cell data, generate representative neighbourhoods
and calculate differential abundances within these neighbourhoods. For
further details on the method, please refer to Dann, et al. This function
will take an \code{single_cell_exp} class, run the neighbourhood detection,
count the occurrences of a sample and return a \code{sc_miloR} class for
subsequent differential abundance testing and further annotations.
}
\references{
Dann, et al., Nat Biotechnol, 2022
}
