% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_model_fit}
\alias{get_model_fit}
\alias{get_model_fit.sc_miloR}
\title{Get the fitted model}
\usage{
get_model_fit(x)

\method{get_model_fit}{sc_miloR}(x)
}
\arguments{
\item{x}{An object from which to get the differential abundance results
from.}
}
\value{
The model object, please refer to \code{\link[edgeR:glmQLFTest]{edgeR::glmQLFTest()}}.
}
\description{
Get the fitted model
}
