% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_nmf.R
\name{get_nmf_gene_loadings}
\alias{get_nmf_gene_loadings}
\title{Get the NMF gene loadings}
\usage{
get_nmf_gene_loadings(object)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}
}
\value{
A features x k numeric matrix (if found) or \code{NULL}.
}
\description{
Getter function to extract the gene loadings matrix (features x k) from a
bulk NMF fit stored in \code{\link[=BulkCoExp]{BulkCoExp()}}. If no NMF fit is present,
returns \code{NULL} with a warning.
}
