% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_nmf.R
\name{get_nmf_sample_activity}
\alias{get_nmf_sample_activity}
\title{Get the NMF sample activity}
\usage{
get_nmf_sample_activity(object)
}
\arguments{
\item{object}{The class, see \code{\link[=BulkCoExp]{BulkCoExp()}}.}
}
\value{
A samples x k numeric matrix (if found) or \code{NULL}.
}
\description{
Getter function to extract the sample activity matrix (samples x k) from a
bulk NMF fit. If no NMF fit is present, returns \code{NULL} with a warning.
}
