% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{get_ontology_ancestry}
\alias{get_ontology_ancestry}
\title{Return ancestry terms from an ontology}
\usage{
get_ontology_ancestry(parent_child_dt)
}
\arguments{
\item{parent_child_dt}{data.table. The data.table with column parent and
child.}
}
\value{
A list with
\itemize{
\item ancestors A list with all ancestor terms.
\item descendants A list with all descendant terms.
}
}
\description{
This function will return all ancestors and descendants based on
a provided data.table with parent-child terms
}
