% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_coexp.R
\name{get_outputs}
\alias{get_outputs}
\title{Return the outputs from bulk_coexp}
\usage{
get_outputs(object)
}
\arguments{
\item{object}{The underlying \code{bulk_coexp} class, see
\code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
Returns the outputs stored in the class.
}
\description{
Getter function to extract the outputs from the \code{\link[=bulk_coexp]{bulk_coexp()}}
class.
}
