% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{get_outputs}
\alias{get_outputs}
\title{Return the outputs}
\usage{
get_outputs(object, ...)
}
\arguments{
\item{object}{The underlying object, either \code{\link[=BulkCoExp]{BulkCoExp()}} or
\code{\link[=BulkDge]{BulkDge()}}.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
Returns the outputs stored in the class.
}
\description{
Getter function to extract the outputs from the \code{\link[=BulkCoExp]{BulkCoExp()}} or
\code{\link[=BulkDge]{BulkDge()}}.
}
