% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_diff_class.R
\name{get_params}
\alias{get_params}
\title{Get the parameters that were used.}
\usage{
get_params(network_diffusions, ...)

get_params(
 network_diffusions,
 to_json = FALSE,
 pretty_json = FALSE
)
}
\arguments{
\item{network_diffusions}{The underlying \code{network_diffusions} class.}

\item{to_json}{Shall the params be returned as a JSON string.}

\item{pretty_json}{Shall the params be returned as a pretty JSON string.}
}
\value{
Depending on parameters either the R list or a (pretty) JSON string.
}
\description{
This method accesses the params slot and can return R lists or JSON strings.

Extracts params from the \code{network_diffusion} class and has options
to return (pretty) JSONs
}
