% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_base_class.R
\name{get_params}
\alias{get_params}
\title{Get the parameters that were used.}
\usage{
get_params(bixverse_generic_class, ...)

get_params(
 bixverse_generic_class,
 to_json = FALSE,
 pretty_json = FALSE
)
}
\arguments{
\item{bixverse_generic_class}{The underlying \code{bixverse_generic_class} class.}

\item{to_json}{Shall the params be returned as a JSON string.}

\item{pretty_json}{Shall the params be returned as a pretty JSON string.}
}
\value{
Depending on parameters either the R list or a (pretty) JSON string.
}
\description{
This method accesses the params slot and can return R lists or JSON strings.

Extracts params from the \code{bixverse_generic_class} (or child)
class and has options to return (pretty) JSONs
}
