% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R, R/classes_single_cell_groups.R
\name{get_pca_factors}
\alias{get_pca_factors}
\alias{get_pca_factors.MetaCells}
\alias{get_pca_factors.ScCache}
\alias{get_pca_factors.SingleCells}
\alias{get_pca_factors.SingleCellsSubset}
\title{Get the PCA factors}
\usage{
get_pca_factors(x, ...)

## S7 method for class <bixverse::MetaCells>
get_pca_factors(x, ...)

\method{get_pca_factors}{ScCache}(x)

## S7 method for class <bixverse::SingleCells>
get_pca_factors(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
get_pca_factors(x, ...)
}
\arguments{
\item{x}{An object to get PCA factors from.}

\item{...}{Other parameters.}
}
\value{
The PCA factors from the object (if found).
}
\description{
Returns the PCA factors (sample-based scores). This function is used for the
single cell-related classes and methods.
}
