% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{get_pca_factors}
\alias{get_pca_factors}
\alias{get_pca_factors.ScCache}
\alias{get_pca_factors.SingleCells}
\title{Get the PCA factors}
\usage{
get_pca_factors(x)

\method{get_pca_factors}{ScCache}(x)
}
\arguments{
\item{x}{An object to get PCA factors from.}
}
\value{
The PCA factors from the object (if found).
}
\description{
Get the PCA factors

Get the PCA factors from a \code{SingleCells}.
}
