% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_pca_factors}
\alias{get_pca_factors}
\alias{get_pca_factors.sc_cache}
\alias{get_pca_factors.single_cell_exp}
\title{Get the PCA factors}
\usage{
get_pca_factors(x)

\method{get_pca_factors}{sc_cache}(x)
}
\arguments{
\item{x}{An object to get PCA factors from.}
}
\description{
Get the PCA factors

Get the PCA factors from a \code{single_cell_exp}.
}
