% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell.R
\name{get_pca_loadings}
\alias{get_pca_loadings}
\alias{get_pca_loadings.MetaCells}
\alias{get_pca_loadings.ScCache}
\alias{get_pca_loadings.SingleCells}
\title{Get the PCA loadings}
\usage{
get_pca_loadings(x)

\method{get_pca_loadings}{ScCache}(x)
}
\arguments{
\item{x}{An object to get PCA loadings from.}
}
\value{
The PCA feature loadings from the object (if found).
}
\description{
Get the PCA loadings
}
