% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_pca_loadings}
\alias{get_pca_loadings}
\title{Get the PCA loadings}
\usage{
get_pca_loadings(x)
}
\arguments{
\item{x}{An object to get PCA loadings for.}
}
\description{
Get the PCA loadings
}
