% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_pca_singular_val}
\alias{get_pca_singular_val}
\alias{get_pca_singular_val.ScCache}
\alias{get_pca_singular_val.SingleCells}
\title{Get the PCA singular values}
\usage{
get_pca_singular_val(x)

\method{get_pca_singular_val}{ScCache}(x)
}
\arguments{
\item{x}{An object to get PCA singular values from.}
}
\value{
The PCA singular values from the object (if found).
}
\description{
Get the PCA singular values

Get the PCA singular values from a \code{SingleCells}.
}
