% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_pca_singular_val}
\alias{get_pca_singular_val}
\alias{get_pca_singular_val.sc_cache}
\alias{get_pca_singular_val.single_cell_exp}
\title{Get the PCA singular values}
\usage{
get_pca_singular_val(x)

\method{get_pca_singular_val}{sc_cache}(x)
}
\arguments{
\item{x}{An object to get PCA singular values from.}
}
\description{
Get the PCA singular values

Get the PCA singular values from a \code{single_cell_exp}.
}
