% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R
\name{get_per_sample_nhood_enrichment}
\alias{get_per_sample_nhood_enrichment}
\alias{get_per_sample_nhood_enrichment.SpCache}
\alias{get_per_sample_nhood_enrichment.SpatialSpot}
\title{Get the per-sample neighbourhood enrichment results}
\usage{
get_per_sample_nhood_enrichment(x, exp_id, label_col, ...)

\method{get_per_sample_nhood_enrichment}{SpCache}(x, exp_id, label_col, ...)

\method{get_per_sample_nhood_enrichment}{SpatialSpot}(x, exp_id, label_col, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{label_col}{String. The obs column name the enrichment was
computed against.}

\item{...}{Other parameters.}
}
\description{
Get the per-sample neighbourhood enrichment results
}
