% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_aggregations.R
\name{get_pseudobulked_sc}
\alias{get_pseudobulked_sc}
\title{Generate pseudo-bulked matrices}
\usage{
get_pseudobulked_sc(
  object,
  cell_list,
  return_format = c("dense", "sparse"),
  assay = c("raw", "norm"),
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{cell_list}{Named list.}

\item{return_format}{String. One of \code{c("dense", "sparse")}.}

\item{assay}{String. One of \code{c("raw", "norm")}.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
Pending on your setting in return_format a dense matrix or sparse
CSR matrix with aggregated cells x genes.
}
\description{
This is a wrapper function to generate pseudo-bulked counts from a list of
cell identifiers. You have the option to return sparse or dense matrices and
if you wish to sum raw counts or calculate the mean of the normalised counts.
}
