% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_rbh_res}
\alias{get_rbh_res}
\title{Get the RBH results}
\usage{
get_rbh_res(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=rbh_graph]{rbh_graph()}}.}
}
\value{
The data.table with the RBH result if found, otherwise NULL.
}
\description{
Pulls out the RBH results if you ran
\code{\link[=generate_rbh_graph]{generate_rbh_graph()}}
}
