% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{get_resolution_res}
\alias{get_resolution_res}
\title{Return the resolution results}
\usage{
get_resolution_res(object)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
If resolution results were found, returns the data.table. Otherwise,
throws a warning and returns NULL.
}
\description{
Getter function to get the resolution results (if available).
}
