% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{get_resolution_res}
\alias{get_resolution_res}
\title{Get the resolution results}
\usage{
get_resolution_res(bulk_coexp)
}
\arguments{
\item{`bulk_coexp`}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
If resolution results were found, returns the data.table. Otherwise,
throws a warning and returns NULL.
}
\description{
Getter function to get the
}
