% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R
\name{get_results}
\alias{get_results}
\title{Get the diffusion results}
\usage{
get_results(bixverse_base_class, ...)

get_results(bixverse_base_class)
}
\arguments{
\item{bixverse_base_class}{The underlying \code{bixverse_base_class} class.}
}
\value{
Returns the final results if any have been stored in the class.
}
\description{
This method returns the final results from a given module class.

Get the final results from the
}
