% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R
\name{get_results}
\alias{get_results}
\title{Get the final results from the class}
\usage{
get_results(object)
}
\arguments{
\item{object}{The underlying \code{\link[=bixverse_base_class]{bixverse_base_class()}}
class. The class functionality is usually inherited by other S7 classes in
\code{bixverse}.}
}
\value{
Returns the final results if any have been stored in the class.
}
\description{
Get the final results from \code{bixverse_base_class} class (or child
classes).
}
