% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_diff_class.R
\name{get_results}
\alias{get_results}
\title{Get the diffusion results}
\usage{
get_results(network_diffusions, ...)

get_results(network_diffusions)
}
\arguments{
\item{network_diffusions}{The underlying \code{network_diffusions} class.}
}
\value{
Returns the community detection results if any can be found.
}
\description{
This method returns the community detection results from the class

Get the community detection results from the class
}
