% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R
\name{get_results}
\alias{get_results}
\title{Get the final results from the class}
\usage{
get_results(bixverse_base_class)
}
\arguments{
\item{bixverse_base_class}{The underlying \code{\link[=bixverse_base_class]{bixverse_base_class()}}
class. This class is usually inherited by other classes.}
}
\value{
Returns the final results if any have been stored in the class.
}
\description{
Get the final results from the a
\code{\link[=bixverse_base_class]{bixverse_base_class()}} class.
}
