% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_rust_count_cell_f_path}
\alias{get_rust_count_cell_f_path}
\title{Get the file path to the counts in cell format}
\usage{
get_rust_count_cell_f_path(object)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}
}
\value{
The path to the \code{counts_cells.bin}
}
\description{
Get the file path to the counts in cell format
}
