% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_sc_counts}
\alias{get_sc_counts}
\title{Getter the counts}
\usage{
get_sc_counts(
  object,
  assay = c("raw", "norm"),
  return_format = c("cell", "gene"),
  cell_indices = NULL,
  gene_indices = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{assay}{String. Which slot to return. One of \code{c("raw", "norm")}.
Defaults to \code{"raw"}.}

\item{return_format}{String. One of \code{c("cell", "gene")}. Return data in
cell-centric compressed format (CSR) or gene-centric compressed format (CSC).
Defaults to \code{"cell"}.}

\item{cell_indices}{Optional cell indices.}

\item{gene_indices}{Optional gene indices.}
}
\value{
The obs table
}
\description{
Getter the counts
}
