% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{get_sc_counts}
\alias{get_sc_counts}
\title{Getter the counts}
\usage{
get_sc_counts(
  object,
  assay = c("raw", "norm"),
  return_format = c("cell", "gene"),
  cell_indices = NULL,
  gene_indices = NULL,
  use_cells_to_keep = TRUE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp}, \code{meta_cells} class.}

\item{assay}{String. Which slot to return. One of \code{c("raw", "norm")}.
Defaults to \code{"raw"}.}

\item{return_format}{String. One of \code{c("cell", "gene")}. Return data in
cell-centric compressed format (CSR) or gene-centric compressed format (CSC).
Defaults to \code{"cell"}. Not relevant for \code{meta_cells}.}

\item{cell_indices}{Optional cell indices.}

\item{gene_indices}{Optional gene indices.}

\item{use_cells_to_keep}{Boolean. Shall cells to keep be found in the class,
shall the counts be reduced to these. Not relevant for \code{meta_cells}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The counts table
}
\description{
Getter the counts
}
