% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{get_sc_obs}
\alias{get_sc_obs}
\title{Getter the obs table}
\usage{
get_sc_obs(object, indices = NULL, cols = NULL, filtered = FALSE)
}
\arguments{
\item{object}{\code{single_cell_exp}, \code{meta_cells} class.}

\item{indices}{Optional integer vector. The integer positions of the cells
to return.}

\item{cols}{Optional string vector. The columns from the obs table to return.}

\item{filtered}{Boolean. Whether to return all cells or filtered to \code{to_keep}
cells. Not relevant for \code{meta_cells}.}
}
\value{
The obs table
}
\description{
Getter the obs table
}
