% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_sc_rust_ptr}
\alias{get_sc_rust_ptr}
\title{Getter for the single cell Rust pointer}
\usage{
get_sc_rust_ptr(object)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}
}
\value{
The Rust structure
}
\description{
Getter for the single cell Rust pointer
}
