% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{get_sc_var}
\alias{get_sc_var}
\title{Getter the var table}
\usage{
get_sc_var(object, indices = NULL, cols = NULL, modality = c("rna", "adt"))
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells}, \code{SingleCellsMultiModal} class.}

\item{indices}{Optional integer vector. The integer positions of the genes
to return.}

\item{cols}{Optional string vector. The columns from the var table to return.}

\item{modality}{String. The modality to return. One of \code{c("rna", "adt")}.}
}
\value{
The vars table
}
\description{
Getter the var table
}
