% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{get_sc_var}
\alias{get_sc_var}
\title{Getter the var table}
\usage{
get_sc_var(object, indices = NULL, cols = NULL)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{indices}{Optional integer vector. The integer positions of the genes
to return.}

\item{cols}{Optional string vector. The columns from the var table to return.}
}
\value{
The vars table
}
\description{
Getter the var table
}
