% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R,
%   R/classes_single_cell_others.R
\name{get_scores}
\alias{get_scores}
\alias{get_scores.ScDblFinderRes}
\alias{get_scores.ScTypeResults}
\title{Get scores}
\usage{
get_scores(x, ...)

\method{get_scores}{ScDblFinderRes}(x, ..., score_type = c("weighted", "cxds_scores"))

\method{get_scores}{ScTypeResults}(x, ...)
}
\arguments{
\item{x}{An object to get scores from.}

\item{...}{Additional arguments passed to methods.}

\item{score_type}{Either \code{"weighted"} or \code{"cxds_scores"}.}
}
\value{
A numeric matrix of cells x cell types.
}
\description{
Get scores
}
