% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_scores}
\alias{get_scores}
\alias{get_scores.ScDblFinderRes}
\title{Get either the cxds or weighted scores}
\usage{
get_scores(x, ..., score_type = c("weighted", "cxds_scores"))

\method{get_scores}{ScDblFinderRes}(x, ..., score_type = c("weighted", "cxds_scores"))
}
\arguments{
\item{x}{An object to get the weighted or cxds scores from.}
}
\description{
Get either the cxds or weighted scores
}
