% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_scores}
\alias{get_scores}
\alias{get_scores.ScDblFinderRes}
\alias{get_scores.ScTypeResults}
\title{Get either the cxds or weighted scores}
\usage{
get_scores(x)

\method{get_scores}{ScDblFinderRes}(x, ..., score_type = c("weighted", "cxds_scores"))

get_scores(x)

\method{get_scores}{ScTypeResults}(x)
}
\arguments{
\item{x}{\code{ScTypeResults} object.}
}
\value{
A numeric matrix of cells x cell types.
}
\description{
Get either the cxds or weighted scores

Get the ScType score matrix
}
