% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{get_sim_matrix}
\alias{get_sim_matrix}
\title{Get the similarity matrix}
\usage{
get_sim_matrix(object, as_data_table = FALSE, .verbose = TRUE)
}
\arguments{
\item{object}{\verb{ontology class}. See \code{\link[=ontology]{ontology()}}.}

\item{as_data_table}{Boolean. Shall the data be returned as a long
data.table.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Returns the semantic similarity data.table from the class
}
\description{
Get the similarity matrix
}
