% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_snf_adjcacency_mat}
\alias{get_snf_adjcacency_mat}
\title{Get an individual affinity matrix}
\usage{
get_snf_adjcacency_mat(object, name)
}
\arguments{
\item{object}{The underlying class \code{\link[=snf]{snf()}}.}

\item{name}{String. The name of the individual data modality affinity
matrix to return.}
}
\value{
Returns adjcacency matrix if found.
}
\description{
Get an individual affinity matrix
}
