% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_snf_final_mat}
\alias{get_snf_final_mat}
\title{Get the final SNF matrix}
\usage{
get_snf_final_mat(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=snf]{snf()}}.}
}
\value{
Returns the SNF adjacency/similarity matrix.
}
\description{
Get the final SNF matrix
}
