% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{get_snf_params}
\alias{get_snf_params}
\title{Get the SNF params}
\usage{
get_snf_params(object)
}
\arguments{
\item{object}{The underlying class \code{\link[=snf]{snf()}}.}
}
\value{
Returns the stored SNF params
}
\description{
Get the SNF params
}
