% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{get_snn_graph}
\alias{get_snn_graph}
\alias{get_snn_graph.ScCache}
\alias{get_snn_graph.SingleCells}
\title{Get the sNN graph}
\usage{
get_snn_graph(x)

\method{get_snn_graph}{ScCache}(x)
}
\arguments{
\item{x}{An object to get the sNN graph from.}
}
\value{
The igraph that has the shared nearest neighbours.
}
\description{
Get the sNN graph

Get the SNN graph from a \code{SingleCells}.
}
