% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_references.R
\name{get_symphony_hvg_names}
\alias{get_symphony_hvg_names}
\title{Getter for the HVG gene names of a Symphony reference}
\usage{
get_symphony_hvg_names(object, ...)
}
\arguments{
\item{object}{\code{SymphonyReference} class.}
}
\value{
Character vector of HVG gene names in reference loading order.
}
\description{
Getter for the HVG gene names of a Symphony reference
}
