% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_references.R
\name{get_symphony_labels}
\alias{get_symphony_labels}
\title{Getter for the stored labels of a Symphony reference}
\usage{
get_symphony_labels(object)
}
\arguments{
\item{object}{\code{SymphonyReference} class.}
}
\value{
A \code{data.table} of reference cell labels in \code{z_corr} row order, or
\code{NULL} if no labels are stored.
}
\description{
Getter for the stored labels of a Symphony reference
}
