% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_references.R
\name{get_symphony_loadings}
\alias{get_symphony_loadings}
\title{Getter for the PCA loadings of a Symphony reference}
\usage{
get_symphony_loadings(object, ...)
}
\arguments{
\item{object}{\code{SymphonyReference} class.}
}
\value{
The PCA gene loadings matrix (n_hvgs x d).
}
\description{
Getter for the PCA loadings of a Symphony reference
}
