% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_symphony.R
\name{get_symphony_z_corr}
\alias{get_symphony_z_corr}
\title{Getter for the corrected embedding of a Symphony reference}
\usage{
get_symphony_z_corr(object, ...)
}
\arguments{
\item{object}{\code{SymphonyReference} class.}
}
\value{
The post-Harmony corrected embedding (N x d), or \code{NULL} for slim
references.
}
\description{
Getter for the corrected embedding of a Symphony reference
}
