% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_tf_to_gene}
\alias{get_tf_to_gene}
\alias{get_tf_to_gene.ScenicGrn}
\title{Extract the TF to gene data from the ScenicGrn object}
\usage{
get_tf_to_gene(x)

\method{get_tf_to_gene}{ScenicGrn}(x)
}
\arguments{
\item{x}{\code{ScenicGrn} object from which to extract the TF to gene data.table.}
}
\value{
data.table with TF to gene information
}
\description{
Extract the TF to gene data from the ScenicGrn object
}
