% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{get_w}
\alias{get_w}
\alias{get_w.NmfResult}
\alias{get_w.StabilisedNmfResult}
\alias{get_w.ConsensusNmfResult}
\title{Get the W (gene loadings) matrix}
\usage{
get_w(x)

\method{get_w}{NmfResult}(x)

\method{get_w}{StabilisedNmfResult}(x)

\method{get_w}{ConsensusNmfResult}(x)
}
\arguments{
\item{x}{An object holding NMF results.}
}
\description{
Get the W (gene loadings) matrix
}
