% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_sc.R
\name{\%>>\%}
\alias{\%>>\%}
\title{Append a step to a pipeline}
\usage{
lhs \%>>\% rhs
}
\arguments{
\item{lhs}{\code{ScPipeline} or \code{ScStep}.}

\item{rhs}{\code{ScStep}.}
}
\value{
A \code{ScPipeline}.
}
\description{
\verb{\%>>\%} chains pipeline steps. Either side can be a \code{ScStep} or a
\code{ScPipeline}; the result is always a \code{ScPipeline}.
}
