% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_go_gse.R
\name{gse_go_elim_method}
\alias{gse_go_elim_method}
\title{Run gene ontology enrichment with elimination method.}
\usage{
gse_go_elim_method(
  object,
  target_genes,
  minimum_overlap = 3L,
  fdr_threshold = 0.05,
  elim_threshold = 0.05,
  min_genes = NULL,
  .debug = FALSE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=gene_ontology_data]{gene_ontology_data()}}.}

\item{target_genes}{String. The target genes you wish to apply the GSEA over.}

\item{minimum_overlap}{Integer. Threshold for the minimal overlap.}

\item{fdr_threshold}{Float. Threshold for maximum fdr to include in the
output.}

\item{elim_threshold}{Float. Threshold from which p-value onwards the
elimination on the ancestors shall be conducted.}

\item{min_genes}{Integer. Minimum number of genes that have to be included in
the gene ontology term. If NULL, it will default to the number of minimum
genes stored in \code{gene_ontology_data}.}

\item{.debug}{Boolean. Shall information from the Rust function be displayed.
For debugging purposes.}
}
\value{
data.table with enrichment results.
}
\description{
This method takes the gene_ontology_data and a target gene set and performs
an GSE enrichment leveraging ontological information. It starts at the lowest
levels of the ontology and tests if there is significant enrichment for any
GO terms. If the threshold of the p-value is below the elimination threshold,
the genes from this term will be removed from all its ancestors. The function
then proceeds to the next level of the ontology and repeats the process.
}
