% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_go_obj_methods.R
\name{GSE_GO_elim_method}
\alias{GSE_GO_elim_method}
\title{GSE on a single set of target genes with the GO elimination method}
\usage{
GSE_GO_elim_method(S7_obj, ...)
}
\arguments{
\item{S7_obj}{\code{gene_ontology_data} class.}

\item{target_genes}{String. ...}

\item{minimum_overlap}{Integer. ...}

\item{fdr_threshold}{Float. ...}

\item{elim_threshold}{Float. ...}

\item{min_genes}{Integer. ...}

\item{.debug}{Boolean. ...}
}
\description{
...
}
