% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_batch_corr.R
\name{harmony_sc}
\alias{harmony_sc}
\title{Run Harmony}
\usage{
harmony_sc(
  object,
  batch_column,
  additional_batch_columns = NULL,
  harmony_params = params_sc_harmony(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{...}

\item{batch_column}{...}

\item{additional_batch_columns}{...}

\item{harmony_params}{...}

\item{seed}{...}

\item{.verbose}{...}
}
\value{
...
}
\description{
A version of Harmony by Korsunsky et al., implemented in Rust.
}
