% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{hedges_g_dge}
\alias{hedges_g_dge}
\title{Calculate the effect}
\usage{
hedges_g_dge(
  meta_data,
  main_contrast,
  contrast_list = NULL,
  normalised_counts,
  small_sample_correction = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{meta_data}{data.table. The meta information about the experiment in
which the contrast info can be found.}

\item{main_contrast}{String. Which column contains the main groups you want
to calculate the Hedge's G effect for. Every permutation of the groups
will be tested.}

\item{normalised_counts}{Numeric Matrix. The normalized count matrix.}

\item{small_sample_correction}{Can be NULL (automatic determination if a
small sample size correction should be applied) or Boolean.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
A data.table with the effect sizes and standard errors based on the
Hedge's G effect size for all found permutations of the groups.
}
\description{
Calculate the effect
}
