% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bulk.R
\name{hedges_g_dge}
\alias{hedges_g_dge}
\title{Calculate the effect size}
\usage{
hedges_g_dge(
  meta_data,
  main_contrast,
  normalised_counts,
  contrast_list = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{meta_data}{data.table. The meta information about the experiment in
which the contrast info can be found.}

\item{main_contrast}{String. Which column contains the main groups you want
to calculate the Hedge's G effect for. Every permutation of the groups
will be tested if \code{contrast_list} is \code{NULL}.}

\item{normalised_counts}{Numeric Matrix. The normalized count matrix.}

\item{contrast_list}{String vector or NULL. Optional string vector of
contrast formatted as \code{"contrast1-contrast2"}. Default NULL will create all
contrasts automatically.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
A data.table with the effect sizes and standard errors based on the
Hedge's G effect size for the groups.
}
\description{
Calculate the effect size
}
