% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_pathways.R
\name{hotspot_autocor_sc}
\alias{hotspot_autocor_sc}
\title{Calculate the local auto-correlation of a gene}
\usage{
hotspot_autocor_sc(
  object,
  embd_to_use = "pca",
  hotspot_params = params_sc_hotspot(),
  no_embd_to_use = NULL,
  cells_to_take = NULL,
  genes_to_take = NULL,
  streaming = FALSE,
  random_seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{embd_to_use}{String. The embedding to use. Defaults to \code{"pca"}.}

\item{hotspot_params}{List with vision parameters, see
\code{\link[=params_sc_hotspot]{params_sc_hotspot()}} with the following elements:
\itemize{
\item model - String. Which of the available models to use for the
gene expression. Choices are one of \code{c("danb", "normal", "bernoulli")}.
\item normalise - Boolean. Shall the data be normalised.
\item k - Number of neighbours for the kNN search. Only relevant if you
set regenerate_knn to \code{TRUE}.
\item knn_method - String. Which kNN algorithm to use. One of
\code{c("annoy", "hnsw", "nndescent")}. Defaults to \code{"annoy"}. Only relevant if
you set regenerate_knn to \code{TRUE}.
\item ann_dist - String. Distance metric for the approximate neighbour
search. One of \code{c("cosine", "euclidean")}. Defaults to \code{"cosine"}. Only
relevant if you set regenerate_knn to \code{TRUE}.
\item n_trees - Integer. Number of trees to use for the annoy algorithm.
Only relevant if you set regenerate_knn to \code{TRUE}.
\item search_budget - Integer. Search budget per tree for the annoy
algorithm. Only relevant if you set regenerate_knn to \code{TRUE}.
\item nn_max_iter - Integer. Maximum iterations for NN Descent. Only
relevant if you set regenerate_knn to \code{TRUE} and use
\code{"nndescent"}.
\item rho - Numeric. Sampling rate for NN Descent. Only relevant if you
set regenerate_knn to \code{TRUE} and use \code{"nndescent"}.
\item delta - Numeric. Early termination criterion for NN Descent. Only
relevant if you set regenerate_knn to \code{TRUE} and use \code{"nndescent"}.
}}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{cells_to_take}{Optional string vector. If you want to only use
selected cells. If \code{NULL} will default to all cells_to_keep in the class.}

\item{genes_to_take}{Optional string vector. If you wish to limit the
search to a subset of genes. If \code{NULL} will default to all genes in the
class.}

\item{streaming}{Boolean. Shall the data be streamed in. Useful for larger
data sets.}

\item{random_seed}{Integer. Used for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A data.table with the auto-correlations on a per gene basis and
various statistics.
}
\description{
This method implements the HotSpot approach (see DeTomaso, et al.) to
calculate the auto-correlation of a given gene in the kNN graph based on
the chosen embedding. This can be used to identify genes that have strong
local correlations and vary across the kNN graph.
}
