% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_pathways.R
\name{hotspot_gene_cor_sc}
\alias{hotspot_gene_cor_sc}
\title{Calculate the local pairwise gene-gene correlation}
\usage{
hotspot_gene_cor_sc(
  object,
  embd_to_use = "pca",
  hotspot_params = params_sc_hotspot(),
  no_embd_to_use = NULL,
  cells_to_take = NULL,
  genes_to_take = NULL,
  streaming = FALSE,
  random_seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{embd_to_use}{String. The embedding to use. Defaults to \code{"pca"}.}

\item{hotspot_params}{List with hotspot parameters, see
\code{\link[=params_sc_hotspot]{params_sc_hotspot()}} with the following elements:
\itemize{
\item model - String. Which of the available models to use for the
gene expression. Choices are one of \code{c("danb", "normal", "bernoulli")}.
\item normalise - Boolean. Shall the data be normalised.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
}}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{cells_to_take}{Optional string vector. If you want to only use
selected cells. If \code{NULL} will default to all cells_to_keep in the class.}

\item{genes_to_take}{Optional string vector. If you wish to limit the
search to a subset of genes. If \code{NULL} will default to all genes in the
class.}

\item{streaming}{Boolean. Shall the data be streamed in. Useful for larger
data sets.}

\item{random_seed}{Integer. Used for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A \code{sc_hotspot} class that can be used for subsequent analysis.
}
\description{
This method implements the HotSpot approach (see DeTomaso, et al.) to
calculate the local gene-gene correlations and their Z-scores.
}
\details{
Should a gene not be found in sufficient cells, the pairs with this gene
will be set to 0. Please ensure prior to running the function that you
are only calculating gene-gene auto-correlations that occur in sufficient
cells.
}
