% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{ica_processing}
\alias{ica_processing}
\title{Prepare class for ICA}
\usage{
ica_processing(object, fast_svd = TRUE, random_seed = 123L, .verbose = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. Ideally, you
should run \code{\link[=preprocess_bulk_coexp]{preprocess_bulk_coexp()}} before applying this function.}

\item{fast_svd}{Boolean. Shall randomised SVD be used for the whitening.
This is faster and usually causes little precision loss.}

\item{random_seed}{Integer. Seed for the randomised SVD. Only relevant, if
fast_svd = \code{TRUE}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} with the needed data for ICA in the
properties of the class.
}
\description{
This is the generic function for doing the necessary preprocessing for
running independent component analysis.
}
