% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_methods_ica.R
\name{ica_processing}
\alias{ica_processing}
\title{Prepare class for ICA}
\usage{
ica_processing(bulk_coexp, ...)

...
}
\arguments{
\item{bulk_coexp}{\code{bulk_coexp} class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} with the needed data for ICA in the
properties of the class.
}
\description{
This is the generic function for doing the necessary preprocessing for
running independent component analysis.

This function will prepare the \code{bulk_coexp} for subsequent usage of the
contrastive ICA functions.
}
