% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_methods_ica.R
\name{ica_processing}
\alias{ica_processing}
\title{Prepare class for ICA}
\usage{
ica_processing(object, .verbose = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. Ideally, you
should run \code{\link[=preprocess_bulk_coexp]{preprocess_bulk_coexp()}} before applying this function.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
\code{bulk_coexp} with the needed data for ICA in the
properties of the class.
}
\description{
This is the generic function for doing the necessary preprocessing for
running independent component analysis.
}
