% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{identify_tf_to_genes}
\alias{identify_tf_to_genes}
\alias{identify_tf_to_genes.ScenicGrn}
\title{Identify the TF to gene regulation}
\usage{
identify_tf_to_genes(x, k_tfs, k_genes, min_importance = NULL)

\method{identify_tf_to_genes}{ScenicGrn}(x, k_tfs, k_genes, min_importance = NULL)
}
\value{
Adds a data.table with the first tf to gene results to the class.
}
\description{
The function will generate the TF to gene associations that can be further
subfiltered subsequently.
}
\details{
You have three options to extract the information
based on the importance scores generated by the tree-based regression model.
\itemize{
\item \code{k_tfs} - This defines how many of the Top TFs per given gene you wish
to include in a given analysis. If you assume that there are 5 TFs
controlling your gene, you would set this to \code{5L}. For each gene, the Top 5
TFs are identified and subsequently the TF -> c(gene_1, gene_2, gene_3)
associations are generated.
\item k_genes - This defines how many target genes per given TF you wish
to include. If you assume that a TF has a 100 potential target genes, you
would set this to \code{100L}. For each TF, the Top 100 potential downstream
targets are being included in this case.
\item min_importance - If you want to set a threshold for the minimum
importance score to filter out noisy genes.
}
You can provide all three parameters at once, in this case you will get a
union of the TF -> gene, gene <- TF approach, filtered by min_importance.
}
