% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{knn_graph_label_propagation}
\alias{knn_graph_label_propagation}
\title{kNN-based graph label propagation}
\usage{
knn_graph_label_propagation(
  edge_list,
  labels,
  iterations = 100L,
  alpha = 0.9,
  tolerance = 1e-06
)
}
\arguments{
\item{edge_list}{Integer vector. In form of node_1, node_2, node_3, ...
which indicates alternating pairs (node_1, node_2), etc in terms of edges.}

\item{labels}{String. The labels with \code{NA} indicating unlabelled.}

\item{iterations}{Integer. Maximum iterations for the algorithm.}

\item{alpha}{Numeric. Parameter that controls the spreading. Usually between
0.9 to 0.95. Larger values drive further labelling, smaller values are more
conversative.}

\item{tolerance}{Tolerance parameter for early stopping.}
}
\value{
A list with the following elements
\itemize{
\item assignment_probs - Matrix with the assignment probabilities.
\item final_labels - Final labels in the graph
}
}
\description{
In case of a kNN graph with a subset of nodes having labels, this function
can be used to propagate the labels through the graph. This function uses
the label spreading version of the algorithm.
}
