% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{load_existing}
\alias{load_existing}
\title{Load an existing SingleCells from disk}
\usage{
load_existing(object)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} or \code{SingleCellsMultiModal} class.}
}
\value{
The object with added information on the data on disk.
}
\description{
Helper function that can load the parameters to access the on-disk stored
data into the class.
}
