% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell_processing.R
\name{load_existing}
\alias{load_existing}
\title{Load an existing single_cell_exp from disk}
\usage{
load_existing(object)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}
}
\value{
The object with added information on the data on disk.
}
\description{
Helper function that can load the parameters to access the on-disk stored
data into the class.
}
