% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell.R
\name{load_h5ad}
\alias{load_h5ad}
\title{Load in h5ad to \code{single_cell_exp} (nightly!)}
\usage{
load_h5ad(
  object,
  h5_path,
  min_genes = 200L,
  min_cells = 10L,
  target_size = 1e+05,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{h5_path}{File path to the h5ad object you wish to load in.}

\item{min_genes}{Integer. Minimum number of genes per cell to be considered
in subsequent analysis.}

\item{min_cells}{Integer. Minimum number of cells per gene to be considered
in subsequent analysis.}

\item{target_size}{Numeric. The target size. 1e6 -> CPM, 1e5 is more commonly
used in single cell analysis.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
It will populate the files on disk and return the class with updated
shape information.
}
\description{
This function takes an h5ad file and loads the obs and var data into the
DuckDB of the \code{single_cell_exp} class and the counts into a Rust-binarised
format for rapid access. During the reading in of the counts, the log CPM
transformation will occur automatically.
}
