% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{load_h5ad}
\alias{load_h5ad}
\title{Load h5ad files (deprecated)}
\usage{
load_h5ad(
  object,
  h5_path,
  sc_qc_param = params_sc_min_quality(),
  streaming = TRUE,
  batch_size = 1000L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{h5_path}{File path to the h5 file. Supports h5ad (AnnData),
CellRanger v3 (10x Genomics), and CellRanger v2 formats. Format is
automatically detected.}

\item{sc_qc_param}{List. Output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}. A
list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of genes to be detected
in the cell to be included.
\item min_lib_size - Integer. Minimum library size in the cell to be
included.
\item min_cells - Integer. Minimum number of cells a gene needs to be
detected to be included.
\item target_size - Float. Target size to normalise to. Defaults to \code{1e5}.
}}

\item{streaming}{Boolean. Shall the data be streamed during the conversion
of CSR to CSC. Defaults to \code{TRUE} and should be used for larger data sets.}

\item{batch_size}{Integer. If \code{streaming = TRUE}, how many cells to process
in one batch. Defaults to \code{1000L}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
A \code{single_cell_exp} object with loaded data
}
\description{
\lifecycle{deprecated}

\code{load_h5ad()} has been renamed to \code{\link[=load_h5]{load_h5()}} to reflect support for
multiple h5 formats (h5ad, CellRanger v2, CellRanger v3).
}
