% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{load_mtx}
\alias{load_mtx}
\title{Load in mtx/plain text files to \code{single_cell_exp} (nightly!)}
\usage{
load_mtx(
  object,
  sc_mtx_io_param = params_sc_mtx_io(),
  sc_qc_param = params_sc_min_quality(),
  streaming = TRUE,
  batch_size = 1000L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{sc_mtx_io_param}{List. Please generate this one via
\code{\link[=params_sc_mtx_io]{params_sc_mtx_io()}}. Needs to contain:
\itemize{
\item path_mtx - String. Path to the .mtx file
\item path_obs - String. Path to the file containing cell/barcode info.
\item path_var - String. String. Path to the file containing gene/variable
info.
\item cells_as_rows - Boolean. Do cells represent the rows or columns.
}}

\item{sc_qc_param}{List. Output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}. A
list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of genes to be detected
in the cell to be included.
\item min_lib_size - Integer. Minimum library size in the cell to be
included.
\item min_cells - Integer. Minimum number of cells a gene needs to be
detected to be included.
\item target_size - Float. Target size to normalise to. Defaults to \code{1e5}.
}}

\item{streaming}{Boolean. Shall the data be streamed during the conversion
of CSR to CSC. Defaults to \code{TRUE} and should be used for larger data sets.}

\item{batch_size}{Integer. If \code{streaming = TRUE}, how many cells to process
in one batch. Defaults to \code{1000L}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
It will populate the files on disk and return the class with updated
shape information.
}
\description{
This is a helper function to load in mtx files and corresponding plain text
files. It will automatically filter out low quality cells and only keep
high quality cells. Under the hood DucKDB and high performance Rust binary
files are being used to store the counts.
}
