% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_single_cell.R
\name{load_mtx}
\alias{load_mtx}
\title{Load in mtx/plain text files to \code{single_cell_exp} (nightly!)}
\usage{
load_mtx(
  object,
  mtx_path,
  obs_path,
  var_path,
  sc_qc_param = params_sc_min_quality(),
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{mtx_path}{File path to the mtx file you wish to load in.}

\item{obs_path}{File path}

\item{var_path}{File path}

\item{sc_qc_param}{List. Output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}. A
list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of genes to be detected
in the cell to be included.
\item min_lib_size - Integer. Minimum library size in the cell to be
included.
\item min_cells - Integer. Minimum number of cells a gene needs to be
detected to be included.
\item target_size - Float. Target size to normalise to. Defaults to \code{1e5}.
}}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
It will populate the files on disk and return the class with updated
shape information.
}
\description{
This is a helper function to load in mtx files and corresponding plain text
files. It will automatically filter out low quality cells and only keep
high quality cells. Under the hood DucKDB and high performance Rust binary
files are being used to store the counts.
}
