% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_io.R
\name{load_r_data}
\alias{load_r_data}
\title{Load in data directly from R objects.}
\usage{
load_r_data(
  object,
  counts,
  obs,
  var,
  sc_qc_param = params_sc_min_quality(),
  batch_size = 1000L,
  streaming = TRUE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{counts}{Sparse matrix. The cells represent the rows, the genes the
indices. Needs to be \code{"dgRMatrix"}.}

\item{obs}{data.table. The data.table representing the observations, i.e.,
cell information.}

\item{var}{data.table. The data.table representing the features, i.e., the
feature information.}

\item{sc_qc_param}{List. Output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}. A
list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of genes to be detected
in the cell to be included.
\item min_lib_size - Integer. Minimum library size in the cell to be
included.
\item min_cells - Integer. Minimum number of cells a gene needs to be
detected to be included.
\item target_size - Float. Target size to normalise to. Defaults to \code{1e5}.
}}

\item{batch_size}{Integer. If \code{streaming = TRUE}, how many cells to process
in one batch. Defaults to \code{1000L}.}

\item{streaming}{Boolean. Shall the data be streamed during the conversion
of CSR to CSC. Defaults to \code{TRUE} and should be used for larger data sets.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
It will populate the files on disk and return the class with updated
shape information.
}
\description{
This function loads in data directly from R objects.
}
