% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{load_seurat}
\alias{load_seurat}
\title{Load in Seurat to \code{single_cell_exp}}
\usage{
load_seurat(
  object,
  seurat,
  sc_qc_param = params_sc_min_quality(),
  batch_size = 1000L,
  streaming = TRUE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{seurat}{\code{Seurat} class you want to transform.}

\item{sc_qc_param}{List. Output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}. A
list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of genes to be detected
in the cell to be included.
\item min_lib_size - Integer. Minimum library size in the cell to be
included.
\item min_cells - Integer. Minimum number of cells a gene needs to be
detected to be included.
\item target_size - Float. Target size to normalise to. Defaults to \code{1e5}.
}}

\item{batch_size}{Integer. If \code{streaming = TRUE}, how many cells to process
in one batch. Defaults to \code{1000L}.}

\item{streaming}{Boolean. Shall the data be streamed during the conversion
of CSR to CSC. Defaults to \code{TRUE} and should be used for larger data sets.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
It will populate the files on disk and return the class with updated
shape information.
}
\description{
This function takes a Seurat file and generates \code{single_cell_exp} class
from it.
}
