% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_meta_cell.R
\name{mc_counts_to_list}
\alias{mc_counts_to_list}
\title{Transform the counts to a Rust-specific list}
\usage{
mc_counts_to_list(object, assay = c("raw", "norm"))
}
\arguments{
\item{object}{\code{MetaCells} class.}

\item{assay}{String. One of \code{c("raw", "norm")}}
}
\description{
Helper function to transform the counts from the \code{MetaCells} into a Rust
specific list.
}
\keyword{internal}
