% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_meta_cell.R
\name{mc_counts_to_list}
\alias{mc_counts_to_list}
\title{Transform the counts to a Rust-specific list}
\usage{
mc_counts_to_list(
  object,
  cell_indices = NULL,
  gene_indices = NULL,
  assay = c("raw", "norm")
)
}
\arguments{
\item{object}{\code{MetaCells} class.}

\item{cell_indices}{Optional integer. Defines the indices of the (meta)cells
to extract.}

\item{gene_indices}{Optional integer. Defines the indices of the genes to
extract.}

\item{assay}{String. One of \code{c("raw", "norm")}}
}
\description{
Helper function to transform the counts from the \code{MetaCells} into a Rust
specific list.
}
\keyword{internal}
