% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_meta_cell.R
\name{mc_get_clr_offsets}
\alias{mc_get_clr_offsets}
\title{Get the offsets for the CLR/PFlogPF transformation prior PCA}
\usage{
mc_get_clr_offsets(object, cell_indices = NULL)
}
\arguments{
\item{object}{\code{MetaCells} class.}

\item{cell_indices}{Optional integer. Defines the indices of the (meta)cells
to use for the calculation.}
}
\value{
A vector of length cell_indices which contains the CLR offsets
}
\description{
Helper function to get the offsets for the CLR/PFlogPF transformation.
}
\keyword{internal}
