% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_meta_cell.R
\name{meta_cells}
\alias{meta_cells}
\title{bixverse meta cell class}
\usage{
meta_cells(meta_cell_data, var_data, meta_cell_method)
}
\arguments{
\item{meta_cell_data}{Named list. Output of meta-cell generation Rust
functions. Will contain the aggregated raw and normalised counts, plus
additional information on the origin of the meta cells.}
}
\value{
Returns the \code{meta_cells} class for further operations.
}
\description{
This is the \code{bixverse}-based meta cell class. It is generated by running
meta-cell generating functions on top of the main class. Due to the
aggregation of the count matrices to a way smaller number of meta cells,
the data is held in memory for this class specifically. Additional methods
that make specific usage of meta cells (especially co-expression module
detection ones will be made available).
}
\section{Properties}{

\describe{
\item{obs_table}{The meta cell observation table.}
\item{var_table}{The meta cell variable table.}
\item{data}{List with the raw and normalised counts.}
\item{original_assignment}{List with original assignment information.}
\item{dims}{Dimensions of the new meta cell matrices.}
\item{other_data}{Potential other data returned from the meta-cell
generating methods.}
}
}

