% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_analysis.R
\name{module_scores_sc}
\alias{module_scores_sc}
\title{Calculate module activity scores}
\usage{
module_scores_sc(
  object,
  gs_list,
  n_bins = 24L,
  n_ctrl = 100L,
  seed = 42L,
  streaming = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{gs_list}{Named list. The elements have the gene identifiers of the
respective gene sets.}

\item{n_bins}{Integer. Number of bins to use. Defaults to \code{24L}.}

\item{n_ctrl}{Integer. Number of control genes to use per gene for each gene
set. Defaults to \code{100L}.}

\item{seed}{Integer. The random seed.}

\item{streaming}{Boolean. Shall the cell and gene data be streamed in.
Useful for larger data sets.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
description
}
\description{
Implements the approach from Tirosh et al into bixverse, i.e., the
AddModuleScore functionality from Seurat. For each module (gene set),
computes the average expression of genes in the set minus the average
expression of randomly selected control genes from the same expression bins.
Genes are binned based on their average expression across cells to ensure
controls are expression-matched.
}
\references{
Tirosh et al, Science (2016)
}
