% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{network_diffusions}
\alias{network_diffusions}
\title{Network diffusion class}
\usage{
network_diffusions(edge_data_frame, weighted, directed)
}
\arguments{
\item{edge_data_frame}{data.table that contains the edge information. It is
expected to have the columns 'from' and 'to'.}

\item{weighted}{Boolean. Is the graph weighted. If set to TRUE, the
\code{edge_data_frame} needs to have a weight column.}

\item{directed}{Boolean. Shall the graph be stored as directed.}
}
\value{
Returns the \code{network_diffusions} class for further operations.
}
\description{
This class helps to do diffusion of seed nodes in a single or tied version
over a network, measure the ability of these diffusion vectors to recall
against a gold standard set of nodes and do community detection within the
subset of the network that received the the most heat from the initial seed
genes.
}
\section{Properties}{

\describe{
\item{graph}{igraph. The underlying graph.}
\item{diffusion_res}{Numeric vector. Contains contains the single or tied
diffusion results.}
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{data.table. Contains final results.}
}
}

