% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_diff_class.R
\name{network_diffusions}
\alias{network_diffusions}
\title{Network diffusion class}
\usage{
network_diffusions(edge_data_frame, weighted, directed)
}
\arguments{
\item{edge_data_frame}{data.table with the columns}

\item{weighted}{Boolean. Is the graph weighted.}

\item{directed}{Boolean. Shall the graph be stored as directed.}
}
\value{
Returns the S7 object for further operations.
}
\description{
...
}
