% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_multimodal.R
\name{new_adt_counts_clr}
\alias{new_adt_counts_clr}
\title{Generates a new \code{ADTCounts} class}
\usage{
new_adt_counts_clr(
  raw_counts,
  cell_info,
  clean_clr_counts = TRUE,
  percentile = 0.01
)
}
\arguments{
\item{raw_counts}{Numeric matrix. The raw ADT counts.}

\item{cell_info}{Named integer vector. Output of \code{\link[=get_cell_info]{get_cell_info()}}. Defines
as elements the cell indices (R-based) and as names the barcodes.}

\item{clean_clr_counts}{Boolean. Shall the per-protein 1st percentile be
removed from the CLR normalised counts.}

\item{percentile}{Numeric. The percentile to remove to reduce background
effects.}
}
\value{
\code{ADTCounts} that contains the raw and normalised ADT counts.
}
\description{
This function generates a new \code{ADTCounts} class which uses CLR normalisation
under the hood.
}
