% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_add.R
\name{new_sc_knn}
\alias{new_sc_knn}
\title{Helper function to generate kNN data with distances}
\usage{
new_sc_knn(knn_data, used_cells)
}
\arguments{
\item{knn_data}{Named list with the following items:
\itemize{
\item indices - Integer matrix containing the indices of the nearest
neighbours (0-indexed).
\item dist - Numerical matrix containing the distances to the nearest
neighbours.
\item dist_metric - String. Distance metric used.
}}

\item{used_cells}{Character vector. The cells used to generate the kNN graph
with the distances.}
}
\value{
Generates the \code{sc_knn} class.
}
\description{
Wrapper class that stores kNN data for subsequent usage in various other
functions, methods and helpers.
}
