% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{new_sc_miloR_res}
\alias{new_sc_miloR_res}
\title{Generate a new miloR result}
\usage{
new_sc_miloR_res(nhoods, sample_counts, spatial_dist, params)
}
\arguments{
\item{nhoods}{Sparse dgCMatrix with cells x neighbourhoods.}

\item{sample_counts}{Integer matrix. Represents neighbourhoods x cells from
sample of interest.}

\item{spatial_dist}{Numeric. The spatial distance between the neighbourhoods
to calculate the spatial FDR.}

\item{params}{Named list. The parameters that were used to generate these
results.}
}
\value{
An \code{sc_miloR} class that contains the provided data and has
subsequent methods to calculate differential abundance statistics.
}
\description{
This is a helper class that wraps the miloR results together. The general
idea of the approach is to use the kNN graph generated in the single cell
data, generate representative neighbourhoods and calculate differential
abundances within these neighbourhoods. For further details on the method,
please refer to Dann, et al.
}
\references{
Dann, et al., Nat Biotechnol, 2022
}
