% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{normalise_bulk_dge}
\alias{normalise_bulk_dge}
\title{Normalise the count data for DGE.}
\usage{
normalise_bulk_dge(
  object,
  group_col,
  norm_method = c("TMM", "TMMwsp", "RLE", "upperquartile", "none"),
  calc_tpm = FALSE,
  calc_fpkm = FALSE,
  gene_lengths = NULL,
  .verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_dge]{bulk_dge()}}.}

\item{group_col}{String. The column in the metadata that will contain the
contrast groups. Needs to be part of the metadata stored in the class.}

\item{norm_method}{String. One of
\code{c("TMM", "TMMwsp", "RLE", "upperquartile", "none")}. Please refer to
\code{\link[edgeR:calcNormFactors]{edgeR::normLibSizes()}}.}

\item{calc_tpm}{Boolean. Output TPM calculation (default = FALSE).}

\item{calc_fpkm}{Boolean. Output FPKM calculation (default = FALSE).}

\item{gene_lengths}{Optional named numeric. If you want to calculate TPM
or FPKM you need to provide this one. The names need to be the same
identifier as used in the counts.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
Returns the class with the \code{processed_data} data slot populated and
applied parameters added to the \code{params} slot.
}
\description{
This function will apply the CPM + Voom normalisation and can additionally
calculate TPM and FPKM values for plotting purposes.
}
