% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_boost}
\alias{params_boost}
\title{Wrapper function for Boost parameters}
\usage{
params_boost(
  boost_rate = 0.25,
  replace = FALSE,
  resolution = 1,
  n_iters = 20L,
  p_thresh = 1e-07,
  voter_thresh = 0.9,
  fast_cluster = FALSE,
  normalisation = list(),
  hvg = list(),
  pca = list(),
  knn = list(k = 0L),
  fast_cluster_params = list()
)
}
\arguments{
\item{boost_rate}{Numeric. Boosting rate for the algorithm. Must be between 0
and 1. Defaults to \code{0.25}.}

\item{replace}{Boolean. Whether to use replacement during boosting. Defaults
to \code{FALSE}.}

\item{resolution}{Numeric. Resolution parameter for graph-based clustering.
Higher values lead to more clusters. Defaults to \code{1.0}.}

\item{n_iters}{Integer. Number of iterations to run the algorithm. Defaults
to \code{20L}.}

\item{p_thresh}{Numeric. P-value threshold for significance testing. Defaults
to \code{1e-07}.}

\item{voter_thresh}{Numeric. Voter threshold across iterations. Proportion of
iterations a cell must be assigned to a cluster to be considered a member.
Must be between 0 and 1. Defaults to \code{0.9}.}

\item{fast_cluster}{Boolean. Shall fast Louvain clustering be applied, i.e.,
k-means clustering and use the centroids for kNN graph generation and Louvain
clustering with then backpropagating the membership based on centroid
proximity. Defaults to \code{FALSE}.}

\item{normalisation}{List. Optional overrides for normalisation parameters.
See \code{\link[=params_norm_doublets_defaults]{params_norm_doublets_defaults()}} for available parameters:
\code{log_transform}, \code{mean_center}, \code{normalise_variance}, \code{target_size}. See
\code{\link[=params_norm_doublets_defaults]{params_norm_doublets_defaults()}} for the available elements. Defaults to
\code{list()}.}

\item{hvg}{List. Optional overrides for highly variable gene selection
parameters. See \code{\link[=params_hvg_defaults]{params_hvg_defaults()}} for available parameters:
\code{min_gene_var_pctl}, \code{hvg_method}, \code{loess_span}, \code{clip_max}. See
\code{\link[=params_hvg_defaults]{params_hvg_defaults()}} for the available elements. Defaults to \code{list()}.}

\item{pca}{List. Optional overrides for PCA parameters. See
\code{\link[=params_pca_defaults]{params_pca_defaults()}} for available parameters: \code{no_pcs},
\code{random_svd}. See \code{\link[=params_pca_defaults]{params_pca_defaults()}} for the available elements.
Defaults to \code{list()}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{extract_knn}, \code{m}, \code{ef_construction},
\code{ef_search}, \code{n_list} and \code{n_probe}. Note: this function defaults to \code{k = 0L}
(automatic neighbour detection). See \code{\link[=params_knn_defaults]{params_knn_defaults()}} for the
available elements. Defaults to \code{list(k = 0L)}.}

\item{fast_cluster_params}{List. Optional overrides for the fast clustering
parameters. Only relevant if \code{fast_cluster = TRUE}. See
\code{\link[=params_fast_cluster_default]{params_fast_cluster_default()}} for available parameters: \code{km_type},
\code{n_centroids}, \code{kmeans_iters} and \code{batch_size}. See
\code{\link[=params_fast_cluster_default]{params_fast_cluster_default()}} for the available elements. Defaults to
\code{list()}.}
}
\value{
A named list with the following elements:
\itemize{
\item The elements of \code{\link[=params_norm_doublets_defaults]{params_norm_doublets_defaults()}}, overridden by
\code{normalisation}, spliced in at this position.
\item The elements of \code{\link[=params_hvg_defaults]{params_hvg_defaults()}}, overridden by \code{hvg}, spliced
in at this position.
\item The elements of \code{\link[=params_pca_defaults]{params_pca_defaults()}}, overridden by \code{pca}, spliced
in at this position.
\item The elements of \code{\link[=params_knn_defaults]{params_knn_defaults()}}, overridden by \code{knn}, spliced
in at this position.
\item The elements of \code{\link[=params_fast_cluster_default]{params_fast_cluster_default()}}, overridden by
\code{fast_cluster_params}, spliced in at this position.
\item boost_rate - Numeric. Boosting rate for the algorithm. Must be between
0 and 1. Defaults to \code{0.25}.
\item replace - Boolean. Whether to use replacement during boosting.
Defaults to \code{FALSE}.
\item resolution - Numeric. Resolution parameter for graph-based clustering.
Higher values lead to more clusters. Defaults to \code{1.0}.
\item fast_cluster - Boolean. Shall fast Louvain clustering be applied,
i.e., k-means clustering and use the centroids for kNN graph generation and
Louvain clustering with then backpropagating the membership based on
centroid proximity. Defaults to \code{FALSE}.
\item n_iters - Integer. Number of iterations to run the algorithm. Defaults
to \code{20L}.
\item p_thresh - Numeric. P-value threshold for significance testing.
Defaults to \code{1e-07}.
\item voter_thresh - Numeric. Voter threshold across iterations. Proportion
of iterations a cell must be assigned to a cluster to be considered a
member. Must be between 0 and 1. Defaults to \code{0.9}.
}
}
\description{
Wrapper function for Boost parameters
}
