% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_community_detection}
\alias{params_community_detection}
\title{Wrapper function to generate community detection parameters}
\usage{
params_community_detection(
  max_nodes = 300L,
  min_nodes = 10L,
  min_seed_nodes = 2L,
  initial_res = 0.5,
  threshold_type = c("prop_based", "pval_based"),
  network_threshold = 0.5,
  pval_threshold = 0.1
)
}
\arguments{
\item{max_nodes}{Integer. Maximum number of nodes in a given community.}

\item{min_nodes}{Integer. Minimum number of nodes in a given community.}

\item{min_seed_nodes}{Integer. Minimum number of seed nodes within a
community.}

\item{initial_res}{Float. Initial resolution parameter to start with.}

\item{threshold_type}{String. One of \code{c("prop_based", "pval_based")}.
You can chose to include a certain proportion of the network with the highest
diffusion scores, or use p-values based on permutations.}

\item{network_threshold}{Float. The proportion of the network to include.
Used if \code{threshold_type = "prop_based"}.}

\item{pval_threshold}{Float. The maximum p-value for nodes to be included.
Used if \code{threshold_type = "pval_based"}.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function to generate community detection parameters
}
