% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_cor_graph}
\alias{params_cor_graph}
\title{Wrapper function for graph generation}
\usage{
params_cor_graph(
  epsilon = 2,
  min_cor = 0.2,
  fdr_threshold = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{epsilon}{Float. Defines the epsilon parameter for the radial basis
function. Defaults to 2, but should be ideally optimised.}

\item{min_cor}{Float. Minimum absolute correlation that needs to be
observed in either data set. Only relevant for differential correlation-based
graphs.}

\item{fdr_threshold}{Float. Maximum FDR for the differential correlation
p-value.}

\item{verbose}{Boolean. Controls verbosity of the graph generation function.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function for graph generation
}
