% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_coremo}
\alias{params_coremo}
\title{Wrapper function to generate CoReMo parameters}
\usage{
params_coremo(
  epsilon = 2,
  k_min = 2L,
  k_max = 150L,
  min_size = NULL,
  junk_module_threshold = 0.05,
  rbf_func = c("gaussian", "inverse_quadratic", "bump"),
  cor_method = c("spearman", "pearson")
)
}
\arguments{
\item{epsilon}{Float. Epsilon parameter for the chosen RBF function, see
\code{rbf_func}. The higher, the more aggressively low correlations will be
shrunk.}

\item{k_min, k_max}{Integer. Minimum and maximum number of cuts to use for the
hierarchical clustering.}

\item{min_size}{Optional integer. Minimum size of the clusters. Smaller
clusters will be combined together.}

\item{junk_module_threshold}{Float. Threshold for the minimum correlation
to be observed in a module. Defaults to \code{0.05}.}

\item{rbf_func}{String. Type of RBF you wish to apply to down-weigh weak
correlations. Defaults to \code{"gaussian"}.}

\item{cor_method}{String. The type of correlation to use. Defaults to
\code{"spearman"}.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function to generate CoReMo parameters
}
