% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_dgrdl}
\alias{params_dgrdl}
\title{Wrapper function to generate DGRDL parameters}
\usage{
params_dgrdl(
  sparsity = 5L,
  dict_size = 5L,
  alpha = 1,
  beta = 1,
  max_iter = 20L,
  k_neighbours = 5L,
  admm_iter = 5L,
  rho = 1
)
}
\arguments{
\item{sparsity}{Integer. Sparsity constraint (max non-zero coefficients per
signal)}

\item{dict_size}{Integer. Dictionary size}

\item{alpha}{Float. Sample context regularisation weight.}

\item{beta}{Float. Feature effect regularisation weight.}

\item{max_iter}{Integer. Maximum number of iterations for the main algorithm.}

\item{k_neighbours}{Integer. Number of neighbours in the KNN graph.}

\item{admm_iter}{Integer. ADMM iterations for sparse coding.}

\item{rho}{Float. ADMM step size.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function to generate DGRDL parameters
}
