% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_fast_cluster_default}
\alias{params_fast_cluster_default}
\title{Helper function to generate default parameters for the fast clustering for
the doublet detection methods}
\usage{
params_fast_cluster_default()
}
\value{
A list with the following parameters for fast clustering:
\itemize{
\item km_type - The type of k-means clustering. Defaults to \code{"minibatch"}
\item n_centroids - The number of centroids to use. Default to \code{NULL} and
the function will use \code{sqrt(N_cells) * 4} for the number of n_centroids.
\item kmeans_iters - Number of maximum k-means iterations. Defaults to
\code{100L}
\item batch_size - Max batch size will be set to \code{4098L}, but pending data
set set to \code{N_cells / 2}.
}
}
\description{
Helper function to generate default parameters for the fast clustering for
the doublet detection methods
}
