% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_graph_resolution}
\alias{params_graph_resolution}
\title{Wrapper function to generate resolution parameters for Leiden or Louvain
clustering.}
\usage{
params_graph_resolution(min_res = 0.1, max_res = 10, number_res = 15L)
}
\arguments{
\item{min_res}{Float. Minimum resolution to test.}

\item{max_res}{Float. Maximum resolution to test.}

\item{number_res}{Integer. Number of resolutions to test between the \code{max_res}
and \code{min_res.}}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function to generate resolution parameters for Leiden or Louvain
clustering.
}
