% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_gsva}
\alias{params_gsva}
\title{Wrapper function to generate GSVA parameters}
\usage{
params_gsva(
  tau = 1,
  min_size = 5L,
  max_size = 500L,
  max_diff = TRUE,
  abs_rank = FALSE
)
}
\arguments{
\item{tau}{Float. Tau parameter, usual recommendation is to use \code{1.0} here.
Larger values emphasise the tails more.}

\item{min_size}{Integer. Minimum number of genes per gene set.}

\item{max_size}{Integer. Maximum number of genes per gene set.}

\item{max_diff}{Boolean. Scoring mode for GSVA, if \code{TRUE} = difference; if
\code{FALSE} = larger absolute value.}

\item{abs_rank}{Boolean. If \code{TRUE} = pos - neg, \code{FALSE} = pos + neg.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function to generate GSVA parameters
}
