% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_hvg_defaults}
\alias{params_hvg_defaults}
\title{Helper function to generate HVG defaults}
\usage{
params_hvg_defaults()
}
\value{
A list with default parameters for kNN searches. Following
parameters:
\itemize{
\item min_gene_var_pctl - Which percentile of the highly variable genes
to include. Defaults to \code{0.7}.
\item hvg_method - Which method to use to identify HVG. Defaults to \code{"vst"}.
\item loess_span - In case of
\item search_budget - Search budget per tree for Annoy. Defaults to \code{100L}.
\item n_trees - Number of trees to generate for Annoy. Defaults to \code{100L}.
}
}
\description{
Helper function to generate HVG defaults
}
