% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_ica_ncomp}
\alias{params_ica_ncomp}
\title{Wrapper function for ICA ncomp iterations}
\usage{
params_ica_ncomp(max_no_comp = 75L, steps = 5L, custom_seq = NULL)
}
\arguments{
\item{max_no_comp}{Integer. Maximum number of ncomp to test.}

\item{steps}{Integer. In which steps to move from 5 onwards.}

\item{custom_seq}{An integer vector. If you wish to provide a custom version
of no_comp to iterate through. If NULL, you will iterate through
\verb{c(2, 3, 4, 5, 5 + step, ... max_no_comp - step, max_no_comp)}}
}
\value{
A list with the parameters for usage in subsequent functions.
}
\description{
Wrapper function to provide parameters through which ncomps to
iterate through.
}
