% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_ica_randomisation}
\alias{params_ica_randomisation}
\title{Wrapper function for ICA randomisation}
\usage{
params_ica_randomisation(
  cross_validate = FALSE,
  random_init = 50L,
  folds = 10L
)
}
\arguments{
\item{cross_validate}{Boolean. Do you want to apply a cross-validation type
approach and split the data into \code{folds} folds to assess within data
stability of the component.}

\item{random_init}{Integer. Number of random initialisations to use.}

\item{folds}{Integer. Number of folds to use if \code{cross_validate} is set to
\code{TRUE}. To note, you will be running \code{random_init * folds} ICA runs.}
}
\value{
A list with the parameters for usage in the subsequent functions.
}
\description{
Wrapper function for ICA randomisation
}
