% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_kmeans_defaults}
\alias{params_kmeans_defaults}
\title{K-mean parameter defaults.}
\usage{
params_kmeans_defaults()
}
\value{
A list with the following parameters
\itemize{
\item k_means_iter - Integer. The number of iterations to use for the
clustering.
\item k_means_init - String. The initialisation. Options are \code{"random"} and
\code{"parallel"}. Defaults to \code{"parallel"}.
\item gemm - Optional boolean. Controls which CPU implementation is used
by the method. GEMM is faster with large dimensionality.
\item hamerly - Optional boolean. Shall a faster exact method be used
leveraging the triangle inequality. Faster on large data sets with large
numbers of centroids.
}
}
\description{
Helper function to generate defaults for the k-mean clustering were more
control is needed.
}
