% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_pca_defaults}
\alias{params_pca_defaults}
\title{Helper function to generate default parameters for PCA}
\usage{
params_pca_defaults()
}
\value{
A list with the following parameters for PCA.
\itemize{
\item no_pcs - Integer. Number of PCs to consider. Defaults to \code{30L}.
\item random_svd - Boolean. Shall randomised SVD be used. Defaults to
\code{TRUE}.
\item sparse - Boolean. Shall sparse solvers be used that do not do
scaling. If set to yes, in the case of \code{random_svd = FALSE}, Lanczos
iterations are used to solve the sparse SVD. With \code{random_svd = TRUE}, the
sparse initial matrix is multiplied with the random matrix, yielding a
much smaller dense matrix that does not increase the memory pressure
massively.
}
}
\description{
Helper function to generate default parameters for PCA
}
