% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_pca_defaults}
\alias{params_pca_defaults}
\title{Helper function to generate default parameters for PCA}
\usage{
params_pca_defaults()
}
\value{
A list with the following parameters for PCA.
\itemize{
\item no_pcs - Integer. Number of PCs to consider. Defaults to \code{30L}.
\item random_svd - Boolean. Shall randomised SVD be used. Defaults to
\code{TRUE}.
}
}
\description{
Helper function to generate default parameters for PCA
}
