% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_bbknn}
\alias{params_sc_bbknn}
\title{Wrapper function for the BBKNN parameters}
\usage{
params_sc_bbknn(
  neighbours_within_batch = 3L,
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  trim = NULL,
  knn = list()
)
}
\arguments{
\item{neighbours_within_batch}{Integer. Number of neighbours to consider
per batch. Defaults to \code{3L}.}

\item{set_op_mix_ratio}{Numeric. Mixing ratio between union (1.0) and
intersection (0.0). Defaults to \code{1.0}.}

\item{local_connectivity}{Numeric. UMAP connectivity computation parameter,
how many nearest neighbours of each cell are assumed to be fully connected.
Defaults to \code{1.0}.}

\item{trim}{Optional integer. Trim the neighbours of each cell to these many
top connectivities. May help with population independence and improve the
tidiness of clustering. If \code{NULL}, it defaults to
\code{10 * neighbours_within_batch}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{extract_knn}, \code{m}, \code{ef_construction},
\code{ef_search}, \code{n_list} and \code{n_probe}.}
}
\value{
A list with the BBKNN parameters.
}
\description{
Wrapper function for the BBKNN parameters
}
