% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_bbknn}
\alias{params_sc_bbknn}
\title{Wrapper function for the BBKNN parameters}
\usage{
params_sc_bbknn(
  neighbours_within_batch = 5L,
  knn_method = c("annoy", "hnsw"),
  ann_dist = c("cosine", "euclidean"),
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  annoy_n_trees = 100L,
  search_budget = 100L,
  trim = NULL
)
}
\arguments{
\item{neighbours_within_batch}{Integer. Number of neighbours to consider
per batch.}

\item{knn_method}{String. One of \code{c("annoy", "hnsw")}. Defaults to
\code{"annoy"}.}

\item{ann_dist}{String. One of \code{c("cosine", "euclidean")}. The distance
metric to be used for the approximate neighbour search.}

\item{set_op_mix_ratio}{Numeric. TO BE WRITTEN.}

\item{local_connectivity}{Numeric. TO BE WRITTEN.}

\item{annoy_n_trees}{Integer. Number of trees to use in the generation of
the Annoy index.}

\item{search_budget}{Integer. Search budget per tree for the \code{annoy}
algorithm.}

\item{trim}{Optional integer. TO BE WRITTEN.}
}
\value{
A list with the BBKNN parameters.
}
\description{
Wrapper function for the BBKNN parameters
}
