% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_bt_metacells}
\alias{params_sc_bt_metacells}
\title{Wrapper function for parameters for bootstrapped meta cell generation}
\usage{
params_sc_bt_metacells(
  max_shared = 15L,
  target_no_metacells = 1000L,
  max_iter = 5000L,
  knn = list()
)
}
\arguments{
\item{max_shared}{Integer. Maximum number of allowed shared neighbours for
the meta cell to be considered. Defaults to \code{15L}.}

\item{target_no_metacells}{Integer. Target number of meta-cells to generate.
Defaults to \code{1000L}.}

\item{max_iter}{Integer. Maximum number of iterations for the algorithm.
Defaults to \code{5000L}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{extract_knn}, \code{m}, \code{ef_construction},
\code{ef_search}, \code{n_list} and \code{n_probe}.}
}
\value{
A list with the metacell parameters.
}
\description{
This function generates parameters for the bootstrapped meta cell generation
based on hdWGCNA, see Morabito, et al., Cell Rep. Methods, 2023.
}
