% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_fast_cluster}
\alias{params_sc_fast_cluster}
\title{Fast single cell clustering parameters}
\usage{
params_sc_fast_cluster(
  kmeans_iters = 100L,
  batch_size = 4096L,
  drift_threshold = 1e-04,
  lr_alpha = 1,
  full_snn = FALSE,
  pruning = NULL,
  snn_similarity = c("jaccard", "rank"),
  louvain_iters = 10L,
  knn = list(k = 5L)
)
}
\arguments{
\item{kmeans_iters}{Integer. Number of iterations for k-means clustering.}

\item{batch_size}{Integer. Batch size for mini batch k-means clustering.}

\item{drift_threshold}{Numeric. The drift for the mini batch k-means
clustering. If the centroid drift is below this, the mini batch k-means
terminates.}

\item{lr_alpha}{Numeric. Learning rate alpha parameter for mini batch
k-means.}

\item{full_snn}{Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours.}

\item{pruning}{Optional numeric. Weights below this threshold will be set to
0 in the generation of the sNN graph. If not provided, defaults to
\code{1 / ceil(k * 0.8)}.}

\item{snn_similarity}{String. One of \code{c("rank", "jaccard")}. The Jaccard
similarity calculates the Jaccard between the neighbours, whereas the rank
method calculates edge weights based on the ranking of shared neighbours.
For the rank method, the weight is determined by finding the shared
neighbour with the lowest combined rank across both cells, where
lower-ranked (closer) shared neighbours result in higher edge weights
Both methods produce weights normalised to the range \verb{[0, 1]}.}

\item{louvain_iters}{Integer. Number of iterations for Louvain clustering.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{extract_knn}, \code{m}, \code{ef_construction},
\code{ef_search}, \code{n_list} and \code{n_probe}. Sets the default \code{k = 5L}.}
}
\value{
A named list with the single cell fast clustering parameters.
}
\description{
Fast single cell clustering parameters
}
