% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_fastmnn}
\alias{params_sc_fastmnn}
\title{Wrapper function for the fastMNN parameters}
\usage{
params_sc_fastmnn(
  ndist = 3,
  cos_norm = TRUE,
  no_pcs = 30L,
  random_svd = TRUE,
  knn = list(k = 20L)
)
}
\arguments{
\item{ndist}{Numeric. Number of median distances for the tricube kernel
bandwidth. Defaults to \code{3.0}.}

\item{cos_norm}{Logical. Apply cosine normalisation before computing
distances. Defaults to \code{TRUE}.}

\item{no_pcs}{Integer. Number of PCs to use for MNN calculations.
Defaults to \code{30L}.}

\item{random_svd}{Logical. Use randomised SVD. Defaults to \code{TRUE}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{m}, \code{ef_construction}, \code{ef_search}, \code{n_list}
and \code{n_probe}.}
}
\value{
A list with the fastMNN parameters.
}
\description{
Wrapper function for the fastMNN parameters
}
