% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_fastmnn}
\alias{params_sc_fastmnn}
\title{Wrapper function for the fastMNN parameters}
\usage{
params_sc_fastmnn(
  k = 20L,
  sigma = 0.1,
  knn_method = c("annoy", "hnsw"),
  dist_metric = c("cosine", "euclidean"),
  n_trees = 100L,
  search_budget = 100L,
  m = 32L,
  ef_construction = 100L,
  ef_search = 100L,
  cos_norm = TRUE,
  var_adj = TRUE,
  no_pcs = 30L,
  random_svd = TRUE
)
}
\arguments{
\item{k}{Integer. Number of mutual nearest neighbours to identify.
Defaults to \code{20L}.}

\item{sigma}{Numeric. Bandwidth of the Gaussian smoothing kernel
(as proportion of space radius). Defaults to \code{0.1}.}

\item{knn_method}{String. One of \code{c("annoy", "hnsw")}. Defaults to
\code{"annoy"}.}

\item{dist_metric}{String. One of \code{c("cosine", "euclidean")}. Defaults to
\code{"cosine"}.}

\item{n_trees}{Integer. Number of trees to use in the generation of
the Annoy index. Relevant whe \code{knn_method = "annoy"}.}

\item{search_budget}{Integer. Search budget per tree for the \code{annoy}
algorithm.}

\item{m}{Integer. Number of connections between layers for HNSW. Relevant
when \code{knn_method = "hnsw"}.}

\item{ef_construction}{Integer. Size of dynamic candidate list during
construction. Relevant when \code{knn_method = "hnsw"}.}

\item{ef_search}{Integer. Size of candidate list (higher = better recall,
slower). Relevant when \code{knn_method = "hnsw"}.}

\item{cos_norm}{Logical. Apply cosine normalisation before computing
distances. Defaults to \code{TRUE}.}

\item{var_adj}{Logical. Apply variance adjustment to avoid kissing effects.
Defaults to \code{TRUE}.}

\item{no_pcs}{Integer. Number of PCs to use for MNN calculations.
Defaults to \code{30L}.}

\item{random_svd}{Logical. Use randomised SVD. Defaults to \code{TRUE}.}
}
\value{
A list with the fastMNN parameters.
}
\description{
Wrapper function for the fastMNN parameters
}
