% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_hotspot}
\alias{params_sc_hotspot}
\title{Wrapper function for parameters for HotSpot}
\usage{
params_sc_hotspot(
  model = c("danb", "normal", "bernoulli"),
  normalise = TRUE,
  knn = list(ann_dist = "cosine")
)
}
\arguments{
\item{model}{String. Model to use for modelling the GEX. One of
\code{c("danb", "bernoulli", "normal")}. Defaults to \code{"danb"}.}

\item{normalise}{Boolean. Shall the data be normalised. Defaults to \code{TRUE}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{delta},
\code{diversify_prob}, \code{ef_budget}, \code{m}, \code{ef_construction}, \code{ef_search},
\code{n_bits}, \code{n_tables}, \code{max_candidates}, \code{n_centroids} and \code{n_probes}.}
}
\value{
A list with the HotSpot parameters.
}
\description{
Wrapper function for parameters for HotSpot
}
