% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_hvg}
\alias{params_sc_hvg}
\title{Wrapper function to identify highly variable genes in single cell}
\usage{
params_sc_hvg(
  method = "vst",
  loess_span = 0.3,
  num_bin = 20L,
  bin_method = "equal_width"
)
}
\arguments{
\item{method}{String. One of \code{c("vst", "meanvarbin", "dispersion")}.}

\item{loess_span}{Numeric. The span parameter for the loess function that is
used to standardise the variance for \code{method = "vst"}.}

\item{num_bin}{Integer. Not yet implemented.}

\item{bin_method}{String. One of \code{c("equal_width", "equal_freq")}.}
}
\value{
A list with the HVG parameters
}
\description{
Wrapper function to identify highly variable genes in single cell
}
