% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_knn}
\alias{params_sc_knn}
\title{Wrapper function for kNN generation parameters for single cell}
\usage{
params_sc_knn(
  k = 10L,
  n_trees = 100L,
  search_budget = 100L,
  knn_algorithm = c("annoy", "hnsw")
)
}
\arguments{
\item{k}{Integer. Number of neighbours to return.}

\item{n_trees}{Integer. Number of trees to use for the \code{annoy} algorithm.}

\item{search_budget}{Integer. Search budget per tree for the \code{annoy}
algorithm.}

\item{knn_algorithm}{String. One of \code{c("annoy", "hnsw")}. Defaults to
\code{"annoy"}}
}
\value{
A list with the HVG parameters
}
\description{
Wrapper function for kNN generation parameters for single cell
}
