% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_metacells}
\alias{params_sc_metacells}
\title{Wrapper function for parameters for meta cell generation}
\usage{
params_sc_metacells(
  max_shared = 15L,
  target_no_metacells = 1000L,
  max_iter = 5000L,
  k = 25L,
  knn_method = c("hnsw", "annoy"),
  ann_dist = c("cosine", "euclidean"),
  n_trees = 100L,
  search_budget = 100L
)
}
\arguments{
\item{max_shared}{Integer. Maximum number of allowed shared neighbours for
the meta cell to be considered.}

\item{target_no_metacells}{Integer. Target number of meta-cells to generate.}

\item{max_iter}{Integer. Maximum number of iterations for the algorithm.}

\item{k}{Integer. Number of neighbours to return.}

\item{knn_method}{String. One of \code{c("annoy", "hnsw")}. Defaults to
\code{"annoy"}}

\item{ann_dist}{String. One of \code{c("cosine", "euclidean")}. The distance
metric to be used for the approximate neighbour search.}

\item{n_trees}{Integer. Number of trees to use for the \code{annoy} algorithm.}

\item{search_budget}{Integer. Search budget per tree for the \code{annoy}
algorithm.}
}
\value{
A list with the neighbour parameters.
}
\description{
Wrapper function for parameters for meta cell generation
}
