% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_miloR}
\alias{params_sc_miloR}
\title{Wrapper function for parameters for MiloR}
\usage{
params_sc_miloR(
  prop = 0.2,
  k_refine = 20L,
  refinement_strategy = c("index", "approximate", "bruteforce"),
  index_type = c("annoy", "hnsw"),
  knn = list()
)
}
\arguments{
\item{prop}{Numeric. Proportion of cells to sample as neighbourhood indices.
Defaults to \code{0.2}. Must be in (0,1).}

\item{k_refine}{Integer. Number of neighbours to use for refinement.
Defaults to \code{20L}.}

\item{refinement_strategy}{String. Strategy for refining sampled indices.
One of \code{c("approximate", "bruteforce", "index")}. Defaults to
\code{"index"}.}

\item{index_type}{String. Type of kNN index to use. One of
\code{c("annoy", "hnsw")}. Defaults to \code{"annoy"}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{nn_max_iter},
\code{rho}, \code{delta}. Note: \code{knn_method} cannot be \code{"nndescent"} for MiloR as
it doesn't generate an index!}
}
\value{
A list with the MiloR parameters.
}
\description{
Wrapper function for parameters for MiloR
}
