% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_min_quality}
\alias{params_sc_min_quality}
\title{Wrapper function to generate QC metric params for single cell}
\usage{
params_sc_min_quality(
  min_unique_genes = 100L,
  min_lib_size = 250L,
  min_cells = 10L,
  target_size = 1e+05
)
}
\arguments{
\item{min_unique_genes}{Integer. Minimum number of unique genes per cell/spot
to be included.}

\item{min_lib_size}{Integer. Minimum library size per cell/spot to be
included.}

\item{min_cells}{Integer. Minimum number of cells a gene has to be
expressed to be included.}

\item{target_size}{Float. The target size for the normalisation. Defaults
to \code{1e5}.}
}
\value{
A list with the minimum quality parameters + target size.
}
\description{
Wrapper function to generate QC metric params for single cell
}
