% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_mtx_io}
\alias{params_sc_mtx_io}
\title{Wrapper function to provide data for mtx-based loading}
\usage{
params_sc_mtx_io(path_mtx, path_obs, path_var, cells_as_rows)
}
\arguments{
\item{path_mtx}{String. Path to the .mtx file}

\item{path_obs}{String. Path to the file containing cell/barcode info.}

\item{path_var}{String. Path to the file containing gene/variable info.}

\item{cells_as_rows}{Boolean. Do cells represent the rows or columns.}
}
\value{
A list with the mtx loading parameters for usage in subsequent
functions.
}
\description{
Wrapper function to provide data for mtx-based loading
}
