% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_neighbours}
\alias{params_sc_neighbours}
\title{Wrapper function for parameters for neighbour identification in single cell}
\usage{
params_sc_neighbours(
  full_snn = FALSE,
  pruning = 1/15,
  snn_similarity = c("rank", "jaccard"),
  knn = list()
)
}
\arguments{
\item{full_snn}{Boolean. Shall the full shared nearest neighbour graph
be generated that generates edges between all cells instead of between
only neighbours.}

\item{pruning}{Numeric. Weights below this threshold will be set to 0 in
the generation of the sNN graph.}

\item{snn_similarity}{String. One of \code{c("rank", "jaccard")}. The Jaccard
similarity calculates the Jaccard between the neighbours, whereas the rank
method calculates edge weights based on the ranking of shared neighbours.
For the rank method, the weight is determined by finding the shared
neighbour with the lowest combined rank across both cells, where
lower-ranked (closer) shared neighbours result in higher edge weights
Both methods produce weights normalised to the range \verb{[0, 1]}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{nn_max_iter},
\code{rho}, \code{delta}. Note: This function uses \code{k = 15L}, \code{ann_dist = "cosine"},
and \code{nn_max_iter = 25L} as defaults.}
}
\value{
A list with the neighbour parameters.
}
\description{
Wrapper function for parameters for neighbour identification in single cell
}
