% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_pca}
\alias{params_sc_pca}
\title{Wrapper for PCA specifically designed for single cells}
\usage{
params_sc_pca(
  mean_center = TRUE,
  normalise_variance = TRUE,
  randomised = TRUE,
  clr = FALSE,
  size_factor = 10000
)
}
\arguments{
\item{mean_center}{Boolean. Shall the data be mean centered}

\item{normalise_variance}{Boolean. Shall the data have normalised variance}

\item{randomised}{Boolean. Shall fast, approximate randomised SVD be used.}

\item{clr}{Boolean. Shall the CLR-type \code{PFlogPF} be applied, see Booeshaghi,
et al.}

\item{size_factor}{Numeric. The used size factor during I/O. It needs to be
the same as during I/O to have correct results when using the \code{PFlogPF}
transformation.}
}
\value{
A list with the parameters
}
\description{
Wrapper for PCA specifically designed for single cells
}
