% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_seacells}
\alias{params_sc_seacells}
\title{Wrapper function for the SEACells parameters}
\usage{
params_sc_seacells(
  n_sea_cells,
  max_fw_iters = 50L,
  convergence_epsilon = 0.001,
  max_iter = 100L,
  min_iter = 10L,
  greedy_threshold = 20000L,
  graph_building = "union",
  pruning = FALSE,
  pruning_threshold = 1e-07,
  k = 25L,
  knn_method = c("annoy", "hnsw", "nndescent"),
  n_trees = 100L,
  search_budget = 100L,
  nn_max_iter = 15L,
  rho = 1,
  delta = 0.001
)
}
\arguments{
\item{n_sea_cells}{Integer. Number of SEA cells to detect.}

\item{max_fw_iters}{Integer. Maximum iterations for the Franke-Wolfe
algorithm. Defaults to \code{50L}.}

\item{convergence_epsilon}{Numeric. Convergence threshold. Algorithm stops
when RSS change < epsilon * RSS(0). Defaults to \code{1e-3}.}

\item{max_iter}{Integer. Maximum iterations to run SEACells for. Defaults to
\code{100L}.}

\item{min_iter}{Integer. Minimum iterations to run SEACells for. Defaults to
\code{10L}.}

\item{greedy_threshold}{Integer. Maximum number of cells before defaulting to
rapid random selection of archetypes. Defaults to \code{20000L}.}

\item{graph_building}{String. Graph building method. Defaults to \code{"union"}.}

\item{pruning}{Boolean. Shall tiny values be pruned during Franke-Wolfe
updates. This will reduce memory pressure and can be a good option on
large data sets. Defaults to \code{FALSE}.}

\item{pruning_threshold}{Float. If \code{pruning = TRUE} values below which
threshold shall be pruned.}

\item{k}{Integer. Number of neighbours for the kNN algorithm. Defaults to
\code{25L}.}

\item{knn_method}{String. One of \code{c("annoy", "hnsw", "nndescent")}. Defaults
to \code{"annoy"}.}

\item{n_trees}{Integer. Number of trees for Annoy index. Defaults to \code{100L}.}

\item{search_budget}{Integer. Search budget during querying. Defaults to
\code{100L}.}

\item{nn_max_iter}{Integer. Maximum iterations for NN Descent. Defaults to
\code{15L}.}

\item{rho}{Numeric. Sampling rate for NN Descent. Defaults to \code{1.0}.}

\item{delta}{Numeric. Early termination criterion for NN Descent. Defaults to
\code{0.001}.}
}
\value{
A list with the SEACells parameters.
}
\description{
Wrapper function for the SEACells parameters
}
