% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_supercell}
\alias{params_sc_supercell}
\title{Wrapper function for parameters for SuperCell generation}
\usage{
params_sc_supercell(
  walk_length = 3L,
  graining_factor = 20,
  linkage_dist = c("complete", "average"),
  k = 5L,
  knn_method = c("annoy", "hnsw", "nndescent"),
  ann_dist = c("cosine", "euclidean"),
  n_trees = 100L,
  search_budget = 100L,
  nn_max_iter = 15L,
  rho = 1,
  delta = 0.001
)
}
\arguments{
\item{walk_length}{Integer. Walk length for the Walktrap algorithm. Defaults
to \code{3L}.}

\item{graining_factor}{Numeric. Graining level of data (proportion of number
of single cells in the initial dataset to the number of metacells in the
final dataset). Defaults to \code{20.0}. (One meta cell per 20 cells.)}

\item{linkage_dist}{String. Which type of distance metric to use for the
linkage. Defaults to \code{"average"}.}

\item{k}{Integer. Number of neighbours to return. Defaults to \code{5L}.}

\item{knn_method}{String. One of \code{c("annoy", "hnsw", "nndescent")}. Defaults
to \code{"annoy"}.}

\item{ann_dist}{String. One of \code{c("cosine", "euclidean")}. The distance
metric to be used for the approximate neighbour search. Defaults to
\code{"cosine"}.}

\item{n_trees}{Integer. Number of trees to use for the \code{annoy} algorithm.
Defaults to \code{100L}.}

\item{search_budget}{Integer. Search budget per tree for the \code{annoy}
algorithm. Defaults to \code{100L}.}

\item{nn_max_iter}{Integer. Maximum iterations for NN Descent. Defaults to
\code{15L}.}

\item{rho}{Numeric. Sampling rate for NN Descent. Defaults to \code{1.0}.}

\item{delta}{Numeric. Early termination criterion for NN Descent. Defaults to
\code{0.001}.}
}
\value{
A list with the SuperCell parameters.
}
\description{
Wrapper function for parameters for SuperCell generation
}
