% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_supercell}
\alias{params_sc_supercell}
\title{Wrapper function for parameters for SuperCell generation}
\usage{
params_sc_supercell(
  walk_length = 3L,
  graining_factor = 20,
  linkage_dist = c("complete", "average"),
  knn = list()
)
}
\arguments{
\item{walk_length}{Integer. Walk length for the Walktrap algorithm. Defaults
to \code{3L}.}

\item{graining_factor}{Numeric. Graining level of data (proportion of number
of single cells in the initial dataset to the number of metacells in the
final dataset). Defaults to \code{20.0}. (One meta cell per 20 cells.)}

\item{linkage_dist}{String. Which type of distance metric to use for the
linkage. Defaults to \code{"average"}.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{nn_max_iter},
\code{rho}, \code{delta}. Note: This function uses \code{k = 5L} and \code{ann_dist = "cosine"}
as defaults.}
}
\value{
A list with the SuperCell parameters.
}
\description{
Wrapper function for parameters for SuperCell generation
}
