% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_synthetic_data}
\alias{params_sc_synthetic_data}
\title{Default parameters for generation of synthetic data}
\usage{
params_sc_synthetic_data(
  n_cells = 1000L,
  n_genes = 100L,
  n_batches = 1L,
  marker_genes = list(cell_type_1 = list(marker_genes = 0:9L), cell_type_2 =
    list(marker_genes = 10:19L), cell_type_3 = list(marker_genes = 20:29L)),
  batch_effect_strength = c("strong", "medium", "weak"),
  n_samples = NULL,
  sample_bias = NULL
)
}
\arguments{
\item{n_cells}{Integer. Number of cells.}

\item{n_genes}{Integer. Number of genes.}

\item{n_batches}{Integer. Number of batches.}

\item{marker_genes}{List. A nested list that indicates which gene indices
are markers for which cell.}

\item{batch_effect_strength}{String. One of
\code{c("strong", "medium", "weak")}. The strength of the batch effect to add.}

\item{n_samples}{Optional integer. Shall sample membership be added to the
synthetic data. If you want sample information you need to provide
\code{n_samples} and \code{sample_bias}.}

\item{sample_bias}{Optional string. One of
\code{c("even", "slightly_uneven", "very_uneven")}}
}
\value{
A list with the parameters.
}
\description{
Default parameters for generation of synthetic data
}
