% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_synthetic_data}
\alias{params_sc_synthetic_data}
\title{Default parameters for generation of synthetic data}
\usage{
params_sc_synthetic_data(
  n_cells = 1000L,
  n_genes = 100L,
  marker_genes = list(cell_type_1 = list(marker_genes = 0:9L), cell_type_2 =
    list(marker_genes = 10:19L), cell_type_3 = list(marker_genes = 20:29L)),
  n_batches = 1L
)
}
\arguments{
\item{n_cells}{Integer. Number of cells.}

\item{n_genes}{Integer. Number of genes.}

\item{marker_genes}{List. A nested list that indicates which gene indices
are markers for which cell.}

\item{n_batches}{Integer. Number of batches.}
}
\value{
A list with the parameters.
}
\description{
Default parameters for generation of synthetic data
}
