% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_sc_vision}
\alias{params_sc_vision}
\title{Wrapper function for parameters for VISION with auto-correlation}
\usage{
params_sc_vision(
  k = 15L,
  knn_algorithm = c("annoy", "hnsw", "nndescent"),
  n_trees = 100L,
  search_budget = 100L,
  max_iter = 25L,
  rho = 1,
  delta = 0.001,
  n_perm = 500L,
  n_cluster = 5L
)
}
\arguments{
\item{k}{Integer. Number of neighbours to return.}

\item{knn_algorithm}{String. One of \code{c("annoy", "hnsw", "nndescent")}.
Defaults to \code{"annoy"}.}

\item{n_trees}{Integer. Number of trees to use for the \code{annoy} algorithm.}

\item{search_budget}{Integer. Search budget per tree for the \code{annoy}
algorithm.}

\item{max_iter}{Integer. Maximum iterations for the \code{nndescent} algorithm.}

\item{rho}{Numeric. Sampling rate for the \code{nndescent} algorithm.}

\item{delta}{Numeric. Early termination criterium for the \code{nndescent}
algorithm.}

\item{n_perm}{Integer. Number of random gene sets to generate per cluster.}

\item{n_cluster}{Integer. Number of clusters for the random gene set
clustering generation.}
}
\value{
A list with the VISION parameters when you wish to use the
auto-correlation version.
}
\description{
Wrapper function for parameters for VISION with auto-correlation
}
