% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_sc_vision}
\alias{params_sc_vision}
\title{Wrapper function for parameters for VISION with auto-correlation}
\usage{
params_sc_vision(n_perm = 500L, n_cluster = 5L, knn = list())
}
\arguments{
\item{n_perm}{Integer. Number of random gene sets to generate per cluster.}

\item{n_cluster}{Integer. Number of clusters for the random gene set
clustering generation.}

\item{knn}{List. Optional overrides for kNN parameters. See
\code{\link[=params_knn_defaults]{params_knn_defaults()}} for available parameters: \code{k},
\code{knn_method}, \code{ann_dist}, \code{search_budget}, \code{n_trees}, \code{nn_max_iter},
\code{rho}, \code{delta}. Note: This function uses \code{k = 15L} and \code{nn_max_iter = 25L}
as defaults.}
}
\value{
A list with the VISION parameters when you wish to use the
auto-correlation version.
}
\description{
Wrapper function for parameters for VISION with auto-correlation
}
