% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_constructors.R
\name{params_scenic}
\alias{params_scenic}
\title{Constructor for SCENIC parameters}
\usage{
params_scenic(
  min_counts = 50L,
  min_cells = 0.03,
  learner_type = "grnboost2",
  gene_batch_strategy = "correlated",
  gene_batch_size = NULL,
  n_pcs = 50L,
  n_subsample = 100000L,
  learner_params = list()
)
}
\arguments{
\item{min_counts}{Integer. Minimum total counts a gene needs to be included
in the analysis. Defaults to \code{50L}.}

\item{min_cells}{Numeric. Minimum proportion of cells (between 0 and 1) that
must express a gene for it to be considered. Defaults to \code{0.03}.}

\item{learner_type}{Character. Regression learner to use. One of
\code{"randomforest"}, \code{"extratrees"}, or \code{"grnboost2"}. Defaults to
\code{"grnboost2"}.}

\item{gene_batch_strategy}{Character. Strategy for grouping target genes into
batches. One of \code{"random"} or \code{"correlated"}. Only used for \code{"randomforest"}
and \code{"extratrees"} learners; ignored for \code{"grnboost2"}. Defaults to
\code{"correlated"}.}

\item{gene_batch_size}{Optional integer. Number of genes per batch. If \code{NULL}
(default), the batch size is determined automatically. Ignored for
\code{"grnboost2"}.}

\item{n_pcs}{Integer. Number of PCs to use for the correlated gene batch
strategy. Defaults to \code{50L}.}

\item{n_subsample}{Integer. Cell subsampling threshold for the correlated
gene batch strategy. If the number of cells meets or exceeds this value,
\code{n_subsample} cells are randomly selected prior to running randomised SVD.
Defaults to \code{100000L}.}

\item{learner_params}{List. Optional overrides for the regression learner
parameters. For \code{"randomforest"}, see
\code{\link[=params_scenic_random_forest_defaults]{params_scenic_random_forest_defaults()}}. For \code{"extratrees"}, see
\code{\link[=params_scenic_extra_trees_defaults]{params_scenic_extra_trees_defaults()}}. For \code{"grnboost2"}, see
\code{\link[=params_scenic_gradient_boosting_defaults]{params_scenic_gradient_boosting_defaults()}}.}
}
\value{
A named flat list with all SCENIC parameters.
}
\description{
Constructor for SCENIC parameters
}
