% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_snf}
\alias{params_snf}
\title{Wrapper function to generate SNF parameters}
\usage{
params_snf(
  k = 20L,
  t = 20L,
  mu = 0.5,
  alpha = 1,
  normalise = TRUE,
  distance_metric = c("euclidean", "manhattan", "canberra", "cosine")
)
}
\arguments{
\item{k}{Integer. Number of neighbours to consider.}

\item{t}{Integer. Number of iterations for the SNF algorithm.}

\item{mu}{Float. Normalisation factor for the Gaussian kernel width.}

\item{alpha}{Float. Normalisation parameter controlling the fusion strength.}

\item{normalise}{Boolean. Shall continuous values be Z-scored.}

\item{distance_metric}{String. One of
\code{c("euclidean", "manhattan", "canberra", "cosine")}. Which distance metric
to use for the continuous calculations. In case of pure categorical, Hamming
will be used, for mixed data types Gower distance is used.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function to generate SNF parameters
}
