% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_ssgsea}
\alias{params_ssgsea}
\title{Wrapper function to generate ssGSEA parameters}
\usage{
params_ssgsea(alpha = 0.25, min_size = 5L, max_size = 500L, normalise = TRUE)
}
\arguments{
\item{alpha}{Float. The exponent defining the weight of the tail in the
random walk performed by ssGSEA.}

\item{min_size}{Integer. Minimum number of genes per gene set.}

\item{max_size}{Integer. Maximum number of genes per gene set.}

\item{normalise}{Boolean. Shall the scores be normalised.}
}
\value{
List with parameters for usage in subsequent function.
}
\description{
Wrapper function to generate ssGSEA parameters
}
