% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{per_cell_qc_outlier}
\alias{per_cell_qc_outlier}
\title{Use MAD outlier detection on per-cell QC metrics}
\usage{
per_cell_qc_outlier(
  metric,
  threshold = 3,
  direction = c("twosided", "below", "above")
)
}
\arguments{
\item{metric}{Numerical vector. The QC metric to check for.}

\item{threshold}{Numeric. How many MADs in either direction to consider for
outlier detection.}

\item{direction}{String. One of \code{c("twosided", "below", "above")}. Which
directionality to consider}
}
\value{
A list with:
\itemize{
\item outlier - Boolean vector indicating which cell is an outlier
\item metrics - The applied thresholds.
}
}
\description{
Use MAD outlier detection on per-cell QC metrics
}
