% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_plots_helpers.R
\name{phate_sc}
\alias{phate_sc}
\title{Run PHATE on a SingleCells object}
\usage{
phate_sc(
  object,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  n_dim = 2L,
  k = 5L,
  knn_method = c("balltree", "hnsw", "annoy", "nndescent", "exhaustive"),
  nn_params = manifoldsR::params_nn(),
  phate_params = manifoldsR::params_phate(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{embd_to_use}{String. The embedding to use for PHATE. Must be available
in the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{n_dim}{Integer. Number of PHATE dimensions. Currently only \code{2L} is
supported. Defaults to \code{2L}.}

\item{k}{Integer. Number of nearest neighbours for graph construction.
Defaults to \code{5L}.}

\item{knn_method}{String. Approximate nearest neighbour algorithm. One of
\code{"balltree"}, \code{"hnsw"}, \code{"annoy"}, \code{"nndescent"}, or \code{"exhaustive"}.}

\item{nn_params}{Named list. See \code{\link[manifoldsR:params_nn]{manifoldsR::params_nn()}}.}

\item{phate_params}{Named list. See \code{\link[manifoldsR:params_phate]{manifoldsR::params_phate()}}.}

\item{seed}{Integer. For reproducibility.}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
The object with a \code{"phate"} embedding added.
}
\description{
Wrapper around \code{\link[manifoldsR:phate]{manifoldsR::phate()}} for the \code{SingleCells} class.
}
