% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{plot.cpca_synthetic_data}
\alias{plot.cpca_synthetic_data}
\title{Plot the contrastive PCA example data}
\usage{
\method{plot}{cpca_synthetic_data}(cpca_synthetic_data, ...)
}
\arguments{
\item{cpca_synthetic_data}{\code{cpca_synthetic_data} class. Output from
\code{\link[=synthetic_c_pca_data]{synthetic_c_pca_data()}}.}

\item{...}{Additional params}
}
\value{
A ggplot showing the two heatmaps from the target and background
matrix.
}
\description{
Plot the contrastive PCA example data
}
