% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{plot.synthetic_matrix_simple}
\alias{plot.synthetic_matrix_simple}
\title{Plot the (simple) synthetic gene expression}
\usage{
\method{plot}{synthetic_matrix_simple}(x, ...)
}
\arguments{
\item{x}{\code{synthetic_matrix_simple} class. Output from
\code{\link[=synthetic_signal_matrix]{synthetic_signal_matrix()}}.}

\item{...}{Additional params}
}
\value{
A plotted heatmap showing the DEG.
}
\description{
Plot the (simple) synthetic gene expression
}
\examples{
\dontrun{

synthetic_gex <- synthetic_signal_matrix()

plot(synthetic_gex)
}
}
