% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{plot_boxplot_normalization}
\alias{plot_boxplot_normalization}
\title{Helper plot function for boxplot of normalized data}
\usage{
plot_boxplot_normalization(samples, voom_object)
}
\arguments{
\item{samples}{data.frame. sample metadata}

\item{voom_object}{Voom object}
}
\value{
ggplot object, i.e., box plot with expression per sample.
}
\description{
Helper plot function for boxplot of normalized data
}
