% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_coexp.R
\name{plot_hvgs}
\alias{plot_hvgs}
\title{Plot the highly variable genes}
\usage{
plot_hvgs(object, bins = 50L)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{bins}{Integer. Number of bins to plot.}
}
\description{
Plots the median-absolute deviation of the genes and applied thresholds.
Expects that \code{\link[=preprocess_bulk_coexp]{preprocess_bulk_coexp()}} was run and will throw an
error otherwise.
}
