% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{plot_ica_ncomp_params}
\alias{plot_ica_ncomp_params}
\title{Plot various parameters with no comp}
\usage{
plot_ica_ncomp_params(object)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. You need to apply
\code{\link[=ica_evaluate_comp]{ica_evaluate_comp()}} before running this function.}
}
\value{
The plot with no comp ~ vs. various parameters.
}
\description{
Helper function to plot various parameters with the no of components. You
have:
\itemize{
\item Median stability - The median stability of the clusters at this
no of components given random initialisations.
\item \% Converged - The percentage of ICA runs that converged at this
no of components.
\item IC Orthogonality - The orthogonality (measured as \code{1 - abs(cos)})
indicating how orthogonal the signals detected at this level are.
\item Combined score - The product of the three other scores.
}
If found, the function will also add the optimal number of components based
on \code{\link[=ica_optimal_ncomp]{ica_optimal_ncomp()}} (if a loess function could be fitted).
}
