% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_ica.R
\name{plot_ica_stability_individual}
\alias{plot_ica_stability_individual}
\title{Plot the stability of the ICA components}
\usage{
plot_ica_stability_individual(object)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}. You need to apply
\code{\link[=ica_evaluate_comp]{ica_evaluate_comp()}} before running this function.}
}
\description{
Helper function to plot the individual stability profiles over the tested
ncomps.
}
