% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{plot_optimal_cuts}
\alias{plot_optimal_cuts}
\title{Plot the k cuts vs median R2}
\usage{
plot_optimal_cuts(object)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
If optimal cuts results were found, returns the ggplot. Otherwise,
throws a warning and returns NULL.
}
\description{
Plots the optimal k vs. median of median R2 graph to identify the optimal
number of cuts.
}
