% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bulk.R
\name{plot_pca}
\alias{plot_pca}
\title{Helper plot function for pca with contrasts}
\usage{
plot_pca(pca_dt, grps)
}
\arguments{
\item{pca_dt}{data.table. Datatable with PCA and contrast information}

\item{grps}{Factor or character vector. The group vector.}
}
\value{
ggplot object for the pca
}
\description{
Helper plot function for pca with contrasts
}
