% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bulk.R
\name{plot_preprocessing_genes}
\alias{plot_preprocessing_genes}
\title{Helper plot function of distribution of genes by samples}
\usage{
plot_preprocessing_genes(samples, group_col)
}
\arguments{
\item{samples}{data.table with sample information with nb_detected_genes and
a column specifying the cohort.}

\item{group_col}{String specifying the column with cohort information}
}
\value{
ggplot object, i.e. boxplot with number of genes by cohort
}
\description{
Boxplot of the number of detected genes per sample, split by cohort. Used in
the bulk DGE pre-processing report to spot samples with poor library
complexity.
}
