% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_coexp_cor.R
\name{plot_rbf_impact}
\alias{plot_rbf_impact}
\title{Helper function to plot distance to affinity relationship}
\usage{
plot_rbf_impact(rbf_type, epsilon)
}
\arguments{
\item{rbf_type}{String. The RBF type. One of
\code{c("gaussian", "bump", "inverse_quadratic")}}

\item{epsilon}{Numerical. The epsilon parameter.}
}
\value{
A plot depicting the distance to affinity relationship after
applying the RBF function.
}
\description{
This function plots the distance to affinity relationship after applying
a given radial basis function (RBF) with a given epsilon value.
}
