% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_methods_cor.R
\name{plot_resolution_res}
\alias{plot_resolution_res}
\title{Plot the resolution results.}
\usage{
plot_resolution_res(object, print_head = TRUE)
}
\arguments{
\item{object}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{print_head}{Boolean. Print the Top5 resolution parameters and their
meta data.}
}
\value{
Plots the result, if the results were found in the class. Otherwise,
throws a warning and returns NULL.
}
\description{
Plots the resolution results (if they can be found in the class). The x-axis
reflects the different resolutions and the y axis the modularity observed
with that resolution.
}
