% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_methods.R
\name{plot_resolution_res}
\alias{plot_resolution_res}
\title{Get the resolution results}
\usage{
plot_resolution_res(bulk_coexp, print_head = TRUE)
}
\arguments{
\item{`bulk_coexp`}{The class, see \code{\link[=bulk_coexp]{bulk_coexp()}}.}
}
\value{
If resolution results were found, returns the data.table. Otherwise,
throws a warning and returns NULL.
}
\description{
Helper function to get a correlation-based igraph from the class
}
