% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bulk.R
\name{plot_voom_normalization}
\alias{plot_voom_normalization}
\title{Helper plot function for Voom normalisation}
\usage{
plot_voom_normalization(voom_object)
}
\arguments{
\item{voom_object}{\code{EList}. Voom object with normalised counts.}
}
\value{
ggplot object, i.e., voom normalisation plot.
}
\description{
Helper plot function for Voom normalisation
}
\examples{
# mean variance trend after voom
syn <- synthetic_bulk_cor_matrix()
grp <- rep(c("case", "control"), each = 50)
dge_list <- edgeR::normLibSizes(edgeR::DGEList(counts = syn$counts))
voom_obj <- limma::voom(dge_list, stats::model.matrix(~grp))
plot_voom_normalization(voom_obj)
}
