% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bulk.R
\name{plot_voom_normalization}
\alias{plot_voom_normalization}
\title{Helper plot function for Voom normalisation}
\usage{
plot_voom_normalization(voom_object)
}
\arguments{
\item{voom_object}{\code{EList}. Voom object with normalised counts.}
}
\value{
ggplot object, i.e., voom normalisation plot.
}
\description{
Helper plot function for Voom normalisation
}
