% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ontology.R
\name{pre_process_sim_onto}
\alias{pre_process_sim_onto}
\title{Pre-process data for subsequent ontology similarity}
\usage{
pre_process_sim_onto(object, .verbose = TRUE)
}
\arguments{
\item{object}{\verb{ontology class}. See \code{\link[=ontology]{ontology()}}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The class with added pre-processed data for semantic similarities to
the properties.
}
\description{
This function calculates needed information for semantic
similiary calculations
}
