% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_data_gower_hamming_dist}
\alias{prep_data_gower_hamming_dist}
\title{Transform data.tables into matrices for distance calculations}
\usage{
prep_data_gower_hamming_dist(dt, sample_names)
}
\arguments{
\item{dt}{data.table. The data.table you want to prepare for Gower distance
calculation in Rust. Assumes samples x features.}

\item{sample_names}{String. The sample names. Needs to be same length as
\code{nrow(dt)}.}
}
\value{
List with the following items:
\itemize{
\item dat - A numerical matrix ready for Gower distance calculations
across samples based on mixed features.
\item is_cat - Boolean vector, indicating which columns are categorical.
}
}
\description{
This function can take in data.tables with categorical and continuous data
and will transform them into matrices for subsequent distance calculations
based on Hamming distance (everything categorical) or for Gower distance
(mixed types).
}
