% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{prepare_cell_markers}
\alias{prepare_cell_markers}
\title{Helper function to prepare cell markers}
\usage{
prepare_cell_markers(obj, marker_df)
}
\arguments{
\item{obj}{A single cell class, i.e. one of \code{SingleCells} or
\code{SingleCellsMultiModal}.}

\item{marker_df}{A data.table with the columns \code{"cell_type"} and \code{"gene_id"}.
You need to ensure that the gene id type matches the obj gene id.}
}
\value{
A list of cell type to marker associations ready for subsequent
usage. Genes not found in the object will be automatically removed.
}
\description{
This function is a helper to generate a list of cell type to cell marker
annotations that can be fed into for example \verb{[REF]}.
}
