% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_classes.R
\name{preprocess_bulk_coexp}
\alias{preprocess_bulk_coexp}
\title{Process the raw data}
\usage{
preprocess_bulk_coexp(bulk_coexp, ...)
}
\arguments{
\item{bulk_coexp}{The underlying \code{bixverse_base_class} class, see
\code{\link[=bulk_coexp]{bulk_coexp()}}.}

\item{hvg}{...}

\item{mad_threshold}{...}

\item{scaling}{...}

\item{scaling_type}{...}
}
\value{
Returns the class with the processed_data data slot populated.
}
\description{
Function to do general pre-processing on top of the \code{\link[=bulk_coexp]{bulk_coexp()}}.
Options to do scaling, HVG selection, etc.

Do pre-processing on top of the data. This can include selection
of highly variable genes and/or scaling.
}
