% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{preprocess_bulk_dge}
\alias{preprocess_bulk_dge}
\title{QC on the bulk dge data (DEPRECATED!)}
\usage{
preprocess_bulk_dge(
  object,
  group_col,
  norm_method = c("TMM", "TMMwsp", "RLE", "upperquartile", "none"),
  outlier_threshold = 2,
  min_prop = 0.2,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_dge]{bulk_dge()}}.}

\item{group_col}{String. The column in the metadata that will contain the
contrast groups. Needs to be part of the metadata stored in the class.}

\item{norm_method}{String. One of
\code{c("TMM", "TMMwsp", "RLE", "upperquartile", "none")}. Please refer to
\code{\link[edgeR:calcNormFactors]{edgeR::normLibSizes()}}.}

\item{outlier_threshold}{Float. Number of standard deviations in terms of
percentage genes detected you allow before removing a sample. Defaults to \code{2}.}

\item{min_prop}{Float. Minimum proportion of samples in which the gene has
to be identified in.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
Throws an error, as it is deprecated.
}
\description{
This is a deprecated method and will raise an error. Please use
\code{\link[=qc_bulk_dge]{qc_bulk_dge()}} and \code{\link[=normalise_bulk_dge]{normalise_bulk_dge()}} instead.
}
