% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gse.R
\name{print.BlitzGseaNull}
\alias{print.BlitzGseaNull}
\title{Print a blitzGSEA null model}
\usage{
\method{print}{BlitzGseaNull}(x, ...)
}
\arguments{
\item{x}{\code{BlitzGseaNull} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a blitzGSEA null model
}
\keyword{internal}
