% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{print.BoostRes}
\alias{print.BoostRes}
\title{Print a BoostRes object}
\usage{
\method{print}{BoostRes}(x, ...)
}
\arguments{
\item{x}{A \code{BoostRes} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print a BoostRes object
}
\keyword{internal}
