% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{print.CellQc}
\alias{print.CellQc}
\title{Print a CellQc object}
\usage{
\method{print}{CellQc}(x, ...)
}
\arguments{
\item{x}{A \code{CellQc} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print a CellQc object
}
