% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_meta_cell.R
\name{print.MetaCells}
\alias{print.MetaCells}
\title{print Method for MetaCells object}
\arguments{
\item{x}{An object of class \code{MetaCells}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a MetaCells object.
}
\keyword{internal}
