% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ontology.R
\name{print.ontology}
\alias{print.ontology}
\title{print Method for ontology object}
\arguments{
\item{x}{An object of class \code{ontology}, see \code{\link[=ontology]{ontology()}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a ontology object.
}
