% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{print.ScDblFinderRes}
\alias{print.ScDblFinderRes}
\title{Print a ScDblFinderRes object}
\usage{
\method{print}{ScDblFinderRes}(x, ...)
}
\arguments{
\item{x}{A \code{ScDblFinderRes} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print a ScDblFinderRes object
}
\keyword{internal}
