% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{print.ScenicGrn}
\alias{print.ScenicGrn}
\title{Print a ScenicGrn object}
\usage{
\method{print}{ScenicGrn}(x, ...)
}
\arguments{
\item{x}{A \code{ScenicGrn} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a ScenicGrn object
}
\keyword{internal}
