% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{print.ScrubletRes}
\alias{print.ScrubletRes}
\title{Print a ScrubletRes object}
\usage{
\method{print}{ScrubletRes}(x, ...)
}
\arguments{
\item{x}{A \code{ScrubletRes} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print a ScrubletRes object
}
\keyword{internal}
