% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_spatial_sample.R
\name{print.SpatialSample}
\alias{print.SpatialSample}
\title{Print method for SpatialSample objects}
\usage{
\method{print}{SpatialSample}(x, ...)
}
\arguments{
\item{x}{A \code{SpatialSample} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for SpatialSample objects
}
