% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{qc_bulk_dge}
\alias{qc_bulk_dge}
\title{QC on the bulk dge data}
\usage{
qc_bulk_dge(
  object,
  group_col,
  outlier_threshold = 2,
  min_prop = 0.2,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=bulk_dge]{bulk_dge()}}.}

\item{group_col}{String. The column in the metadata that will contain the
contrast groups. Needs to be part of the metadata stored in the class.}

\item{outlier_threshold}{Float. Number of standard deviations in terms of
percentage genes detected you allow before removing a sample. Defaults to \code{2}.}

\item{min_prop}{Float. Minimum proportion of samples in which the gene has
to be identified in.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
Returns the class with the \code{processed_data} data slot populated and
applied parameters added to the \code{params} slot.
}
\description{
This function will do QC on the bulk data and remove outlier samples that
show substantially lower expression of genes compared to the rest of the
data and remove lowly expressed genes.
}
