% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_graphs.R
\name{rbh_graph}
\alias{rbh_graph}
\title{Reciprocal best hit graph}
\usage{
rbh_graph(
  module_results,
  rbh_type = c("cor", "set"),
  dataset_col = NULL,
  module_col = NULL,
  value_col = NULL
)
}
\arguments{
\item{module_results}{data.table or list. If data.table (you set the class
to \code{"set"}, i.e., set similarities will be used) it cointains all of
the gene modules for which you wish to generate the RBH graph and you need
to have the three column names. If list (you set the class to `"cor"``,
i.e., correlations between for example gene loadings will be used), the
list contains all of the matrices (must have col and row names).}

\item{rbh_type}{String. One of \code{c("cor", "set")}.}

\item{dataset_col}{The column (name) which indicates from which data set/
method the gene module was derived. Only needed if you want to use set
similarities.}

\item{module_col}{The column (name) which stores the names of the modules.
Only needed if you want to use set similarities.}

\item{value_col}{The column (name) which stores the genes that are part of
the modules. Only needed if you want to use set similarities.}
}
\value{
Returns the \code{rbh_graph} class for further operations.
}
\description{
This class can be used to generate reciprocal best hit graphs between
gene modules from different origins.
}
\section{Properties}{

\describe{
\item{module_data}{Nested list of the modules generated by different
methods.}
\item{rbh_graph}{igraph. The reciprocal best hit graph.}
\item{rbh_edge_df}{data.table. Contains the RBH graph data as an edge df.}
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
\item{final_results}{data.table. Contains final results.}
}
}

