% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_graph_classes.R
\name{rbh_graph}
\alias{rbh_graph}
\title{Reciprocal best hit graph}
\usage{
rbh_graph(module_results, dataset_col, module_col, value_col)
}
\arguments{
\item{module_results}{data.table with the all of the gene modules for which you
wish to generate the RBH graph.}

\item{dataset_col}{The column (name) which indicates from which data set/method
the gene module was derived.}

\item{module_col}{The column (name) which stores the names of the modules.}

\item{value_col}{The column (name) which stores the genes that are part of the modules.}
}
\value{
Returns the \code{rbh_graph} class for further operations.
}
\description{
This class can be used to generate reciprocal best hit graphs between
gene modules from different origins.
}
