% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_h5ad_metadata}
\alias{read_h5ad_metadata}
\title{Read obs and var tables and metadata from an h5ad file}
\usage{
read_h5ad_metadata(f_path)
}
\arguments{
\item{f_path}{File path to the \code{.h5ad} file.}
}
\value{
A list with:
\itemize{
\item obs - data.table of cell-level metadata
\item var - data.table of gene-level metadata
\item dims - named integer vector c(obs, var)
\item type - "CSR" or "CSC"
}
}
\description{
Read obs and var tables and metadata from an h5ad file
}
