% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_h5ad.R
\name{read_tenx_h5_adt}
\alias{read_tenx_h5_adt}
\title{Read in 10x h5 ADT data}
\usage{
read_tenx_h5_adt(f_path, feature_type = "Antibody Capture")
}
\arguments{
\item{f_path}{String. File to the h5 file from which to read the ADT counts.}

\item{feature_type}{String. The feature type to return. Defaults here to
\code{"Antibody Capture"}}
}
\value{
A dense matrix of cells x features
}
\description{
Helper function to load in ADT counts from h5. Leverages Rust under the
hood for faster filtering and reading.
}
