% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell_.R
\name{remove_knn}
\alias{remove_knn}
\alias{remove_knn.MetaCells}
\alias{remove_knn.ScCache}
\alias{remove_knn.SingleCells}
\title{Remove the KNN data}
\usage{
remove_knn(x, ...)

## S7 method for class <bixverse::MetaCells>
remove_knn(x, ...)

\method{remove_knn}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
remove_knn(x, ...)
}
\arguments{
\item{x}{An object from which to remove the kNN data.}

\item{...}{Other parameters.}
}
\description{
Remove the KNN data
}
\keyword{internal}
