% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{remove_knn}
\alias{remove_knn}
\alias{remove_knn.ScCache}
\alias{remove_knn.SingleCells}
\title{Remove the KNN data}
\usage{
remove_knn(x)

\method{remove_knn}{ScCache}(x)
}
\arguments{
\item{x}{An object from which to remove the kNN data}
}
\description{
Remove the KNN data

Remove the KNN matrix from a \code{SingleCells} class.
}
\keyword{internal}
