% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{remove_knn}
\alias{remove_knn}
\alias{remove_knn.sc_cache}
\alias{remove_knn.single_cell_exp}
\title{Remove the KNN data}
\usage{
remove_knn(x)

\method{remove_knn}{sc_cache}(x)
}
\arguments{
\item{x}{An object from which to remove the kNN data}
}
\description{
Remove the KNN data

Remove the KNN matrix from a \code{single_cell_exp} class.
}
