% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R,
%   R/classes_single_cell_subset.R
\name{remove_magic}
\alias{remove_magic}
\alias{remove_magic.ScCache}
\alias{remove_magic.SingleCells}
\alias{remove_magic.SingleCellsSubset}
\title{Remove the MAGIC imputed layer}
\usage{
remove_magic(x, ...)

\method{remove_magic}{ScCache}(x, ...)

## S7 method for class <bixverse::SingleCells>
remove_magic(x, ...)

## S7 method for class <bixverse::SingleCellsSubset>
remove_magic(x, ...)
}
\arguments{
\item{x}{An object from which to remove the imputed layer.}

\item{...}{Other parameters.}
}
\value{
The object with the imputed layer dropped.
}
\description{
Remove the MAGIC imputed layer
}
