% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples from object}
\usage{
remove_samples(object, samples_to_remove, ...)
}
\arguments{
\item{object}{The underlying object, either \code{bixverse::bulk_coexp} or
\code{bixverse::bulk_dge}.}

\item{samples_to_remove}{Character vector. The sample identifiers to remove.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
Returns the object with the samples removed. This will regenerated
the object from the start and remove any data in it.
}
\description{
This function allows to remove certain samples from the object
}
