% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell.R
\name{remove_snn_graph}
\alias{remove_snn_graph}
\alias{remove_snn_graph.MetaCells}
\alias{remove_snn_graph.ScCache}
\alias{remove_snn_graph.SingleCells}
\title{Remove the sNN graph}
\usage{
remove_snn_graph(x)

\method{remove_snn_graph}{ScCache}(x)
}
\arguments{
\item{x}{An object from which to remove the sNN graph}
}
\description{
Remove the sNN graph
}
\keyword{internal}
