% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{remove_snn_graph}
\alias{remove_snn_graph}
\alias{remove_snn_graph.sc_cache}
\alias{remove_snn_graph.single_cell_exp}
\title{Remove the sNN graph}
\usage{
remove_snn_graph(x)

\method{remove_snn_graph}{sc_cache}(x)
}
\arguments{
\item{x}{An object from which to remove the sNN graph}
}
\description{
Remove the sNN graph

Remove the sNN graph from a \code{single_cell_exp} class.
}
