% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc_lr.R
\name{resolve_weights}
\alias{resolve_weights}
\title{Resolve prioritisation weights from the scenario or validate a user vector}
\usage{
resolve_weights(weights = NULL, scenario, has_condition_de)
}
\arguments{
\item{weights}{Optional named numeric.}

\item{scenario}{String. One of \code{c("case_control", "one_condition")}.}

\item{has_condition_de}{Boolean. If \code{scenario = "case_control"}, validates
that the corresponding weights are non-zero.}
}
\value{
Final weights
}
\description{
Resolve prioritisation weights from the scenario or validate a user vector
}
\keyword{internal}
