% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_2d_loess}
\alias{rs_2d_loess}
\title{Rust implementation of a Loess function}
\usage{
rs_2d_loess(x, y, span, degree)
}
\arguments{
\item{x}{Numeric. The x values to fit.}

\item{y}{Numeric. The y values to fit.}

\item{span}{Numeric. The span parameter. Needs to be between 0.1 and 1.}

\item{degree}{Integer. Either 1 (linear) or 2 (quadratic). Other values
will cause an error.}
}
\value{
A list with the following items
\itemize{
\item predicted - The predicted values.
\item residuals - The residuals for every data point.
\item valid_idx - Which data indices were included.
}
}
\description{
Rust implementation of a Loess function
}
