% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_adt_clr}
\alias{rs_adt_clr}
\title{Applies CLR normalisation on ADT counts (Seurat-style, per cell)}
\usage{
rs_adt_clr(counts)
}
\arguments{
\item{counts}{R matrix of shape cells x features.}
}
\value{
CLR-transformed matrix.
}
\description{
Applies CLR normalisation on ADT counts (Seurat-style, per cell)
}
