% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_aucell}
\alias{rs_aucell}
\title{Calculate AUCell in Rust}
\usage{
rs_aucell(f_path, gs_list, cells_to_keep, aucell_params, streaming, verbose)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{gs_list}{List. List with the gene set indices (0-indexed!) of the
genes of interest.}

\item{cells_to_keep}{Integer. Vector of indices of the cells to keep.}

\item{aucell_params}{List. The AUCell parameters, see
\code{\link[=params_sc_aucell]{params_sc_aucell()}}.}

\item{streaming}{Boolean. Shall the data be streamed.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A matrix of cells x gene sets with the values representing the
AUC.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The function will take in a list of gene set indices (0-indexed!) and
calculate an AUCell type statistic. Three options here: the recovery-curve
AUC of Aibar, et al. (the actual AUCell statistic), an AUC derived from the
Mann-Whitney statistic, or average precision. Data can be streamed in
chunks of 50k cells per or loaded in in one go.
}
\keyword{internal}
