% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_aucell}
\alias{rs_aucell}
\title{Calculate AUCell in Rust}
\usage{
rs_aucell(f_path, gs_list, cells_to_keep, auc_type, streaming, verbose)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{gs_list}{List. List with the gene set indices (0-indexed!) of the
genes of interest.}

\item{cells_to_keep}{Integer. Vector of indices of the cells to keep.}

\item{auc_type}{String. One of \code{"wilcox"} or \code{"auroc"}, pending on
which statistic you wish to calculate.}

\item{streaming}{Boolean. Shall the data be streamed.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A matrix of cells x gene sets with the values representing the
AUC.
}
\description{
The function will take in a list of gene set indices (0-indexed!) and
calculate an AUCell type statistic. Two options here: calculate this
with proper AUROC calculations (useful for marker gene expression) or
based on the Mann-Whitney statistic (useful for pathway activity
measurs). Data can be streamed in chunks of 50k cells per or loaded in
in one go.
}
