% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_bbknn}
\alias{rs_bbknn}
\title{BBKNN implementation in Rust}
\usage{
rs_bbknn(embd, batch_labels, bbknn_params, seed, verbose)
}
\arguments{
\item{embd}{Numerical matrix. The embedding matrix to use to generate the
BBKNN parameters. Usually PCA. Rows represent cells.}

\item{batch_labels}{Integer vector. These represent to which batch a given
cell belongs.}

\item{bbknn_params}{List. Contains all of the BBKNN parameters.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list of two lists representing the sparse matrix representation
of the distances and the connectivities.
}
\description{
This function implements the BBKNN algorithm from TO ADD
}
