% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_bbknn_filtering}
\alias{rs_bbknn_filtering}
\title{Reduce BBKNN matrix to Top X neighbours}
\usage{
rs_bbknn_filtering(indptr, indices, no_neighbours_to_keep)
}
\arguments{
\item{indptr}{Integer vector. The index pointers of the underlying data.}

\item{indices}{Integer vector. The indices of the nearest neighbours.}

\item{no_neighbours_to_keep}{Integer. Number of nearest neighbours to keep.}
}
\value{
A numerical matrix with the Top X neighbours per row. If
\code{no_neighbours_to_keep} is larger than the number of neighbours in the data,
these positions will be \code{NA}.
}
\description{
Reduce BBKNN matrix to Top X neighbours
}
