% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_calc_gsea_stat_cumulative_batch}
\alias{rs_calc_gsea_stat_cumulative_batch}
\title{Helper function to generate fgsea simple-based permutations}
\usage{
rs_calc_gsea_stat_cumulative_batch(
  stats,
  pathway_scores,
  pathway_sizes,
  iters,
  gsea_param,
  seed
)
}
\arguments{
\item{stats}{Numeric vector. The gene level statistic. Needs to
sorted in descending nature.}

\item{pathway_scores}{Numeric vector. The enrichment scores for the
pathways}

\item{pathway_sizes}{Integer vector. The sizes of the pathways.}

\item{iters}{Integer. Number of permutations.}

\item{gsea_param}{Float. The Gene Set Enrichment parameter.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
List with the following elements
\itemize{
\item es The enrichment scores for the pathway
\item nes The normalised enrichment scores for the pathway
\item pvals The p-values for this pathway based on permutation
testing
\item size The pathway sizes.
}
}
\description{
Helper function to generate fgsea simple-based permutations
}
