% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_calc_multi_level}
\alias{rs_calc_multi_level}
\title{Calculates p-values for pre-processed data}
\usage{
rs_calc_multi_level(es, stats, pathway_size, sample_size, seed, eps, sign)
}
\arguments{
\item{es}{Numerical vector. The enrichment scores of the pathways of that specific size}

\item{stats}{Named numerical vector. Needs to be sorted. The gene level statistics.}

\item{pathway_size}{Integer. The size of the pathways to test.}

\item{sample_size}{Integer. The size of the random gene sets to test against.}

\item{seed}{Integer. Random seed.}

\item{eps}{Float. Boundary for calculating the p-value.}

\item{sign}{Boolean. Bit unclear what this is supposed to do. Original documentation says
\verb{This option will be used in future implementations.}, but is used in the function.}
}
\value{
List with the following elements:
\itemize{
\item pvals The pvalues.
\item is_cp_ge_half Flag indicating if conditional probability is ≥0.5. Indicates
overesimation of the p-values.
}
}
\description{
Calculates p-values for pre-processed data
}
