% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cluster_stability}
\alias{rs_cluster_stability}
\title{Helper function to assess cluster stability}
\usage{
rs_cluster_stability(data)
}
\arguments{
\item{data}{Integer matrix. Assumes that each column represents a given
resampling/bootstrap and the rows represent the features, while each integer
indicates cluster membership.}
}
\value{
A list containing:
\itemize{
\item mean_jaccard - mean Jaccard similarities for this feature across all
the bootstraps, resamplings.
\item std_jaccard - the standard deviation of the Jaccard similarities for
this feature across all the bootstraps, resamplings.
}
}
\description{
Helper function to assess cluster stability
}
