% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_compare_knn}
\alias{rs_compare_knn}
\title{Helper to compare kNN graphs}
\usage{
rs_compare_knn(knn_data_a, knn_data_b)
}
\arguments{
\item{knn_mat_a}{Integer matrix. The indices of the first kNN graph to
compare. Should be samples x neighbours. This will be treated as ground
truth.}

\item{knn_mat_b}{Integer matrix. The indices of the second kNN graph to
compare. Should be samples x neighbours.}

\item{knn_dist_a}{Numeric matrix.}
}
\value{
A list with the following elements:
\itemize{
\item all_matches - Matching neighbours for this sample.
\item all_ratios - Distance ratio for this sample (with b / a).
\item final_recall - The final recall of assuming a being the ground truth
across all samples
\item final_ratio - The final distance ratio across all samples
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Compare two kNN graphs and return the distance ratios and overlaps of
k-nearest neighbours between them.
}
