% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_compare_knn}
\alias{rs_compare_knn}
\title{Helper to compare kNN graphs}
\usage{
rs_compare_knn(knn_data_a, knn_data_b)
}
\arguments{
\item{knn_data_a}{Named list. This contains the kNN data (including
distances) of the first kNN graph. This one will be treated as the ground
truth}

\item{knn_data_b}{Named list. This contains the kNN data (including
distances) of the second kNN graph.}
}
\value{
A list with the following elements:
\itemize{
\item all_matches - Matching neighbours for this sample.
\item all_ratios - Distance ratio for this sample (with b / a).
\item final_recall - The final recall of assuming a being the ground truth
across all samples
\item final_ratio - The final distance ratio across all samples
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Compare two kNN graphs and return the distance ratios and overlaps of
k-nearest neighbours between them.
}
