% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_compare_knn}
\alias{rs_compare_knn}
\title{Helper to compare kNN graphs}
\usage{
rs_compare_knn(knn_mat_a, knn_mat_b)
}
\arguments{
\item{knn_mat_a}{Integer matrix. The first kNN graph to compare.}

\item{knn_mat_b}{Integer matrix. The second kNN graph to compare.}
}
\value{
Vector of number of overlaps per sample.
}
\description{
Helper to compare kNN graphs
}
