% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_constrained_page_rank}
\alias{rs_constrained_page_rank}
\title{Calculate a constrained page-rank score}
\usage{
rs_constrained_page_rank(
  node_names,
  node_types,
  from,
  to,
  weights,
  edge_type,
  personalised,
  sink_nodes,
  sink_edges
)
}
\arguments{
\item{node_names}{String vector. Name of the graph nodes.}

\item{node_types}{String vector. The node types.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{weights}{Numerical vector. The edge weights from the edge list.}

\item{edge_type}{String vector. The edge types.}

\item{personalised}{Numerical vector. The reset values. They must sum to 1
and be of same length of \code{node_names}!}

\item{sink_nodes}{Optional string vector. Should these node types be seen as
sinks, i.e., the reset occurs when this node is reached.}

\item{sink_edges}{Optional string vector. Shall an automatic reset occur
when this edge type is traversed.}
}
\value{
The personalised constrained page rank values.
}
\description{
This function can be used to get constrainted personalised
page-rank scores akin to Ruiz, et al. You can provide optionally
\code{sink_nodes} (node types that will force a reset) and/or \code{sink_edges}
(edge types that will force a reset).
}
\references{
Ruiz, et al., Nat Commun, 2021
}
