% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_contrastive_pca}
\alias{rs_contrastive_pca}
\title{Calculate the contrastive PCA}
\usage{
rs_contrastive_pca(
  target_covar,
  background_covar,
  target_mat,
  alpha,
  n_pcs,
  return_loadings
)
}
\arguments{
\item{target_covar}{The co-variance matrix of the target data set.}

\item{background_covar}{The co-variance matrix of the background data set.}

\item{target_mat}{The original values of the target matrix.}

\item{alpha}{How much of the background co-variance should be removed.}

\item{n_pcs}{How many contrastive PCs to return}

\item{return_loadings}{Shall the loadings be returned from the contrastive
PCA}
}
\value{
A list containing:
\itemize{
\item factors - The factors of the contrastive PCA.
\item loadings - The loadings of the contrastive PCA. Will be NULL if
return_loadings is set to FALSE.
}
}
\description{
This function calculate the contrastive PCA given a target
covariance matrix and the background covariance matrix you wish to subtract.
The alpha parameter controls how much of the background covariance you wish
to remove. You have the options to return the feature loadings and you can
specificy the number of cPCAs to return.
WARNING! Incorrect use can cause kernel crashes. Wrapper around the Rust functions
with type checks are provided in the package.
}
