% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_contrastive_pca}
\alias{rs_contrastive_pca}
\title{Calculate the contrastive PCA}
\usage{
rs_contrastive_pca(
  target_covar,
  background_covar,
  target_mat,
  alpha,
  n_pcs,
  return_loadings
)
}
\arguments{
\item{target_covar}{The co-variance matrix of the target data set.}

\item{background_covar}{The co-variance matrix of the background data set.}

\item{target_mat}{The original values of the target matrix.}

\item{alpha}{How much of the background co-variance should be removed.}

\item{n_pcs}{How many contrastive PCs to return}

\item{return_loadings}{Shall the loadings be returned from the contrastive
PCA}
}
\value{
An R list with loadings and factors. If return_loadings == FALSE,
then loadings will be NULL.
}
\description{
Calculate the contrastive PCA
}
