% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor}
\alias{rs_cor}
\title{Calculate the column wise correlations.}
\usage{
rs_cor(x, spearman)
}
\arguments{
\item{x}{R matrix with doubles.}

\item{spearman}{Shall the Spearman correlation be calculated instead of
Pearson.}
}
\value{
The correlation matrix.
}
\description{
Calculates the correlation matrix of the columns.
WARNING! Incorrect use can cause kernel crashes. Wrapper around the Rust
functions with type checks are provided in the package.
}
