% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor2}
\alias{rs_cor2}
\title{Calculate the column wise correlations.}
\usage{
rs_cor2(x, y, spearman)
}
\arguments{
\item{x}{R matrix with doubles.}

\item{y}{R matrix with doubles.}

\item{spearman}{Shall the Spearman correlation be calculated instead of
Pearson.}
}
\value{
The correlation matrix.
}
\description{
Calculates the correlation between the columns of two matrices.
The number of rows need to be the same!
}
