% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor_upper_triangle}
\alias{rs_cor_upper_triangle}
\title{Calculate the column wise correlations and returns the upper triangle}
\usage{
rs_cor_upper_triangle(x, spearman, shift)
}
\arguments{
\item{x}{R matrix with doubles.}

\item{spearman}{Shall the Spearman correlation be calculated instead of
Pearson.}

\item{shift}{Boolean. If you applied a shift, i.e. included the diagonal
values. If \code{true}, assumes the diagonal values are \code{1}, otherwise derives
them from the data.}
}
\value{
The upper triangle of the correlation matrix iterating through the
rows, shifted by one (the diagonal will not be returned).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the correlation matrix of the columns, but returns the upper
triangle only as a flat vector.
}
