% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cor_upper_triangle}
\alias{rs_cor_upper_triangle}
\title{Calculate the column wise correlations.}
\usage{
rs_cor_upper_triangle(x, spearman, shift)
}
\arguments{
\item{x}{R matrix with doubles.}

\item{spearman}{Shall the Spearman correlation be calculated instead of
Pearson.}

\item{shift}{Shall a shift be applied to the matrix. 0 = the diagonal will
be included. 1 = the diagonal will not be included.}
}
\value{
The upper triangle of the correlation matrix iterating through the
rows, shifted by one (the diagonal will not be returned).
}
\description{
Calculates the correlation matrix of the columns. This function
will return the upper triangle. WARNING! Incorrect use can cause kernel
crashes. Wrapper around the Rust functions with type checks are provided in
the package.
}
