% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_coremo_quality}
\alias{rs_coremo_quality}
\title{Helper function to assess CoReMo cluster quality}
\usage{
rs_coremo_quality(cluster_genes, cor_mat, row_names, seed)
}
\arguments{
\item{cluster_genes}{A list. Contains the cluster and their respective genes.}

\item{cor_mat}{Numerical matrix. Contains the correlation coefficients.}

\item{row_names}{String vector. The row names (or column names) of the
correlation matrix.}

\item{seed}{Integer. Random seed for the sub sampling of genes.}
}
\value{
A list containing:
\itemize{
\item r2med - median R2 of the cluster.
\item r2mad - median absolute deviation of the R2 in the cluster.
\item size - size of the cluster.
}
}
\description{
This function assesses the quality of the clusters
with a given cut \code{k}. Returns the median R2 (cor^2) and the median absolute
deviation (MAD) of the clusters. Large clusters (≥1000) are subsampled
to a random set of 1000 genes.
}
