% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_coremo_stability}
\alias{rs_coremo_stability}
\title{Helper function to assess CoReMo cluster stability}
\usage{
rs_coremo_stability(data, indices, epsilon, rbf_type, spearman)
}
\arguments{
\item{data}{Numeric matrix. The original processed matrix.}

\item{indices}{Integer vector. The sample indices to remove to re-calculate
the distances.}

\item{epsilon}{Float. Epsilon parameter for the RBF.}

\item{rbf_type}{String. Needs to be from
\code{c("gaussian", "bump", "inverse_quadratic")}.}

\item{spearman}{Boolean. Shall Spearman correlation be used.}
}
\value{
A list with \code{length(indices)} elements, each containing the distance
minus the given sample.
}
\description{
This function is a helper for the leave-on-out stability
assessment of CoReMo clusters. The function will generate the distance
vectors based on leaving out the samples defined in indices one by one.
}
