% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_covariance}
\alias{rs_covariance}
\title{Calculate the column-wise co-variance.}
\usage{
rs_covariance(x)
}
\arguments{
\item{x}{R matrix with doubles.}
}
\value{
The co-variance matrix.
}
\description{
Calculates the co-variance of the columns.
WARNING! Incorrect use can cause kernel crashes. Wrapper around the Rust
functions with type checks are provided in the package.
}
