% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_create_random_aucs}
\alias{rs_create_random_aucs}
\title{Create random AUCs}
\usage{
rs_create_random_aucs(score_vec, size_pos, random_iters, auc_iters, seed)
}
\arguments{
\item{score_vec}{The overall vector of scores.}

\item{size_pos}{The size of the hits represented in the score_vec.}

\item{random_iters}{Number of random AUCs to generate.}

\item{auc_iters}{Number of random iterations to approximate the AUCs.
Recommended size: 10000L.}

\item{seed}{Seed.}
}
\value{
A vector of random AUCs based the score vector and size of the
positive set.
}
\description{
This function creates a random set of AUCs based on a score
vector and a size of the positive set. This can be used for permutation-
based estimation of Z-scores and subsequently p-values.
}
