% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_critval_mat}
\alias{rs_critval_mat}
\title{Calculate the critical value}
\usage{
rs_critval_mat(mat, iters, alpha, seed)
}
\arguments{
\item{mat}{Numeric matrix. The (symmetric matrix with all of the values).}

\item{iters}{Integer. Number of random permutations to use.}

\item{alpha}{Float. The alpha value. For example, 0.001 would mean that the
critical value is smaller than 0.1 percentile of the random permutations.}

\item{seed}{Integer. For reproducibility purposes}
}
\value{
The critical value for the given parameters.
}
\description{
This function calculates the critical value for a given set based on random
permutations and a given alpha value.
}
