% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_dense_to_upper_triangle}
\alias{rs_dense_to_upper_triangle}
\title{Generate a vector-based representation of the upper triangle of a matrix}
\usage{
rs_dense_to_upper_triangle(x, shift)
}
\arguments{
\item{x}{Numeric vector. The vector of correlation coefficients that you
want to use to go back to a dense matrix.}

\item{shift}{Integer. If you want to apply a shift, i.e. included the diagonal
values = 0; or excluded the diagonal values = 1.}
}
\value{
The dense R matrix.
}
\description{
This function generates a vector from the upper triangle of
a given symmetric matrix. You have the option to remove the diagonal with
setting shift to 1.
}
