% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_differential_cor}
\alias{rs_differential_cor}
\title{Calculate the column wise differential correlation between two sets of data.}
\usage{
rs_differential_cor(x_a, x_b, spearman)
}
\arguments{
\item{x_a}{R matrix a to be used for the differential correlation analysis.}

\item{x_b}{R matrix a to be used for the differential correlation analysis.}

\item{spearman}{Shall the Spearman correlation be calculated instead of
Pearson.}
}
\value{
A list containing:
\itemize{
\item r_a - The correlation coefficients in the upper triangle of
matrix a.
\item r_b - The correlation coefficients in the upper triangle of
matrix b.
\item z_score - The z-scores of the difference in correlation
coefficients.
\item p_val - The z-scores transformed to p-values.
}
}
\description{
This function calculates the differential correlation based on
the Fisher method. For speed purposes, the function will only calculate the
differential correlation on the upper triangle of the two correlation
matrices.
WARNING! Incorrect use can cause kernel crashes. Wrapper around the Rust
functions with type checks are provided in the package.
}
