% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_dist}
\alias{rs_dist}
\title{Calculate the pairwise column distance in a matrix}
\usage{
rs_dist(x, distance_type)
}
\arguments{
\item{x}{Numerical matrix. The matrix for which to calculate the pairwise
column distances.}

\item{distance_type}{String. One of
\code{c("euclidean", "manhattan", "canberra", "cosine")}.}
}
\value{
The calculated distance matrix
}
\description{
This function allows to calculate pairwise between all columns the specified
distance metric.
}
