% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_extract_counts_plots}
\alias{rs_extract_counts_plots}
\title{Helper to extract single cell counts as a dense vector for plotting}
\usage{
rs_extract_counts_plots(f_path, cell_indices, gene_index, norm, scale, clip)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep.}

\item{gene_index}{Integer. Gene index position to return (0-indexed!).}

\item{norm}{Boolean. Shall normalised counts be returned.}

\item{scale}{Boolean. Shall the normalised counts be scaled.}

\item{clip}{Optional float. Clipping for the Z-scores if scale is set to
\code{TRUE}}
}
\value{
The dense vector of expression values for this gene.
}
\description{
Helper to extract single cell counts as a dense vector for plotting
}
