% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_extract_grouped_gene_stats}
\alias{rs_extract_grouped_gene_stats}
\title{Calculates the gene statistics for a set of cell groups and genes}
\usage{
rs_extract_grouped_gene_stats(
  f_path,
  cell_indices,
  gene_indices,
  group_ids,
  group_levels
)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep.}

\item{gene_indices}{Integer. Gene index position to return (0-indexed!).}

\item{group_ids}{Integer. The levels of the data. (0-indexed!)}

\item{group_levels}{String. Name of the factors.}
}
\value{
A list with the following elements:
\itemize{
\item grp_label - The label of that group
\item mean_exp - Vector of mean expression values in row major (genes x
n_levels)
\item perc_exp - Vector of proportions of cells with expression in row
major (genes x n_levels)
}
}
\description{
Helper function to extract data for dot plots and/or heatmaps.
}
