% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_extract_several_genes_plots}
\alias{rs_extract_several_genes_plots}
\title{Helper to extract single cell counts for several genes}
\usage{
rs_extract_several_genes_plots(f_path, cell_indices, gene_indices, scale, clip)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep.}

\item{scale}{Boolean. Shall the normalised counts be scaled.}

\item{clip}{Optional float. Clipping for the Z-scores if scale is set to
\code{TRUE}}

\item{gene_index}{Integer. Gene index position to return (0-indexed!).}
}
\value{
A list of dense vectors of the normalised counts.
}
\description{
Helper to extract single cell counts for several genes
}
