% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_fast_auc}
\alias{rs_fast_auc}
\title{Fast AUC calculation}
\usage{
rs_fast_auc(pos_scores, neg_scores, iters, seed)
}
\arguments{
\item{pos_scores}{The scores of your hits.}

\item{neg_scores}{The scores of your non-hits.}

\item{iters}{Number of iterations to run the function for. Recommended size: 10,000.}

\item{random_seed}{Seed.}
}
\description{
This function calculates rapidly AUCs based on an approximation.
}
