% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_fast_cluster_sc}
\alias{rs_fast_cluster_sc}
\title{Runs fast Louvain cluster on the data}
\usage{
rs_fast_cluster_sc(
  embd,
  km_type,
  resolutions,
  n_centroids,
  fc_params,
  seed,
  verbose
)
}
\arguments{
\item{embd}{Numeric matrix. The original embedding.}

\item{km_type}{String. One of \code{c("kmeans", "minibatch")} for the type of
k means clustering to run.}

\item{resolutions}{Numeric vector. The Louvain resolutions to iterate
through.}

\item{n_centroids}{Optional integer. The number of clusters to find. If
not provided, defaults to \code{sqrt(nrow(embd))}.}

\item{fc_params}{Named list. The fast clustering parameters.}

\item{seed}{Integer. For reproducibility.}

\item{verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
A list with the memberships per resolution.
}
\description{
Runs first k-means clustering, followed by a kNN detection on the centroids
to then run Louvain clustering on the graph and propagate the membership
back to the original data.
}
