% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_fast_cluster_sc_grid}
\alias{rs_fast_cluster_sc_grid}
\title{Runs fast Louvain cluster on the data (with multiple seeds)}
\usage{
rs_fast_cluster_sc_grid(
  embd,
  km_type,
  resolutions,
  n_centroids,
  fc_params,
  snn,
  return_kmeans,
  no_seeds,
  seed,
  verbose
)
}
\arguments{
\item{embd}{Numeric matrix. The original embedding.}

\item{km_type}{String. One of \code{c("kmeans", "minibatch")} for the type of
k means clustering to run.}

\item{resolutions}{Numeric vector. The Louvain resolutions to iterate
through.}

\item{n_centroids}{Optional integer. The number of clusters to find. If
not provided, defaults to \code{sqrt(nrow(embd))}.}

\item{fc_params}{Named list. The fast clustering parameters.}

\item{snn}{Boolean. Shall the kNN graph be additionally transformed into
an sNN graph.}

\item{return_kmeans}{Boolean. Shall the k-means centroid assignments be
returned alongside the memberships.}

\item{no_seeds}{Integer. Number of additional seeds to use. Should be >=2.}

\item{seed}{Integer. For reproducibility.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following elements:
\itemize{
\item memberships - The memberships across the different resolutions. The
membership from the random seed with the best conductance is returned.
\item stats - The statistics per given resolution run.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Runs first k-means clustering, followed by a kNN detection on the centroids
to then run Louvain clustering with several seeds (based on the original
one) on the graph and propagate the membership back to the original data.
Returns additional metrics around cluster stability and community
conductance.
}
\keyword{internal}
