% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_fast_ica}
\alias{rs_fast_ica}
\title{Run the Rust implementation of fast ICA.}
\usage{
rs_fast_ica(whiten, w_init, maxit, alpha, tol, ica_type, verbose)
}
\arguments{
\item{whiten}{The whitened matrix.}

\item{w_init}{The w_init matrix. ncols need to be equal to nrows of whiten.}

\item{maxit}{Maximum number of iterations to try if algorithm does not
converge.}

\item{alpha}{The alpha parameter for the LogCosh implementation of ICA.}

\item{tol}{Tolerance parameter.}

\item{ica_type}{One of 'logcosh' or 'exp'.}

\item{verbose}{Controls the verbosity of the function.}

\item{debug}{Additional messages if desired.}

\item{x}{The matrix to whiten. The whitening will happen over the columns.}
}
\value{
A list containing:
\itemize{
\item mixing - The mixing matrix for subsequent usage.
\item converged - Boolean if the algorithm converged.
}
}
\description{
This function serves as a wrapper over the fast ICA
implementations in Rust. It assumes a pre-whiten matrix and also an
intialised w_init. WARNING! Incorrect use can cause kernel crashes. Wrapper
around the Rust functions with type checks are provided in the package.
}
