% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_fast_ica}
\alias{rs_fast_ica}
\title{Run the Rust implementation of fast ICA.}
\usage{
rs_fast_ica(whiten, w_init, ica_type, ica_params)
}
\arguments{
\item{whiten}{Numerical matrix. The whitened matrix.}

\item{w_init}{Numerical matrix. The initial unmixing matrix. ncols need to
be equal to nrows of whiten.}

\item{ica_type}{String. One of 'logcosh' or 'exp'.}

\item{ica_params}{A list containing:
\itemize{
\item maxit - Integer. Maximum number of iterations for ICA.
\item alpha - Float. The alpha parameter for the logcosh version of ICA.
Should be between 1 to 2.
\item max_tol - Maximum tolerance of the algorithm
\item verbose - Verbosity of the function, i.e., shall individual iters
be shown.
}
If the list is empty or the expected elements are not found, default values
are used.}
}
\value{
A list with the following items:
\itemize{
\item mixing - The mixing matrix for subsequent usage.
\item converged - Boolean if the algorithm converged.
}
}
\description{
This function serves as a wrapper over the fast ICA
implementations in Rust. It assumes a whitened matrix and also an
intialised w_init. WARNING! Incorrect use can cause kernel crashes. Wrapper
around the Rust functions with type checks are provided in the package.
}
