% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_filter_onto_sim}
\alias{rs_filter_onto_sim}
\title{Filter the term similarities for a specific critical value}
\usage{
rs_filter_onto_sim(sim_vals, names, threshold)
}
\arguments{
\item{sim_vals}{Numerical vector. The upper triangle of the similarity matrix
as a flattened vector.}

\item{names}{String vector. The row/col names of the similarity matrix.}

\item{threshold}{Float. The filtering threshold.}
}
\value{
A list with:
\itemize{
\item t1 - name of term 1.
\item t2 - name of term 2.
\item sim - the similarity between the two terms.
}
}
\description{
This function takes the similarity values as the upper triangle,
the row/column names and filtering the values down based on the threshold.
}
