% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_generate_bulk_rnaseq}
\alias{rs_generate_bulk_rnaseq}
\title{Generation of bulkRNAseq-like data with optional correlation structure}
\usage{
rs_generate_bulk_rnaseq(
  num_samples,
  num_genes,
  seed,
  add_modules,
  module_sizes
)
}
\arguments{
\item{num_samples}{Integer. Number of samples to simulate.}

\item{num_genes}{Integer. Number of genes to simulate.}

\item{seed}{Integer. Seed for reproducibility.}

\item{add_modules}{Boolean. Shall correlation structures be added to the
data.}

\item{module_sizes}{\code{NULL} or vector of sizes of the gene modules. When
\code{NULL} defaults to \code{c(300, 250, 200, 300, 500)}. Warning! The sum of this
vector must be ≤ num_genes!}
}
\value{
List with the following elements
\itemize{
\item counts The matrix of simulated counts.
\item module_membership Vector defining the module membership.
}
}
\description{
Function generates synthetic bulkRNAseq data with heteroskedasticity (lowly
expressed genes show higher variance) and can optionally add correlation
structures for testing purposes.
}
