% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_generate_ligand_target_influence}
\alias{rs_generate_ligand_target_influence}
\title{Generate the ligand to target influence matrices}
\usage{
rs_generate_ligand_target_influence(
  ligand_seeds,
  ppi_network,
  grn_network,
  n_nodes,
  params
)
}
\arguments{
\item{ligand_seeds}{List. Contains the indices of the seeds, i.e., ligands.}

\item{ppi_network}{Named list. Contains the PPI network with the ligand
to receptor to signalling to TFs. Must contain from (indices), to
(indices), and edge weights.}

\item{grn_network}{Named list. Contains the gene regulatory network with the
TF to target gene network. Must contain from (indices), to (indices), and
edge weights.}

\item{n_nodes}{Integer. Number of total nodes.}

\item{params}{Named list.}
}
\value{
A dense matrix of ligands x genes that contains the influence
scores of each
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Helper function to generate the ligand to target influence matrix for the
NicheNet like approach.
}
