% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_geom_elim_fgsea_simple}
\alias{rs_geom_elim_fgsea_simple}
\title{Run fgsea simple method for gene ontology with elimination method}
\usage{
rs_geom_elim_fgsea_simple(
  stats,
  levels,
  go_obj,
  gsea_param,
  elim_threshold,
  min_size,
  max_size,
  iters,
  seed,
  debug
)
}
\arguments{
\item{stats}{Named numerical vector. Needs to be sorted. The gene level statistics.}

\item{levels}{A character vector representing the levels to iterate through.
The order will be the one the iterations are happening in.}

\item{go_obj}{The gene_ontology_data S7 class. See \code{\link[=gene_ontology_data]{gene_ontology_data()}}.}

\item{gsea_param}{Float. The GSEA parameter. Usually defaults to 1.0.}

\item{elim_threshold}{p-value below which the elimination procedure shall be
applied to the ancestors.}

\item{min_size}{Minimum size of the gene ontology term for testing.}

\item{max_size}{Maximum size of the gene ontology term for testing. Setting this
parameter to large values will slow the function down.}

\item{iters}{Integer. Number of random permutations for the fgsea simple method
to use}

\item{seed}{Integer. For reproducibility purposes.}

\item{debug}{Boolean that will provide additional console information for
debugging purposes.}
}
\value{
List with the following elements
\itemize{
\item go_ids The name of the tested gene ontology identifer.
\item es The enrichment scores for the pathway
\item nes The normalised enrichment scores for the pathway
\item size The pathway sizes (after elimination!).
\item pvals The p-values for this pathway based on permutation
testing
\item leading_edge A list of the index positions of the leading edge
genes for this given GO term.
}
}
\description{
Run fgsea simple method for gene ontology with elimination method
}
