% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_get_gs_indices}
\alias{rs_get_gs_indices}
\title{Helper function to rapidly retrieve the indices of the gene set members}
\usage{
rs_get_gs_indices(gene_universe, pathway_list)
}
\arguments{
\item{gene_universe}{Character Vector. The genes represented in the gene universe.}

\item{pathway_list}{List. A named list with each element containing the genes for this
pathway.}
}
\value{
Returns a list with the index positions of the gene set genes in the gene universe.
Importantly, these are indexed to R's 1-indexing!
}
\description{
Helper function to rapidly retrieve the indices of the gene set members
}
