% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_get_metacells_bootstrapped}
\alias{rs_get_metacells_bootstrapped}
\title{Generate meta cells (hdWGCNA method)}
\usage{
rs_get_metacells_bootstrapped(
  f_path,
  knn_mat,
  embd,
  cells_to_keep,
  cells_to_use,
  meta_cell_params,
  target_size,
  seed,
  verbose
)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{knn_mat}{Optional integer matrix. The kNN matrix you wish to use
for the generation of the meta cells. This function expects 0-indices!}

\item{embd}{Optional numerical matrix. The embedding matrix (for example
PCA embedding) you wish to use for the generation of the kNN graph that
is used subsequently for aggregation of the meta cells.}

\item{cells_to_keep}{Optional indices of the cells to keep, i.e., the
cells used for the generation of the embedding.}

\item{cells_to_use}{Optional indices of cells to use for meta cell
generation. Useful if you wish to generate meta cells in specific cell
types. If this is provided, the kNN graph will be regenerated.}

\item{meta_cell_params}{A list containing the meta cell parameters.}

\item{target_size}{Numeric. Target library size for re-normalisation of
the meta cells. Typically \code{1e4}.}

\item{seed}{Integer. For reproducibility.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following elements:
\itemize{
\item assignments - A list containing assignment information with elements:
assignments (vector), metacells (list), unassigned (vector), n_metacells,
n_cells, n_unassigned
\item aggregated - A list with indptr, indices, raw_counts, norm_counts,
nrow, ncol in sparse format.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function implements the approach from Morabito, et al. to generate meta
cells. You can provide an already pre-computed kNN matrix or an embedding to
regenerate the kNN matrix with specified parameters in the meta_cell_params.
If \code{knn_mat} is provided, this one will be used. You need to at least
provide \code{knn_mat} or \code{embd}!
}
\keyword{internal}
