% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_get_seacells}
\alias{rs_get_seacells}
\title{Generate SEACells}
\usage{
rs_get_seacells(f_path, embd, seacells_params, target_size, seed, verbose)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{embd}{Numerical matrix. The embedding matrix (for example PCA embedding)
used for the generation of the kNN graph and kernel matrix.}

\item{seacells_params}{A list containing the SEACells parameters.}

\item{target_size}{Numeric. Target library size for re-normalisation of
the meta cells. Typically \code{1e4}.}

\item{seed}{Integer. For reproducibility purposes.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item assignments - A list containing assignment information with elements:
assignments (vector), metacells (list), unassigned (vector), n_metacells,
n_cells, n_unassigned
\item aggregated - A list with indptr, indices, raw_counts, norm_counts,
nrow, ncol in sparse format.
\item rss - Vector of RSS values from each iteration.
\item archetypes - Vector of cell indices selected as archetypes.
}
}
\description{
This function implements the SEACells algorithm for generating
meta cells from Persad et al. An embedding matrix must be provided which is
used to construct the kNN graph and kernel matrix for the SEACells
algorithm. This version is highly memory and speed-optimised and will
truncate small values during matrix operations which can affect convergence.
}
\references{
Persad, et al., Nat. Biotechnol., 2023.
}
