% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gower_dist}
\alias{rs_gower_dist}
\title{Calculates the Gower distance for a given matrix}
\usage{
rs_gower_dist(x, is_cat)
}
\arguments{
\item{x}{Numerical matrix. Converted matrix of continuous and categorical
variables as numerical values.}

\item{is_cat}{Boolean. Which of the columns represent categorical values.}
}
\value{
The Gower distance matrix.
}
\description{
Calculates the Gower distance for a given matrix
}
