% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gse_geom_elim}
\alias{rs_gse_geom_elim}
\title{Run hypergeometric enrichment over the gene ontology}
\usage{
rs_gse_geom_elim(
  target_genes,
  go_to_genes,
  ancestors,
  levels,
  gene_universe_length,
  min_genes,
  elim_threshold,
  debug
)
}
\description{
This function implements a Rust version of the gene ontology enrichment with elimination:
the starting point are the leaves of the ontology and hypergeometric tests will first conducted there.
Should the hypergeometric test
}
