% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gse_geom_elim_list}
\alias{rs_gse_geom_elim_list}
\title{Run hypergeometric enrichment a list of target genes over the gene ontology}
\usage{
rs_gse_geom_elim_list(
  target_genes_list,
  levels,
  go_obj,
  gene_universe_length,
  min_genes,
  elim_threshold,
  debug
)
}
\arguments{
\item{target_genes_list}{A list of target genes against which to run the
method.}

\item{levels}{A character vector representing the levels to iterate through.
The order will be the one the iterations are happening in.}

\item{go_obj}{The gene_ontology_data S7 class. See \code{\link[=gene_ontology_data]{gene_ontology_data()}}.}

\item{gene_universe_length}{The length of the gene universe.}

\item{min_genes}{number of minimum genes for the gene ontology term to be
tested.}

\item{elim_threshold}{p-value below which the elimination procedure shall
be applied to the ancestors.}

\item{debug}{boolean that will provide additional console information for
debugging purposes.}
}
\value{
A list containing:
\itemize{
\item go_ids - The gene ontology identifier.
\item pvals - The calculated odds ratios.
\item odds_ratios - The calculated odds ratios.
\item overlap - The size of the overlap.
\item gene_set_lengths - The length of the gene sets.
\item no_test - The number of tests that were conducted against
target_gene_list. First element indicates how many values belong to the
first target_genes set in the list, etc.
}
}
\description{
This function implements a Rust version of the gene ontology enrichment with
elimination: the starting point are the leafs of the ontology and
hypergeometric tests will first conducted there. Should the hypergeometric
test p-value be below a certain threshold, the genes of that gene ontology
term will be removed from all ancestors. This function is designed to
leverage Rust-based threading for parallel processing of a list of target
genes. WARNING! Incorrect use can cause kernel crashes. Wrapper around the
Rust functions with type checks are provided in the package.
}
