% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gse_geom_elim_list}
\alias{rs_gse_geom_elim_list}
\title{Run hypergeometric enrichment a list of target genes over the gene ontology}
\usage{
rs_gse_geom_elim_list(
  target_genes_list,
  go_to_genes,
  ancestors,
  levels,
  gene_universe_length,
  min_genes,
  elim_threshold,
  debug
)
}
\arguments{
\item{target_genes_list}{A list of target genes against which to run the method.}

\item{go_to_genes}{A named list with the gene identifers as elements and gene ontology identifiers as
names.}

\item{ancestors}{A named list with the go identifiers of all ancestors as elements and the gene ontology
identifiers as names.}

\item{levels}{A named list with the go identifiers of that ontology level as elements and the level name
as names. IMPORTANT! This list needs to be ordered in the right way!}

\item{gene_universe_length}{The length of the gene universe.}

\item{min_genes}{number of minimum genes for the gene ontology term to be tested.}

\item{debug}{boolean that will provide additional console information for debugging purposes.}

\item{elim_threshold:}{p-value below which the elimination procedure shall be applied to the ancestors.}
}
\description{
This function implements a Rust version of the gene ontology enrichment with elimination:
the starting point are the leaves of the ontology and hypergeometric tests will first conducted there.
Should the hypergeometric test p-value be below a certain threshold, the genes of that gene ontology
term will be removed from all ancestors. This function is designed to leverage Rust-based threading
for parallel processing of a list of target genes.
}
