% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_h5ad_data}
\alias{rs_h5ad_data}
\title{Load in h5ad data via Rust}
\usage{
rs_h5ad_data(f_path, cs_type, nrows, ncols, cell_quality, verbose)
}
\arguments{
\item{f_path}{File path. The path to the h5ad file.}

\item{cs_type}{String. Is the data stored in CSC or CSR.}

\item{nrows}{Integer. Number of rows in the file.}

\item{ncols}{Integer. Number of columns in the file.}

\item{cell_quality}{List. Specifiying the cell quality. Please refer
to \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}.}

\item{verbose}{Boolean. Controls verbosity of the function}
}
\value{
A list with:
\itemize{
\item data - The data of the sparse matrix stored on the h5ad file.
\item indices - The indices of the sparse matrix stored in the h5ad file.
\item indptr - The indptr of the sparse matrix stored in the h5ad file.
\item no_genes - No of genes in the sparse matrix (i.e., ncol).
\item no_cells - No of cells in the sparse matrix (i.e., nrow).
}
}
\description{
Loads in h5ad data within Rust and automatically converts the data into
CSR with cells x genes.
}
