% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hamming_dist}
\alias{rs_hamming_dist}
\title{Calculates the Hamming distance between categorical columns}
\usage{
rs_hamming_dist(x)
}
\arguments{
\item{x}{Integer matrix. The integers represent the factor data.}
}
\value{
The Hamming distance matrix
}
\description{
Calculates the Hamming distance between categorical columns
}
