% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_harmony_v2}
\alias{rs_harmony_v2}
\title{Harmony batch correction in Rust (version 2)}
\usage{
rs_harmony_v2(pca, harmony_params, batch_labels, seed, verbose)
}
\arguments{
\item{pca}{Numerical matrix, i.e., the PCA matrix you want to correct.}

\item{harmony_params}{List. The parameters for the Harmony (v2) algorithm.}

\item{batch_labels}{List. Each element in the list needs to be a 0-indexed
integer that represents the batch effects you wish to regress out.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
The batch-corrected Harmony (v2) embedding space.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function implements the version 2 Harmony algorithm from Patikas, et
al., 2026.
}
