% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_harmony_v2}
\alias{rs_harmony_v2}
\title{Harmony batch correction in Rust (version 2)}
\usage{
rs_harmony_v2(pca, harmony_params, batch_labels, seed, verbose)
}
\arguments{
\item{pca}{Numerical matrix, i.e., the PCA matrix you want to correct.}

\item{harmony_params}{List. The parameters for the Harmony (v2) algorithm.}

\item{batch_labels}{List. Each element in the list needs to be a 0-indexed
integer that represents the batch effects you wish to regress out.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
The batch-corrected Harmony (v2) embedding space.
}
\description{
This function implements the version 2 Harmony algorithm from Patikas, et
al., 2026.
}
