% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hedges_g}
\alias{rs_hedges_g}
\title{Calculate the OT harmonic sum}
\usage{
rs_hedges_g(mat_a, mat_b, small_sample_correction)
}
\arguments{
\item{mat_a}{The matrix of samples and features in grp A for which to
calculate the Hedge's G effect.}

\item{mat_b}{The matrix of samples and features in grp B for which to
calculate the Hedge's G effect.}

\item{small_sample_correction}{Shall the small sample correction be applied.}
}
\value{
Returns the harmonic sum according to the OT calculation.
}
\description{
Calculates the Hedge's G effect for two sets of matrices. The
function assumes that rows = samples and columns = features.
WARNING! Incorrect use can cause kernel crashes. Wrapper around the Rust
functions with type checks are provided in the package.
}
