% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hotspot_autocor}
\alias{rs_hotspot_autocor}
\title{Calculate gene spatial auto-correlations}
\usage{
rs_hotspot_autocor(
  f_path_genes,
  f_path_cells,
  embd,
  knn_data,
  hotspot_params,
  cells_to_keep,
  genes_to_use,
  streaming,
  verbose,
  seed
)
}
\arguments{
\item{f_path_genes}{Path to the \code{counts_genes.bin} file.}

\item{f_path_cells}{Path to the \code{counts_cells.bin} file.}

\item{embd}{Numerical matrix. The embedding matrix from which to generate
the kNN graph.}

\item{knn_data}{Optional list. This contains pre-computed kNN data
(including distances) and the \code{dist_metric} it was built with. The user has
to ensure consistency! If provided, this will be used rather than a graph
built from the parameter list.}

\item{hotspot_params}{List. The HotSpot parameter list.}

\item{cells_to_keep}{Integer vector. 0-index vector indicating which cells
to include in the analysis. Ensure that this is of same order/length
as the embedding matrix.}

\item{genes_to_use}{Integer vector. 0-index vector indicating which genes
to include.}

\item{streaming}{Boolean. Shall the data be streamed in chunks. Useful
for large data sets.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{seed}{Integer. Random seed for reproducibility.}
}
\value{
A list with the following elements.
\itemize{
\item gene_idx - 0-based integer indicating the gene index.
\item gaerys_c - Gaery's C calculation for the autocorrelation
coefficient.
\item z_score - Z-score of the auto-correlation.
\item pval - P-value derived from the Z-score.
\item fdr - False discovery rate based on the p-value.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function implements the HotSpot auto-correlation functionality and
will return to what extent a given gene shows auto-correlation in the
generated kNN-graph from the embeddings. For details see DeTomaso, et al.
}
\references{
DeTomaso, et al., Cell Systems, 2021
}
\keyword{internal}
