% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hotspot_cluster_genes}
\alias{rs_hotspot_cluster_genes}
\title{Cluster the genes by Z-score together}
\usage{
rs_hotspot_cluster_genes(z_matrix, fdr_threshold, min_size)
}
\arguments{
\item{z_matrix}{Numerical matrix representing the Z-scores.}

\item{fdr_threshold}{Float. The FDR thresholds in terms of the Z-scores.}

\item{min_size}{Integer. Minimum cluster size.}
}
\value{
An assignment vector. NA indicates that the gene did not pass the
thresholds and has not been assigned.
}
\description{
Cluster the genes by Z-score together
}
