% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hotspot_gene_cor}
\alias{rs_hotspot_gene_cor}
\title{Calculate gene<>gene spatial correlations}
\usage{
rs_hotspot_gene_cor(
  f_path_genes,
  f_path_cells,
  embd,
  hotspot_params,
  cells_to_keep,
  genes_to_use,
  streaming,
  verbose,
  seed
)
}
\arguments{
\item{f_path_genes}{Path to the \code{counts_genes.bin} file.}

\item{f_path_cells}{Path to the \code{counts_cells.bin} file.}

\item{embd}{Numerical matrix. The embedding matrix from which to generate
the kNN graph.}

\item{hotspot_params}{List. The HotSpot parameter list.}

\item{cells_to_keep}{Integer vector. 0-index vector indicating which cells
to include in the analysis. Ensure that this is of same order/length
as the embedding matrix.}

\item{genes_to_use}{Integer vector. 0-index vector indicating which genes
to include.}

\item{streaming}{Boolean. Shall the data be streamed in chunks. Useful
for large data sets.}

\item{verbose}{Boolean. Controls verbosity of the function.}

\item{seed}{Integer. Random seed for reproducibility.}
}
\value{
A list with the following elements.
\itemize{
\item cor - A matrix of the N x N genes_to_use length with the auto-
correlation coefficients.
\item z - A matrix of N x N genes_to_use length with the Z-scores of the
local correlations between two genes.
}
}
\description{
This function implements the HotSpot gene <> gene local correlation
functionality from HotSpot, see DeTomaso, et al.
}
\references{
DeTomaso, et al., Cell Systems, 2021
}
