% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hypergeom_test}
\alias{rs_hypergeom_test}
\title{Run a single hypergeometric test.}
\usage{
rs_hypergeom_test(target_genes, gene_sets, gene_universe)
}
\description{
Run a single hypergeometric test.
}
