% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hypergeom_test}
\alias{rs_hypergeom_test}
\title{Run a single hypergeometric test.}
\usage{
rs_hypergeom_test(target_genes, gene_sets, gene_universe)
}
\arguments{
\item{target_genes}{A character vector representing the target gene set.}

\item{gene_sets}{A list of strings that represent the gene sets to test against.}

\item{gene_universe}{A character vector representing the gene universe from which the target genes
and gene sets are sampled from.}
}
\value{
A list with the following elements:
\itemize{
\item pvals, the p-values from the hypergeometric test
\item odds ratios, the calculated odds ratios
\item overlap, the size of the overlap,
\item gene_set_lengths, the length of the gene sets.
}
}
\description{
Given a set of target genes, this is a Rust implementation of an hypergeometric test testing for overenrichment
of the target genes in the gene sets.
}
