% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_hypergeom_test_list}
\alias{rs_hypergeom_test_list}
\title{Run a single hypergeometric test.}
\usage{
rs_hypergeom_test_list(target_genes, gene_sets, gene_universe)
}
\description{
Run a single hypergeometric test.
}
