% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_ica_iters_cv}
\alias{rs_ica_iters_cv}
\title{Run ICA with cross-validation and random initialsiation}
\usage{
rs_ica_iters_cv(
  x_raw,
  no_comp,
  no_folds,
  no_random_init,
  ica_type,
  random_seed,
  ica_params
)
}
\arguments{
\item{x_raw}{Numeric matrix. The processed data (no whitening function has
been applied yet.)}

\item{no_comp}{Integer. Number of components to test for.}

\item{no_random_init}{Integer. Number of random initialisations.}

\item{ica_type}{String. Which type of ICA shall be run.}

\item{random_seed}{Integer. For reproducibility.}

\item{ica_params}{A list containing:
\itemize{
\item maxit - Integer. Maximum number of iterations for ICA.
\item alpha - Float. The alpha parameter for the logcosh version of ICA.
Should be between 1 to 2.
\item max_tol - Maximum tolerance of the algorithm
\item verbose - Verbosity of the function, i.e., shall individual iters
be shown.
}
If the list is empty or the expected elements are not found, default values
are used.}
}
\value{
A list containing:
\itemize{
\item s_combined - The combined matrices for S. Dimensions are nrows =
features; and ncols = ncomp * no_random_init.
\item converged - Boolean vector indicating if the respective run reached
convergence. Length = no_random_init
}
}
\description{
This function will split the data into \code{no_folds} and apply
ICA with \code{no_random_inits} over that fold.
}
