% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_jaccard_row_integers}
\alias{rs_jaccard_row_integers}
\title{Calculate rapidbly Jaccard similarities between rows}
\usage{
rs_jaccard_row_integers(data_1, data_2)
}
\arguments{
\item{data_1}{Integer matrix. The first matrix to compare.}

\item{data_2}{Integer matrix. The second matrix to compare.}
}
\value{
The average Jaccard similarity.
}
\description{
Helper function to quickly calculate the Jaccard similarity
between the rows across the two matrices.
}
