% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_kbet}
\alias{rs_kbet}
\title{Calculate kBET type scores}
\usage{
rs_kbet(knn_mat, batch_vector)
}
\arguments{
\item{knn_mat}{Integer matrix. The rows represent the cells and the
columns the neighbour indices.}

\item{batch_vector}{Integer vector. The integers indicate to which
batch a given cell belongs.}
}
\value{
A vector of p-values based on the ChiSquare statistic per cell.
}
\description{
The function takes in a kNN matrix and a batch vector indicating which
cell belongs to which batch. The function will check for the neighbourhood
of each cell if the proportion of represented batches are different from
the overall batch proportions. Good mixing of batches would mean very
cells have significant differences; bad mixing a lot of the batches
have bad mixing.
}
