% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_knn_mat_to_edge_pairs}
\alias{rs_knn_mat_to_edge_pairs}
\title{Flatten kNN matrix to edge list}
\usage{
rs_knn_mat_to_edge_pairs(knn_mat, one_index)
}
\arguments{
\item{knn_mat}{Integer matrix. Rows represent the samples and the columns
the indices of the k-nearest neighbours.}

\item{one_index}{Boolean. If the original data is 0-index, shall 1-indexed
data be returned.}
}
\value{
A list with the following elements
\itemize{
\item from - the from indices
\item to - the to indices
}
}
\description{
Helper function to leverage Rust to transform a kNN matrix into an edge
list.
}
