% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_ligand_activity_scores}
\alias{rs_ligand_activity_scores}
\title{Calculate the NicheNet ligand activity scores}
\usage{
rs_ligand_activity_scores(ligand_influence, in_gene_sets)
}
\arguments{
\item{ligand_influence}{A ligand x background genes matrix that measures the
ligand to target gene influence.}

\item{in_gene_sets}{A list of logicals with the genes of interest being set
to \code{TRUE} and the background genes set to \code{FALSE}.}
}
\value{
A list with internal lists with:
\itemize{
\item \code{auroc} - The Area Under the Receiver Operating Characteristic for
that ligand
\item \code{aupr} - The Area Under the Precision-Recall curve for that ligand.
\item \code{aupr_corrected} - The corrected AUPR
\item \code{pearson} - The Pearson correlations
\item \code{spearman} - The Spearman correlations
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
