% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mad_outlier}
\alias{rs_mad_outlier}
\title{Calculate MAD outlier detection in Rust.}
\usage{
rs_mad_outlier(x, threshold, direction)
}
\arguments{
\item{x}{Numerical vector to test.}

\item{threshold}{Numeric. Number of MADs in either direction that is
acceptable.}

\item{direction}{String. One of \code{c("below", "above", "twosided")}. Shall
the outlier direction be done for values below the threshold, above the
threshold or in both directions.}
}
\value{
A list with the following items:
\itemize{
\item outlier - Boolean vector if element is an outlier
\item threshold - Applied final threshold
}
}
\description{
Calculate MAD outlier detection in Rust.
}
\details{
Should you provide too short vectors, the function will return an empty
boolean and a threshold of 0.
}
