% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_make_milor_nhoods}
\alias{rs_make_milor_nhoods}
\title{Generate the neighbourhoods akin to the miloR approach}
\usage{
rs_make_milor_nhoods(embd, knn_indices, milor_params, seed, verbose)
}
\arguments{
\item{embd}{Numeric matrix. Represents the matrix used to generate the kNN
graph and will be used to refine the neighbourhoods.}

\item{knn_indices}{Integer matrix. Each row represents a given cell and
the columns the neighbours. (0-indexed!)}

\item{milor_params}{Named list. Contains the parameters for running the
miloR approach.}

\item{seed}{Integer. Seed for reproducibility.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item index_cell - Integer. 0-indexed positions of the cells defining the
neighbourhood.
\item nhoods_i - Integer. 0-indexed positions of the cells in the
neighbourhood.
\item nhoods_j - Integer. To which neighbourhood the cell belongs.
\item nhoods_x - Numeric. The x-value of the COO type matrix, i.e.,
defaults to \code{1.0}.
\item nrows - Integer. Number of cells in the matrix
\item ncols - Integer. Number of refined neighbourhoods.
\item kth_distances - The k-th distances for spatial FDR calculations.
}
}
\description{
Rust version of the
}
