% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_hvg}
\alias{rs_mc_hvg}
\title{Meta cells highly variable genes}
\usage{
rs_mc_hvg(sparse_data, hvg_method, loess_span, binning, n_bins, clip_max)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{format}.}

\item{hvg_method}{String. Which HVG detection method to use. Options
are \code{c("vst", "meanvarbin", "dispersion")}.}

\item{loess_span}{Numeric. The span parameter for the loess function
(only used for \code{"vst"}).}

\item{binning}{String. The binning strategy for the \code{meanvarbin} and
\code{dispersion} methods. One of \code{c("equal_width", "equal_frequency")}.}

\item{n_bins}{Integer. Number of bins for the \code{meanvarbin} and
\code{dispersion} methods.}

\item{clip_max}{Optional clipping number. Defaults to \code{sqrt(no_cells)} if
not provided (only used for \code{"vst"}).}
}
\value{
A list with the HVG statistics. If \code{hvg_method == "vst"}:
\itemize{
\item mean - The average expression of the gene.
\item var - The variance of the gene.
\item var_exp - The expected variance of the gene.
\item var_std - The standardised variance of the gene.
}
For \code{"meanvarbin"} and \code{"dispersion"}:
\itemize{
\item mean - The average expression of the gene.
\item dispersion - The dispersion of the gene.
\item dispersion_scaled - The scaled dispersion per bin per gene.
\item bin - The bin of the gene.
}
}
\description{
Meta cells highly variable genes
}
