% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_pca}
\alias{rs_mc_pca}
\title{PCA on MetaCells (sparse data)}
\usage{
rs_mc_pca(sparse_data, no_pcs, pca_params, clr_offsets, seed, verbose)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{format}.}

\item{no_pcs}{Integer. Number of PCs to return.}

\item{pca_params}{Named list. Contains the parameters to use for this PCA
run.}

\item{clr_offsets}{Optional numeric. If you wish to use the \code{PFlogPF}
normalisation prior to PCA from Booeshaghi, et al.}

\item{seed}{Integer. Random seed for the randomised SVD.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with with the following items
\itemize{
\item scores - The samples projected on the PCA space (solved via sparse
SVD).
\item loadings - The loadings of the features for the PCA (solved via
sparse SVD).
\item singular_values - The singular values for the PCA (solved via sparse
SVD).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates PCA for MetaCells or more generally speaking sparse
data. This is happening in-memory compared to the (usually much) larger
single cell data sets.
}
\references{
Booeshaghi, et al., bioRxive, 2026.
}
