% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_pca}
\alias{rs_mc_pca}
\title{PCA on MetaCells (sparse data)}
\usage{
rs_mc_pca(sparse_data, no_pcs, random_svd, seed)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{format}.}

\item{no_pcs}{Integer. Number of PCs to return.}

\item{random_svd}{Boolean. Shall randomised SVD be used.}

\item{seed}{Integer. Random seed for the randomised SVD.}
}
\value{
A list with with the following items
\itemize{
\item scores - The samples projected on the PCA space (solved via sparse
SVD).
\item loadings - The loadings of the features for the PCA (solved via
sparse SVD).
\item singular_values - The singular values for the PCA (solved via sparse
SVD).
}
}
\description{
Calculates PCA for MetaCells or more generally speaking sparse
data.
}
