% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_scenic}
\alias{rs_mc_scenic}
\title{SCENIC on MetaCells}
\usage{
rs_mc_scenic(sparse_data, tf_indices, scenic_params, seed, verbose)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{format}.}

\item{tf_indices}{Integer vector. The indices of the transcription factors.}

\item{scenic_params}{Named list. Contains all of the parameters need for
SCENIC.}

\item{seed}{Integer. Controls reproducibility of the function.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A gene x TF importance matrix
}
\description{
Assumes that the sparse data is pre-filtered for the genes you wish to
include. The indices need to be 0-indexed.
}
