% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mc_vision_with_autocorrelation}
\alias{rs_mc_vision_with_autocorrelation}
\title{Calculate VISION pathway scores with auto-correlation (for meta cells)}
\usage{
rs_mc_vision_with_autocorrelation(
  sparse_data,
  embd,
  knn_data,
  gs_list,
  random_gs_list,
  vision_params,
  cluster_membership,
  verbose,
  seed
)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{cs_type}. Shape is (metacells, genes) and the
data need to be the \strong{normalised} counts.}

\item{embd}{Numerical matrix. The embedding matrix to use to generate the
kNN graph. Needs to be of the same order/length as the meta cells in
\code{sparse_data}.}

\item{knn_data}{Optional list. This contains pre-computed kNN data
(\code{indices}, \code{dist}, \code{k}) and the \code{dist_metric} it was built with. The user
has to ensure consistency! If provided, this will be used rather than a
graph built from the parameter list.}

\item{gs_list}{Nested list. Each sublist contains the (0-indexed!) positive
and negative gene indices of that specific gene set.}

\item{random_gs_list}{Double-nested list. The outer list represents the
clusters of gene sets and the inner list represents the permutations
within that cluster.}

\item{vision_params}{List. The kNN parameters, only read when no
\code{knn_data} is provided.}

\item{cluster_membership}{Integer vector. 1-indexed(!) position in
\code{random_gs_list} of the permuted cluster each gene set belongs to.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{seed}{Integer. Random seed for reproducibility.}
}
\value{
A list with the following items:
\itemize{
\item autocor_res - List with \code{auto_cor} (1 - Geary's C), \code{p_val} and
\code{fdr}, one entry per gene set.
\item vision_mat - A matrix of meta cells x vision scores per gene set.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The function will take in a list of gene sets that contains lists of \code{"pos"}
and \code{"neg"} gene indices (0-indexed). You don't have to provide the \code{"neg"},
but it can be useful to classify the delta of two stats (EMT, Th1; Th2) etc.
Additionally, it will take a random gene list and calculate an
auto-correlation score based on Geary's C to identify pathways that show
significant patterns on the kNN graph generated on the provided embedding.
This version works on MetaCell counts which are stored in memory directly.
}
\references{
DeTomaso, et al., Nat. Commun., 2019
}
