% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_meld_sc}
\alias{rs_meld_sc}
\title{Run MELD}
\usage{
rs_meld_sc(embd, knn_data, meld_params, labels, n_labels, seed, verbose)
}
\arguments{
\item{embd}{Numeric matrix. The original embedding that was used to generate
the kNN graph.}

\item{knn_data}{Optional named list. This contains pre-computed kNN data
(including distances). The user has to ensure consistency! If provided, this
will be used.}

\item{meld_params}{Named list. Contains the parameters to use for MELD.}

\item{labels}{Integer. The labels of the different groups. (1-indexed!)}

\item{seed}{Integer. For reproducibility.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{landmark}{Boolean. Shall a landmark method be used for accelerated
MELD.}

\item{n_landmarks}{Integer. If \code{landmark = TRUE}, how many landmarks to use.}
}
\value{
A list with the following items
\itemize{
\item raw_scores - The raw MELD scores
\item norm_scores - Negative values were clamped to 0 and the rows L1
normalised. This yields probability-like values.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This implements a Rust-based version of the MELD algorithm, see Burkhardt,
et al. Nat. Biotechnol., 2021.
}
\keyword{internal}
