% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_meld_sc}
\alias{rs_meld_sc}
\title{Run MELD}
\usage{
rs_meld_sc(
  embd,
  knn_data,
  meld_params,
  landmark,
  n_landmark,
  labels,
  n_labels,
  seed,
  verbose
)
}
\arguments{
\item{embd}{Numeric matrix. The original embedding that was used to generate
the kNN graph.}

\item{knn_data}{Optional named list. This contains pre-computed kNN data
(including distances). The user has to ensure consistency! If provided, this
will be used.}

\item{meld_params}{Named list. Contains the parameters to use for MELD.}

\item{landmark}{Boolean. Shall a landmark method be used for accelerated
MELD.}

\item{labels}{Integer. The labels of the different groups.}

\item{seed}{Integer. For reproducibility.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{n_landmarks}{Integer. If \code{landmark = TRUE}, how many landmarks to use.}
}
\value{
A numeric matrix with the MELD values per given condition/cell
}
\description{
This implements a Rust-based version of the MELD algorithm,
see Burkhardt, et al.
}
