% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_metacell_compactness}
\alias{rs_metacell_compactness}
\title{Calculates the compactness of the MetaCells based on diffusion map
coordinates}
\usage{
rs_metacell_compactness(dc, meta_cells)
}
\arguments{
\item{dc}{Numerical matrix. The diffusion map coordinates.}

\item{meta_cells}{List. The cell indices of the meta cells.}
}
\value{
The compactness results
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the meta cell compactness based on the diffusion map coordinates.
}
\keyword{internal}
