% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_metacell_density}
\alias{rs_metacell_density}
\title{Calculates diffusion maps for density calculations for meta cells}
\usage{
rs_metacell_density(knn_data, n_dcs, k_density, knn_params, verbose, seed)
}
\arguments{
\item{knn_data}{Named list. Needs to have the relevant data from the kNN
graph.}

\item{n_dcs}{Integer. The number of diffusion coordinates to return.
Typically \code{10}.}

\item{k_density}{Integer. The k-nearest neighbour to use for the density
estimation. Typically \code{150}.}

\item{knn_params}{List. The kNN parameters defined by
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{seed}{Integer. For reproducibility.}
}
\value{
A list with the following items
\itemize{
\item dcs - Density coordinates
\item density_distances - Density distances at \code{k_density} neighbours.
\item regions - Region of the manifold where this given cell is.
}
}
\description{
Calculates diffusion maps for density calculations for meta cells
}
