% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_metacell_separation}
\alias{rs_metacell_separation}
\title{Calculates the separation of the centroids of the MetaCells based on
diffusion map coordinates.}
\usage{
rs_metacell_separation(dc, meta_cells)
}
\arguments{
\item{dc}{Numerical matrix. The diffusion map coordinates.}

\item{meta_cells}{List. The cell indices of the meta cells.}
}
\value{
The separation results
}
\description{
Calculates the separation of the centroids of the MetaCells based on
diffusion map coordinates.
}
