% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mitch_calc}
\alias{rs_mitch_calc}
\title{Calculate mitch enrichment leveraging Rust under the hood}
\usage{
rs_mitch_calc(x, pathway_list, min_size)
}
\arguments{
\item{x}{Numerical matrix. Each column represents on the contrasts you
wish to test for and the rows represent the gene statistics per contrast.}

\item{pathway_list}{Named list. Each element represents one of the pathways
to test for.}

\item{min_size}{Integer. Minimum size of gene the gene set to be tested for.}
}
\value{
A list with the following elements:
\itemize{
\item pathway_names - The name of the pathway.
\item pathway_sizes The size of the pathway.
\item manova_pvals - The p-value of the MANOVA test.
\item anova_pvals The p-values of the ANOVA test on top of the MANOVA
results. Total length = \code{ncol(x)} * number of pathways.
\item scores - The scores for each pathway set, contrast. Same length
as \code{anova_pvals}.
\item s_dist - Calculated distances from the hypotenuse.
\item sd - SDs of the scores.
}
}
\description{
Calculate mitch enrichment leveraging Rust under the hood
}
