% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mnn}
\alias{rs_mnn}
\title{FastMNN batch correction in Rust}
\usage{
rs_mnn(
  f_path_gene,
  cell_indices,
  gene_indices,
  batch_indices,
  mnn_params,
  verbose,
  seed
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer. The cell indices to use. (0-indexed!)}

\item{gene_indices}{Integer. The gene indices to use. (0-indexed!) Ideally
these are batch-aware highly variable genes.}

\item{batch_indices}{Integer vector. These represent to which batch a given
cell belongs.}

\item{mnn_params}{List. Contains all of the fastMNN parameters.}

\item{verbose}{Boolean. Controls verbosity of the function.}

\item{seed}{Integer. Seed for reproducibility purposes.}
}
\value{
The batch-corrected embedding space.
}
\description{
This function implements the (fast) MNN algorithm from Haghverdi, et al.
Instead of working on the full matrix, it uses under the hood PCA and
generates an aligned embedding space.
}
