% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_mutual_info}
\alias{rs_mutual_info}
\title{Calculates the mutual information matrix}
\usage{
rs_mutual_info(x, n_bins, normalise)
}
\arguments{
\item{x}{R matrix with doubles for which to calculate the mutual information}

\item{n_bins}{Optional integer. Number of bins to use. If \code{NULL} is provided
the function will default to \code{sqrt(nrows(x))}.}

\item{normalise}{Boolean. Shall the normalised mutual information be
calculated via joint entropy.}
}
\value{
The mutual information matrix.
}
\description{
Calculates the mutual information across all columns in the
data.
}
