% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_nmf_single_mc}
\alias{rs_nmf_single_mc}
\title{Run NMF (HALS) on MetaCells}
\usage{
rs_nmf_single_mc(
  sparse_data,
  k,
  preprocessing,
  use_second_layer,
  nmf_hals_params,
  seed,
  verbose
)
}
\arguments{
\item{sparse_data}{A named list with \code{data}, \code{indptr}, \code{indices}, \code{nrow},
\code{ncol} and \code{format}.}

\item{k}{Integer. Number of latent factors to return.}

\item{preprocessing}{String. One of \code{c("none", "sd", "sqrt_sd")}.}

\item{use_second_layer}{Boolean. If \code{TRUE}, runs NMF on normalised counts.}

\item{nmf_hals_params}{Named list. Contains the NMF parameters.}

\item{seed}{Integer. Random seed for initialisation.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with \code{w}, \code{h}, \code{final_loss}, \code{n_iter}, \code{converged}.
}
\description{
Assumes that the sparse data is pre-filtered for the cells/genes you wish
to include. Indices in the sparse data need to be 0-indexed.
}
