% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_nmf_single_sc}
\alias{rs_nmf_single_sc}
\title{Run NMF (HALS) over a set of single cells and genes}
\usage{
rs_nmf_single_sc(
  f_path_gene,
  gene_indices,
  cell_indices,
  k,
  preprocessing,
  use_second_layer,
  nmf_hals_params,
  seed,
  verbose
)
}
\arguments{
\item{f_path_gene}{Path to the \code{counts_genes.bin} file.}

\item{gene_indices}{Integer vector. 0-indexed(!) positions of the genes
to include.}

\item{cell_indices}{Integer vector. 0-indexed(!) positions of cells to
include in the analysis.}

\item{k}{Integer. Number of latent factors to return.}

\item{preprocessing}{String. One of \code{c("none", "sd", "sqrt_sd")}. Takes the
data as is, or scales by standard deviation or squared standard deviation
per feature.}

\item{use_second_layer}{Boolean. If \code{TRUE}, runs NMF on the normalised
counts; if \code{FALSE}, on the raw counts.}

\item{nmf_hals_params}{Named list. Contains the NMF parameters.}

\item{seed}{Integer. Random seed for initialisation.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following items
\itemize{
\item w - The left factor matrix (n_features x k)
\item h - The right factor matrix (k x n_samples)
\item final_loss - Loss at the final iteration
\item n_iter - Number of iterations the algorithm run for
\item converged - Did the NMF algorithm converge
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Runs a single NMF (HALS) run with the specified initialisation.
}
\keyword{internal}
