% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_onto_semantic_sim}
\alias{rs_onto_semantic_sim}
\title{Calculate the semantic similarity in an ontology}
\usage{
rs_onto_semantic_sim(terms, sim_type, ancestor_list, ic_list)
}
\arguments{
\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}

\item{sim_type}{String. Must be one of \code{c("resnik", "lin", "combined")}.}

\item{ancestor_list}{R list with names being the term and the elements in the
list the names of the ancestors.}

\item{ic_list}{R list with the names being the term and the elements the
information content of this given term. Needs to be a single float!}
}
\value{
A vector containing all the desired similarity scores. This is
equivalent of the upper triangle of the similarity matrix.
}
\description{
This function calculates the specified semantic similarity and
returns the full vector (only calculating the upper triangle) for the given
similarity.
}
