% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_onto_semantic_sim_mat}
\alias{rs_onto_semantic_sim_mat}
\title{Calculate the semantic similarity in an ontology}
\usage{
rs_onto_semantic_sim_mat(sim_type, ancestor_list, ic_list, flat_matrix)
}
\arguments{
\item{sim_type}{String. Must be one of \code{c("resnik", "lin", "combined")}.}

\item{ancestor_list}{R list with names being the term and the elements in the
list the names of the ancestors.}

\item{ic_list}{R list with the names being the term and the elements the
information content of this given term. Needs to be a single float!}

\item{flat_matrix}{Boolean. Shall only the upper triangle be returned.}
}
\value{
A list with:
\itemize{
\item sim_mat - the semantic similarity matrix (flat or as matrix.)
\item names - the row and column names for the calculated matrix.
}
}
\description{
This function calculates the specified semantic similarity and
returns the full vector (only calculating the upper triangle) for the given
similarity.
}
