% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_onto_sim_wang}
\alias{rs_onto_sim_wang}
\title{Calculate the Wang similarity for an ontology}
\usage{
rs_onto_sim_wang(parents, children, w, flat_matrix)
}
\arguments{
\item{parents}{String vector. The names of the parents.}

\item{children}{String vector. The names of the childs. The length of
\code{parents} needs to be equal to \code{children}.}

\item{w}{Float. The w parameter for the ontology. Needs to be between
\verb{0 < w < 1}.}

\item{flat_matrix}{Boolean. Shall only the upper triangle be returned.}
}
\value{
A list with:
\itemize{
\item sim_mat - the Wang similarity matrix.
\item names - the row and column names for the calculated matrix.
}
}
\description{
This function calculates the Wang similarity for a given
ontology.
}
