% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_onto_similarity}
\alias{rs_onto_similarity}
\title{Calculate the semantic similarity in an ontology}
\usage{
rs_onto_similarity(terms, ancestor_list, ic_list)
}
\arguments{
\item{terms}{Vector of strings. The terms in the ontology you wish to screen.}

\item{ancestor_list}{R list with names being the term and the elements in the list the names
of the ancestors.}

\item{ic_list}{R list with the names being the term and the elements the information content
of this given term. Needs to be a single float!}
}
\value{
A list with:
\itemize{
\item term1 - String, the first term.
\item term2 - String, the second term.
\item resnik_sim - Float, the unnormalised Resnik similarity.
\item lin_sim - Float, the Lin similarity.
}
}
\description{
This function calculates the Resnik and Lin similarity for a given ontology.
}
