% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_ot_harmonic_sum}
\alias{rs_ot_harmonic_sum}
\title{Calculate the OT harmonic sum}
\usage{
rs_ot_harmonic_sum(x)
}
\arguments{
\item{x}{The numeric vector (should be between 0 and 1) for which to
calculate the harmonic sum}
}
\value{
Returns the harmonic sum according to the OT calculation.
}
\description{
Calculate the OT harmonic sum
}
