% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_page_rank_parallel}
\alias{rs_page_rank_parallel}
\title{Calculate massively parallelised personalised page rank scores}
\usage{
rs_page_rank_parallel(node_names, from, to, diffusion_scores, undirected)
}
\arguments{
\item{node_names}{String vector. Name of the graph nodes.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{diffusion_scores}{List. The personalised vectors for the page rank reset
values. Each element must sum to 1 and be of same length of \code{node_names}!}

\item{undirected}{Boolean. Is this an undirected graph.}
}
\value{
A matrix of the scores with each row representing an element in the
\code{diffusion_scores} list (in order), and each column representing the value
of the personalised page rank diffusion for this node.
}
\description{
Helper function to calculate in parallel on the same (unweighted)
network the personalised page rank as fast as possible. Can be used for permutations
type approaches.
}
