% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_pairwise_gene_cors}
\alias{rs_pairwise_gene_cors}
\title{Calculates pairwise gene correlations in single cell}
\usage{
rs_pairwise_gene_cors(
  f_path,
  gene_indices_1,
  gene_indices_2,
  cells_to_keep,
  spearman
)
}
\arguments{
\item{f_path}{Path to the \code{counts_genes.bin} file.}

\item{gene_indices_1}{Integer vector. The first set of gene indices to
correlate against \code{gene_indices_2} (0-indexed!)}

\item{gene_indices_2}{Integer vector. The second set of gene indices. Needs
to be of same length as \code{gene_indices_1}. (0-indexed!)}

\item{cells_to_keep}{Integer. The indices of the cells to include in this
analysis. (0-indexed!)}

\item{spearman}{Boolean. Shall spearman correlation be used.}
}
\value{
The vector of correlations between the pairs of gene_indices_1
and gene_indices_2
}
\description{
Calculates pairwise gene correlations in single cell
}
