% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_pairwise_gene_cors_mc}
\alias{rs_pairwise_gene_cors_mc}
\title{Calculate the pairwise gene-correlation for meta cells}
\usage{
rs_pairwise_gene_cors_mc(
  sparse_data,
  gene_indices_1,
  gene_indices_2,
  spearman,
  verbose
)
}
\arguments{
\item{sparse_data}{A named list that needs to have \code{data}, \code{indptr},
\code{indices}, \code{nrow}, \code{ncol} and \code{format}.}

\item{gene_indices_1}{Integer. The gene indices for the first set of genes.
Must be 0-indexed!}

\item{gene_indices_2}{Integer. The gene indices for the first set of genes.
Must be 0-indexed!}

\item{spearman}{Boolean. Shall the spearman correlation be calculated.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
The vector of correlations between the pairs of gene_indices_1
and gene_indices_2
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
