% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_phyper}
\alias{rs_phyper}
\title{Calculate the hypergeometric rest in Rust}
\usage{
rs_phyper(q, m, n, k)
}
\arguments{
\item{q}{Number of white balls drawn out of urn.}

\item{m}{Number of white balls in the urn.}

\item{n}{Number of black balls in the urn.}

\item{k}{The number of balls drawn out of the urn.}
}
\value{
P-value (with lower.tail set to False)
}
\description{
Calculate the hypergeometric rest in Rust
}
