% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_prepare_whitening}
\alias{rs_prepare_whitening}
\title{Prepare the data for whitening}
\usage{
rs_prepare_whitening(x, fast_svd, seed, rank, oversampling, n_power_iter)
}
\arguments{
\item{x}{The matrix to whiten. The whitening will happen over the columns.}

\item{fast_svd}{Boolean. Shall a randomised SVD be used. This is way faster
on larger data sets.}

\item{seed}{Integer. Only relevant with fast_svd is set to \code{TRUE}.}

\item{rank}{Integer. How many ranks to use for the fast SVD approximation.
If you supply \code{NULL}, it will default to \code{10L}. Only relevant with
fast_svd is set to \code{TRUE}.}

\item{oversampling}{Integer. Oversampling parameter to make the approximation
more precise. If you supply \code{NULL}, it will default to \code{10L}. Only relevant
with fast_svd is set to \code{TRUE}.}

\item{n_power_iter}{Integer. How much shall the QR low rank approximation be
powered. If you supply \code{NULL}, it will default to \code{2L}.}
}
\value{
A list containing:
\itemize{
\item x - The preprocessed matrix.
\item k - The pre-whitening matrix k.
}
}
\description{
Prepares the data for subsequent usag in ICA.
WARNING! Incorrect use can cause kernel crashes. Wrapper around the Rust
functions with type checks are provided in the package.
}
