% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_prepare_whitening}
\alias{rs_prepare_whitening}
\title{Prepare the data for whitening}
\usage{
rs_prepare_whitening(x)
}
\arguments{
\item{x}{The matrix to whiten. The whitening will happen over the columns.}
}
\value{
A list containing:
\itemize{
\item whiten - The whitened matrix.
\item k - The K matrix.
}
}
\description{
Whitens the matrix for subsequent usage. This is a need pre-
processing step for ICA.
}
