% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_prepare_whitening}
\alias{rs_prepare_whitening}
\title{Prepare the data for whitening}
\usage{
rs_prepare_whitening(x)
}
\arguments{
\item{x}{The matrix to whiten. The whitening will happen over the columns.}
}
\value{
A list containing:
\itemize{
\item x - The transposed and scaled data for subsequent usage in ICA.
\item k - The K matrix.
}
}
\description{
Prepares the data for subsequent usag in ICA.
WARNING! Incorrect use can cause kernel crashes. Wrapper around the Rust functions
with type checks are provided in the package.
}
