% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_pseudobulk_cells_dense}
\alias{rs_pseudobulk_cells_dense}
\title{Pseudo-bulk a set of cells (dense)}
\usage{
rs_pseudobulk_cells_dense(f_path, cell_indices_ls, assay, verbose)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{cell_indices_ls}{List. Must contains 0-indexed positions of the
cells to aggregate per element.}

\item{assay}{String. One of \code{c("raw", "norm")}. Which counts to normalise.}

\item{verbose}{Controls verbosity of the function}
}
\value{
A dense matrix with the pseudo-bulked data.
}
\description{
This function will return a dense matrix of
\verb{length(cell_indices_ls) x number of genes}. The function has the option
to return the sum of the sum of the raw counts or the average of the
normalised counts.
}
