% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_pseudobulk_cells_sparse}
\alias{rs_pseudobulk_cells_sparse}
\title{Pseudo-bulk a set of cells (sparse)}
\usage{
rs_pseudobulk_cells_sparse(f_path, cell_indices_ls, assay, verbose)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{cell_indices_ls}{List. Must contains 0-indexed positions of the
cells to aggregate per element.}

\item{assay}{String. One of \code{c("raw", "norm")}. Which counts to normalise.}

\item{verbose}{Controls verbosity of the function}
}
\value{
A list with the following elements (easy to convert into CSR in R)
\itemize{
\item indptr - The index pointers (representing cells)
\item indices - The indices (representing genes)
\item data - The pseudo-bulked data
\item nrow - Number of rows (i.e., pseudo-bulked samples)
\item ncol - Number of columns
}
}
\description{
This function will return a sparse matrix of
\verb{length(cell_indices_ls) x number of genes} (in list form in CSR).
The function has the option to return the sum of the sum of the raw counts
or the average of the normalised counts.
}
