% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_random_svd}
\alias{rs_random_svd}
\title{Run randomised SVD over a matrix}
\usage{
rs_random_svd(x, scale, rank, seed, oversampling, n_power_iter)
}
\arguments{
\item{x}{Numeric matrix. Rows = samples, columns = features.}

\item{scale}{Boolean. Shall the columns be variance normalised. (Mean
centring will automatically occur.)}

\item{rank}{Integer. The rank to use.}

\item{seed}{Integer. Random seed for reproducibility.}

\item{oversampling}{Integer. Defaults to \code{10L} if nothing is provided.}

\item{n_power_iter}{Integer. How often shall the QR decomposition be
applied. Defaults to \code{2L} if nothing is provided.}
}
\value{
A list with:
\itemize{
\item u - u matrix of the SVD.
\item v - v matrix of the SVD.
\item s - Eigenvalues of the SVD.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Runs a randomised singular value decomposition over a matrix. This
implementation is faster than the full SVD on large data sets, with slight
loss in precision.
}
