% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_range_norm}
\alias{rs_range_norm}
\title{Apply a range normalisation on a vector.}
\usage{
rs_range_norm(x, max_val, min_val)
}
\arguments{
\item{x}{Numerical vector. The data to normalise.}

\item{max_val}{Numeric. The upper bound value to normalise into. If set to 1,
the function will be equal to a min-max normalisation.}

\item{min_val}{Numeric. The lower bound value to normalise into. If set to 0,
the function will equal a min-max normalisation.}
}
\value{
Normalised values
}
\description{
Applies a range normalisation on an R vector.
}
