% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_rbf_function}
\alias{rs_rbf_function}
\title{Apply a Gaussian affinity kernel to a distance metric}
\usage{
rs_rbf_function(x, epsilon, rbf_type)
}
\arguments{
\item{x}{Numeric vector. The distances you wish to apply the Gaussian kernel
onto.}

\item{bandwidth}{The bandwidth of the kernel. Smaller values will yield
smaller affinities.}
}
\value{
The affinities after the Kernel was applied.
}
\description{
Applies a Gaussian kernel to a vector of distances.
}
