% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_rbh_cor}
\alias{rs_rbh_cor}
\title{Generate reciprocal best hits based on correlations}
\usage{
rs_rbh_cor(module_matrices, spearman, min_similarity)
}
\arguments{
\item{module_matrices}{A list of named matrices. Rows represent features
and columns the samples you wish to calculate the correlations for.}

\item{spearman}{Shall Spearman correlation be used.}

\item{min_similarity}{Minimum (absolute) correlations that needs to exist
between two terms.}
}
\value{
A list containing:
\itemize{
\item origin - The name of the origin of the gene modules.
\item target - The name of the target of the gene modules.
\item comparisons - Integer vector indicating how many RBH hits were
identified in this comparison
\item origin_modules - Names of the gene modules from the origin.
\item target_modules - Names of the gene modules from the target.
\item similarity - The similarities between the two respective gene
modules.
}
}
\description{
This function takes list of (named) matrices which represent
for example matrix factorisation results you wish to identify reciprocal
best hits (RBH) for. The rows need to represent the features and the columns
the parts you wish to calculate the RBH for.
}
