% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_read_tenx_h5_modality}
\alias{rs_read_tenx_h5_modality}
\title{Loads in a modality from a 10x h5 file}
\usage{
rs_read_tenx_h5_modality(f_path, version, feature_type)
}
\arguments{
\item{f_path}{String. The path to the h5 file}

\item{version}{String. The 10x version. If \code{"auto"} uses the automatic
detection.}

\item{feature_type}{String. The feature type to return.}
}
\value{
A list with:
\itemize{
\item counts - Numerical matrix of cells x features
\item barcodes - The barcodes as a string
\item features - The features as a string
}
}
\description{
Loads in a modality from a 10x h5 file
}
