% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_doublet_detection}
\alias{rs_sc_doublet_detection}
\title{Detect Doublets via BoostClassifier (in Rust)}
\usage{
rs_sc_doublet_detection(
  f_path_gene,
  f_path_cell,
  cells_to_keep,
  boost_params,
  seed,
  streaming,
  verbose
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file.}

\item{cells_to_keep}{Integer vector. The indices (0-indexed!) of the cells
to include in this analysis.}

\item{boost_params}{List. Parameter list, see
\code{\link[=params_boost]{params_boost()}}.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{streaming}{Boolean. Shall the data be streamed for the HVG
calculations.}

\item{verbose}{Boolean. Controls verbosity}
}
\value{
A list with
\itemize{
\item predicted_doublets - Boolean vector indicating which observed cells
predicted as doublets (TRUE = doublet, FALSE = singlet).
\item doublet_scores_obs - Numerical vector with the likelihood of being
a doublet for the observed cells.
\item voting_avg - Voting average across the different iterations.
}
}
\description{
Detect Doublets via BoostClassifier (in Rust)
}
