% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_get_gene_set_perc}
\alias{rs_sc_get_gene_set_perc}
\title{Calculate the percentage of gene sets in the cells}
\usage{
rs_sc_get_gene_set_perc(
  f_path_cell,
  gene_set_idx,
  cell_indices,
  streaming,
  verbose
)
}
\arguments{
\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file.}

\item{gene_set_idx}{Named list with integer(!) positions (0-indexed!) as
elements of the genes of interest.}

\item{cell_indices}{Integer. The indices of the cells for which to calculate
the proportions. (0-indexed!)}

\item{streaming}{Boolean. Shall the data be worked on in chunks.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the percentages of counts per gene set group detected
in the cells.
}
\description{
This function allows to calculate for example the proportion of
mitochondrial genes, or ribosomal genes in the cells for QC purposes.
}
