% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_get_top_genes_perc}
\alias{rs_sc_get_top_genes_perc}
\title{Calculates the cumulative proportion of the top X genes}
\usage{
rs_sc_get_top_genes_perc(
  f_path_cell,
  top_n_vals,
  cell_indices,
  streaming,
  verbose
)
}
\arguments{
\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file.}

\item{top_n_vals}{Integer. The different Top X to look for.}

\item{cell_indices}{Integer. The indices of the cells for which to calculate
the proportions. (0-indexed!)}

\item{streaming}{Boolean. Shall the data be worked on in chunks.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the cumulative percentages of the Top X genes defined
as in \code{top_n_vals}.
}
\description{
This calculates the cumulative proportion of the top X genes, for example
Top10, 50, 100. High values here indicate low complexity samples, i.e.,
bad quality.
}
