% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_hvg}
\alias{rs_sc_hvg}
\title{Calculate the percentage of gene sets in the cells}
\usage{
rs_sc_hvg(
  f_path_gene,
  hvg_method,
  cell_indices,
  loess_span,
  binning,
  n_bins,
  clip_max,
  streaming,
  verbose
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{hvg_method}{String. Which HVG detection method to use. One of
\code{c("vst", "meanvarbin", "dispersion")}.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep. Must be unique and within the store; duplicates or out-of-range
positions raise an error.}

\item{loess_span}{Numeric. The span parameter for the loess function. Must
be within \verb{(0, 1]}.}

\item{binning}{String. The binning strategy for the \code{meanvarbin} method. One
of \code{c("equal_width", "equal_frequency")}.}

\item{n_bins}{Integer. Number of bins for the \code{meanvarbin} method.}

\item{clip_max}{Optional clipping number. Defaults to \code{sqrt(no_cells)} if
not provided.}

\item{streaming}{Boolean. Shall the genes be streamed in to reduce memory
pressure.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the highly variable genes. If \code{hvg_method == "vst"}, the
following elements can be found:
\itemize{
\item mean - The average expression of the gene.
\item var - The variance of the gene.
\item var_exp - The expected variance of the gene.
\item var_std - The standardised variance of the gene.
}
For the other two methods, these elements can be found:
\itemize{
\item mean - The average expression of the gene.
\item dispersion - The dispersion of the gene
\item dispersion_scaled - The scaled dispersion per bin per gene
\item bin - The bin of the gene
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function identifies highly variable genes with the three methods known
in Seurat.
}
\keyword{internal}
