% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_hvg}
\alias{rs_sc_hvg}
\title{Calculate the percentage of gene sets in the cells}
\usage{
rs_sc_hvg(f_path_gene, hvg_method, cell_indices, loess_span, clip_max)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{hvg_method}{String. Which HVG detection method to use. Options
are \code{c("vst", "meanvarbin", "dispersion")}. So far, only the first is
implemented.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep.}

\item{loess_span}{Numeric. The span parameter for the loess function.}

\item{clip_max}{Optional clipping number. Defaults to \code{sqrt(no_cells)} if
not provided.}
}
\value{
A list with the percentages of counts per gene set group detected
in the cells:
\itemize{
\item mean - The average expression of the gene.
\item var - The variance of the gene.
\item var_exp - The expected variance of the gene.
\item var_std - The standardised variance of the gene.
}
}
\description{
This function allows to calculate for example the proportion of
mitochondrial genes, or ribosomal genes in the cells for QC purposes.
}
