% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_hvg_batch_aware}
\alias{rs_sc_hvg_batch_aware}
\title{Calculate HVG per batch}
\usage{
rs_sc_hvg_batch_aware(
  f_path_gene,
  hvg_method,
  cell_indices,
  batch_labels,
  loess_span,
  binning,
  n_bins,
  clip_max,
  streaming,
  verbose
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{hvg_method}{String. Which HVG detection method to use. One of
\code{c("vst", "meanvarbin", "dispersion")}.}

\item{cell_indices}{Integer positions (0-indexed!) that defines the cells
to keep.}

\item{batch_labels}{Integer vector (0-indexed!) defining batch membership
for each cell. Must be same length as \code{cell_indices}.}

\item{loess_span}{Numeric. The span parameter for the loess function.}

\item{binning}{String. The binning strategy for the \code{meanvarbin} method. One
of \code{c("equal_width", "equal_frequency")}.}

\item{n_bins}{Integer. Number of bins for the \code{meanvarbin} method.}

\item{clip_max}{Optional clipping number. Defaults to \code{sqrt(no_cells)} per
batch if not provided.}

\item{streaming}{Boolean. Shall the genes be streamed in to reduce memory
pressure.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with HVG statistics concatenated across all batches. For
\code{hvg_method == 'vst'}, the following elements can be found:
\itemize{
\item mean - The average expression of each gene in each batch.
\item var - The variance of each gene in each batch.
\item var_exp - The expected variance of each gene in each batch.
\item var_std - The standardised variance of each gene in each batch.
\item batch - Batch index for each gene (length = n_genes * n_batches).
\item gene_idx - Gene index for each entry (0-indexed, length = n_genes *
n_batches).
}
For the other methods
\itemize{
\item mean - The average expression of each gene in each batch.
\item dispersion - The dispersion of the gene in each batch.
\item dispersion_scaled - The scaled dispersion per bin per gene in each
batch.
\item bin - The bin of the gene in each batch.
\item batch - Batch index for each gene (length = n_genes * n_batches).
\item gene_idx - Gene index for each entry (0-indexed, length = n_genes *
n_batches).
}
}
\description{
Batch-aware highly variable gene detection. Calculates HVG statistics
separately for each batch, allowing for downstream selection strategies
such as union of top genes per batch.
}
