% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_knn}
\alias{rs_sc_knn}
\title{Generates the kNN graph}
\usage{
rs_sc_knn(
  embd,
  no_neighbours,
  n_trees,
  search_budget,
  algorithm_type,
  ann_dist,
  verbose,
  seed
)
}
\arguments{
\item{embd}{Numerical matrix. The embedding matrix to use to generate the
kNN graph.}

\item{no_neighbours}{Integer. Number of neighbours to return}

\item{n_trees}{Integer. Number of trees to use for the \code{"annoy"} algorithm.}

\item{search_budget}{Integer. Search budget per tree for the \code{"annoy"}
algorithm.}

\item{algorithm_type}{String. Which of the two implemented algorithms to
use. One of \code{c("annoy", "hnsw")}.}

\item{verbose}{Boolean. Controls verbosity of the function and returns
how long certain operations took.}

\item{seed}{Integer. Seed for reproducibility purposes.}
}
\value{
A integer matrix of N x k with N being the number of cells and k the
number of neighbours.
}
\description{
This function is a wrapper over the Rust-based generation of the approximate
nearest neighbours. You have two options to generate the kNNs. \code{"annoy"} or
\code{"hnsw"}.
}
