% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_knn_w_dist}
\alias{rs_sc_knn_w_dist}
\title{Generates the kNN graph with additional distances}
\usage{
rs_sc_knn_w_dist(embd, knn_params, verbose, seed)
}
\arguments{
\item{embd}{Numerical matrix. The embedding matrix to use to generate the
kNN graph.}

\item{knn_params}{List. The kNN parameters defined by
\code{\link[=params_sc_neighbours]{params_sc_neighbours()}}.}

\item{verbose}{Boolean. Controls verbosity of the function and returns
how long certain operations took.}

\item{seed}{Integer. Seed for reproducibility purposes.}
}
\value{
A list with:
\itemize{
\item indices - An integer matrix representing the indices of the
approximate nearest neighbours.
\item dist - An numerical matrix representing the distances to the nearest
neighbours.
\item dist_metric - String representing the used distance metric.
}
}
\description{
This function is a wrapper over the Rust-based generation of the approximate
nearest neighbours. You have several options to get the approximate nearest
neighbours:
\itemize{
\item \code{"annoy"}: leverages binary trees to generate rapidly in a parallel manner
an index. Good compromise of index generation, querying speed.
\item \code{"hnsw"}: uses a hierarchical navigatable small worlds index under the
hood. The index generation takes more long, but higher recall and ideal
for very large datasets due to subdued memory pressure.
\item \code{"nndescent"}: an index-free approximate nearest neighbour algorithm
that is ideal for small, ephemeral kNN graphs.
}
}
