% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_pca}
\alias{rs_sc_pca}
\title{Calculates PCA for single cell}
\usage{
rs_sc_pca(
  f_path_gene,
  f_path_cell,
  no_pcs,
  pca_params,
  cell_indices,
  gene_indices,
  seed,
  return_scaled,
  verbose
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file. Used if
you wish to use the PFlogPF transformation.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{pca_params}{Named list. Contains the parameters to use for this PCA
run.}

\item{cell_indices}{Integer. The cell indices to use. (0-indexed!)}

\item{gene_indices}{Integer. The gene indices to use. (0-indexed!)}

\item{seed}{Integer. Random seed for the randomised SVD.}

\item{return_scaled}{Boolean. Shall the scaled data be returned.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with with the following items
\itemize{
\item scores - The samples projected on the PCA space.
\item loadings - The loadings of the features for the PCA.
\item singular_values - The singular values for the PCA.
\item scaled - The scaled matrix if you set return_scaled to \code{TRUE}.
}
}
\description{
Helper function that will calculate the PCA for the specified highly
variable genes. You have the option to do mean centering, variance
normalisation and/or apply the new proposed transformation \code{PFlogPF} from
Booeshaghi, et al.
}
\references{
Booeshaghi, et al., bioRxive, 2026.
}
