% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_pca}
\alias{rs_sc_pca}
\title{Calculates PCA for single cell}
\usage{
rs_sc_pca(
  f_path_gene,
  no_pcs,
  random_svd,
  cell_indices,
  gene_indices,
  seed,
  return_scaled,
  verbose
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{no_pcs}{Integer. Number of PCs to calculate.}

\item{random_svd}{Boolean. Shall randomised SVD be used.}

\item{cell_indices}{Integer. The cell indices to use. (0-indexed!)}

\item{gene_indices}{Integer. The gene indices to use. (0-indexed!)}

\item{seed}{Integer. Random seed for the randomised SVD.}

\item{return_scaled}{Boolean. Shall the scaled data be returned.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with with the following items
\itemize{
\item scores - The samples projected on the PCA space.
\item loadings - The loadings of the features for the PCA.
\item scaled - The scaled matrix if you set return_scaled to \code{TRUE}.
}
}
\description{
Helper function that will calculate the PCA for the specified highly
variable genes. Has the option to use randomised SVD for faster solving
of the PCA.
}
