% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_scdblfinder}
\alias{rs_sc_scdblfinder}
\title{Run scDblFinder doublet detection}
\usage{
rs_sc_scdblfinder(
  f_path_gene,
  f_path_cell,
  cell_indices,
  params,
  return_features,
  streaming,
  seed,
  verbose
)
}
\arguments{
\item{f_path_gene}{String. Path to the gene-based binary file.}

\item{f_path_cell}{String. Path to the cell-based binary file.}

\item{cell_indices}{Integer vector (0-indexed).}

\item{params}{List. scDblFinder parameters from R.}

\item{return_features}{Boolean. Return the features for the observed cells
that are used to train the classifier.}

\item{streaming}{Boolean. Shall the gene data be streamed in for the
selection of the top genes.}

\item{seed}{Integer. Seed for reproducibility.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with predicted_doublets, doublet_scores, threshold,
cluster_labels and detected_doublet_rate.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Implementation of scDblFinder in Rust.
}
\keyword{internal}
