% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_scdblfinder}
\alias{rs_sc_scdblfinder}
\title{Run scDblFinder doublet detection}
\usage{
rs_sc_scdblfinder(
  f_path_gene,
  f_path_cell,
  cell_indices,
  params,
  return_features,
  seed,
  verbose,
  debug
)
}
\arguments{
\item{f_path_gene}{String. Path to the gene-based binary file.}

\item{f_path_cell}{String. Path to the cell-based binary file.}

\item{cell_indices}{Integer vector (0-indexed).}

\item{params}{List. scDblFinder parameters from R.}

\item{return_features}{Boolean. Return the features for the observed cells
that are used to train the classifier.}

\item{seed}{Integer. Seed for reproducibility.}

\item{verbose}{Boolean. Controls verbosity.}

\item{debug}{Boolean. Additional verbosity for debugging purposes.}
}
\value{
A list with predicted_doublets, doublet_scores, threshold,
cluster_labels and detected_doublet_rate.
}
\description{
Run scDblFinder doublet detection
}
