% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_scrublet}
\alias{rs_sc_scrublet}
\title{Scrublet Rust interface}
\usage{
rs_sc_scrublet(
  f_path_gene,
  f_path_cell,
  cells_to_keep,
  scrublet_params,
  seed,
  verbose,
  streaming,
  return_combined_pca,
  return_pairs
)
}
\arguments{
\item{f_path_gene}{String. Path to the \code{counts_genes.bin} file.}

\item{f_path_cell}{String. Path to the \code{counts_cells.bin} file.}

\item{cells_to_keep}{Integer vector. The indices (0-indexed!) of the cells
to include in this analysis.}

\item{scrublet_params}{List. Parameter list, see
\code{\link[=params_scrublet]{params_scrublet()}}.}

\item{seed}{Integer. Seed for reproducibility purposes.}

\item{verbose}{Boolean. Controls verbosity}

\item{streaming}{Boolean. Shall the data be streamed for the HVG
calculations.}
}
\value{
A list with
\itemize{
\item predicted_doublets - Boolean vector indicating which observed cells
predicted as doublets (TRUE = doublet, FALSE = singlet).
\item doublet_scores_obs - Numerical vector with the likelihood of being
a doublet for the observed cells.
\item doublet_scores_sim - Numerical vector with the likelihood of being
a doublet for the simulated cells.
\item doublet_errors_obs - Numerical vector with the standard errors of
the scores for the observed cells.
\item z_scores - Z-scores for the observed cells. Represents:
\code{score - threshold / error}.
\item threshold - Used threshold.
\item detected_doublet_rate - Fraction of cells that are called as
doublet.
\item detectable_doublet_fraction - Fraction of simulated doublets with
scores above the threshold.
\item overall_doublet_rate - Estimated overall doublet rate. Should roughly
match the expected doublet rate.
\item pca - Optional PCA embeddings across the original cells and simulated
doublets.
\item pair_1 - Optional integer vector representing the first parent of the
simulated doublets.
\item pair_2 -  Optional integer vector representing the second parent of
the simulated doublets.
}
}
\description{
Scrublet Rust interface
}
