% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_snn}
\alias{rs_sc_snn}
\title{Generates the sNN graph for igraph}
\usage{
rs_sc_snn(knn_mat, snn_method, limited_graph, pruning, verbose)
}
\arguments{
\item{knn_mat}{Integer matrix. Rows represent cells and the columns
represent the neighbours (0-indexed!).}

\item{snn_method}{String. Which method to use to calculate the similarity.
Choice of \code{c("jaccard", "rank")}; anything else errors.}

\item{limited_graph}{Boolean. Shall the sNNs only be calculated between
direct neighbours in the graph, or between all possible combinations.}

\item{pruning}{Float. Below which similarity value to prune the weight
to 0.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with the following items:
\itemize{
\item edges - Integer vector with the flattened sNN edge pairs
(1-indexed!), ready for \code{igraph::add_edges()}.
\item weights - sNN weights of the pairs above.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function takes a kNN matrix and generates the inputs for an SNN
graph based on it.
}
