% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_snn}
\alias{rs_sc_snn}
\title{Generates the sNN graph for igraph}
\usage{
rs_sc_snn(knn_mat, snn_method, limited_graph, pruning, verbose)
}
\arguments{
\item{knn_mat}{Integer matrix. Rows represent cells and the columns
represent the neighbours.}

\item{snn_method}{String. Which method to use to calculate the similarity.
Choice of \code{c("jaccard", "rank")}.}

\item{pruning}{Float. Below which value for the Jaccard similarity to prune
the weight to 0.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A integer matrix of N x k with N being the number of cells and k the
number of neighbours.
}
\description{
This function takes a kNN matrix and generates the inputs for an SNN
graph based on it.
}
