% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_type}
\alias{rs_sc_type}
\title{Run the ScType scoring approach}
\usage{
rs_sc_type(
  f_path,
  cell_indices,
  cell_markers,
  sensitivity,
  weight_floor,
  verbose
)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer vector. 0-indexed(!) positions of cells to
include in the analysis}

\item{cell_markers}{A list with the cell marker gene indices.}

\item{sensitivity}{Boolean. Shall a sensitivity correction be applied that
downweights common cell type markers.}

\item{weight_floor}{Optional numeric. If \code{sensitivity = TRUE}, what is
the weight floor. If not provided, defaults to \code{0.1}.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with
\itemize{
\item cell_types - String vector. The cell types
\item scores - Row-major scores (cells x cell_types).
\item n_cells - Number of cells
\item n_cell_types - Number of cell types
}
}
\description{
This Rust function implements the cell type scoring approach
from Ianevski et al. (2022).
}
