% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sc_type_cluster_assignment}
\alias{rs_sc_type_cluster_assignment}
\title{Score the individual clusters based on ScType}
\usage{
rs_sc_type_cluster_assignment(sc_type_res, cluster_labels)
}
\arguments{
\item{sc_type_res}{List. The ScType results.}

\item{cluster_labels}{Integer. Cluster assignment. Needs to be of length
of scored cells.}
}
\value{
A list with
\itemize{
\item cluster_id - The cluster id/integer
\item cell_type - String; the predicted cell type
\item score - The final score for the clsuter.
\item n_cells - The number of cells in the cluster.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This Rust function implements the cell type scoring approach from Ianevski
et al. (2022).
}
\keyword{internal}
