% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_scenic_gene_filter}
\alias{rs_scenic_gene_filter}
\title{Identifies genes to include into a SCENIC analysis}
\usage{
rs_scenic_gene_filter(f_path_genes, cell_indices, scenic_params, verbose)
}
\arguments{
\item{f_path_genes}{Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer vector. 0-indexed(!) positions of cells to
include in the analysis}

\item{scenic_params}{Named list. Contains all of the parameters need for
SCENIC.}

\item{verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
The 0-indexed positions of the genes to include in the scenic
analysis.
}
\description{
Identifies genes to include into a SCENIC analysis
}
