% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_scenic_grn}
\alias{rs_scenic_grn}
\title{SCENIC to generate gene-regulatory networks}
\usage{
rs_scenic_grn(
  f_path_genes,
  cell_indices,
  gene_indices,
  tf_indices,
  scenic_params,
  seed,
  verbose
)
}
\arguments{
\item{f_path_genes}{Path to the \code{counts_genes.bin} file.}

\item{cell_indices}{Integer vector. 0-indexed(!) positions of cells to
include in the analysis}

\item{gene_indices}{Integer vector. 0-indexed(!) positions of the genes
to include.}

\item{tf_indices}{Integer vector. 0-indexed(!) positions of the TF
predictor variables to use in the generation of the regression learners.}

\item{scenic_params}{Named list. Contains all of the parameters need for
SCENIC.}

\item{Seed.}{Controls reproducibility of the function.}

\item{Boolean.}{Controls the verbosity of the function.}
}
\value{
A gene x TF importance matrix
}
\description{
SCENIC to generate gene-regulatory networks
}
