% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_similarity}
\alias{rs_set_similarity}
\title{Set similarities}
\usage{
rs_set_similarity(s_1, s_2, overlap_coefficient)
}
\arguments{
\item{s_1}{The String vector against which to calculate the set similarities.}

\item{s_2}{The String vector against which to calculate the set similarities.}

\item{overlap_coefficient}{Boolean. Use the overlap coefficient instead of the Jaccard similarity be calculated.}
}
\description{
This function calculates the Jaccard or similarity index between a two given
string vector and a  of other string vectors.
}
