% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_similarity_list}
\alias{rs_set_similarity_list}
\title{Set similarities over one list}
\usage{
rs_set_similarity_list(list, overlap_coefficient)
}
\arguments{
\item{list}{A named R list.}

\item{overlap_coefficient}{Boolean. Use the overlap coefficient instead of
the Jaccard similarity be calculated.}
}
\value{
A list with the following items:
\itemize{
\item from - Name of element i
\item to - Name of element j
\item sim - Similarity between the two elements
}
}
\description{
This function calculates the set similarity via Jaccard or
overlap coefficient across all permutations of one list.
}
