% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_similarity_list}
\alias{rs_set_similarity_list}
\title{Set similarities over list}
\usage{
rs_set_similarity_list(s_1_list, s_2_list, overlap_coefficient)
}
\arguments{
\item{s_2_list}{A List of vector against which to calculate the set similarities.}

\item{overlap_coefficient}{Boolean. Use the overlap coefficient instead of the Jaccard similarity be calculated.}

\item{s_1}{The String vector against which to calculate the set similarities.}
}
\description{
This function calculates the Jaccard or similarity index between a one given
string vector and list of vectors.
}
