% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_simple_and_multi_err}
\alias{rs_simple_and_multi_err}
\title{Calculates the simple and multi error for fgsea multi level}
\usage{
rs_simple_and_multi_err(n_more_extreme, nperm, sample_size)
}
\arguments{
\item{n_more_extreme}{Integer vector. The number of times the ES was larger than the
permutations.}

\item{nperm}{Integer. Number of permutations.}

\item{sample_size}{Integer. Number of samples.}
}
\value{
List with the following elements:
\itemize{
\item simple_err Vector of simple errors.
\item multi_err Vector of multi errors.
}
}
\description{
Calculates the simple and multi error for fgsea multi level
}
