% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_singscore_multi}
\alias{rs_singscore_multi}
\title{Rust version of singscore for many gene sets}
\usage{
rs_singscore_multi(
  ranks,
  up_list,
  down_list,
  center_score,
  known_direction,
  stable
)
}
\arguments{
\item{ranks}{Numerical matrix. The ranked expression matrix.}

\item{up_list}{List. Up gene sets as zero-indexed indices. See
\code{\link[=rs_prepare_gsva_gs]{rs_prepare_gsva_gs()}}.}

\item{down_list}{List or NULL. Paired down gene sets, same length and
ordering as \code{up_list}.}

\item{center_score}{Boolean.}

\item{known_direction}{Boolean.}

\item{stable}{Boolean.}
}
\value{
A named list with
\itemize{
\item \code{scores} - Numerical matrix with the scores
\item \code{dispersion} - Numerical matrix with the dispersions
}
Both matrices are of shape gene_sets × samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Rust-based implementation of singscore over many gene sets with optional
paired down sets.
}
