% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_snf_affinity_continuous}
\alias{rs_snf_affinity_continuous}
\title{Calculate the SNF affinity matrix for continuous values}
\usage{
rs_snf_affinity_continuous(data, distance_type, k, mu, normalise)
}
\arguments{
\item{data}{Numerical matrix. Needs to be oriented features x samples!}

\item{distance_type}{String. One of
\code{c("euclidean", "manhattan", "canberra", "cosine")}. Which distance metric
to use here.}

\item{k}{Integer. Number of neighbours to consider.}

\item{mu}{Float. Normalisation factor for the Gaussian kernel width.}

\item{normalise}{Boolean. Shall continuous values be Z-scored.}
}
\value{
The affinity matrix based on continuous values.
}
\description{
Calculate the SNF affinity matrix for continuous values
}
