% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sparse_dict_dgrdl}
\alias{rs_sparse_dict_dgrdl}
\title{Generate a sparse dictionary with DGRDL}
\usage{
rs_sparse_dict_dgrdl(dat, dgrdl_params, seed, verbose)
}
\arguments{
\item{dat}{...}

\item{dgrdl_params}{...}

\item{seed}{...}

\item{verbose}{...}
}
\value{
A list ...
}
\description{
Generate a sparse dictionary with DGRDL
}
