% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_sparse_dict_dgrdl_grid_search}
\alias{rs_sparse_dict_dgrdl_grid_search}
\title{Generate a sparse dictionary with DGRDL}
\usage{
rs_sparse_dict_dgrdl_grid_search(
  x,
  dgrdl_params,
  seeds,
  dict_sizes,
  k_neighbours_vec,
  verbose
)
}
\arguments{
\item{x}{Numerical matrix. Rows = samples, columns = features.}

\item{dgrdl_params}{A list with the parameters for the algorithm. Expects
the following items.
\itemize{
\item sparsity - Sparsity constraint (max non-zero coefficients per signal).
\item dict_size - Size of the dictionary. This parameter will be ignored
for this function and \code{dict_sizes} will be used.
\item alpha - Float. Sample context regularisation weight. The higher the stronger
the regularisation.
\item beta - Float. Feature context regularisation weight. The higher the stronger
the regularisation.
\item max_iter - Integer. Maximum iteration for the algorithm.
\item k_neighbours - Integer. Number of k neighbours for the sample and feature
Laplacian matrix for the regularisation. This parameter will be ignored and
\code{k_neighbours_vec} will be used.
\item admm_iter Integer. Number of iterations for using alternating direction
method of multipliers (ADMM).
\item rho Float. ADMM step size.
}}

\item{seeds}{Integer vectors. The random seeds to include in the grid search.}

\item{dict_sizes}{Integer vector. The dictionary sizes to test in the grid
search.}

\item{k_neighbours_vec}{Integer vector. The number of neighbours for the KNN
graph generation to test in the grid search.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list with the following elements:
\itemize{
\item seed - The tested seeds.
\item dict_size - The tested dictionary sizes.
\item reconstruction_errs - The reconstruction errors for these hyper
parameters.
\item feature_laplacian_objective - The objective values of the feature
Laplacian term for these hyperparameters.
\item sample_laplacian_objective - The objective values of the sample
Laplacian term for these hyperparameters.
}
}
\description{
This is the Rust implementation of dual graph regularised
dictionary learning in the implementation of Pan, et al., Cell Systems,
2022. This helper function is designed to run a grid search over the data.
}
