% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_split_cor_signs}
\alias{rs_split_cor_signs}
\title{Helper function to split correlation matrices by sign}
\usage{
rs_split_cor_signs(data)
}
\arguments{
\item{data}{The correlation matrix to split by sign.}
}
\value{
A vector of 1 and -1 indicating the respective sign of the
correlation matrix.
}
\description{
Helper function to split correlation matrices by sign
}
