% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_synthetic_sc_data_with_cell_types}
\alias{rs_synthetic_sc_data_with_cell_types}
\title{Generates synthetic data for single cell}
\usage{
rs_synthetic_sc_data_with_cell_types(
  n_cells,
  n_genes,
  n_batches,
  cell_configs,
  batch_effect_strength,
  seed
)
}
\arguments{
\item{n_cells}{Integer. Number of cells to generate.}

\item{n_genes}{Integer. Number of genes to generate.}

\item{n_batches}{Integer. Number of the batches to generated.}

\item{cell_configs}{A nested list that indicates which gene indices
are markers for which cell.}

\item{batch_effect_strength}{String. One of \code{c("strong", "medium", "low")}.
Defines the strength of the added batch effect.}

\item{seed}{Integer. Random seed for reproducibility.}
}
\value{
A list with the following items.
\itemize{
\item data - The synthetic raw counts.
\item indptr - The index pointers of the cells.
\item indices - The indices of the genes for the given cells.
\item nrow - Number of rows.
\item ncol - Number of columns
\item cell_type_indices - Vector indicating which cell type this is.
\item batch_indices - Vector indicating the batch.
}
}
\description{
Helper function to generate synthetic single cell data with optional
}
