% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_tied_diffusion_parallel}
\alias{rs_tied_diffusion_parallel}
\title{Calculate massively parallelised tied diffusion scores}
\usage{
rs_tied_diffusion_parallel(
  node_names,
  from,
  to,
  diffusion_scores_1,
  diffusion_scores_2,
  summarisation_fun,
  undirected
)
}
\arguments{
\item{node_names}{String vector. Name of the graph nodes.}

\item{from}{String vector. The names of the \code{from} edges from the edge list.}

\item{to}{String vector. The names of the \code{to} edges from the edge list.}

\item{diffusion_scores_1}{List. The first set of personalised vectors for
the page rank reset values. Each element must sum to 1 and be of same length
of \code{node_names}!}

\item{diffusion_scores_2}{List. The second set of personalised vectors for
the page rank reset values. Each element must sum to 1 and be of same length
of \code{node_names}!}

\item{summarisation_fun}{String. One of \code{c("min", "max", "avg")}. Which type
of summarisation function to use to calculate the tied diffusion.}

\item{undirected}{Boolean. Is this an undirected graph.}
}
\value{
A matrix of the scores with each row representing a tied diffusion of
of \code{diffusion_scores_1} and  \code{diffusion_scores_2} lists (in order), and each
column representing the value of the tied diffusion for this node.
}
\description{
Helper function to calculate in parallel on the same (unweighted)
network the tied diffusions as fast as possible. Can be used for permutation.
}
