% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_top_k_targets}
\alias{rs_top_k_targets}
\title{SCENIC: Select the Top TF <> Gene pairs}
\usage{
rs_top_k_targets(matrix, k, margin, min_value)
}
\arguments{
\item{matrix}{Numeric matrix with genes x TF importance values}

\item{k}{Integer. Number of top genes / TFs to extract.}

\item{margin}{If set to 1, the top k TFs per gene are used. If set to 2, the
top k genes per TF are used. Both versions were used in the original paper.}

\item{min_value}{Float. An}
}
\value{
A list with three vectors: tf, gene, importance
}
\description{
SCENIC: Select the Top TF <> Gene pairs
}
