% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_upper_triangle_to_dense}
\alias{rs_upper_triangle_to_dense}
\title{Reconstruct a matrix from a flattened upper triangle vector}
\usage{
rs_upper_triangle_to_dense(cor_vector, shift, n)
}
\arguments{
\item{cor_vector}{Numeric vector. The vector of correlation coefficients
that you want to use to go back to a dense matrix.}

\item{shift}{Integer. If you applied a shift, i.e. included the diagonal
values = 0; or excluded the diagonal values = 1.}

\item{n}{Integer. Original dimension (i.e., ncol/nrow) of the matrix to be
reconstructed.}
}
\value{
The dense R matrix.
}
\description{
This function takes a flattened vector of the upper triangle
from a symmetric matrix (think correlation matrix) and reconstructs the full
dense matrix for you.
}
