% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_upper_triangle_to_sparse}
\alias{rs_upper_triangle_to_sparse}
\title{Generate sparse data from an upper triangle}
\usage{
rs_upper_triangle_to_sparse(value, shift, n)
}
\arguments{
\item{value}{Numeric vector. The upper triangle values.}

\item{shift}{Integer Did you apply a shift to remove the diagonal values?}

\item{n}{Integer. The number of columns/rows in the symmetric matrix.}
}
\value{
A list containing:
\itemize{
\item data - A vector of lists with the elements. (Related to the way
Robj are stored in Rust.)
\item row_indices - A vector of integers with the row indices.
\item col_ptr - A vector of integers with the column pointers.
}
}
\description{
This function takes the values from an upper triangle matrix
the shift and the nrows/ncols and returns a list.
}
