% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_vision}
\alias{rs_vision}
\title{Calculate VISION pathway scores in Rust}
\usage{
rs_vision(f_path, gs_list, cells_to_keep, streaming, verbose)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{gs_list}{Nested list. Each sublist contains the (0-indexed!) positive
and negative gene indices of that specific gene set.}

\item{cells_to_keep}{Integer. Vector of indices of the cells to keep.}

\item{streaming}{Boolean. Shall the data be streamed.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A matrix of cells x vision scores per gene set.
}
\description{
The function will take in a list of gene sets that contains lists of \code{"pos"}
and \code{"neg"} gene indices (0-indexed). You don't have to provide the \code{"neg"},
but it can be useful to classify the delta of two stats (EMT, Th1; Th2) etc.
}
