% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_vision_with_autocorrelation}
\alias{rs_vision_with_autocorrelation}
\title{Calculate VISION pathway scores in Rust with auto-correlation}
\usage{
rs_vision_with_autocorrelation(
  f_path,
  embd,
  knn_data,
  gs_list,
  random_gs_list,
  vision_params,
  cells_to_keep,
  cluster_membership,
  streaming,
  verbose,
  seed
)
}
\arguments{
\item{f_path}{String. Path to the \code{counts_cells.bin} file.}

\item{embd}{Numerical matrix. The embedding matrix to use to generate the
kNN graph.}

\item{knn_data}{Optional list. This contains pre-computed kNN data
(including distances) and the \code{dist_metric} it was built with. The user has
to ensure consistency! If provided, this will be used and whether the
distances are treated as squared is derived from \code{dist_metric} rather than
from the parameter list.}

\item{gs_list}{Nested list. Each sublist contains the (0-indexed!) positive
and negative gene indices of that specific gene set.}

\item{random_gs_list}{Double-nested list. The outer list represents the
clusters of clusters and the inner list represents the permutations within
that cluster.}

\item{vision_params}{List. Contains various parameters to use in terms
of the kNN generation.}

\item{cells_to_keep}{Integer. Vector of indices of the cells to keep.}

\item{cluster_membership}{Integer. Vector that indicates to which of the
permuted gene set clusters the given gene set belongs.}

\item{streaming}{Boolean. Shall the data be streamed.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}

\item{seed}{Integer. Random seed for reproducibility.}
}
\value{
A list with the following items:
\itemize{
\item autocor_res - Auto-correlation results, i.e., 1 - C, p-value and
FDR.
\item vision_mat - A matrix of cells x vision scores per gene set.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The function will take in a list of gene sets that contains lists of \code{"pos"}
and \code{"neg"} gene indices (0-indexed). You don't have to provide the \code{"neg"},
but it can be useful to classify the delta of two stats (EMT, Th1; Th2) etc.
Additionally, it will take a random gene list and calculate an
auto-correlation score based on Gaery's C to identify pathways that show
significant patterns on the kNN graph generate on the provided embedding.
}
\keyword{internal}
