% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_wnn}
\alias{rs_wnn}
\title{Run the weighted nearest neighbour algorithm}
\usage{
rs_wnn(modality_emb_one, modality_emb_two, wnn_params, seed, verbose)
}
\arguments{
\item{modality_emb_one}{Numerical matrix of the first modality. For example
the PCA (or other embeddings) from the transcriptomics.}

\item{modality_emb_two}{Numerical matrix of the second modality. For example
the PCA (or other embeddings) from the ADT counts.}

\item{wnn_params}{Named list. The weighted nearest neighbour parameters.}

\item{seed}{Integer. For reproducibility purposes.}

\item{verbose}{Integer. \code{0L} - quiet; \code{1L} - normal verbosity; \code{2L} -
detailed verbosity.}
}
\value{
A list with
\itemize{
\item indices An integer matrix representing the indices of the
approximate nearest neighbours.
\item dist - An numerical matrix representing the distances to the nearest
neighbours.
\item modality_one_weights - The weights of the first modality.
\item modality_two_weights - The weights of the second modality.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This provides a Rust-based implementation of the WNN algorithm from Hao,
et al.
}
\references{
Hao et al., Cell, 2021
}
