% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bulk.R
\name{run_limma_voom}
\alias{run_limma_voom}
\title{Run Limma Voom}
\usage{
run_limma_voom(
  meta_info,
  main_contrast,
  dge_list,
  co_variates = NULL,
  ...,
  .verbose = TRUE
)
}
\arguments{
\item{meta_info}{data.table. The meta information.}

\item{main_contrast}{String. Which column contains the main groups you want
to test dge for.}

\item{dge_list}{DGEList, see \code{\link[edgeR:DGEList]{edgeR::DGEList()}}.}

\item{co_variates}{String or NULL. Optional co-variates you wish to consider
during model fitting.}

\item{...}{Additional parameters to forward to \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr()}}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A data.table with all the DGE results by the identified contrasts.
}
\description{
Wrapper function to run Limma Voom workflows.
}
