% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_graphs.R
\name{run_snf}
\alias{run_snf}
\title{Run the SNF algorithm}
\usage{
run_snf(object, to_include = NULL, params = NULL)
}
\arguments{
\item{object}{The underlying class, see \code{\link[=snf]{snf()}}.}

\item{to_include}{Optional string, if you wish to only use a subset of the
generated adjacency matrices. If \code{NULL} all matrices will be used for the
fusion process.}

\item{params}{Optional List. If you wish to overwite the already set up
parameters for SNF, see \code{\link[=params_snf]{params_snf()}}. If \code{NULL}, the settings
from within the object will be used. If not NULL, the new parameters will
be used for this modality specifically and only for this modality!}
}
\value{
The class with added adjacency matrix based on the SNF algorithm.
}
\description{
This function will run the SNF algorithm on top of the adjacency matrices
found in the object. You can also optionally specify which adjacency matrices
to use via the \code{to_include} parameter.
}
