% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_io.R
\name{save_sc_exp_to_disk}
\alias{save_sc_exp_to_disk}
\title{Save memory-bound data to disk}
\usage{
save_sc_exp_to_disk(object, type = c("qs2", "rds"))
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{type}{String. One of \code{c("qs2", "rds")}. Defines which binary format to
use. Will default to \code{"qs2"} for speed.}
}
\value{
The object with added information on the data on disk.
}
\description{
Helper function that stores the memory-bound data to disk for checkpointing
or when you close the session for quick recovery of prior work. You have the
option to save as \code{".rds"} or \code{".qs2"} (you need to have the package \code{"qs2"}
installed for this option!).
}
