% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{score_clusters}
\alias{score_clusters}
\alias{score_clusters.ScTypeResults}
\title{Score clusters based on ScType}
\usage{
score_clusters(x, cluster_labels)

\method{score_clusters}{ScTypeResults}(x, cluster_labels)
}
\arguments{
\item{x}{\code{ScTypeResults} object.}

\item{cluster_labels}{Integer vector. Cluster assignment, of length of the
scored cells.}
}
\value{
A \code{data.table} with cluster_id, cell_type, scores and n_cells.
}
\description{
Score clusters based on ScType
}
