% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{scrublet_sc}
\alias{scrublet_sc}
\title{Doublet detection with Scrublet}
\usage{
scrublet_sc(
  object,
  scrublet_params = params_scrublet(),
  seed = 42L,
  streaming = FALSE,
  return_combined_pca = FALSE,
  return_pairs = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{scrublet_params}{A list with the final scrublet parameters, see
\code{\link[=params_scrublet]{params_scrublet()}} for full details.}

\item{seed}{Integer. Random seed.}

\item{streaming}{Boolean. Shall streaming be used during the HVG
calculations. Slower, but less memory usage.}

\item{return_combined_pca}{Boolean. Shall the PCA of the observed cells and
simulated doublets be returned.}

\item{return_pairs}{Boolean. Shall the pairs be returned.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A \code{scrublet_res} class that has with the following items:
\itemize{
\item predicted_doublets - Boolean vector indicating which observed cells
predicted as doublets (TRUE = doublet, FALSE = singlet).
\item doublet_scores_obs - Numerical vector with the likelihood of being
a doublet for the observed cells.
\item doublet_scores_sim - Numerical vector with the likelihood of being
a doublet for the simulated cells.
\item doublet_errors_obs - Numerical vector with the standard errors of
the scores for the observed cells.
\item z_scores - Z-scores for the observed cells. Represents:
\code{score - threshold / error}.
\item threshold - Used threshold.
\item detected_doublet_rate - Fraction of cells that are called as
doublet.
\item detectable_doublet_fraction - Fraction of simulated doublets with
scores above the threshold.
\item overall_doublet_rate - Estimated overall doublet rate. Should roughly
match the expected doublet rate.
\item pca - Optional PCA embeddings across the original cells and simulated
doublets.
\item pair_1 - Optional index of the parent cell 1 of the simulated
doublets.
\item pair_2 - Optional index of the parent cell 2 of the simulated
doublets.
}
}
\description{
This function implements the doublet detection from Scrublet,
see Wolock, et al. Briefly, arteficial doublets are being generated from
the data via random combination of initial cells. Highly variable genes (HVG)
are being identified and the observed cells are being projected on a PCA
space. Subsequently, the simulated doublets are being projected on the same
PCA space given the same HVGs; kNN graphs are being generated and a kNN
classifier is used to assign a probability that a given cell in the original
data is a doublet. For more details, please check the publication.
}
\details{
The function will be executed on the cells that are defined by
\code{\link[=set_cells_to_keep]{set_cells_to_keep()}}. If you have not set this parameter, all
cells in the object will be initially used. This is the recommend way of
using this function.
}
\references{
Wollock, et al., Cell Syst, 2020
}
