% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{set_cell_mapping}
\alias{set_cell_mapping}
\alias{set_cell_mapping.ScMap}
\alias{set_cell_mapping.SingleCells}
\title{Set cell mapping for ScMap object}
\usage{
set_cell_mapping(x, cell_map)

\method{set_cell_mapping}{ScMap}(x, cell_map)
}
\arguments{
\item{x}{An object to set cell mapping for}

\item{cell_map}{Named integer indicating indices and names of the cells}
}
\description{
Set cell mapping for ScMap object

Set the cell mapping for a \code{SingleCells} class.
}
\keyword{internal}
