% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_cell_mapping}
\alias{set_cell_mapping}
\title{Set cell mapping for sc_mapper object}
\usage{
set_cell_mapping(x, cell_map)
}
\arguments{
\item{x}{An object to set cell mapping for}

\item{cell_map}{Named integer indicating indices and names of the cells}
}
\description{
Set cell mapping for sc_mapper object
}
