% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_cell_to_keep}
\alias{set_cell_to_keep}
\title{Set cells to keep for sc_mapper object}
\usage{
set_cell_to_keep(x, cell_indices)
}
\arguments{
\item{x}{An object to set cells to keep for}

\item{cell_indices}{Integer. The indices of the cells to keep.}
}
\description{
Set cells to keep for sc_mapper object
}
