% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_cell_to_keep.sc_mapper}
\alias{set_cell_to_keep.sc_mapper}
\title{Set cells to keep for sc_mapper}
\usage{
\method{set_cell_to_keep}{sc_mapper}(x, cells_to_keep)
}
\arguments{
\item{x}{An \code{sc_mapper} object}

\item{cells_to_keep}{String or integer. The names or indices of the cells
to keep in downstream analysis.}
}
\value{
Updated \code{sc_mapper} object with column index mapping set
}
\description{
Set cells to keep for sc_mapper
}
