% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R, R/classes_single_cell.R
\name{set_cells_to_keep}
\alias{set_cells_to_keep}
\alias{set_cells_to_keep.ScMap}
\alias{set_cells_to_keep.SingleCells}
\title{Set cells to keep for ScMap object}
\usage{
set_cells_to_keep(x, cells_to_keep)

\method{set_cells_to_keep}{ScMap}(x, cells_to_keep)
}
\arguments{
\item{x}{An object to set cells to keep for}

\item{cells_to_keep}{String or integer. The names or indices of the cells
to keep in downstream analysis.}
}
\description{
Set cells to keep for ScMap object

Set the cell mapping for a \code{SingleCells} class.
}
\keyword{internal}
