% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_embedding}
\alias{set_embedding}
\alias{set_embedding.sc_cache}
\alias{set_embedding.single_cell_exp}
\title{Add additional embeddings to the class}
\usage{
set_embedding(x, embd, name, cells_used)

\method{set_embedding}{sc_cache}(x, embd, name, cells_used)
}
\arguments{
\item{x}{An object to add the singular values for.}

\item{embd}{Numerical matrix representing the additional embedding.}

\item{name}{String. Name of the embedding.}

\item{cells_used}{Integer vector. The cells used for the generation of this
embedding.}
}
\value{
The object with the added embedding.
}
\description{
Add additional embeddings to the class

Set additional embeddings to \code{single_cell_exp}.
}
