% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_gene_mapping}
\alias{set_gene_mapping}
\alias{set_gene_mapping.ScMap}
\alias{set_gene_mapping.SingleCells}
\title{Set gene mapping for ScMap object}
\usage{
set_gene_mapping(x, gene_map)

\method{set_gene_mapping}{ScMap}(x, gene_map)
}
\arguments{
\item{x}{An object to set gene mapping for}

\item{gene_map}{Named integer indicating indices and names of the genes}
}
\description{
Set gene mapping for ScMap object

Set the gene mapping for a \code{SingleCells} class.
}
\keyword{internal}
