% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_gene_mapping}
\alias{set_gene_mapping}
\alias{set_gene_mapping.sc_mapper}
\alias{set_gene_mapping.single_cell_exp}
\title{Set gene mapping for sc_mapper object}
\usage{
set_gene_mapping(x, gene_map)

\method{set_gene_mapping}{sc_mapper}(x, gene_map)
}
\arguments{
\item{x}{An object to set gene mapping for}

\item{gene_map}{Named integer indicating indices and names of the genes}
}
\description{
Set gene mapping for sc_mapper object

Set the gene mapping for a \code{single_cell_exp} class.
}
