% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_hvg}
\alias{set_hvg}
\alias{set_hvg.sc_mapper}
\alias{set_hvg.single_cell_exp}
\title{Set the HVG genes}
\usage{
set_hvg(x, hvg)

\method{set_hvg}{sc_mapper}(x, hvg)
}
\arguments{
\item{x}{An object to set the HVGs for}

\item{hvg}{String or integer. The names or indices of the highly variable
genes.}
}
\description{
Set the HVG genes

Set the highly variable genes for a \code{single_cell_exp} class.
}
