% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_knn}
\alias{set_knn}
\alias{set_knn.ScCache}
\alias{set_knn.SingleCells}
\title{Set/add KNN}
\usage{
set_knn(x, knn)

\method{set_knn}{ScCache}(x, knn)
}
\arguments{
\item{x}{An object to add the KNN data to}

\item{knn}{\code{SingleCellNearestNeighbour} class to add to the classes.}
}
\description{
Set/add KNN

Set the KNN matrix for a \code{SingleCells} class.
}
\keyword{internal}
