% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_knn}
\alias{set_knn}
\title{Set/add KNN}
\usage{
set_knn(x, knn_matrix)
}
\arguments{
\item{x}{An object to add the KNN data to}

\item{knn_mat}{Numerical matrix. The matrix with the KNN data}
}
\description{
Set/add KNN
}
