% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_knn}
\alias{set_knn}
\alias{set_knn.sc_cache}
\alias{set_knn.single_cell_exp}
\title{Set/add KNN}
\usage{
set_knn(x, knn_mat)

\method{set_knn}{sc_cache}(x, knn_mat)
}
\arguments{
\item{x}{An object to add the KNN data to}

\item{knn_mat}{Numerical matrix. The matrix with the KNN data}
}
\description{
Set/add KNN

Set the KNN matrix for a \code{single_cell_exp} class.
}
