% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_single_cell_.R,
%   R/classes_single_cell_subset.R
\name{set_magic}
\alias{set_magic}
\alias{set_magic.ScCache}
\alias{set_magic.SingleCells}
\alias{set_magic.SingleCellsSubset}
\title{Set/add the MAGIC imputed layer}
\usage{
set_magic(x, magic, ...)

\method{set_magic}{ScCache}(x, magic, ...)

## S7 method for class <bixverse::SingleCells>
set_magic(x, magic, ...)

## S7 method for class <bixverse::SingleCellsSubset>
set_magic(x, magic, ...)
}
\arguments{
\item{x}{An object to add the imputed layer to.}

\item{magic}{\code{ScMagic} class with the imputed counts.}

\item{...}{Other parameters.}
}
\description{
Set/add the MAGIC imputed layer
}
\keyword{internal}
