% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R, R/classes_meta_cell.R,
%   R/classes_single_cell.R
\name{set_pca_factors}
\alias{set_pca_factors}
\alias{set_pca_factors.MetaCells}
\alias{set_pca_factors.ScCache}
\alias{set_pca_factors.SingleCells}
\title{Set/add PCA factors}
\usage{
set_pca_factors(x, pca_factor)

\method{set_pca_factors}{ScCache}(x, pca_factor)
}
\arguments{
\item{x}{An object to add the PCA factors for.}

\item{pca_factor}{Numerical matrix. The matrix with the PCA factors.}
}
\description{
Set/add PCA factors
}
\keyword{internal}
