% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_pca_factors}
\alias{set_pca_factors}
\alias{set_pca_factors.sc_cache}
\alias{set_pca_factors.single_cell_exp}
\title{Set/add PCA factors}
\usage{
set_pca_factors(x, pca_factor)

\method{set_pca_factors}{sc_cache}(x, pca_factor)
}
\arguments{
\item{x}{An object to add the PCA factors for.}

\item{pca_factor}{Numerical matrix. The matrix with the PCA factors.}
}
\description{
Set/add PCA factors

Set the PCA factors for a \code{single_cell_exp} class.
}
