% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_pca_loadings}
\alias{set_pca_loadings}
\title{Set/add PCA loadings}
\usage{
set_pca_loadings(x, pca_loading)
}
\arguments{
\item{x}{An object to add the PCA loadings for.}

\item{pca_loading}{Matrix with the PCA loadings.}
}
\description{
Set/add PCA loadings
}
