% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_pca_singular_vals}
\alias{set_pca_singular_vals}
\alias{set_pca_singular_vals.ScCache}
\alias{set_pca_singular_vals.SingleCells}
\title{Set/add PCA singular values}
\usage{
set_pca_singular_vals(x, singular_vals)

\method{set_pca_singular_vals}{ScCache}(x, singular_vals)
}
\arguments{
\item{x}{An object to add the singular values for.}

\item{singular_vals}{Numerical vector. The singular values.}
}
\description{
Set/add PCA singular values

Set the PCA singular values for a \code{SingleCells} class.
}
