% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R
\name{set_per_sample_sparkx}
\alias{set_per_sample_sparkx}
\alias{set_per_sample_sparkx.SpCache}
\alias{set_per_sample_sparkx.SpatialSpot}
\title{Set the per-sample SPARK-X results}
\usage{
set_per_sample_sparkx(x, exp_id, sparkx, ...)

\method{set_per_sample_sparkx}{SpCache}(x, exp_id, sparkx, ...)

\method{set_per_sample_sparkx}{SpatialSpot}(x, exp_id, sparkx, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{sparkx}{\code{data.table} of SPARK-X results for this sample.}

\item{...}{Other parameters.}
}
\description{
Set the per-sample SPARK-X results
}
\keyword{internal}
