% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_spatial.R,
%   R/classes_spatial_spot.R
\name{set_per_sample_spatial_graph}
\alias{set_per_sample_spatial_graph}
\alias{set_per_sample_spatial_graph.SpCache}
\alias{set_per_sample_spatial_graph.SpatialSpot}
\title{Set the per-sample spatial graph}
\usage{
set_per_sample_spatial_graph(x, exp_id, graph, ...)

\method{set_per_sample_spatial_graph}{SpCache}(x, exp_id, graph, ...)

\method{set_per_sample_spatial_graph}{SpatialSpot}(x, exp_id, graph, ...)
}
\arguments{
\item{x}{An object with a \code{SpCache} slot or an \code{SpCache} itself.}

\item{exp_id}{String. The experiment identifier.}

\item{graph}{A \code{dgCMatrix} (CSR-style sparse matrix) representing the
spatial graph for this sample.}

\item{...}{Other parameters.}
}
\description{
Set the per-sample spatial graph
}
\keyword{internal}
