% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_sc_new_obs_col}
\alias{set_sc_new_obs_col}
\title{Add a new column to the obs table}
\usage{
set_sc_new_obs_col(object, col_name, new_data)
}
\arguments{
\item{object}{\code{bixverse::single_cell_exp} class.}

\item{col_name}{String. The name of the column to add.}

\item{new_data}{Atomic vector. The data to add to the column. Needs to be
of same length as nrow obs table.}
}
\value{
The class with updated obs table in the DuckDB
}
\description{
Add a new column to the obs table
}
