% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_sc_new_obs_col_multiple}
\alias{set_sc_new_obs_col_multiple}
\title{Add multiple new columns to the obs table}
\usage{
set_sc_new_obs_col_multiple(object, new_data)
}
\arguments{
\item{object}{\code{bixverse::single_cell_exp} class.}

\item{new_data}{Named list. The names will be the column names and the
elements will be added to the obs table. Needs to be of same length as
\code{\link[=get_cells_to_keep]{get_cells_to_keep()}} and have the same order!}
}
\value{
The class with updated obs table in the DuckDB
}
\description{
Add multiple new columns to the obs table
}
