% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_sc_new_var_cols}
\alias{set_sc_new_var_cols}
\title{Add a new column to the obs table}
\usage{
set_sc_new_var_cols(object, data_list)
}
\arguments{
\item{object}{\code{bixverse::single_cell_exp} class.}

\item{data_list}{Named list with the data to add.}
}
\value{
The class with updated obs table in the DuckDB
}
\description{
Add a new column to the obs table
}
