% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{set_snn_graph}
\alias{set_snn_graph}
\alias{set_snn_graph.sc_cache}
\alias{set_snn_graph.single_cell_exp}
\title{Set/add KNN}
\usage{
set_snn_graph(x, snn_graph)

\method{set_snn_graph}{sc_cache}(x, snn_graph)
}
\arguments{
\item{x}{An object to add the KNN data to}

\item{snn_graph}{Igraph. The sNN graph for subsequent clustering.}
}
\description{
Set/add KNN

Set the sNN graph for a \code{single_cell_exp} class.
}
