% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_db.R
\name{single_cell_duckdb_con}
\alias{single_cell_duckdb_con}
\title{Class for storing single cell experimental data in DuckDB (nightly!)}
\value{
Invisible self and populates the internal obs table.
Function to populate the var table from plain text files

Invisible self and populates the internal var table.
Function to populate the obs table from R

Invisible self and populates the internal obs table.
Function to populate the var table from R

Invisible self and populates the internal obs table.
}
\description{
This class wraps up the DB connection and methods to interact with the
observation/cell metadata and the feature/gene metadata.
}
\section{Super class}{
\code{\link[bixverse:single_cell_duckdb_base]{bixverse::single_cell_duckdb_base}} -> \code{single_cell_duckdb_con}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-single_cell_duckdb_con-populate_obs_from_h5}{\code{single_cell_duckdb_con$populate_obs_from_h5()}}
\item \href{#method-single_cell_duckdb_con-populate_vars_from_h5}{\code{single_cell_duckdb_con$populate_vars_from_h5()}}
\item \href{#method-single_cell_duckdb_con-populate_obs_from_plain_text}{\code{single_cell_duckdb_con$populate_obs_from_plain_text()}}
\item \href{#method-single_cell_duckdb_con-populate_var_from_plain_text}{\code{single_cell_duckdb_con$populate_var_from_plain_text()}}
\item \href{#method-single_cell_duckdb_con-populate_obs_from_data.table}{\code{single_cell_duckdb_con$populate_obs_from_data.table()}}
\item \href{#method-single_cell_duckdb_con-populate_var_from_data.table}{\code{single_cell_duckdb_con$populate_var_from_data.table()}}
\item \href{#method-single_cell_duckdb_con-clone}{\code{single_cell_duckdb_con$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="add_data_obs"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-add_data_obs'><code>bixverse::single_cell_duckdb_base$add_data_obs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="add_data_var"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-add_data_var'><code>bixverse::single_cell_duckdb_base$add_data_var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="filter_var_table"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-filter_var_table'><code>bixverse::single_cell_duckdb_base$filter_var_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="get_cells_to_keep"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-get_cells_to_keep'><code>bixverse::single_cell_duckdb_base$get_cells_to_keep()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="get_obs_index_map"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-get_obs_index_map'><code>bixverse::single_cell_duckdb_base$get_obs_index_map()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="get_obs_table"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-get_obs_table'><code>bixverse::single_cell_duckdb_base$get_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="get_var_index_map"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-get_var_index_map'><code>bixverse::single_cell_duckdb_base$get_var_index_map()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="get_vars_table"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-get_vars_table'><code>bixverse::single_cell_duckdb_base$get_vars_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="initialize"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-initialize'><code>bixverse::single_cell_duckdb_base$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="join_data_obs"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-join_data_obs'><code>bixverse::single_cell_duckdb_base$join_data_obs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="set_cells_to_keep"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-set_cells_to_keep'><code>bixverse::single_cell_duckdb_base$set_cells_to_keep()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bixverse" data-topic="single_cell_duckdb_base" data-id="set_to_keep_column"><a href='../../bixverse/html/single_cell_duckdb_base.html#method-single_cell_duckdb_base-set_to_keep_column'><code>bixverse::single_cell_duckdb_base$set_to_keep_column()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_obs_from_h5"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_obs_from_h5}{}}}
\subsection{Method \code{populate_obs_from_h5()}}{
This function populates the obs table from an h5 file (if found).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_obs_from_h5(h5_path, filter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h5_path}}{String. Path to the h5 file from which to load in the
observation table.}

\item{\code{filter}}{Optional integer. Positions of obs to read in from file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns invisible self. As a side effect, it will load in the
obs data from the h5ad file into the DuckDB.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_vars_from_h5"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_vars_from_h5}{}}}
\subsection{Method \code{populate_vars_from_h5()}}{
This populates the vars table from an h5 file (if found.)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_vars_from_h5(h5_path, filter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h5_path}}{String. Path to the h5 file from which to load in the
observation table.}

\item{\code{filter}}{Optional integer. Positions of obs to read in from file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns invisible self. As a side effect, it will load in the
obs data from the h5ad file into the DuckDB.
Function to populate the obs table from plain text files
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_obs_from_plain_text"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_obs_from_plain_text}{}}}
\subsection{Method \code{populate_obs_from_plain_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_obs_from_plain_text(
  f_path,
  has_hdr,
  filter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f_path}}{File path to the plain text file.}

\item{\code{has_hdr}}{Boolean. Does the flat file have a header.}

\item{\code{filter}}{Optional integer. Positions of obs to read in from file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_var_from_plain_text"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_var_from_plain_text}{}}}
\subsection{Method \code{populate_var_from_plain_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_var_from_plain_text(
  f_path,
  has_hdr,
  filter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f_path}}{File path to the plain text file.}

\item{\code{has_hdr}}{Boolean. Does the flat file have a header.}

\item{\code{filter}}{Optional integer. Positions of obs to read in from file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_obs_from_data.table"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_obs_from_data.table}{}}}
\subsection{Method \code{populate_obs_from_data.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_obs_from_data.table(obs_dt, filter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs_dt}}{data.table}

\item{\code{filter}}{Optional integer. Row indices to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-populate_var_from_data.table"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-populate_var_from_data.table}{}}}
\subsection{Method \code{populate_var_from_data.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$populate_var_from_data.table(var_dt, filter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_dt}}{data.table}

\item{\code{filter}}{Optional integer. Row indices to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_cell_duckdb_con-clone"></a>}}
\if{latex}{\out{\hypertarget{method-single_cell_duckdb_con-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_cell_duckdb_con$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
