% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell.R
\name{single_cell_exp}
\alias{single_cell_exp}
\title{bixverse single cell class (nightly!)}
\usage{
single_cell_exp(dir_data)
}
\arguments{
\item{dir_data}{String. This is the directory in which the experimental files
will be}
}
\value{
Returns the \code{single_cell_exp} class for further operations.
}
\description{
This is the \code{bixverse}-based single cell class. Under the hood it uses a
DuckDB for obs and vars storing, and a Rust-based binarised file format to
store the raw and normalised counts. In both cases, the idea is not to hold
any data that is not needed at a given point of time in memory, but leverage
speedy on-disk computations and streaming engines powered by Rust and DuckDB
to run the analysis.
}
\section{Properties}{

\describe{
\item{db_connection}{This contains an R6 class with DuckDB pointers and
wrappers to interact with the table-like data for this experiment.}
\item{count_connection}{This contains an R6-like environment that points
to Rust functions that can work on the counts more specifically.}
\item{dir_data}{Path to the directory in which the data will be saved on
disk.}
\item{cache}{List with cached data. Future feature, nothing implemented
yet.}
\item{dims}{Dimensions of the original data.}
}
}

