% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_groups.R
\name{SingleCellsSubset}
\alias{SingleCellsSubset}
\title{bixverse single cell with grouping class}
\usage{
SingleCellsSubset(sc_object, grouping_column, group, other_data = list())
}
\arguments{
\item{dir_data}{String. This is the directory in which the experimental files
will be saved}

\item{obs_data}{data.table with the original obs table.}

\item{var_data}{data.table with the variable/feature informations.}

\item{grouping}{String describing the grouping variable.}
}
\value{
Returns the \code{SingleCellsSubset} class for further operations.
}
\description{
This is the \code{bixverse}-based single cell class with grouping variable.
It is generated internally when requesting grouping based meta-cell generating
functions on top of the main single cell class.
The class is derived from the original single cell class, pointing
to the original count matrix. To prevent issues with state management
of the single cell object, the obs and var are kept in memory.
}
\section{Properties}{

\describe{
\item{count_connection}{This contains an R6-like environment that points
to Rust functions that can work on the RNAseq counts more specifically.}
\item{obs_table}{The single cell observation table.}
\item{var_table}{The single cell variable table.}
\item{grouping_variable}{The grouping variable.}
\item{sc_cache}{Class with embeddings, kNN/sNN graphs, etc. Shared with
\code{\link[=SingleCells]{SingleCells()}}.}
\item{dims}{Dimensions of the original data.}
\item{other_data}{List that contains additional data and results, such
as for example the hvg indices}
}
}

