% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sparse_list_to_mat}
\alias{sparse_list_to_mat}
\title{Helper function to transform the Rust-exported sparse matrices into R ones}
\usage{
sparse_list_to_mat(ls)
}
\arguments{
\item{ls}{List. Needs to represent the (column) sparse data.}
}
\value{
The sparseMatrix from the data.
}
\description{
Helper function to transform the Rust-exported sparse matrices into R ones
}
