% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_sc.R
\name{step_fast_mnn_sc}
\alias{step_fast_mnn_sc}
\title{Pipeline step: fastMNN batch correction}
\usage{
step_fast_mnn_sc(
  batch_column,
  batch_hvg_genes,
  fastmnn_params = params_sc_fastmnn(),
  use_precomputed_pca = FALSE,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{batch_column}{String. The column with the batch information in the
obs data of the class.}

\item{batch_hvg_genes}{Integer vector. These are the highly variable genes,
identified by a batch-aware method. Please refer to
\code{\link[=find_hvg_batch_aware_sc]{find_hvg_batch_aware_sc()}} for more details. These genes have to
be 0-indexed!}

\item{fastmnn_params}{A list, please see \code{\link[=params_sc_fastmnn]{params_sc_fastmnn()}}. The
list has the following parameters:
\itemize{
\item sigma - Numeric. Bandwidth of the Gaussian smoothing kernel (as
proportion of space radius).
\item cos_norm - Logical. Apply cosine normalisation before computing
distances.
\item var_adj - Logical. Apply variance adjustment to avoid kissing
effects.
\item no_pcs - Integer. Number of PCs to use for MNN calculations.
\item knn - List of kNN parameters. See \code{\link[=params_knn_defaults]{params_knn_defaults()}}
for available parameters and their defaults.
\item pca - List of PCA parameters, see \code{\link[=params_sc_pca]{params_sc_pca()}}
for available parameters and their defaults.
}}

\item{use_precomputed_pca}{Boolean. Should the PCA in the object be used
if found. If you decide to do this, make sure that you have run the PCA
on the batch-aware HVG ideally.}

\item{seed}{Integer. Random seed.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
An \code{ScStep}.
}
\description{
Wraps \code{\link[=fast_mnn_sc]{fast_mnn_sc()}} as an \code{ScStep}.
}
