% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{stream_h5ad}
\alias{stream_h5ad}
\title{Stream h5ad files (deprecated)}
\usage{
stream_h5ad(
  object,
  h5_path,
  sc_qc_param = params_sc_min_quality(),
  max_genes_in_memory = 2000L,
  cell_batch_size = 100000L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{h5_path}{File path to the h5 file. Supports h5ad (AnnData),
CellRanger v3 (10x Genomics), and CellRanger v2 formats. Format is
automatically detected.}

\item{sc_qc_param}{List. Output of \code{\link[=params_sc_min_quality]{params_sc_min_quality()}}. A
list with the following elements:
\itemize{
\item min_unique_genes - Integer. Minimum number of genes to be detected
in the cell to be included.
\item min_lib_size - Integer. Minimum library size in the cell to be
included.
\item min_cells - Integer. Minimum number of cells a gene needs to be
detected to be included.
\item target_size - Float. Target size to normalise to. Defaults to \code{1e5}.
}}

\item{max_genes_in_memory}{Integer. How many genes shall be held in memory
at a given point. Defaults to \code{2000L}.}

\item{cell_batch_size}{Integer. How big are the batch sizes for the cells
in the transformation from the cell-based to gene-based format. Defaults to
\code{100000L}.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
A \code{single_cell_exp} object with loaded data
}
\description{
\lifecycle{deprecated}

\code{stream_h5ad()} has been renamed to \code{\link[=stream_h5]{stream_h5()}} to reflect support for
multiple h5 formats.
}
