% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{synthetic_bulk_cor_matrix}
\alias{synthetic_bulk_cor_matrix}
\title{Generates synthetic bulk RNAseq data}
\usage{
synthetic_bulk_cor_matrix(
  num_samples = 100L,
  num_genes = 1000L,
  add_modules = TRUE,
  module_sizes = c(100L, 100L, 100L),
  seed = 123L
)
}
\arguments{
\item{num_samples}{Integer. Number of samples.}

\item{num_genes}{Integer. Number of genes.}

\item{add_modules}{Boolean. Shall gene modules with correlation structures
be generated.}

\item{module_sizes}{Integer vector. Sizes of the different correlation
modules. The sum needs be smaller than \code{num_genes}.}

\item{seed}{Integer. Seed for reproducibility purposes.}
}
\value{
A \code{synthetic_bulk_data} class containing:
\itemize{
\item counts - The count matrix
\item sparse_counts - A slot for sparse counts that can be added later, see
\code{\link[=simulate_dropouts]{simulate_dropouts()}}.
\item module_data - The module membership of the genes
}
}
\description{
Function generates synthetic bulkRNAseq data with heteroskedasticity (lowly
expressed genes show higher variance) and can optionally add gene ~ gene
correlations for testing purposes of module detection methods.
}
