% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synthetic_bulk.R
\name{synthetic_c_pca_data}
\alias{synthetic_c_pca_data}
\title{Generates synthetic data for contrastive PCA exploration.}
\usage{
synthetic_c_pca_data(seed = 10101L)
}
\arguments{
\item{seed}{Integer. Initial random seed for generation of the synthetic
data. Default: 10101L.}
}
\value{
A \code{cpca_synthetic_data} class with the following elements:
\itemize{
\item target - The target matrix.
\item background - The background matrix.
\item target_labels - The target labels
}
}
\description{
Generates three elements: target matrix, background matrix and labels for the
target matrix.
}
