% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{test_nhoods}
\alias{test_nhoods}
\alias{test_nhoods.sc_hotspot}
\title{Test neighbourhoods for differential abundance}
\usage{
test_nhoods(
  x,
  design,
  design_df,
  norm_method = c("TMM", "RLE", "logMS"),
  min_mean = 0,
  robust = TRUE,
  fdr_weighting = c("k-distance", "graph-overlap", "none")
)

\method{test_nhoods}{sc_hotspot}(
  x,
  design,
  design_df,
  norm_method = c("TMM", "RLE", "logMS"),
  min_mean = 0,
  robust = TRUE,
  fdr_weighting = c("k-distance", "graph-overlap", "none")
)
}
\arguments{
\item{x}{sc_miloR x}

\item{design}{Formula for the experimental design}

\item{design_df}{data.frame. Contains the metadata to be used for the
generation of the model matrix.}

\item{norm_method}{String. Normalisation method to use. One of
\code{c("TMM", "RLE", "logMS")}}

\item{min_mean}{Numeric. Minimum mean count threshold for filtering
neighbourhoods.}

\item{robust}{Use robust dispersion estimation}

\item{fdr_weighting}{String. Spatial FDR weighting scheme. One of
\code{c("k-distance", "graph-overlap", "none")}.}
}
\value{
data.table with test results per neighbourhood
}
\description{
Test neighbourhoods for differential abundance
}
