% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_single_cell_others.R
\name{tf_to_genes_correlations}
\alias{tf_to_genes_correlations}
\alias{tf_to_genes_correlations.ScenicGrn}
\title{Generate TF to gene correlations}
\usage{
tf_to_genes_correlations(
  x,
  object,
  cor_filter = NULL,
  remove_self = TRUE,
  spearman = TRUE,
  .verbose = TRUE
)

\method{tf_to_genes_correlations}{ScenicGrn}(
  x,
  object,
  cor_filter = NULL,
  remove_self = TRUE,
  spearman = TRUE,
  .verbose = TRUE
)
}
\arguments{
\item{x}{\code{ScenicGrn} object for which to generate the TF to gene
associations.}

\item{object}{\code{SingleCells} object that was used to generate the original
GRNs.}

\item{cor_filter}{Optional float. If you wish to filter out TF genes below
a certain correlation. If \code{NULL} all genes will be kept.}

\item{remove_self}{Boolean. Shall self loops (where TF controls its own
expression) be removed. Defaults to \code{TRUE}.}

\item{spearman}{Boolean. Shall Spearman correlation be used. Defaults to
\code{TRUE}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Adds the correlations coefficients between to the TF to gene.
}
\description{
This function will calculate the correlations between the identified TF to
gene pairs. You need to have run \code{\link[=identify_tf_to_genes]{identify_tf_to_genes()}}!
}
