% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_diff_class_methods.R
\name{tied_diffusion}
\alias{tied_diffusion}
\title{Use tied diffusion for two sets of seed genes.}
\usage{
diffuse_seed_genes(
 network_diffusions,
 diffusion_vector.1,
 diffusion_vector.2,
 summarisation = c("max", "mean", "harmonic_sum"),
 score_aggregation = c("min", "max", "mean"),
 verbose = TRUE
)
}
\arguments{
\item{network_diffusions}{The underlying \code{network_diffusions} class.}

\item{diffusion_vector.1}{The first named vector with values to use for the reset parameter in the
personalised page-rank diffusion. Names should represent node names of the graph.}

\item{diffusion_vector.2}{The second named vector with values to use for the reset parameter in the
personalised page-rank diffusion. Names should represent node names of the graph.}

\item{summarisation}{If there are duplicated names in the \code{diffusion_vector} how to summarise
these.}

\item{score_aggregation}{How to summarise the tied scores.}

\item{verbose}{Controls verbosity of the function.}
}
\description{
This function takes two sets of diffusion vector and leverages tied diffusion to identify an intersection
of influential nodes.
}
