% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_diff_class_methods.R
\name{tied_diffusion}
\alias{tied_diffusion}
\title{Diffuse seed genes in a tied manner over a network}
\usage{
tied_diffusion(network_diffusions, ...)

tied_diffusion(
 network_diffusions,
 diffusion_vector.1,
 diffusion_vector.2,
 summarisation = c("max", "mean", "harmonic_sum"),
 score_aggregation = c("min", "max", "mean"),
 verbose = TRUE
)
}
\arguments{
\item{network_diffusions}{The underlying \code{network_diffusions} class.}

\item{diffusion_vector.1}{The first named vector with values to use for the reset parameter in the
personalised page-rank diffusion. Names should represent node names of the graph.}

\item{diffusion_vector.2}{The second named vector with values to use for the reset parameter in the
personalised page-rank diffusion. Names should represent node names of the graph.}

\item{summarisation}{If there are duplicated names in the \code{diffusion_vector} how to summarise
these.}

\item{score_aggregation}{How to summarise the tied scores.}

\item{verbose}{Controls verbosity of the function.}
}
\value{
The class with added diffusion score based on a two sets of seed genes. Additionally,
the seed genes are stored in the class.
}
\description{
This is the generic function for diffusion two sets of seed genes over
a network.

This function takes two sets of diffusion vector and leverages tied diffusion to identify an intersection
of influential nodes.
}
