% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Helper function to transform strings to snake_case}
\usage{
to_snake_case(x, ignore_na = FALSE)
}
\arguments{
\item{x}{String. The string (vector) you wish to transform to snake_case.}

\item{ignore_na}{Boolean. Shall the function just ignore \code{NA} values and
return \code{NA} at this position.
Defaults to \code{FALSE}.}
}
\value{
Returns the string in snake_case format.
}
\description{
Helper function to transform strings to snake_case
}
