% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_processing.R
\name{top_genes_perc_sc}
\alias{top_genes_perc_sc}
\title{Calculate the proportions of reads for the Top N genes}
\usage{
top_genes_perc_sc(
  object,
  top_n_vals = c(25L, 50L, 100L),
  streaming = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{top_n_vals}{Integer. The Top N thresholds to test.}

\item{streaming}{Boolean. Shall the cells be streamed in. Useful for larger
data sets where you wish to avoid loading in the whole data. Default to
\code{FALSE}.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
It will add the columns based on the names in the \code{gene_set_list} to
the obs table.
}
\description{
This is a helper function that calculates proportions of reads to the Top N
genes by expression in a given cell. High values here can indicate low
complexity, quality cells. The values will be automatically added to the
obs table.
}
