% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_plots_helpers.R
\name{tsne_sc}
\alias{tsne_sc}
\title{Run t-SNE on a SingleCells object}
\usage{
tsne_sc(
  object,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  n_dim = 2L,
  perplexity = 30,
  approx_type = c("bh", "fft"),
  knn_method = c("hnsw", "balltree", "annoy", "nndescent", "exhaustive"),
  nn_params = manifoldsR::params_nn(),
  tsne_params = manifoldsR::params_tsne(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{embd_to_use}{String. The embedding to use for t-SNE. Must be available
in the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{n_dim}{Integer. Number of t-SNE dimensions. Currently only \code{2L} is
supported. Defaults to \code{2L}.}

\item{perplexity}{Numeric. Perplexity parameter. Typical values between 5
and 50. Defaults to \code{30.0}.}

\item{approx_type}{String. Approximation method. One of \code{"bh"} (Barnes-Hut)
or \code{"fft"}. Defaults to \code{"bh"}.}

\item{knn_method}{String. Approximate nearest neighbour algorithm. One of
\code{"hnsw"}, \code{"balltree"}, \code{"annoy"}, \code{"nndescent"}, or \code{"exhaustive"}.}

\item{nn_params}{Named list. See \code{\link[manifoldsR:params_nn]{manifoldsR::params_nn()}}.}

\item{tsne_params}{Named list. See \code{\link[manifoldsR:params_tsne]{manifoldsR::params_tsne()}}.}

\item{seed}{Integer. For reproducibility.}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
The object with a \code{"tsne"} embedding added.
}
\description{
Wrapper around \code{\link[manifoldsR:tsne]{manifoldsR::tsne()}} for the \code{SingleCells} class.
}
