% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_plots_helpers.R
\name{umap_sc}
\alias{umap_sc}
\title{Run UMAP on a SingleCells object}
\usage{
umap_sc(
  object,
  use_knn = TRUE,
  embd_to_use = "pca",
  no_embd_to_use = NULL,
  n_dim = 2L,
  k = 15L,
  min_dist = 0.5,
  spread = 1,
  knn_method = c("hnsw", "balltree", "annoy", "nndescent", "exhaustive"),
  nn_params = manifoldsR::params_nn(),
  umap_params = manifoldsR::params_umap(),
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{SingleCells} class.}

\item{use_knn}{Boolean. Use the kNN graph found in the object. Defaults to
\code{TRUE}. If not available, will default to the embedding.}

\item{embd_to_use}{String. The embedding to use for UMAP. Must be available
in the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{n_dim}{Integer. Number of UMAP dimensions. Defaults to \code{2L}.}

\item{k}{Integer. Number of nearest neighbours. Defaults to \code{15L}.}

\item{min_dist}{Numeric. Minimum distance between embedded points. Defaults
to \code{0.5}.}

\item{spread}{Numeric. Effective scale of embedded points. Defaults to \code{1.0}.}

\item{knn_method}{String. Approximate nearest neighbour algorithm. One of
\code{"hnsw"}, \code{"balltree"}, \code{"annoy"}, \code{"nndescent"}, or \code{"exhaustive"}.}

\item{nn_params}{Named list. See \code{\link[manifoldsR:params_nn]{manifoldsR::params_nn()}}.}

\item{umap_params}{Named list. See \code{\link[manifoldsR:params_umap]{manifoldsR::params_umap()}}.}

\item{seed}{Integer. For reproducibility.}

\item{.verbose}{Boolean. Controls verbosity.}
}
\value{
The object with a \code{"umap"} embedding added.
}
\description{
Wrapper around \code{\link[manifoldsR:umap]{manifoldsR::umap()}} for the \code{SingleCells} class.
}
