% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_bulk.R
\name{update_metadata_values}
\alias{update_metadata_values}
\title{Replace values in a metadata column}
\usage{
update_metadata_values(object, column, replacement, ...)
}
\arguments{
\item{object}{The underlying object, either \code{bixverse::bulk_coexp} or
\code{bixverse::bulk_dge}.}

\item{column}{Character vector. The columns for which to replace the
values.}

\item{replacement}{Named character vector. The values with which to replace
the data.}

\item{...}{Additional arguments to parse to the functions.}
}
\value{
Returns the object with the respective metadata updated.
}
\description{
This function will update the values in a given metadata column based on
what you are providing in terms of replacement.
}
