% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{upper_triangle_to_sparse}
\alias{upper_triangle_to_sparse}
\title{Transform an upper triangle-stored matrix to a sparse one}
\usage{
upper_triangle_to_sparse(upper_triangle_vals, shift, n)
}
\arguments{
\item{upper_triangle_vals}{Numerical vector. The values of the upper triangle
stored in a row major format.}

\item{shift}{Integer. Did you shift the diagonal up.}

\item{n}{Integer. Number of columns and rows of the symmetric matrix.}
}
\value{
The sparse matrix.
}
\description{
Transform an upper triangle-stored matrix to a sparse one
}
