% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics_sc.R
\name{vision_sc}
\alias{vision_sc}
\title{Calculate VISION scores}
\usage{
vision_sc(object, gs_list, streaming = NULL, .verbose = TRUE)
}
\arguments{
\item{object}{\code{SingleCells}, \code{MetaCells} (or potentially other) class.}

\item{gs_list}{Named nested list. The elements have the gene identifiers of
the respective gene sets and have the option to have a \code{"pos"} and \code{"neg"}
gene sets. The names need to be part of the variables of the object.}

\item{streaming}{Optional Boolean. Shall the data be streamed in. Useful for
larger data sets where you wish to avoid loading in the whole data. If
\code{NULL}, will automatically detect. Ignored when applied to \code{MetaCells}.}

\item{.verbose}{Boolean or integer. Controls verbosity and returns run times.
\code{FALSE} -> quiet, \code{TRUE} or \code{1L} -> normal verbosity, \code{2L} -> detailed
verbosity.}
}
\value{
The VISION scores in form of a matrix that is cells x gene sets
or as \code{ScMatrixRes} pending the input.
}
\description{
Calculates an VISION-type scores for pathways based on DeTomaso, et al.
Compared to other score types, you can also calculate delta-type scores
between positive and negative gene indices, think epithelial vs mesenchymal
gene signature, etc.
}
\references{
DeTomaso, et al., Nat. Commun., 2019
}
