% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_analysis.R
\name{vision_sc}
\alias{vision_sc}
\title{Calculate VISION scores}
\usage{
vision_sc(object, gs_list, streaming = FALSE, .verbose = TRUE)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{gs_list}{Named nested list. The elements have the gene identifiers of
the respective gene sets and have the option to have a \code{"pos"} and \code{"neg"}
gene sets. The names need to be part of the variables of the
\code{single_cell_exp} class.}

\item{streaming}{Boolean. Shall the cell data be streamed in. Useful for
larger data sets.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
Matrix of cells x signatures with the VISION pathway scores as
values.
}
\description{
Calculates an VISION-type scores for pathways based on DeTomaso, et al.
Compared to other score types, you can also calculate delta-type scores
between positive and negative gene indices, think epithelial vs mesenchymal
gene signature, etc.
}
\references{
DeTomaso, et al., Nat. Commun., 2019
}
