% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sc_analysis.R
\name{vision_w_autocor_sc}
\alias{vision_w_autocor_sc}
\title{Calculate VISION scores (with auto-correlation scores)}
\usage{
vision_w_autocor_sc(
  object,
  gs_list,
  embd_to_use,
  no_embd_to_use = NULL,
  vision_params = params_sc_vision(),
  streaming = FALSE,
  random_seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{\code{single_cell_exp} class.}

\item{gs_list}{Named nested list. The elements have the gene identifiers of
the respective gene sets and have the option to have a \code{"pos"} and \code{"neg"}
gene sets. The names need to be part of the variables of the
\code{single_cell_exp} class.}

\item{embd_to_use}{String. The embedding to use. Whichever you chose, it
needs to be part of the object.}

\item{no_embd_to_use}{Optional integer. Number of embedding dimensions to
use. If \code{NULL} all will be used.}

\item{vision_params}{List with vision parameters, see
\code{\link[=params_sc_vision]{params_sc_vision()}} with the following elements:
\itemize{
\item n_perm - Integer. Number of random permutations
\item n_cluster - Integer. Number of random clusters to generate to
associate each set with.
\item k - Number of neighbours for the kNN search. Only relevant if you
set regenerate_knn to \code{TRUE}.
\item knn_method - String. Which kNN algorithm to use. One of
\code{c("annoy", "hnsw", "nndescent")}. Defaults to \code{"annoy"}. Only relevant if
you set regenerate_knn to \code{TRUE}.
\item ann_dist - String. Distance metric for the approximate neighbour
search. One of \code{c("cosine", "euclidean")}. Defaults to \code{"cosine"}. Only
relevant if you set regenerate_knn to \code{TRUE}.
\item n_trees - Integer. Number of trees to use for the annoy algorithm.
Only relevant if you set regenerate_knn to \code{TRUE}.
\item search_budget - Integer. Search budget per tree for the annoy
algorithm. Only relevant if you set regenerate_knn to \code{TRUE}.
\item nn_max_iter - Integer. Maximum iterations for NN Descent. Only
relevant if you set regenerate_knn to \code{TRUE} and use
\code{"nndescent"}.
\item rho - Numeric. Sampling rate for NN Descent. Only relevant if you
set regenerate_knn to \code{TRUE} and use \code{"nndescent"}.
\item delta - Numeric. Early termination criterion for NN Descent. Only
relevant if you set regenerate_knn to \code{TRUE} and use \code{"nndescent"}.
}}

\item{streaming}{Boolean. Shall the cell data be streamed in. Useful for
larger data sets.}

\item{random_seed}{Integer. The random seed.}

\item{.verbose}{Boolean. Controls the verbosity of the function.}
}
\value{
Matrix of cells x signatures with the VISION pathway scores as
values.
}
\description{
Calculates an VISION-type scores for pathways based on DeTomaso, et al.
Compared to other score types, you can also calculate delta-type scores
between positive and negative gene indices, think epithelial vs mesenchymal
gene signature, etc. Additionally, this function also calculates the auto-
correlation values, answering the question if a given signature shows non-
random enrichment on the kNN graph. The kNN graph (and distance measures)
will be generated on-the-fly based on the embedding you wish to use.
}
\references{
DeTomaso, et al., Nat. Commun., 2019
}
