% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synthetic_sc.R
\name{write_cellranger_output}
\alias{write_cellranger_output}
\title{Helper function to write data to a cell ranger like output}
\usage{
write_cellranger_output(
  f_path,
  counts,
  obs,
  var,
  format_type = c("csv", "tsv"),
  rows = c("cells", "genes"),
  overwrite = TRUE,
  .verbose = TRUE
)
}
\arguments{
\item{f_path}{String. The filepath to which to save the data}

\item{counts}{Sparse matrix. Needs to be of class \code{dgRMatrix} or
\code{dgCMatrix}.}

\item{obs}{data.table. The observations. Needs to have
\code{nrow(obs) == nrow(counts)}.}

\item{var}{data.table. The variable data. Needs to have
\code{ncol(var) == ncol(counts)}.}

\item{overwrite}{Boolean. Shall any found h5ad file be overwritten.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Returns invisible
}
\description{
This is a helper to write synthetic data to cell ranger like
output, i.e., an .mtx file, an barcodes.csv (or .tsv) and a features.csv (or
.tsv).
}
