% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_h5.R
\name{write_h5ad_sc}
\alias{write_h5ad_sc}
\title{Helper function to write data to h5ad format}
\usage{
write_h5ad_sc(f_path, counts, obs, var, overwrite = TRUE, .verbose = TRUE)
}
\arguments{
\item{f_path}{String. The filepath to which to save the data}

\item{counts}{Sparse matrix. Needs to be of class \code{dgRMatrix} or
\code{dgCMatrix}.}

\item{obs}{data.table. The observations. Needs to have
\code{nrow(obs) == nrow(counts)}.}

\item{var}{data.table. The variable data. Needs to have
\code{ncol(var) == ncol(counts)}.}

\item{overwrite}{Boolean. Shall any found h5ad file be overwritten.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Returns invisible
}
\description{
Helper function to write data to h5ad format
}
